% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accent_table.R
\name{accent_table}
\alias{accent_table}
\title{aaccent_table: table() analytics as data frame, mosaicplot, usign add_ratio}
\usage{
accent_table(
  data,
  cols = "a1",
  rows = "성조",
  title = "",
  trans = TRUE,
  cex = 1.3,
  color = TRUE,
  ylab = "onset",
  xlab = "accent",
  sub = NULL,
  plot = FALSE,
  type = "res",
  raw = FALSE
)
}
\arguments{
\item{data}{table data}

\item{cols}{column names in data}

\item{rows}{row names in data}

\item{title}{table main}

\item{trans}{transpose}

\item{cex}{graph text size}

\item{color}{graph color 1:2, mosaicplot}

\item{ylab}{ylab}

\item{xlab}{xlab}

\item{sub}{subtitle}

\item{plot}{output option}

\item{type}{res, res_df, ratio, g, all}

\item{raw}{raw data, or table data}
}
\value{
matrix type data.frame
}
\description{
aaccent_table: table() analytics as data frame, mosaicplot, usign add_ratio
}
\examples{

\dontrun{
mtcars \%>\% select(am, vs) \%>\% table() \%>\% accent_table("am","vs")

mtcars \%>\% select(am, vs) \%>\% table() \%>\% accent_table("am","vs", plot=TRUE)


mtcars \%>\% select(am, vs) \%>\% table() \%>\% accent_table("am","vs", plot=TRUE, title = "Mtcars")

}



}
