% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cor.R
\name{add_cor}
\alias{add_cor}
\title{correlation analysis}
\usage{
add_cor(selectdata, type = "res_all", digits = 3, plot = TRUE)
}
\arguments{
\item{selectdata}{seledted daa}

\item{type}{res_all, res, p, r, res_p, res_md. g}

\item{digits}{default 3}

\item{plot}{graph performanceAnalytis::chart.Correlation}
}
\value{
table and picture
}
\description{
correlation analysis
}
\examples{

\dontrun{

mtcars\%>\%select(mpg, disp, wt, hp, qsec)\%>\%add_cor()
}
}
