% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_id.R
\name{add_id}
\alias{add_id}
\title{Add row id}
\usage{
add_id(data, sel = "names", each = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{sel}{select column}

\item{each}{each =TRUE then 111222....FALSE 123123}
}
\value{
add column data
}
\description{
Add row id
}
\examples{

\dontrun{

mtcars \%>\% mutate(MPG= which(colnames(mtcars) =="mpg"))\%>\%add_id()
mtcars \%>\% mutate(MPG= which(colnames(mtcars) =="mpg"))\%>\%add_id("cyl")
mtcars[1:3, 1:3] \%>\% add_id()
mtcars[1:3, 1:3] \%>\% add_id("cyl")
mtcars[1:4, 1:4] \%>\% to_long() \%>\% add_id()
mtcars[1:4, 1:4] \%>\% to_long() \%>\% add_id(each=FALSE)



}
}
