% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_partial_sum.R
\name{add_partial_sum}
\alias{add_partial_sum}
\title{add_partial_sum}
\usage{
add_partial_sum(df, col, freq_col)
}
\arguments{
\item{df}{df}

\item{col}{select col}

\item{freq_col}{select count frequency col}
}
\value{
df
}
\description{
add_partial_sum
}
\examples{
\dontrun{
#'
# 함수 사용 예시
df <- tibble::tribble(
  ~학교급, ~term, ~Freq, ~`prop(\%)`,
  "-", "합계", 3110, 100,
  "고등학교", "고1", 341, 11.0,
  "고등학교", "고2", 322, 10.4,
  "고등학교", "고3", 192, 6.17,
  "중학교", "중1", 591, 19.0,
  "중학교", "중2", 527, 16.9,
  "중학교", "중3", 290, 9.32,
  "초등학교", "초1", 8, 0.257,
  "초등학교", "초2", 48, 1.54,
  "초등학교", "초3", 47, 1.51,
  "초등학교", "초4", 191, 6.14,
  "초등학교", "초5", 207, 6.66,
  "초등학교", "초6", 346, 11.1
)

# 함수 실행
result_df <- add_partial_sum(df, col = "학교급", freq_col = "Freq")
print(result_df)
}

}
