% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ratio.R
\name{add_ratio}
\alias{add_ratio}
\title{Convert matrix data to a ratio by using the sum of columns to get a ratio}
\usage{
add_ratio(data, type = "res2", digits = 3)
}
\arguments{
\item{data}{matrix}

\item{type}{res2: *100, res1}

\item{digits}{defalut 3}
}
\value{
ratio tabl e
}
\description{
Convert matrix data to a ratio by using the sum of columns to get a ratio
}
\examples{

\dontrun{


mat = matrix(c(36, 67, 11,
          0, 35, 44,
         41, 108, 1,
          56, 87, 54),
           nrow = 4, byrow = TRUE,
 dimnames = list(c("공명음", "마찰음",
                   "유기음_경음", "평파열음_평파찰음"),
                    c("H", "H(H)", "L")))

## step1
  mat \%>\% add_ratio()
##conbind
 mat \%>\% add_ratio_df()
##add margins
 mat \%>\% add_sum()

}

}
