% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_sum.R
\name{add_row_sum}
\alias{add_row_sum}
\title{add_row_sum, row sum}
\usage{
add_row_sum(df)
}
\arguments{
\item{df}{df}
}
\value{
df
}
\description{
add_row_sum, row sum
}
\examples{
\dontrun{
mtcars$cyl \%>\% Freq_table(prop = TRUE) \%>\% addrow_sum()
}
}
