% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows.R
\name{add_rows}
\alias{add_rows}
\title{Insert data into the required act in the dataframe}
\usage{
add_rows(df, ..., row = NULL, char = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{...}{data,  c(100, 200), c(200, 300),}

\item{row}{row position}

\item{char}{transpose character column}
}
\value{
dagta
}
\description{
Insert data into the required act in the dataframe
}
\examples{
\dontrun{
#'
#'


df <- data.frame(A = 1:5, B = 6:10)
df
# 2번째 행부터 새 데이터(벡터 c(1,2,3,4)와 c(2,3,4,5))를 삽입

add_rows(df, c(100, 200), c(200, 300), row = 2)

df <- data.frame(
  측정변수 = c("진로목표(L01)", "준비여부(L02)", "직업결정(L03)", "부모추천(L04)", "학업성적(L05)"),
  Class.1 = c(0.2178, 0.3533, 0.1552, 0.2708, 0.1605),
  Class.2 = c(0.9888, 0.9769, 0.9910, 1.0000, 0.9804),
  Class.3 = c(0.9122, 0.8704, 0.9667, 0.0000, 0.6761)
)

# 1행에 새 데이터를 삽입
new_df <- add_rows(df, c("class name", "불명확 진로인식형", "적극적 진로준비형", "조건분 진로지향형"), row = 1)

print(new_df)


df <- data.frame(
  측정변수 = c("진로목표(L01)", "준비여부(L02)", "직업결정(L03)", "부모추천(L04)", "학업성적(L05)"),
  Class.1 = c(0.2178, 0.3533, 0.1552, 0.2708, 0.1605),
  Class.2 = c(0.9888, 0.9769, 0.9910, 1.0000, 0.9804),
  Class.3 = c(0.9122, 0.8704, 0.9667, 0.0000, 0.6761)
)

# 첫 번째 행에 새 데이터를 삽입, char = TRUE로 문자형 변환
new_df_1 <- add_rows(df, c("class name", "불명확 진로인식형", "적극적 진로준비형", "조건분 진로지향형"), row = 1, char = TRUE)

print(new_df_1)

# 마지막 행에 새 데이터를 삽입, char = FALSE로 기존 타입 유지
new_df_2 <- add_rows(df, c("class name", "불명확 진로인식형", "적극적 진로준비형", "조건분 진로지향형"), row = 7, char = FALSE)

print(new_df_2)


}

}
