% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows_freq.R
\name{add_rows_freq}
\alias{add_rows_freq}
\title{Functions that generate rows with a frequency}
\usage{
add_rows_freq(
  data,
  sel = ncol(data),
  remove = TRUE,
  value = TRUE,
  type = "res"
)
}
\arguments{
\item{data}{data}

\item{sel}{frequency column}

\item{remove}{raw freq remove}

\item{value}{frequency 1}

\item{type}{res, all}
}
\value{
data
}
\description{
Functions that generate rows with a frequency
}
\examples{

\dontrun{
dff=data.frame(
  trt=c("r","s"),
  con = c("f","m"),
  fre=c(4,5))

dff
dff \%>\%add_rows_freq()
dff \%>\%add_rows_freq(value=F)
dff \%>\%add_rows_freq(3)
dff \%>\%add_rows_freq("fre")

}


}
