% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sig.R
\name{add_sig}
\alias{add_sig}
\title{mutate data generate}
\usage{
add_sig(var)
}
\arguments{
\item{p.value}{pvalue}
}
\value{
in data
}
\description{
mutate data generate
}
\examples{

\dontrun{

aov_table(data = mtcars, dv_var = "mpg",
iv_vars = c("cyl", "gear", "carb"), grp_mean = T,
unite_F=TRUE) \%>\%
 mutate(STAR= add_sig(p_value))

}
}
