% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_significance_symbols.R
\name{add_significance_symbols}
\alias{add_significance_symbols}
\title{Functions that output statistical significance for matrices and dataframes Simple and exact expressions}
\usage{
add_significance_symbols(p_value_table, simple = FALSE)
}
\arguments{
\item{p_value_table}{A matrix of p-values, or other numerical matrix}

\item{simple}{Whether to represent the * mark simply or precisely with ***.}
}
\value{
Matrices marked with a *.
}
\description{
Functions that output statistical significance for matrices and dataframes Simple and exact expressions
}
\examples{

 \dontrun{

 #' p_value <- matrix(c(0, 0.00001, 0.00000,
                    0, 0.00000, 0.01497,
                    0, 0.00013, 0.00000,
                    0, 0.00000, 0.00102),
                  nrow = 4, byrow = TRUE,
                  dimnames = list(c("공명음", "마찰음", "유기음_경음", "평파열음_평파찰음"),
                                  c("H", "H(H)", "L")))
     # Add significance symbols
     # p_value

     add_significance_symbols(p_value)

 ##                  H     H(H)  L
 ##공명음            "***" "***" "***"
 ##마찰음            "***" "***" "*"
 ##유기음_경음       "***" "***" "***"
 ##평파열음_평파찰음 "***" "***" "**"




 }


}
