% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stat.R
\name{add_stat}
\alias{add_stat}
\title{데이터에서변수를 택하면 함수에 맞는 새로운 열을 추가하는 함수}
\usage{
add_stat(
  data,
  ...,
  fun = mean,
  col_name = NULL,
  na.rm = TRUE,
  margin = 1,
  type = "data"
)
}
\arguments{
\item{data}{data.frame}

\item{...}{cols}

\item{fun}{mean, ad, ....}

\item{col_name}{New variable}

\item{na.rm}{na.rm=TRUE}

\item{margin}{margin 1, 2}

\item{type}{data, col 2type}
}
\value{
data.fram or vector
}
\description{
데이터에서변수를 택하면 함수에 맞는 새로운 열을 추가하는 함수
}
\examples{

\dontrun{

df = data.frame(A = c(1, 2, 3, 4, 5), B = c(6, 7, NA, 9, 10), C = c(11, 12, 13, 14, 15) )

df\%>\% add_stat( "AC_MEAN","A", "C", fun = mean)
df\%>\% add_stat( "AC_MEAN","A", "C", fun = mean, type="col")
df \%>\%add_stat( "A", "C", type="col")
df \%>\%add_stat( "A", "C")
df \%>\%add_stat( "A", "C", fun = mean, "AC_MEAN")

df \%>\%add_stat( "A", "C", "AC_MEAN",fun = sd)

df \%>\%add_stat(col_name = "AC_MEAN","A", "C", fun = mean)
df \%>\%add_stat("A", "C", fun = mean, col_name = "AC_MEAN")
df \%>\%add_stat("A", "C", fun = mean)
df \%>\%add_stat(c("A", "C"))
df \%>\%add_stat("ROWs", c("A", "C"))

mtcars \%>\% add_stat("MEAN", "am","vs")
mtcars \%>\% add_stat("am","vs")
mtcars \%>\% add_stat("MEAN", c("am","vs"))
mtcars \%>\% add_stat("MEAN", c("mpg","qsec"))
mtcars \%>\% add_stat("MEAN", c("mpg","qsec"), col_name="ACME")
df

}


}
