% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ratio.R
\name{add_sum}
\alias{add_sum}
\title{Functions to sum margins after adding ratios}
\usage{
add_sum(
  dataset,
  v1 = "a1",
  v2 = "성조",
  raw = FALSE,
  add_ratio = TRUE,
  plot = FALSE,
  color = TRUE,
  ylab = "onset",
  xlab = "speaker",
  cex = 1.2,
  title = NULL,
  type = "df"
)
}
\arguments{
\item{dataset}{data(rwa=T)or table}

\item{v1}{col}

\item{v2}{row}

\item{raw}{TRUE dataset. not table}

\item{add_ratio}{using add_ration function}

\item{plot}{mosicplot}

\item{color}{color 1:4, etc}

\item{ylab}{yalb}

\item{xlab}{xlab}

\item{cex}{size}

\item{title}{title}

\item{type}{all, df, res}
}
\value{
matrix
}
\description{
Functions to sum margins after adding ratios
}
\examples{

#'
\dontrun{

#' #'mat = matrix(c(36, 67, 11,
          0, 35, 44,
         41, 108, 1,
          56, 87, 54),
           nrow = 4, byrow = TRUE.
   dimnames = list(c("공명음", "마찰음",
     "유기음_경음", "평파열음_평파찰음"),  c("H", "H(H)", "L")))
## step1
  mat \%>\% add_ratio()
##conbind
 mat \%>\% add_ratio_df()
##add margins
 mat \%>\% add_sum()

 #                           H H(H)+H(H)_1         L SUM
 #공명음             26(11.5\%)   35(19.4\%) 36(28.1\%)  97
 #마찰음             49(21.6\%)   62(34.4\%) 35(27.3\%) 146
 #유기음_경음       108(47.6\%)   82(45.6\%)  11(8.6\%) 201
 #평파열음_평파찰음  44(19.4\%)     1(0.6\%) 46(35.9\%)  91
 #SUM                      227         180       128 535






}

}
