% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_na_zero.R
\name{all_na_zero}
\alias{all_na_zero}
\title{all_na_zero}
\usage{
all_na_zero(df, imp = 0)
}
\arguments{
\item{df}{data.frame}

\item{imp}{implement default = 0}
}
\value{
NA to 0
}
\description{
all_na_zero
}
\examples{
\dontrun{


}
}
