% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_na_zero.R
\name{all_na_zero1}
\alias{all_na_zero1}
\title{NA to zero}
\usage{
all_na_zero1(data, fun = "extend")
}
\arguments{
\item{data}{data.frame}

\item{fun}{'extend', 'base' are equal method}
}
\value{
Na data are changed zero
}
\description{
NA to zero
}
