% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_alpha.R
\name{alpha_apa}
\alias{alpha_apa}
\title{alpha_apa report}
\usage{
alpha_apa(data)
}
\arguments{
\item{data}{분석결과}
}
\value{
apa보고서
}
\description{
alpha_apa report
}
\examples{
\dontrun{

#'
# 진로동기 신뢰도 분석
jjstat::bind_alpha_table(jutr2\%>\% select(A01:A15),
                         c(1:5), c(6:10), c(11:15)) \%>\%nice_Var()\%>\% nice_table()\%>\%
  replace_df("a","진로정체감") \%>\%
  replace_df("b","진로통찰력") \%>\%
  replace_df("c","진로탄력성") \%>\%
  alpha_apa()

}
}
