% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_table.R
\name{alpha_table}
\alias{alpha_table}
\title{Function to tabulate reliability}
\usage{
alpha_table(
  alpha_data_res,
  title = "",
  show = "data",
  format = "markdown",
  digits = 3,
  variable = "."
)
}
\arguments{
\item{alpha_data_res}{alpha result}

\item{title}{your title}

\item{show}{'data', 'markdown'}

\item{format}{markdown, html}

\item{digits}{default 3}

\item{variable}{. output}
}
\description{
Function to tabulate reliability
}
\examples{
\dontrun{
bind_rows(
stat_onl \%>\% select(메타인지1:메타인지4) \%>\%
  alpha_table(show="data", variable = "A" ),
stat_onl \%>\% select(도움요청1:도움요청3) \%>\%
  alpha_table(show="data", variable = "B"),
stat_onl \%>\% select(환경관리1:환경관리4) \%>\%
  alpha_table(show="data", variable = "C"),
#' stat_onl \%>\% select(노력지속1:노력지속4) \%>\%
 alpha_table(show="data", variable = "D")
)  \%>\% markdown_table(digits=3, font_size = 20,
                     caption= "total")
}




}
