% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_aov.R
\name{anova_apa}
\alias{anova_apa}
\title{Report aov data (format APA)}
\usage{
anova_apa(
  aov_post_data,
  ns = 7,
  np = 6,
  ndf = 4,
  nt = 5,
  digits = 3,
  md = FALSE,
  posthoc_type = "The Multi Least Significant Difference(MLSD)"
)
}
\arguments{
\item{aov_post_data}{aov_post data}

\item{ns}{ncol sig column 7}

\item{np}{ncol p value column 6}

\item{ndf}{ncol df column 4}

\item{nt}{ncol t value column 5}

\item{digits}{digits 3}

\item{md}{md = TRUE is markdown outpput}
}
\value{
report data
}
\description{
Report aov data (format APA)
}
\examples{
\dontrun{
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post() \%>\% anova_apa()
# This is error
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="contrast")\%>\% anova_apa()

# aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="all")

aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="all")\%>\% anova_apa()
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="res")\%>\% anova_apa()

}

}
