% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_aov.R
\name{aov_apa}
\alias{aov_apa}
\title{Report aov data (format APA)}
\usage{
aov_apa(aov_post_data, md = FALSE)
}
\arguments{
\item{aov_post_data}{aov_post data}

\item{md}{md = TRUE is markdown outpput}
}
\value{
report data
}
\description{
Report aov data (format APA)
}
\examples{
\dontrun{
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post()\%>\% aov_apa()
# This is error
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="contrast")\%>\% aov_apa()

# aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="all")

aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="all")\%>\% aov_apa()
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="res")\%>\% aov_apa()

}

}
