% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_table.R
\name{aov_df}
\alias{aov_df}
\title{Generate an AOV results table for multiple independent variables}
\usage{
aov_df(
  data,
  dv_var,
  iv_vars,
  grp_mean = FALSE,
  unite_p = FALSE,
  unite_F = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{dv_var}{dv}

\item{iv_vars}{c(iv1, iv2....)}

\item{unite}{pvalue unite}

\item{mean}{Option to average by group and add}
}
\value{
table
}
\description{
Generate an AOV results table for multiple independent variables
}
\examples{

\dontrun{
aov_df(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"))

aov_df(data = mtcars, dv_var = "mpg",iv_vars = c("cyl", "gear", "carb"),
           mean = TRUE)

aov_df(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = FALSE)

aov_df(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = TRUE)

}


}
