% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_table.R
\name{aov_table}
\alias{aov_table}
\title{Generate an AOV results table for multiple independent variables}
\usage{
aov_table(
  data,
  dv_var = NULL,
  iv_vars = NULL,
  grp_mean = TRUE,
  unite_p = FALSE,
  unite_F = FALSE,
  digits = 2,
  posthoc = "lsd",
  p.adj = "bonferroni",
  sig = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{dv_var}{dv}

\item{iv_vars}{c(iv1, iv2....)}

\item{grp_mean}{'grooup mean remove}

\item{unite_p}{pvalue unite}

\item{unite_F}{Fvalue unite}

\item{posthoc}{'tukey, scheffe, duncan, lsd}

\item{p.adj}{p.adj=c("none","holm","hommel", "hochberg", "bonferroni", "BH", "BY", "fdr") only lsd test}

\item{sig}{'sig = T' is add star}
}
\value{
table
}
\description{
Generate an AOV results table for multiple independent variables
}
\examples{

\dontrun{
mtcars <- mtcars\%>\%as_trt("cyl", "gear", "carb")
aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"))

aov_table(data = mtcars, dv_var = "mpg",
iv_vars = c("cyl", "gear", "carb"), grp_mean = FALSE)
aov_table(data = mtcars, dv_var = "mpg",
iv_vars = c("cyl", "gear", "carb"), grp_mean = FALSE) \%>\%
aov_table_apaOP

aov_table(data = mtcars, dv_var = "mpg",
 iv_vars = c("cyl", "gear", "carb"), grp_mean = TRUE)

aov_table(data = mtcars, dv_var = "mpg",
 iv_vars = c("cyl", "gear", "carb"), grp_mean = TRUE, unite_F = TRUE)

aov_table(data = mtcars, dv_var = "mpg",
 iv_vars = c("cyl", "gear", "carb"),
  grp_mean = TRUE, unite_F = TRUE) \%>\% aov_rename()
  markdown_table()
}



}
