% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_table.R
\name{aov_table_apa}
\alias{aov_table_apa}
\title{total anova report}
\usage{
aov_table_apa(data, show = TRUE, dv_var = NULL, iv_vars = NULL)
}
\arguments{
\item{data}{aoa_df(grp_mean=F)}

\item{show}{show data}
}
\value{
report
}
\description{
total anova report
}
\examples{
\dontrun{
aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = FALSE)\%>\% aov_table_apa()

aov_table_apa(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"))
}


}
