% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_aov_df.R
\name{apa_aov_df}
\alias{apa_aov_df}
\title{total anova report}
\usage{
apa_aov_df(data, show = TRUE, dv_var = NULL, iv_vars = NULL)
}
\arguments{
\item{data}{aoa_df(grp_mean=F)}

\item{show}{show data}
}
\value{
report
}
\description{
total anova report
}
\examples{
\dontrun{
aov_table(data = mtcars, dv_var = "mpg",
iv_vars = c("cyl", "gear", "carb"),
grp_mean = FALSE)\%>\% apa_aov_df()

apa_aov_df(data = mtcars, dv_var = "mpg",
 iv_vars = c("cyl", "gear", "carb"))
}


}
