% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_aov_df_res.R
\name{apa_aov_df_res}
\alias{apa_aov_df_res}
\title{apa_aov_df_res, aov_df의 분석결과}
\usage{
apa_aov_df_res(data, dv_col = 2, iv_col = 1, show = TRUE, digits = 3)
}
\arguments{
\item{data}{data}

\item{dv_col}{dv}

\item{iv_col}{iv}

\item{show}{TRUE, data result}

\item{digits}{3}
}
\value{
결과해석 text
}
\description{
apa_aov_df_res, aov_df의 분석결과
}
\examples{
\dontrun{
#'
# 테스트 데이터로 실행
data <- data.frame(
  독립변수 = c("LCA_class", "", "", "LCA_class",
  "", "", "LCA_class", "", "", "LCA_class", "", ""),
  종속변수 = c("진로동기", "진로동기", "진로동기",
  "진로태도성숙", "진로태도성숙", "진로태도성숙",
  "진로결정자기효능감", "진로결정자기효능감", "진로결정자기효능감",
  "진로준비행동", "진로준비행동", "진로준비행동"),
  df1 = c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2),
  df2 = c(511, 511, 511, 511, 511, 511,
   511, 511, 511, 511, 511, 511),
  F = c(132.64, NA, NA, 107.45, NA, NA,
   138.02, NA, NA, 106.1, NA, NA),
  p = c("4.00665784818538e-47", NA, NA,
   "1.11197564656087e-39", NA, NA,
   "1.18550277068589e-48", NA, NA,
   "2.8829281865653e-39", NA, NA)
)

# show=TRUE로 결과 출력
apa_results <- apa_aov_df_res(data, dv_col = 2, iv_col = 1, show = TRUE)

}
}
