% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobel_test.R
\name{aroian_sobel_test_general}
\alias{aroian_sobel_test_general}
\title{Aroian Approximation Sobel Test generalized function}
\usage{
aroian_sobel_test_general(a, SE, digits = 5)
}
\arguments{
\item{a}{est1}

\item{b}{est2}

\item{sa}{se1}

\item{sb}{se2}

\item{type}{Sobel, Aroian, Goodman}
}
\value{
test result
}
\description{
Aroian Approximation Sobel Test generalized function
}
\examples{

\dontrun{
#'
# Example usage
a <- 0.5  # Example coefficient a
b <- 0.3  # Example coefficient b
sa <- 0.1 # Standard error of a
sb <- 0.05 # Standard error of b

Sobel_test(a, b, sa, sb)
Sobel_test(a, b, sa, sb, "all")
Sobel_test(a, b, sa, sb, "Sobel")

Sobel_test(a, b, sa, sb,"Aroian")

Sobel_test(a, b, sa, sb,"Goodman")
}
}
