% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartlett_test.R
\name{bartlett_test}
\alias{bartlett_test}
\title{bartlett_test}
\usage{
bartlett_test(
  data,
  formula = NULL,
  Mean = "Mean",
  sd = "SD",
  n = "N",
  var = "variable"
)
}
\arguments{
\item{data}{data.frame or mean. sd. n}

\item{formula}{dv ~ iv}

\item{Mean}{mean selection}

\item{sd}{sd selection}

\item{n}{n selection}

\item{var}{variable}
}
\value{
result
}
\description{
bartlett_test
}
\examples{

\dontrun{

set.seed(20240418)
my_data <- data.frame(
  value = c(group1 <- rnorm(20, mean = 10, sd = 2),
            group2 <- rnorm(20, mean = 12, sd = 2),
            group3 <- rnorm(20, mean = 15, sd = 2)),
  group = rep(c("Group1", "Group2", "Group3"), each = 20)
)
my_data \%>\%mysummaryBy(value ~ group)\%>\%select(1:4) \%>\%bartlett_test()

data = data.frame(row.names = c("1", "2", "3"),
                  var = c('Group1', 'Group2', 'Group3'),
                  Mean = c('10.0782306287367', '11.5824600813951', '14.764623393546'),
                  SD = c('1.88743751774179', '1.9951475292551', '2.07862332265816'),
                  N = c('20', '20', '20'))
data
bartlett_test(data)

# original test
bartlett.test(value ~ group, data=my_data)
#now test
bartlett_test(value ~ group, data=my_data)


#origin
bartlett.test(mpg~cyl, data=mtcars)
#aggregated data
mtcars\%>\%mysummaryBy(mpg~cyl) \%>\%bartlett_test()
# Using function
bartlett_test(mpg~cyl, data=mtcars)
#'
}
}
