% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_recode.R
\name{binary_recode}
\alias{binary_recode}
\title{binary_recode}
\usage{
binary_recode(data, variable, recode_list, new_name = NULL, tail = "_grp")
}
\arguments{
\item{data}{data}

\item{variable}{original variable}

\item{recode_list}{list recode variable}

\item{new_name}{new name}

\item{tail}{tails _grp}
}
\value{
data
}
\description{
binary_recode
}
\examples{

\dontrun{
#'
# 예제 데이터
data <- data.frame(
  학력 = factor(c("고졸", "대학졸", "석박사졸", "고졸", "석박사졸")),
  grade = 1:5
)
data
binary_recode(data, grade, list(미숙자 = 1:3, 숙달자 = 4:5))
binary_recode(data, grade, list(미숙자 = c(1,2,3), 숙달자 = 4:5))
binary_recode(data, "grade", list(미숙자 = 1:3, 숙달자= 4:5), new_name="수준")

binary_recode(data, 학력, list(저학년= 1:2, 고학년= 3))
binary_recode(data, "학력", list(저학년= 1:2, 고학년= 3))
binary_recode(data, "학력", list(저학년= c("고졸","대학졸"), 고학년= 3))
binary_recode(data, 학력, list(저학년= c("고졸","대학졸"), 고학년= 3))
binary_recode(data, 학력, list(저학년= c("고졸","대학졸"), 고학년= "석박사졸"))

}

}
