% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_recode_sep.R
\name{binary_recode_sep}
\alias{binary_recode_sep}
\title{Multi-recoding binary variables}
\usage{
binary_recode_sep(data, ..., new_names = NULL, tail = "_grp")
}
\arguments{
\item{data}{data}

\item{...}{recode list}

\item{new_names}{newnaem list}

\item{tail}{tail _grp}
}
\value{
data
}
\description{
Multi-recoding binary variables
}
