% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_comare_data.R
\name{bind_compare_data}
\alias{bind_compare_data}
\title{Table-building functions that pairwise compare two pieces of data}
\usage{
bind_compare_data(dff1, dff2, digits = 3, pair = TRUE)
}
\arguments{
\item{dff1}{data1}

\item{dff2}{data2}

\item{digits}{rounding}

\item{pair}{pari option default TRUE}
}
\value{
table
}
\description{
Table-building functions that pairwise compare two pieces of data
}
\examples{
\dontrun{


}

}
