% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_lme_model.R
\name{bind_lme_model}
\alias{bind_lme_model}
\title{Tbind_lme_model}
\usage{
bind_lme_model(
  ...,
  model_name = paste("모형", 1:length(model)),
  title = "모형의 비교"
)
}
\arguments{
\item{...}{input}

\item{model_name}{model name}

\item{title}{model title}
}
\value{
table
}
\description{
Tbind_lme_model
}
\examples{
\dontrun{
data(birthwt)
str(birthwt)
fitm <- lm(bwt ~ age+lwt+factor(race)+smoke+ptl+ht+ui, data=birthwt)
fitm2 <- lm(bwt ~ lwt+factor(race)+smoke+ht+ui, data=birthwt)
bind_lme_model(fitm, fitm2)

}
}
