% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMH_test.R
\name{breslowdata_test}
\alias{breslowdata_test}
\title{Breslow-Day Test for Homogeneity of the Odds Ratios}
\usage{
breslowdata_test(data, OR = NA, correct = FALSE, tidy = TRUE)
}
\arguments{
\item{data}{2x2xk table}

\item{OR}{the odds ratio to be tested against. If left undefined (default) the Mantel-Haenszel estimate will be used.}

\item{correct}{If TRUE, the Breslow-Day test with Tarone's adjustment is computed, which subtracts an adjustment factor to make the resulting statistic asymptotically chi-square.}

\item{tidy}{broom::tidy}
}
\value{
result
}
\description{
Breslow-Day Test for Homogeneity of the Odds Ratios
}
\details{
For the Breslow-Day test to be valid, the sample size should be relatively large in each stratum, and at least 80% of the expected cell counts should be greater than 5. Note that this is a stricter sample size requirement than the requirement for the Cochran-Mantel-Haenszel test for tables, in that each stratum sample size (not just the overall sample size) must be relatively large. Even when the Breslow-Day test is valid, it might not be very powerful against certain alternatives, as discussed in Breslow and Day (1980).
Alternatively, it might be better to cast the entire inference problem into the setting of a logistic regression model. Here, the underlying question of the Breslow-Day test can be answered by investigating whether an interaction term with the strata variable is necessary (e.g. using a likelihood ratio test using the anova function).
}
\examples{
\dontrun{
#'
migraine <- xtabs(freq ~ .,
                  cbind(expand.grid(treatment=c("active", "placebo"),
                                    response =c("better", "same"),
                                    gender   =c("female", "male")),
                        freq=c(16, 5, 11, 20, 12, 7, 16, 19))
)
# get rid of gender
tab <- xtabs(Freq ~ treatment + response, migraine)
DescTools::Desc(tab)
# only the women
female <- migraine[,, 1]
DescTools::Desc(female)
# .. and the men
male <- migraine[,, 2]
DescTools::Desc(male)
breslowdata_test(migraine)
breslowdata_test(migraine, correct = TRUE)

}
}
