% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_vif.R
\name{cal_vif}
\alias{cal_vif}
\title{calculation vif}
\usage{
cal_vif(model)
}
\arguments{
\item{model}{lm model data}
}
\value{
vif
}
\description{
calculation vif
}
\examples{

\dontrun{
#'
# Create the linear model
model <- lm(hp ~ wt + disp + drat, data = mtcars)

# Calculate VIF using the custom function
cal_vif(model)

lm(mpg ~ wt + disp + drat + hp , data = mtcars) \%>\%cal_vif()


}
}
