% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chi_sig.R
\name{calculate_chi_sig}
\alias{calculate_chi_sig}
\title{Create oserved/expected tapes after chisq calculation and give them individual significance}
\usage{
calculate_chi_sig(observed, type = "data", simple = FALSE)
}
\arguments{
\item{observed}{contigency table}

\item{type}{data, Res(all), chisq_test, cramers v, observed_over_expected_ratios,cell_p_values, p_sig, data, data2}

\item{simple}{simple sar one *}
}
\value{
matrix, data.frame
}
\description{
Create oserved/expected tapes after chisq calculation and give them individual significance
}
\examples{

\dontrun{
# # calculate_chi_sig(data)
# Example data
matrix(c(36, 67, 11,
          0, 35, 44,
         41, 108, 1,
          56, 87, 54),
           nrow = 4, byrow = TRUE.
   dimnames = list(c("공명음", "마찰음",
     "유기음_경음", "평파열음_평파찰음"),  c("H", "H(H)", "L"))) \%>\%
     calculate_chi_sig("p_sig")


     matrix(c(36, 67, 11,
              40, 35, 44,
              41, 108, 1,
              56, 87, 54), nrow = 4, byrow = TRUE,
                     dimnames = list(c("공명음", "마찰음",
                                    "유기음_경음", "평파열음_평파찰음"),
                                 c("H", "H(H)", "L"))) \%>\%
                calculate_chi_sig(simple = TRUE)



  chisq.test(matrix(c(36, 67, 11,
                      40, 35, 44,
                   41, 108, 1,
              56, 87, 54), nrow = 4, byrow = TRUE,
               dimnames = list(c("공명음", "마찰음",
              "유기음_경음", "평파열음_평파찰음"),
              c("H", "H(H)", "L")))   )



}


}
