% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCentering.R
\name{centering}
\alias{centering}
\title{centering each vector}
\usage{
centering(variable, na.rm = TRUE)
}
\arguments{
\item{variable}{vector variable in data}

\item{na.rm}{missing data chedk}
}
\description{
centering each vector
}
