% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa2.R
\name{cfa2}
\alias{cfa2}
\title{Create CFA result summary by lavaan sem estimated result By Park Joonghee Ph.D}
\usage{
cfa2(
  x,
  format = "markdown",
  htmt_cut = 0.9,
  htmt2 = FALSE,
  cut = 0.7,
  angle = 90,
  cex = 11,
  hjust = 0.9,
  val.size = 4,
  dis.sort = TRUE,
  rename = FALSE,
  var_name = NA,
  digits = 3,
  type = "all"
)
}
\arguments{
\item{x}{lavaan result}

\item{format}{knitr markdown decision}

\item{htmt_cut}{htmt_cut is default =1.
if you want roburst option htmt_cut =0.9}

\item{htmt2}{TRUE using htmt2}

\item{cut}{loading criteran}

\item{angle}{tesxt angle}

\item{cex}{text size}

\item{hjust}{on text value gap}

\item{val.size}{text size}

\item{dis.sort}{item sort}

\item{rename}{item validiation new name apply}

\item{var_name}{new text name default NULL}

\item{digits}{value rounding}

\item{dataset}{this is import name that dataset column name}

\item{model}{lavaan model need indices validation name}

\item{res}{result type}
}
\description{
Create CFA result summary by lavaan sem estimated result By Park Joonghee Ph.D
}
\examples{
\dontrun{
library(lavaan)
data(PoliticalDemocracy)

model1 <- "
  ind60 =~ x1 + x2 + x3
  dem60 =~ y1 + y2 + y3 + y4
  dem65 =~ y5 + y6 + y7 + y8
  dem60 ~ ind60
  dem65 ~ ind60 + dem60
"

fit1 <- sem(model1, data = PoliticalDemocracy)

# 기본 사용법: 전체 결과 출력
cfa2(fit1)

# 특정 결과(res)만 출력 예시
cfa2(fit1, res = "modelfit")        # 모델 적합도
cfa2(fit1, res = "loadings")        # 요인적재량 테이블
cfa2(fit1, res = "alpha")           # 신뢰도(Cronbach alpha)
cfa2(fit1, res = "CR_AVE")          # 수렴타당도(CR, AVE)
cfa2(fit1, res = "Discriminant")    # 판별타당도 (Fornell-Larcker 기준)
cfa2(fit1, res = "htmt", dataset = PoliticalDemocracy, model = model1)  # HTMT 지표
cfa2(fit1, res = "str_cor")         # 잠재변수 간 경로 유의성

# 시각화 결과: 요인적재량 바그래프 출력
cfa2(fit1, res = "loadings_bar")

# 사용자 정의 변수명 적용(rename + var_name)
new_names <- c("Indicator1", "Indicator2", "Indicator3",
               "Democracy1", "Democracy2", "Democracy3", "Democracy4",
               "Democracy5", "Democracy6", "Democracy7", "Democracy8")

# 그래프의 색을 바꾸고 싶은 경우
cfa2(fit, type = "loadings_bar") +
scale_fill_grey()
# rename = TRUE일 경우에만 var_name이 반영됨
cfa2(fit1, rename = TRUE, var_name = new_names)

# rename = FALSE일 경우 var_name은 무시됨
cfa2(fit1, rename = FALSE, var_name = new_names)

# 다양한 출력 포맷 지정
cfa2(fit1, format = "markdown")
cfa2(fit1, format = "html") \%>\% kableExtra::kable_minimal(full_width = FALSE)

# 모델 간 비교
model2 <- "
  ind60 =~ x1 + x2 + x3
  dem60 =~ y1 + y2 + y3 + y4
  dem65 =~ y5 + y6 + y7 + y8
  dem60 ~ ind60
  dem65 ~ ind60 + dem60
"
fit2 <- sem(model2, data = PoliticalDemocracy)
CompareFit_diff(fit1, fit2)

# AVE 직접 확인
AVE(fit1)
AVE(fit2)

# 통합 요약 테이블 출력 (CR, AVE 포함)
library(dplyr)
full_join(
  cfa2(fit1, type = "item") \%>\% dplyr::select(-Accept, -p),
  cfa2(fit1, type = "Convergent") \%>\% row2col("Latent") \%>\%
    dplyr::select(-Cronbach) \%>\% Round(3),
  by = "Latent"
) \%>\% nice_table()
}
}
