% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Char2num_fct.R
\name{Char2num_fct}
\alias{Char2num_fct}
\title{fact varaible to num  in data.frame}
\usage{
Char2num_fct(df, col = 1)
}
\arguments{
\item{df}{df}

\item{col}{select col}
}
\value{
df
}
\description{
fact varaible to num  in data.frame
}
\examples{
\dontrun{


}
}
