% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_chisq.R
\name{chisq_apa}
\alias{chisq_apa}
\title{카이제곱 apa보고}
\usage{
chisq_apa(data, chi_type = NULL, show = TRUE)
}
\arguments{
\item{data}{chisq.test result data}

\item{chi_type}{"적합성 검정", "독립성 검정", "동질성 검정"}

\item{show}{TRUE cat output}
}
\value{
data result
}
\description{
카이제곱 apa보고
}
\examples{

\dontrun{
#'
dfa= matrix(c(22, 26, 55,13, 31, 33,13, 16, 38), ncol=3, byrow=TRUE)
colnames(dfa)=c("가","나","다")
rownames(dfa)=c("A","B","C")
dfa
dfat= t(dfa)
dfat\%>\%chisq.test()\%>\%chisq_apa()
}
}
