% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_contrast.R
\name{chisq_contrast}
\alias{chisq_contrast}
\title{Chisq test Contrast test}
\usage{
chisq_contrast(
  data,
  ...,
  test = chisq.test,
  p.adjust = "bonferroni",
  adj = TRUE,
  name = FALSE
)
}
\arguments{
\item{data}{data}

\item{...}{contrast}

\item{test}{chisq.test, fisher.test}

\item{p.adjust}{c("bonferroni","fdr","BH","BY","holm","hochberg","hommel")}

\item{adj}{cadjust p value TRUE}

\item{name}{rowname, when  pairwise}
}
\value{
data
}
\description{
Chisq test Contrast test
}
\examples{

\dontrun{
#'
d <- matrix(c(584, 24, 1721, 56, 2400, 90, 8164, 289),
            nrow = 4, byrow = TRUE,
            dimnames = list(c("kangnam", "Moksdong", "nowon", "dongtan"), c("Yes", "No")))

d\%>\%chisq.test()
chisq_contrast(d)
chisq_contrast(d, c(1,2,3,4))


chisq_contrast(d, c(1,1,2,3),c(2,1,2,2),c(2,2,1,2),c(2,2,2,1))
# fisher.test
chisq_contrast(d, c(1,2,2,2),c(2,1,2,2),c(2,2,1,2),c(2,2,2,1), test=fisher.test)
# fisher.exact test
chisq_contrast(d, c(1,2,2,3),c(2,1,2,2),c(2,2,1,2),c(2,2,2,1), test=fisher.test)

chisq_contrast(d, c(1,2,2,2),c(2,1,1,2),c(2,2,1,2),c(2,2,2,1), test=fisher.test)
chisq_contrast(d, c(1,2,2,2),c(2,3,1,2),c(2,2,1,2),c(2,2,2,1), test=fisher.test)
chisq_contrast(d, c(1,2,2,2),c(2,3,1,2),c(2,2,1,3),c(2,2,2,1), test=fisher.test)
chisq_contrast(d, c(1,2,2,2),c(2,3,1,2),c(2,2,1,3),c(3,4,2,1), test=fisher.test)


#other data
data_text = "Tobacco_Dependence (0,5] (5,10] (10,15] (15,20] (20,98]
No_Nicotine_Dependence 130 210 43 114 20
Nicotine_Dependence 119 267 91 254 67"

#method2
data_text\%>\% text2df(header=T) \%>\% make_df_text()
nt
nt = data_text\%>\% text2df(header=T)\%>\% jjstat::col2row()
nt \%>\%tibble()
nt=Char2num(nt,1:5)
nt\%>\%t() \%>\%as.data.frame()\%>\%
  chisq_contrast(c(1,2,2,2,2),
                 c(2,1,2,2,2),
                 c(2,2,1,2,2),
                 c(2,3,4,1,2))
}


}
