% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_gof_posthoc.R
\name{chisq_each}
\alias{chisq_each}
\title{chisq_gof_posthoc}
\usage{
chisq_each(counts, type = "all", method = "fdr")
}
\arguments{
\item{counts}{data}

\item{type}{all, res, p, chisq}

\item{method}{p.adust}
}
\value{
data
}
\description{
chisq_gof_posthoc
}
\examples{

\dontrun{
#'
#'
count <- c(49, 30, 63, 59)

chisq_each(count, "res")
chisq_each(count,"all")$gof_test
chisq_each(count,"all")$p
chisq_each(count,"all")$chisq
chisq_each(count, typ="all")
chisq_each(count, typ="res")
chisq_each(count, type="p")
chisq_each(count, type="chisq")

chisq_each(count, method="bonferroni")

count1 <- c(4, 3, 6, 19)
chisq_each(count1)
chisq_each(count1, typ="all")
chisq_each(count1, type="p")
chisq_each(count1, type="chisq")
chisq_each(count1, method="bonferroni")


chisq_each(count, method="bonferroni")

chisq_each(c(49, 30, 63, 59, 40, 60))

#'
}



}
