% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_gof_posthoc.R
\name{chisq_gof_posthoc}
\alias{chisq_gof_posthoc}
\title{chisq_gof_posthoc}
\usage{
chisq_gof_posthoc(counts, type = "all", method = "fdr")
}
\arguments{
\item{counts}{data}

\item{type}{all, res, p, chisq}

\item{method}{p.adust}
}
\value{
data
}
\description{
chisq_gof_posthoc
}
\examples{

\dontrun{
#'

count <- c(49, 30, 63, 59)

chisq_gof_posthoc(count, "res")
chisq_gof_posthoc(count,"all")$gof_test
chisq_gof_posthoc(count,"all")$p
chisq_gof_posthoc(count,"all")$chisq
chisq_gof_posthoc(count, typ="all")
chisq_gof_posthoc(count, typ="res")
chisq_gof_posthoc(count, type="p")
chisq_gof_posthoc(count, type="chisq")

chisq_gof_posthoc(count, method="bonferroni")

count1 <- c(4, 3, 6, 19)
chisq_gof_posthoc(count1)
chisq_gof_posthoc(count1, typ="all")
chisq_gof_posthoc(count1, type="p")
chisq_gof_posthoc(count1, type="chisq")
chisq_gof_posthoc(count1, method="bonferroni")


chisq_gof_posthoc(count, method="bonferroni")

chisq_gof_posthoc(c(49, 30, 63, 59, 40, 60))

#'
}



}
