% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_gof_posthoc_apa.R
\name{chisq_gof_posthoc_apa}
\alias{chisq_gof_posthoc_apa}
\title{chisq_gof_posthoc_apa}
\usage{
chisq_gof_posthoc_apa(df, adj = FALSE)
}
\arguments{
\item{df}{df}

\item{adj}{adj}
}
\value{
df
}
\description{
chisq_gof_posthoc_apa
}
\examples{
\dontrun{


}

}
