% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_posthoc.R
\name{chisq_posthoc}
\alias{chisq_posthoc}
\title{chisq post hoc}
\usage{
chisq_posthoc(
  tbl,
  test = fisher.test,
  Rows = TRUE,
  control = c("bonferroni", "fdr", "BH", "BY", "holm", "hochberg", "hommel"),
  digits = 4,
  ...
)
}
\arguments{
\item{tbl}{tavke}

\item{test}{fusger.test, chisq.test}

\item{Rows}{rows , FLASE = cols}

\item{control}{p adjust"bonferroni","fdr","BH","BY","holm","hochberg","hommel"}

\item{digits}{4}

\item{...}{...}
}
\value{
result
}
\description{
chisq post hoc
}
\examples{

\dontrun{
as.table(cbind("T"=c(30,40,60,120),"S"= c(50,70,60,80)))\%>\%
  chisq_posthoc(Rows  = TRUE)

as.table(cbind("T"=c(30,40,60,120),"S"= c(50,70,60,80)))\%>\%
  chisq_posthoc(Rows  = FALSE)

as.table(cbind("T"=c(30,40,60,120),"S"= c(50,70,60,80)))\%>\%
  chisq_posthoc(test="chisq.test", Rows  = TRUE)
}


}
