% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chisq_retest.R
\name{Chisq_retest}
\alias{Chisq_retest}
\title{Functions to check the chi-square test of a paper}
\usage{
Chisq_retest(
  input,
  ncol = 2,
  colname = NULL,
  rowname = NULL,
  simulate.p.value = FALSE,
  type = "all"
)
}
\arguments{
\item{input}{vectro}

\item{ncol}{matrix col}

\item{colname}{input}

\item{rowname}{input}

\item{simulate.p.value}{default FaLSE fisher exact test}

\item{type}{all, chisq}
}
\value{
chisquare data
}
\description{
Functions to check the chi-square test of a paper
}
\examples{

\dontrun{
##Mina Choi. (2007).
Chisq_retest(c(111, 33, 64, 58), ncol=2 )

Chisq_retest(c(111, 33, 64, 58), ncol=2, type="chisq" )

Chisq_retest(c(111, 33, 64, 58), ncol=2,
colname = c("experience", "Non-experience"),
rowname =c("E-Leanring","FaceToFace"))

}

}
