% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_table.R
\name{chisq_table}
\alias{chisq_table}
\title{hisq table observed/Expected table}
\usage{
chisq_table(
  data,
  v1,
  v2,
  title = "Table",
  type = "res2",
  digits = 3,
  trans = FALSE,
  simulate.p.value = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{v1}{v1}

\item{v2}{v1}

\item{title}{title name}

\item{type}{outtype ct: crosstable, df: data.frame. 'margin','chisq_test','chitable','res1','res2'}

\item{digits}{round 3}

\item{trans}{transpose}

\item{simulate.p.value}{Chi-squared approximation may be incorrect, default= FALSE}
}
\value{
output
}
\description{
hisq table observed/Expected table
}
