% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test.R
\name{chisq_test}
\alias{chisq_test}
\title{chisq_test}
\usage{
chisq_test(
  x,
  y = NULL,
  correct = TRUE,
  p = rep(1/length(x), length(x)),
  rescale.p = FALSE,
  simulate.p.value = FALSE,
  B = 2000
)
}
\arguments{
\item{x}{a numeric vector or matrix. x and y can also both be factors.}

\item{y}{a numeric vector; ignored if x is a matrix. If x is a factor,
y should be a factor of the same length.}

\item{correct}{a logical indicating whether to apply continuity correction when computing
the test statistic for 2 by 2 tables: one half is subtracted from all ∣O−E∣ differences
; however, the correction will not be bigger than the differences themselves.
No correction is done if simulate.p.value = TRUE.}

\item{p}{a vector of probabilities of the same length as x.
An error is given if any entry of p is negative.}

\item{rescale.p}{a logical scalar; if TRUE then p is rescaled (if necessary) to sum to 1.
If rescale.p is FALSE, and p does not sum to 1, an error is given.}

\item{simulate.p.value}{a logical indicating whether to compute p-values by Monte Carlo simulation}

\item{B}{an integer specifying the number of replicates used in the Monte Carlo test.}
}
\value{
A list with class "htest" containing the following components:
}
\description{
chisq_test
}
\examples{
\dontrun{

## From Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(gender = c("F", "M"),
party = c("Democrat","Independent", "Republican"))
(Xsq <- chisq.test(M))  # Prints test summary
Xsq$observed   # observed counts (same as M)
Xsq$expected   # expected counts under the nul
Xsq$residuals  # Pearson residuals
Xsq$stdres     # standardized residuals


x <- matrix(c(12, 5, 7, 7), ncol = 2)
chisq_test(x)$p.value           # 0.4233
chisq_test(x, simulate.p.value = TRUE, B = 10000)$p.value
}
}
