% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test_By.R
\name{chisq_test_By1}
\alias{chisq_test_By1}
\title{one sample chisq test by group}
\usage{
chisq_test_By1(dataset, sel = NULL, correct = FALSE, simulate.p.value = FALSE)
}
\arguments{
\item{dataset}{data.frame}

\item{sel}{select column group}

\item{correct}{chisq correct}

\item{simulate.p.value}{simulate.p.value}
}
\value{
Chi-square results by group
}
\description{
one sample chisq test by group
}
\examples{
\dontrun{
mtcars\%>\%chisq_test_By1("cyl")

mtcars\%>\%chisq_test_By1("am")
iris\%>\%chisq_test_By1("Species")
}
}
