% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test_By.R
\name{chisq_test_By2}
\alias{chisq_test_By2}
\title{twosample chisq.test by group}
\usage{
chisq_test_By2(dataset, v1, v2, sel = NULL, type = "res")
}
\arguments{
\item{v1}{first variable}

\item{v2}{second variable}

\item{sel}{group}

\item{type}{res, long}

\item{data}{data,frame}
}
\value{
chisq result
}
\description{
twosample chisq.test by group
}
\examples{
\dontrun{

mtcars\%>\%  chisq_test_By2("am", "vs", "cyl")

mtcars\%>\% chisq_test_By2("am", "vs", "cyl", type="long")

}


}
