% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test_By.R
\name{chisq_test_By3}
\alias{chisq_test_By3}
\title{그룹별 카이제곱 분석}
\usage{
chisq_test_By3(data, grp, formula, raw = FALSE)
}
\arguments{
\item{data}{data}

\item{grp}{group variable}

\item{formula}{~ vq +v2}

\item{raw}{warning}
}
\value{
table
}
\description{
그룹별 카이제곱 분석
}
\examples{

\dontrun{
#'
chisq_test_By3(mtcars, cyl, ~ am+vs )
chisq_test_By3(mtcars, cyl, ~ am+vs , raw=T)

}
}
