% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test_kge.R
\name{chisq_test_kge}
\alias{chisq_test_kge}
\title{kge_chisq_table chisq table observed/Expected table-}
\usage{
chisq_test_kge(
  dataset,
  v1 = "var1",
  v2 = "var2",
  title = "Table",
  type = "res2",
  digits = 3,
  yadd = 0.4,
  Ncol = NULL,
  trans = FALSE,
  simple = FALSE,
  ko = TRUE,
  simulate.p.value = FALSE,
  correct = FALSE,
  size_bartext = 5,
  strip_size = 16,
  axis_size = 16,
  text_size = 13,
  xlab = "관측기대비율",
  ylab = "var2",
  cramer = "adjust",
  raw = TRUE
)
}
\arguments{
\item{dataset}{data.frame}

\item{v1}{col}

\item{v2}{row}

\item{title}{title}

\item{type}{res2 default}

\item{digits}{3}

\item{yadd}{up yadd}

\item{Ncol}{grid}

\item{trans}{transpose}

\item{simple}{* style}

\item{ko}{graph korean}

\item{simulate.p.value}{exact}

\item{correct}{exactt}

\item{size_bartext}{graph text}

\item{strip_size}{strip}

\item{axis_size}{axis_size}

\item{text_size}{text_size}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{cramer}{cramer cor}

\item{raw}{TRUE data.grame FALSE table data}
}
\value{
multiple data
}
\description{
kge_chisq_table chisq table observed/Expected table-
}
\examples{

\dontrun{
kge_chisq_table(mtcars,"am","cyl")


#chisq test result
vote_chi <- Chisq_retest(c(183, 35, 824, 50), ncol=2,
                        colname = c("Did not vote", "Voted"),
                         rowname =c("All other ages","Youngest"))

vote_chi


# generate matrix
voteMat = matrix(c(183, 35, 824, 50), ncol=2)
colnames(voteMat) = c("Did not vote", "Voted")
rownames(voteMat) = c("All other ages","Youngest")


voteMat\%>\%
  long_df("voted", rowname="youth")\%>\% #long format
  add_rows_freq() \%>\%     # add rows
  kge_chisq_table("voted","youth")    #chisq test

}
}
