% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test.R
\name{chisqTest}
\alias{chisqTest}
\title{chisq_test 2way}
\usage{
chisqTest(Obs, correct = FALSE, digits = 3, cat = TRUE)
}
\arguments{
\item{Obs}{observation}

\item{correct}{Yates' continuity correction}

\item{digits}{3}

\item{cat}{chisq.test form}
}
\value{
chisq
}
\description{
chisq_test 2way
}
\examples{
\dontrun{

table(mtcars$am, mtcars$vs) \%>\% chisq_test()


table(mtcars$am, mtcars$vs) \%>\% chisq_test(correct=F)

## compare original chisq.test
table(mtcars$am, mtcars$vs) \%>\% chisq.test()


##Pearson's Chi-squared test with Yates' continuity correction
table(mtcars$am, mtcars$vs) \%>\% chisq_test(correct=T)
}


}
