% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHM_test.R
\name{cmh_test}
\alias{cmh_test}
\title{cochran_mantel_haenszel_test}
\usage{
cmh_test(
  x,
  y = NULL,
  z = NULL,
  alternative = c("two.sided", "less", "greater"),
  correct = TRUE,
  exact = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{x}{array}

\item{y}{NULL}

\item{z}{NULL}

\item{correct}{TRUE}

\item{exact}{FALSE}

\item{conf.level}{0.95}

\item{alternativec("two.sided", }{"less", "greater")}
}
\value{
result
}
\description{
cochran_mantel_haenszel_test
}
\examples{

\dontrun{
#'
data(Titanic, package = "datasets")
Titanic
cmh_test(temp)
cmh_test(temp)\%>\%tidy()
}
}
