% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_rownames.R
\name{col2row}
\alias{col2row}
\title{column to rownames col2row}
\usage{
col2row(data, col = 1)
}
\arguments{
\item{data}{data.frame}

\item{col}{Columns that should be rows}
}
\value{
data
}
\description{
column to rownames col2row
}
\examples{

\dontrun{
#'
data_text = "Tobacco_Dependence (0,5] (5,10] (10,15] (15,20] (20,98]
No_Nicotine_Dependence 130 210 43 114 20
Nicotine_Dependence 119 267 91 254 67"
#method1
# nt = data_text\%>\% text2df(header=T)\%>\%
#         tibble::column_to_rownames("Tobacco_Dependence")
#method2
nt = data_text\%>\% text2df(header=T)\%>\% jjstat::col2row()
nt=Char2num(nt,1:5)
nt
nt \%>\%tibble()

#NEW data
New = data.frame(
  rownames= c('No_Nicotine_Dependence', 'Nicotine_Dependence'),

  "A" = c(130, 119),
  "B" = c(210, 267),
  "C" = c(43, 91),
  "D" = c(114, 254),
  "E" = c(20, 67)
);New
New\%>\%col2row()
New\%>\%col2row(col="B")
#mtcars
data_mtcars=mtcars\%>\%tibble::rownames_to_column()
data_mtcars\%>\%col2row()



}


}
