% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_capital.R
\name{col_capital}
\alias{col_capital}
\title{col_capital: Convert letters in a selected column to upper or lower case}
\usage{
col_capital(df, col, direction = TRUE)
}
\arguments{
\item{df}{A data.frame. The input dataset.}

\item{col}{A column name (character) or index (numeric) indicating which column to transform.}

\item{direction}{Logical. If TRUE (default), convert to upper case. If FALSE, convert to lower case.}
}
\value{
A data.frame with the selected column transformed to upper or lower case.
}
\description{
col_capital: Convert letters in a selected column to upper or lower case
}
\examples{
\dontrun{
df <- data.frame(
  구분 = c("chisq", "df", "pvalue"),
  criterian = c("Chisq", "df", "p > .05"),
  Value = c("1732.243", "269", "0")
)

# Use column name
col_capital(df, "criterian", TRUE)

# Use column index
col_capital(df, 2, FALSE)
}
}
