% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_change.R
\name{colnames_change}
\alias{colnames_change}
\title{colnames_change}
\usage{
colnames_change(df, ...)
}
\arguments{
\item{df}{df}

\item{...}{input list}
}
\value{
data
}
\description{
colnames_change
}
\examples{
\dontrun{
#'
# 사용 예시
jutLca3 \%>\%
  lca_class_freq() \%>\%
  data.frame() \%>\%
  colnameschange("Class", "N")
}
}
