% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combind_data.R
\name{combine_data}
\alias{combine_data}
\title{A function that combines two matrices of data, for example, to combine observation/expectation tables and p-values.}
\usage{
combine_data(mat_observed_expected, mat_p_value, left = " ", right = "")
}
\arguments{
\item{mat_observed_expected}{matrix  observed/expected}

\item{mat_p_value}{matirx pvalue}

\item{left}{"(" etc input}

\item{right}{")"}
}
\value{
matrix


 \dontrun{

# Example data
observed_expected <- matrix(c(1.0587162, 1.1477347, 0.50877193,
                              1.1269248, 0.5743712, 1.94957983,
                              0.9163776, 1.4060606, 0.03515152,
                              0.9530237, 0.8624314, 1.44531611),
                            nrow = 4, byrow = TRUE,
         dimnames = list(c("공명음", "마찰음", "유기음_경음", "평파열음_평파찰음"),
                           c("H", "H(H)", "L")))

p_value <- matrix(c(0, 0.00001, 0.00000,
                    0, 0.00000, 0.01497,
                    0, 0.00013, 0.00000,
                    0, 0.00000, 0.00102),
                  nrow = 4, byrow = TRUE,
       dimnames = list(c("공명음", "마찰음", "유기음_경음", "평파열음_평파찰음"),
                      c("H", "H(H)", "L")))

sig_sym = add_significance_symbols(p_value)

# Combine the data
combined_data1 <- combine_data(observed_expected, p_value)
 print(combined_data1)
combined_data2 <- combine_data(observed_expected, sig_sym)
 print(combined_data2)

}
}
\description{
A function that combines two matrices of data, for example, to combine observation/expectation tables and p-values.
}
