% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_rows.R
\name{combine_rows}
\alias{combine_rows}
\title{중복된 데이터를 합쳐주는 함수}
\usage{
combine_rows(df, rows = NULL, name_col = 1, value_col = 2, prop_col = 3)
}
\arguments{
\item{df}{df}

\item{rows}{rows =c(1,2,3) 지정된 행을 합침}

\item{name_col}{이름 열 숫자로 입력 가능}

\item{value_col}{빈도 열}

\item{prop_col}{비율 열}
}
\value{
df
}
\description{
중복된 데이터를 합쳐주는 함수
}
\examples{
\dontrun{
#'
# 예제 데이터
library(dplyr)

df <- data.frame(
  term = c("지평선중", "한국게임과학고", "가천초", "감곡초", "감곡초",
           "강호항공고교", "강호항공고", "개야도초", "고", "고부초",
           "고산고", "고산고"),
  Freq = c(1, 1, 1, 2, 1, 1, 1, 1, 4, 1, 1, 2),
  prop.. = c(0.1237624, 0.1237624, 0.1237624, 0.2475248, 0.1237624,
             0.1237624, 0.1237624, 0.1237624, 0.4950495, 0.1237624,
             0.1237624, 0.2475248)
)

# 함수 호출: 디폴트 값 사용 (열 번호)
result <- combine_rows(df)

# 결과 출력
print(result)

# 함수 호출: 열 이름 사용
result_named <- combine_rows(df, name_col = "term", value_col = "Freq", score_col = "prop..")
print(result_named)

#'
}
}
