% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.R
\name{comma}
\alias{comma}
\title{number comma}
\usage{
comma(x, digits = 2, big_mark = ",")
}
\arguments{
\item{x}{vector}

\item{digits}{digits 2}

\item{big_mark}{"," is default}
}
\value{
include comma
}
\description{
number comma
}
\examples{

\dontrun{
comma(3452345)
comma(.12358124331)
}
}
