% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_scale_Df.R
\name{convert_scale_df}
\alias{convert_scale_df}
\title{범위를 변환하는 데이터프레임 전용 함수 정의}
\usage{
convert_scale_df(
  df,
  from_min = -4,
  from_max = 4,
  to_min = 1,
  to_max = 9,
  select_col = everything(),
  remove = FALSE,
  post_fix = "_re"
)
}
\arguments{
\item{df}{df}

\item{from_min}{from_min}

\item{from_max}{from_max}

\item{to_min}{tomin}

\item{to_max}{to max}

\item{select_col}{col}

\item{remove}{FALSE 유지}

\item{post_fix}{_re}
}
\value{
data
}
\description{
범위를 변환하는 데이터프레임 전용 함수 정의
}
\examples{
\dontrun{

convert_scale_df(jen01, from_min = -4, from_max = 4, to_min = 1, to_max = 9,
                 select_col = contains("Q6_"), remove = FALSE, post_fix = "_re") \%>\%
                 str()

# 변환된 데이터 구조 확인
str(jen01_transformed)
}
}
