% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_count_df.R
\name{count_ne}
\alias{count_ne}
\title{0이 아닌 요소의 개수를 세는 함수}
\usage{
count_ne(matrix)
}
\arguments{
\item{matrix}{matrix}
}
\value{
count
}
\description{
0이 아닌 요소의 개수를 세는 함수
}
\examples{

\dontrun{
#'

#  행렬
matrix2 <- matrix(c(0, 1, 2, 0, 0, 0,
                    0, 0, 0, 0, 3, 4), nrow = 6, byrow = TRUE)
matrix2
count_ne(matrix2)

# 주어진 theta 행렬

theta <- matrix(c(5, 0, 0, 0, 0, 0,
                  0, 6, 0, 0, 0, 0,
                  0, 0, 7, 0, 0, 0,
                  0, 0, 0, 8, 0, 0,
                  0, 0, 0, 0, 9, 0,
                  0, 0, 0, 0, 0, 10), nrow = 6, byrow = TRUE)
theta
count_ne(theta)
}

}
