% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_lm_sem.R
\name{cov_lm_sem}
\alias{cov_lm_sem}
\title{Regression analysis with covariance matrix, for paper checking}
\usage{
cov_lm_sem(
  formula = NULL,
  cov,
  n,
  sds = NULL,
  dv,
  ivs,
  method = "none",
  md = FALSE,
  lm = TRUE
)
}
\arguments{
\item{formula}{regression formula}

\item{cov}{cov matrix if your data cor, then you must sds}

\item{n}{sample}

\item{dv}{dv}

\item{ivs}{iv variales}

\item{method}{c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY")}

\item{md}{markdown default FALSE}

\item{lm}{lm option TREW, FALSE sem}
}
\value{
result table, plot, confint sig plot
}
\description{
Regression analysis with covariance matrix, for paper checking
}
\examples{

\dontrun{
#' #method1
cov_lm_sem(formula="mpg ~ cyl+disp+hp+drat+wt" ,n = 32, cov = cov(mtcars[1:6]))
# method 2
cov_lm_sem(dv = "mpg", ivs = c("cyl", "disp", "hp", "drat", "wt"), n = 32, cov = cov(mtcars[1:6]))
# method 2 cor matrix
cov_lm_sem(formula="mpg ~ cyl+disp+hp+drat+wt" ,n = 32, cov = cor(mtcars[1:6]),
           sds = mtcars[1:6]\%>\%sapply(sd))
#est
cov_lm(formula="mpg ~ cyl+disp+hp+drat+wt; qsec ~ mpg +vs;" ,
       n = 32, cov = cov(mtcars[1:8]),lm=FALSE)
#std
cov_lm(formula="mpg ~ cyl+disp+hp+drat+wt; qsec ~ mpg +vs;" ,
       n = 32, cov = cor(mtcars[1:8]), sds = sapply(mtcars[1:8], sd),
       lm=FALSE)

###################
#'
Songkj2024 = lav_matrix_lower2full(
  c(
    1,
    .38,   1,
    .49, .54, 1,
    .58, .33, .45, 1,
    .35, .17, .35, .60, 1,
    .22, .11, .20, .30, .28, 1,
    .46, .25, .46, .61, .67, .31, 1,
    .52, .29, .37, .57, .47, .31, .55, 1,
    -.09, -.03,-.02, -.20,-.05, .03, .00, .01, 1,
    .23, .10, .13, .07, -.03, .06, -.06, .40,-.18, 1   )
)

Songkj2024

colnames(Songkj2024)=c("학습애정열정","학습기회개방성",
                       "솔선수범독립성","효율학습자자아개념",
                       "학업적효능감","사회적효능감","자기조절효능감","경력개발역량",
                       "성별","연령")
rownames(Songkj2024)=c("학습애정열정","학습기회개방성",
                       "솔선수범독립성","효율학습자자아개념",
                       "학업적효능감","사회적효능감","자기조절효능감","경력개발역량",
                       "성별","연령")

#기술통계 데이터
songki2024_stat=  data.frame(
  var = c("학습애정열정","학습기회개방성",
          "솔선수범독립성","효율학습자자아개념",
          "학업적효능감","사회적효능감","자기조절효능감","경력개발역량",
          "성별","연령"),
  mean = c(4.14, 3.86, 4.14, 3.86, 3.61, 3.44, 3.95, 3.89, 0.65, 1.70),
  sd =   c(.52, .72, .70, .59, .64, .82, .54, .60, 0.48, 0.82)

);songki2024_stat

Songkj2024_cov = cor2cov(Songkj2024, sds= songki2024_stat$sd)
# Songkj2024_cov
# original
Songkj2024_model_1="
학업적효능감 ~ 학습애정열정 + 학습기회개방성 + 솔선수범독립성 + 효율학습자자아개념+ 성별 + 연령
"
Songkj2024_sem1 = sem( Songkj2024_model_1,
                       sample.cov = Songkj2024 , sample.nobs =  389  )

#cor, std
cov_lm(formula="학업적효능감 ~ 학습애정열정 + 학습기회개방성 + 솔선수범독립성 + 효율학습자자아개념 + 성별 + 연령" ,n = 389, cov = Songkj2024)
# cov
cov_lm(formula="학업적효능감 ~ 학습애정열정 + 학습기회개방성 + 솔선수범독립성 + 효율학습자자아개념 + 성별 + 연령" ,n = 389, cov = Songkj2024, sds = songki2024_stat$sd)
# cov
cov_lm(formula="학업적효능감 ~ 학습애정열정 + 학습기회개방성 + 솔선수범독립성 + 효율학습자자아개념 + 성별 + 연령" ,n = 389, cov = Songkj2024_cov)
}

}
