% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramers_v.R
\name{cramers_v}
\alias{cramers_v}
\title{Tcramers_v}
\usage{
cramers_v(data, type = "cramer", digits = 3, correct = FALSE)
}
\arguments{
\item{data}{data table}

\item{type}{cramer}

\item{digits}{3}
}
\value{
cramer correlation
}
\description{
Tcramers_v
}
