% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_print.R
\name{cut_print_all}
\alias{cut_print_all}
\title{overall cut}
\usage{
cut_print_all(data, cut = 0, imp = "", digits = 3)
}
\arguments{
\item{data}{data}

\item{cut}{cut =0.5}

\item{imp}{"}

\item{digits}{3}
}
\value{
data
}
\description{
overall cut
}
