% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dall.R
\name{dall}
\alias{dall}
\title{view data.frame}
\usage{
dall(data, type = "all", n = 10)
}
\arguments{
\item{data}{tibble}

\item{type}{all, head}

\item{n}{rows n = 10 (default)}
}
\value{
data.frame
}
\description{
view data.frame
}
