% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapate.R
\name{datapaste}
\alias{datapaste}
\title{exceldata to dataframe}
\usage{
datapaste(type = "matrix", x = "clipboard")
}
\arguments{
\item{type}{"clipboard" to table data. dataframe, tibble}

\item{x}{"clipboard"}
}
\value{
paste data.frame
}
\description{
exceldata to dataframe
}
\examples{
#first you must copy data (Ctrl+C)
## then excute below ctrl+Enter
df = datapaste()
df = datapaste("tibble")
df = datapaste("table")
df = datapaste("data.frame")


datapaste("datapasta")
New = data.frame(row.names = c("1", "2", "3", "4", "5", "6"),
                   응답자 = c("응답1", "응답2", "응답3", "응답4", "응답5", "응답6"),
                   role_v_inst1 = c(5, 5, 5, 5, 4, 5),
                   role_v_inst2 = c(5, 5, 5, 5, 5, 5),
                   role_v_inst3 = c(4, 4, 5, 5, 4, 4),
                   role_v_inst4 = c(5, 4, 5, 5, 3, 4))


}
