% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydescriptive.R
\name{Describe}
\alias{Describe}
\title{Describedescriptive statistics cusum}
\usage{
Describe(myobject, ...)
}
\arguments{
\item{myobject}{data.frame, matrix}

\item{...}{column variable}
}
\value{
size   MEAN    SD  MIN  MAX
}
\description{
Describedescriptive statistics cusum
}
\examples{
\dontrun{
mysummary(mtcars, "mpg","wt","hp")
## size      MEAN         SD    MIN     MAX
## 1   32  20.09062  6.0269481 10.400  33.900
## 2   32   3.21725  0.9784574  1.513   5.424
## 3   32 146.68750 68.5628685 52.000 335.000


}

}
