% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_diagram.R
\name{diagram2}
\alias{diagram2}
\title{diagram sem plot 2}
\usage{
diagram2(
  lav_obj,
  whatLabels = "est",
  rotation = 4,
  nDigits = 2,
  residScale = 18,
  mar = c(2, 8, 3, 10),
  edge.label.cex = 1,
  sizeMan = 8,
  sizeMan2 = 4,
  sizeLat = 10,
  style = "lisrel",
  sig = TRUE,
  groups = FALSE
)
}
\arguments{
\item{lav_obj}{lavaan}

\item{whatLabels}{est, std, model(sig=FALSE)}

\item{rotation}{rotation= 4}

\item{nDigits}{3}

\item{residScale}{residScale=18}

\item{mar}{c(2,8,3,10)}

\item{edge.label.cex}{edge.label.cex= 1}

\item{sizeMan}{sizeMan = 8,}

\item{sizeMan2}{sizeMan2 = 4,}

\item{sizeLat}{sizeLat = 10,}

\item{style}{style="lisrel"}

\item{sig}{sig=TRUE}

\item{groups}{true color}
}
\value{
plot
}
\description{
diagram sem plot 2
}
\examples{

\dontrun{

example(sem)
fit\%>\% diagram2()

}


}
