% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_diagram_model.R
\name{diagram_model}
\alias{diagram_model}
\title{Drawing Structural Equation Models Using the lavaan Syntax (park JH PhD)}
\usage{
diagram_model(
  model,
  residuals = FALSE,
  whatLabels = "model",
  rotation = 2,
  residScale = 12,
  layout = "tree",
  edge.label.cex = 1,
  sizeMan = 8,
  sizeMan2 = 4,
  sizeLat = 10,
  sizeLat2 = 6,
  style = "lisrel",
  sig = FALSE,
  exoVar = FALSE,
  exoCov = TRUE,
  curve = 1.5,
  asize = 2,
  mar = c(3, 8, 3, 10),
  nDigits = 2,
  sample.nobs = 100,
  border.width = 2,
  edge.width = 1.5,
  groups = FALSE,
  growth = FALSE,
  structural = FALSE,
  type = "plot"
)
}
\arguments{
\item{model}{lavaan syntax}

\item{residuals}{residauls TRUE}

\item{whatLabels}{"model}

\item{rotation}{rotation=2}

\item{residScale}{residScale=12}

\item{layout}{layout="tree"}

\item{edge.label.cex}{1}

\item{sizeMan}{8}

\item{sizeMan2}{4}

\item{sizeLat}{10}

\item{sizeLat2}{6}

\item{style}{lisrel}

\item{sig}{FALSE}

\item{exoVar}{FALSE}

\item{exoCov}{TRUE}

\item{curve}{1.5}

\item{asize}{1.5}

\item{mar}{c(2,8,3,10)}

\item{nDigits}{2}

\item{sample.nobs}{sample.nobs = 100}

\item{border.width}{border.width=1}

\item{edge.width}{edge.width=1.5}

\item{groups}{FALSE}

\item{growth}{growth= FALSE, TRUE  model.type ="grouwth"}

\item{structural}{structural = FALSE}

\item{type}{type ="plot, and res is lavaan simdata result}
}
\value{
plot
}
\description{
Drawing Structural Equation Models Using the lavaan Syntax (park JH PhD)
}
\examples{
\dontrun{
#'
model_test <- '
  # latent variable definitions
     ind60 =~ x1 + x2 + x3
     dem60 =~ y1 + a*y2 + b*y3 + c*y4
     dem65 =~ y5 + a*y6 + b*y7 + c*y8

  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60

  # residual correlations
    # y1 ~~ y5
    # y2 ~~ y4 + y6
    # y3 ~~ y7
    # y4 ~~ y8
    # y6 ~~ y8
'

HSmodel <- ' visual  =~ x1 + x2 + x3
             textual =~ x4 + x5 + x6
             speed   =~ x7 + x8 + x9 '


m1a <- "
y1 =~ x1 + x2 +x5
y2 =~ x3 + x4 +x6
y1 ~~ y2"

diagram_model(model= HSmodel, groups=F,  residuals=F)

diagram_model(model= model_test, style="ram", groups=F)
diagram_model(model= model_test, residuals=T)
diagram_model(model= m1a, residuals=T)

#'
}
}
