% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{e}
\alias{e}
\title{Google translaste english  to koean}
\usage{
e(
  text = "hello let's start translate english to korean.",
  slang = "en",
  elang = "ko",
  show = "translate"
)
}
\arguments{
\item{text}{Input text}

\item{slang}{english}

\item{elang}{korean}

\item{show}{'translate', 'browser', 'all' select 3 opttions}
}
\description{
Google translaste english  to koean
}
\examples{
#eexamples code
\dontrun{

e("Translating English to Korean is the default option.")
}



}
