% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAI_index.R
\name{EAI_index}
\alias{EAI_index}
\title{EAI_index: Compute and visualize the Eduteck Acceptance Index (EAI) for given profiles}
\usage{
EAI_index(
  df,
  profile = NULL,
  caption = "",
  sel_col = "Class",
  level_start = 9,
  jump = 5,
  adj = 0,
  size_label = 5,
  size_text = 5,
  color_line = "gray20",
  color_label = "gray20",
  lpos = 2.35,
  rpos = 2.65,
  print = TRUE,
  web = FALSE
)
}
\arguments{
\item{df}{Data frame containing the necessary columns for EAI calculation.}

\item{profile}{A specific profile (e.g., a category from a column) to filter the data. Default is NULL, meaning all data will be used.}

\item{caption}{Optional caption for the plot title. Default is an empty string.}

\item{sel_col}{The column name in the data frame used for selecting profiles. Default is "Class".}

\item{level_start}{Starting level for categorization in the plot. Default is 9.}

\item{jump}{Increment to calculate subsequent levels. Default is 5.}

\item{adj}{Adjustment factor for level calculations. Default is 0.}

\item{size_label}{Font size for level annotations. Default is 4.}

\item{size_text}{Font size for score labels in the plot. Default is 5.}

\item{color_line}{Color for the dashed lines in the plot. Default is "gray20".}

\item{color_label}{Color for the level annotations. Default is "gray20".}

\item{lpos}{Horizontal position for left-side annotations. Default is 2.35.}

\item{rpos}{Horizontal position for right-side annotations. Default is 2.65.}

\item{print}{Logical indicating whether to print the summary table. Default is TRUE.}

\item{web}{Logical indicating whether to print web the summary table. Default is FALSE.}
}
\value{
A ggplot object visualizing the EAI index along with calculated levels and annotations.
}
\description{
EAI_index: Compute and visualize the Eduteck Acceptance Index (EAI) for given profiles
}
\examples{
\dontrun{
df <- data.frame(
  Class = rep(c("A", "B"), each = 10),
  EAI = rnorm(20, 15, 5),
  PU_sum = rnorm(20, 20, 4),
  PEOU_sum = rnorm(20, 18, 3),
  ATT_sum = rnorm(20, 17, 4),
  USE_sum = rnorm(20, 19, 5),
  EAI_check = rnorm(20, 16, 2)
)
g <- EAI_index(df, profile = "A", caption = "EAI Analysis", sel_col = "Class")
print(g)
}
}
