% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect1.R
\name{effect1}
\alias{effect1}
\title{effec1 is the path coefficient of the structural equation.}
\usage{
effect1(
  x,
  format = "markdown",
  op1 = "~",
  op2 = ":=",
  what = "std",
  digits = 3,
  layout = "tree2",
  curve = 1,
  mar = c(8, 8, 8, 8),
  residScale = 15,
  sizeLat = 12,
  sizeLat2 = 6,
  sizeMan = 8,
  sizeMan2 = 4,
  edge.label.cex = 0.9,
  label.cex = 1.5,
  asize = 1.4,
  name_change = FALSE,
  label_list = NULL,
  type = "res"
)
}
\arguments{
\item{x}{lavaan result}

\item{format}{'markdown', 'html'}

\item{op1}{first'~'}

\item{op2}{first ':='  newparameters}

\item{what}{default 'std', Can be changeed to 'est'}

\item{digits}{round 3}

\item{layout}{'tree2'}

\item{curve}{curve =1}

\item{mar}{margins default c(8,8,8,8)}

\item{residScale}{length of residuals. default 15}

\item{sizeLat}{Horizontal length of latent variable}

\item{sizeLat2}{Vertical length of latent variable}

\item{sizeMan}{horizontal length of the item}

\item{sizeMan2}{Vertical length of the item}

\item{edge.label.cex}{edge size}

\item{label.cex}{label size}

\item{asize}{arrow size}

\item{name_change}{When set to default, the variable name of the data is used, and when set to TRUE, a new variable name can be entered. Just enter them all in label_list using the c() function.}

\item{label_list}{label list}

\item{type}{type output}
}
\description{
effec1 is the path coefficient of the structural equation.
}
\examples{

\dontrun{
## change label_list
abel_listnode.name =list(
  list(node="f인정욕구", to="인정신념"),
  list(node="f파국화",  to="불안신념1"),
  list(node="f과잉불안", to="불안신념2"),
  list(node="f무력감",  to="의존신념"),
  list(node="f개인완벽", to="해결신념"),
  list(node="IRB",  to="비합리적신념"),
  list(node="PREACT",  to="진로준비행동"),
  list(node="DECISION",  to="진로결정수준"),
  list(node="f정보수집",  to="정보수집"),
  list(node="f직업체험",  to="직업체험"),
  list(node="f진로탐색",  to="정보탐색"),
  list(node="f구체성",  to="구체성"),
  list(node="f확신성",  to="확신성"),
  list(node="f적합성",  to="적합성"))
}


}
