% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fei_Chisq_test.R
\name{fei_Chisq_test}
\alias{fei_Chisq_test}
\title{fei분석을위한 카이제곱함수}
\usage{
fei_Chisq_test(tabledf, col = "빈도", name = "구분", ns = FALSE, sep = " / ")
}
\arguments{
\item{tabledf}{빈도분석 Freq_table 데이터}

\item{col}{"빈도"}

\item{name}{구분}

\item{ns}{FALSE 생성한 것사용 ,TRUE RVAideMemoire로 행렬로 나타남}

\item{sep}{sep="/"}
}
\value{
df
}
\description{
fei분석을위한 카이제곱함수
}
\examples{
\dontrun{
#'
eduteck_s$Q31\%>\% na_remover() \%>\% Freq_table(prop=TRUE) \%>\%
  fei_Chisq_test()
}
}
