% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{filter and,  or function}
\usage{
filter_rows(data, op = "or", ...)
}
\arguments{
\item{data}{data}

\item{op}{operatror or, and}

\item{...}{condition}
}
\value{
data
}
\description{
filter and,  or function
}
\examples{

\dontrun{
#'
mtcars \%>\% filter_rows(cyl = 6)
mtcars \%>\% filter_rows(vs = 1)
mtcars \%>\% filter_rows(cyl = 6, vs = 1, op ="and")
mtcars \%>\% filter_rows(op="and", cyl = 6, vs = 1)
mtcars \%>\% filter_rows("and", cyl = 6, vs = 1 )

mtcars \%>\% filter_rows("or", cyl = 6, vs = 1 )
mtcars \%>\% filter_rows(op= "or", cyl = 6, vs = 1 )
mtcars \%>\% filter_rows(cyl = 6, vs = 1, op ="or")

mtcars \%>\% filter_rows(cyl = 6, vs = 1, gear = 4, op = "and")
mtcars \%>\% filter_rows(cyl = 6, vs = 1, gear = 4, op = "or")
mtcars \%>\% filter_rows(cyl = 6, vs = 1, gear = 4, op = "or")

mtcars \%>\% filter_rows(mpg > 25) #error
mtcars \%>\% filter(mpg > 25)
}

}
