% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_find_paths.R
\name{find_paths}
\alias{find_paths}
\title{Resolver knows the path}
\usage{
find_paths(data, type = "paths", est = "Original")
}
\arguments{
\item{data}{boot data}

\item{type}{es, all, est, ind, paths(default)}

\item{est}{Original}
}
\value{
result text
}
\description{
Resolver knows the path
}
\examples{

\dontrun{
#'
# 예제 데이터
satpls_boot <- list(boot = list(paths = data.frame(
  Original = c(0.578959128, 0.200724200, 0.275149576, 0.848344408, 0.105477650, -0.002753995, 0.676655529, 0.122144566, 0.589330650, 0.495479322),
  Mean.Boot = c(0.57466537, 0.20926883, 0.28748175, 0.84907549, 0.09629728, -0.02304841, 0.68789406, 0.13300281, 0.59112195, 0.48611617),
  Std.Error = c(0.04941293, 0.05440409, 0.07095248, 0.02022514, 0.05731729, 0.06209752, 0.06286102, 0.08889983, 0.08591013, 0.07920673),
  perc.025 = c(0.46120153, 0.09494506, 0.14255079, 0.80126496, -0.02555946, -0.15182035, 0.57655200, -0.01836564, 0.41297885, 0.34267859),
  perc.975 = c(0.65528114, 0.30696127, 0.41537966, 0.88210331, 0.19678402, 0.09426639, 0.81986383, 0.32270662, 0.75274086, 0.63147051),
  row.names = c("IMAG -> EXPE", "IMAG -> SAT", "IMAG -> LOY", "EXPE -> QUAL", "EXPE -> VAL", "EXPE -> SAT", "QUAL -> VAL", "QUAL -> SAT", "VAL -> SAT", "SAT -> LOY")
)))

# 함수 실행
find_paths_all(satpls_boot$boot$paths)
find_paths_all(satpls_boot$boot$paths,"all")
}

}
