% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_find_paths.R
\name{find_paths1}
\alias{find_paths1}
\title{Resolver knows the path}
\usage{
find_paths1(
  data,
  type = "paths",
  path_col = "paths",
  est = "Original",
  se = "Std.Error"
)
}
\arguments{
\item{data}{boot data}

\item{type}{es, all, est, ind, paths(default)}

\item{path_col}{paths}

\item{est}{Original}

\item{se}{Ste.Error}
}
\value{
paths types
}
\description{
Resolver knows the path
}
\examples{

\dontrun{
#'

# Example usage
sdata <- data.frame(
  paths = c("IMAG -> EXPE", "IMAG -> SAT", "IMAG -> LOY", "EXPE -> QUAL", "EXPE -> VAL", "EXPE -> SAT",
            "QUAL -> VAL", "QUAL -> SAT", "VAL -> SAT", "SAT -> LOY"),
  Original = c(0.578959128, 0.200724200, 0.275149576, 0.848344408, 0.105477650, -0.002753995,
               0.676655529, 0.122144566, 0.589330650, 0.495479322),
  Mean.Boot = c(0.58232765, 0.21209094, 0.27486564, 0.84793899, 0.10992051, -0.01008222,
                0.67780527, 0.15132543, 0.56312717, 0.50757755),
  Std.Error = c(0.04358293, 0.06056988, 0.07729000, 0.01864573, 0.06968826, 0.06507903,
                0.07760527, 0.08440236, 0.08723293, 0.07856694),
  perc.025 = c(0.497644719, 0.107561558, 0.140641502, 0.815794494, -0.025566471, -0.147788188,
               0.519959683, -0.008629516, 0.392664443, 0.351774130),
  perc.975 = c(0.6646198, 0.3275312, 0.4161131, 0.8828875, 0.2545785, 0.1086856,
               0.8166486, 0.2991045, 0.7012494, 0.6556333)
)

#'
find_paths(sdata)
#'
find_paths(sdata)\%>\%find_paths_cal() #OK
find_paths_cal(find_paths(sdata)) #OK
find_paths_cal(sdata) #error
find_paths(satpls_boot)

}



}
