% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_find_paths.R
\name{find_paths_cal}
\alias{find_paths_cal}
\title{real paths calculation est}
\usage{
find_paths_cal(paste_path_result, type = "all", digits = 3, est = "Original")
}
\arguments{
\item{paste_path_result}{find_paths data}

\item{type}{all=res, ind,  est}

\item{digits}{round 3}

\item{est}{Original  or Mean.Boot}
}
\value{
data
}
\description{
real paths calculation est
}
\examples{


\dontrun{
#'

# Example usage
sdata <- data.frame(
  paths = c("IMAG -> EXPE", "IMAG -> SAT", "IMAG -> LOY",
   "EXPE -> QUAL", "EXPE -> VAL", "EXPE -> SAT",
            "QUAL -> VAL", "QUAL -> SAT",
            "VAL -> SAT", "SAT -> LOY"),
  Original = c(0.578959128, 0.200724200, 0.275149576,
   0.848344408, 0.105477650, -0.002753995,
               0.676655529, 0.122144566, 0.589330650, 0.495479322),
  Mean.Boot = c(0.58232765, 0.21209094, 0.27486564,
  0.84793899, 0.10992051, -0.01008222,
                0.67780527, 0.15132543, 0.56312717, 0.50757755),
  Std.Error = c(0.04358293, 0.06056988, 0.07729000,
  0.01864573, 0.06968826, 0.06507903,
                0.07760527, 0.08440236, 0.08723293, 0.07856694),
  perc.025 = c(0.497644719, 0.107561558, 0.140641502,
   0.815794494, -0.025566471, -0.147788188,
               0.519959683, -0.008629516, 0.392664443, 0.351774130),
  perc.975 = c(0.6646198, 0.3275312, 0.4161131,
   0.8828875, 0.2545785, 0.1086856,
               0.8166486, 0.2991045, 0.7012494, 0.6556333)
)

#'

##realdata analysis
data(satisfaction)
sat_path = plspm_paths(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),
  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY"))
)
sat_path

# blokcs
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)

sat_blocks1

satpls_boot = plspm_sem(satisfaction, sat_path, sat_blocks1, scaled = FALSE,
                        boot.val =TRUE, br=100)
#' find_paths(satpls_boot)
find_paths_cal(find_paths(satpls_boot,"all") )

#값을 계산하여 보기
find_paths(jutpls_boot,"all")\%>\%find_paths_cal()
#Mean.Boot자료 보기
find_paths(jutpls_boot,"all", est="Mean.Boot")\%>\%find_paths_cal()



}


}
