% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_select_col.R
\name{format_col}
\alias{format_col}
\title{select_format}
\usage{
format_col(data, ..., nsmall = 3, digits = 3)
}
\arguments{
\item{data}{data}

\item{...}{col selection}

\item{nsmall}{default 3}

\item{digits}{round 3}
}
\value{
data
}
\description{
select_format
}
\examples{


\dontrun{


# Usage example
data2 <- data.frame(a = rnorm(5), b = rnorm(5), c = rnorm(5), d = rnorm(5), e = rnorm(5))
format_col(data2, 1, 2)
format_col(data2, 1:5)
format_col(data2, a)
format_col(data2, a:e)
format_col(data2, "a")
}
}
