% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number.R
\name{format_number}
\alias{format_number}
\title{Define a function: Convert numbers to scientific notation or common number}
\usage{
format_number(number)
}
\arguments{
\item{number}{number}
}
\value{
formatting
}
\description{
Define a function: Convert numbers to scientific notation or common number
}
\examples{
\dontrun{
## number of data
nchar(sub("\\\\d+\\\\.", "", as.character(0.123456)))

numbers <- c(0.123456, 123.456789, 0.0000123456)
sapply(numbers, format_number)

for (i in seq_along(numbers)) {
   cat("Original:", numbers[i], "Formatted:", formatted_numbers[i], "\n")}
}

}
