% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq_rowcounter.R
\name{Freq_rowcounter}
\alias{Freq_rowcounter}
\title{Freq_rowcounter}
\usage{
Freq_rowcounter(df, type = "counter")
}
\arguments{
\item{df}{df}

\item{type}{counter, term, freq_counter, freq_term}
}
\value{
df
}
\description{
Freq_rowcounter
}
\examples{
\dontrun{

 Freq_rowcounter(df = eduteck2, type = "counter")
Freq_rowcounter(df = eduteck2, type = "freq_counter")
}
}
