% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_freq.R
\name{Freq_table}
\alias{Freq_table}
\title{Frequency Table output function}
\usage{
Freq_table(
  data,
  format = "markdown",
  title = "Frequency output function",
  sort = FALSE,
  prop = TRUE,
  arrange = ""
)
}
\arguments{
\item{data}{data.frame}

\item{format}{markdown, html}

\item{title}{your name}

\item{sort}{sorting default FALSE}

\item{prop}{default TRUE}

\item{arrange}{input 'desc' . default none ""}
}
\value{
table
}
\description{
Frequency Table output function
}
\examples{
\dontrun{
mtcars$am\%>\% SEM212::Freq_table()
}


}
