% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_Freq_table.R
\name{Freq_table_apa}
\alias{Freq_table_apa}
\title{Interpreting frequency analysis results from Freq_table}
\usage{
Freq_table_apa(df, title = "", print = TRUE, digits = 2)
}
\arguments{
\item{df}{data}

\item{title}{title =""}

\item{print}{df output}

\item{digits}{2}
}
\value{
apa
}
\description{
Interpreting frequency analysis results from Freq_table
}
\examples{

\dontrun{
#'
Mtcars = mtcars

Mtcars$am = factor(Mtcars$am, levels=c(0,1), labels= c("automatic","manual" ))
Mtcars$vs  = factor(Mtcars$vs, levels=c(0,1), labels= c("V-shaped","straight" ))
Mtcars$cyl  = factor(Mtcars$cyl, levels=c(4,6,8), labels= c("cyl-4","cyl-6","cyl-8"))

Mtcars \%>\% Freq_table("am","vs", prop=T) \%>\% Freq_table_apa()
Mtcars \%>\% table_freq("vs","am","cyl") \%>\% Freq_table_apa()
Mtcars \%>\% table_freq("vs","am","cyl", angle=90)\%>\% Freq_table_apa()
}

}
