% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq_table_colnames.R, R/table_df_colnames.R
\name{Freq_table_colnames}
\alias{Freq_table_colnames}
\title{Freq_table_colnames caange korean names}
\usage{
Freq_table_colnames(df, input = "구분")

Freq_table_colnames(df, input = "구분")
}
\arguments{
\item{df}{df}

\item{input}{"구분}
}
\value{
df

df
}
\description{
Freq_table_colnames caange korean names

table_df_colnames
}
\examples{
\dontrun{
#' table_df_multit(eduteck2 \%>\% select(contains("Q5_"), -Q5_4_TEXT)) \%>\%
arrange(desc(Freq)) \%>\%Freq_table_colnames()
}

\dontrun{


}
}
