% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_scores.R
\name{fs}
\alias{fs}
\title{Add factor score calculation functions and data}
\usage{
fs(data, range = 1:ncol(data), nfactor = NULL, prefix = "", rotate = "varimax")
}
\arguments{
\item{data}{data}

\item{range}{range default 1:ncol(data)}

\item{nfactor}{number of factor}

\item{prefix}{data prefic}

\item{rotate}{varimax}
}
\value{
data
}
\description{
Add factor score calculation functions and data
}
\examples{


\dontrun{
#'
# 활용방법 1
factor_scores(mtcars,
              range = c("disp", "hp", "drat", "wt", "qsec"),
              nfactor = 2, prefix="a")

mtcars \%>\% select("disp", "hp", "drat", "wt", "qsec") \%>\% factor_scores(nfactor=2)

# 활용방법 2
factor_scores(mtcars, range = c(3:7), nfactor = 2)

factor_scores(mtcars, c("disp", "hp", "drat", "wt", "qsec"), nfactor = 2)
factor_scores(mtcars, c(3, 4, 5, 6, 7), nfactor = 2)

factor_scores(mtcars, range = "disp:qsec", nfactor = 2)


mtcars \%>\% select("disp", "hp", "drat", "wt", "qsec") \%>\%
  factor_scores(c("disp","hp"),nfactor=2, prefix="a") \%>\%
  factor_scores(c("drat", "wt", "qsec"),nfactor=3, prefix="b") \%>\%head()

}
}
