% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobel_test.R
\name{general_Sobel_test}
\alias{general_Sobel_test}
\title{Sobeltest toal}
\usage{
general_Sobel_test(a, SE, roburst = FALSE, n = NULL, k = 2, digits = 5)
}
\arguments{
\item{a}{est}

\item{SE}{se}

\item{roburst}{roburst T whide correction}

\item{n}{sample size}

\item{k}{number of regressin coefs, default 2}

\item{digits}{5}

\item{conf}{0.90, 0.95, 0.99, 0.999}

\item{type}{all, res, sobel, aroian, goodman}
}
\value{
result
}
\description{
Sobeltest toal
}
\examples{

\dontrun{
#'
# Example usage
a <- 0.5  # Example coefficient a
b <- 0.3  # Example coefficient b
sa <- 0.1 # Standard error of a
sb <- 0.05 # Standard error of b

Sobel_test(a, b, sa, sb)
Sobel_test(a, b, sa, sb, "all")
Sobel_test(a, b, sa, sb, "Sobel")

Sobel_test(a, b, sa, sb,"aroian")

Sobel_test(a, b, sa, sb,"Goodman")
#'
a <- c(0.5, 0.3, 0.4)
SE <- c(0.1, 0.1, 0.1)
general_Sobel_test(a, SE)
general_Sobel_test(a, SE)\%>\%unite_ci(digits=4)


}
}
