% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCov.R
\name{GetCov}
\alias{GetCov}
\title{text to get covariance matrix}
\usage{
GetCov(
  x,
  lower = TRUE,
  diagonal = TRUE,
  sds = NULL,
  names = paste("V", 1:nvar, sep = "")
)
}
\arguments{
\item{x}{text}

\item{lower}{True}

\item{diagonal}{TRUE}

\item{sds}{NULL if cor is need}

\item{names}{mat colnames,}
}
\value{
matrix
}
\description{
text to get covariance matrix
}
\examples{

\dontrun{
#'
library(jjstat)

lower <- '
 11.834
  6.947   9.364
  6.819   5.091  12.532
  4.783   5.028   7.495   9.986
 -3.839  -3.889  -3.841  -3.625  9.610
-21.899 -18.831 -21.748 -18.775 35.522 450.288 '

GetCov(lower)
#' GetCov(lower, names = c("anomia67", "powerless67", "anomia71", "powerless71","education", "sei"))

wheaton.cov <-
  GetCov(lower, names = c("anomia67", "powerless67",
                          "anomia71", "powerless71",
                          "education", "sei"))
wheaton.cov
wheaton.model <- '
  # latent variables
    ses     =~ education + sei
    alien67 =~ anomia67 + powerless67
    alien71 =~ anomia71 + powerless71
  # regressions
    alien71 ~ alien67 + ses
    alien67 ~ ses
  # correlated residuals
    anomia67 ~~ anomia71
    powerless67 ~~ powerless71
'
fit <- sem(wheaton.model,
           sample.cov = wheaton.cov,
           sample.nobs = 932)
summary(fit, standardized = TRUE)
diagram2(fit, curve=2)
}


}
