% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtl_df.R
\name{gtl}
\alias{gtl}
\title{Googole Translate korean to english}
\usage{
gtl(text = "hi.", slang = "ko", elang = "en", show = "data")
}
\arguments{
\item{text}{input your text}

\item{slang}{start language. default Korean}

\item{elang}{translated language. default English}

\item{show}{show Methode. 'translate' is console, 'browser'  is Print to your computer's browser, 'all' are There are two ways to output it. 'data' is char data}
}
\description{
Googole Translate korean to english
}
\examples{
# example code
\dontrun{
#korean to english : default
g("안녕하세요 번역을 시작합니다!(by 박중희).", slang ="ko", elang = "en")
g("안녕하세요 번역을 시작합니다!(by 박중희).", "ko", "en")
# english to korean
g("Hello, let’s start translating! (by Park Joong-hee", slang ="ko", elang = "en")
g("Hello, let’s start translating! (by Park Joong-hee","ko","en")
#japan to korean
g("こんにちは翻訳を始めます！（byパク・ジュンヒ", slang ="ja", elang = "ko")

# ko: korean , en: English , ja: Japan, china : zh
# https://cloud.google.com/translate/docs/languages?hl=ko
}
}
