% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtl_df.R
\name{gtl_df}
\alias{gtl_df}
\title{Google translate data.frame
#gtl_df() dataframe 번역 ----------
#영어논문을 만들 때, 표에 들어가는 단어를 번역하여 넣기
#너무 많은 수는 오래걸림--> 웹페이지를 갔다와야 하므로...
Title}
\usage{
gtl_df(df, col = 1)
}
\arguments{
\item{df}{data.frame}

\item{col}{Select the column you want to translate}
}
\description{
Google translate data.frame
#gtl_df() dataframe 번역 ----------
#영어논문을 만들 때, 표에 들어가는 단어를 번역하여 넣기
#너무 많은 수는 오래걸림--> 웹페이지를 갔다와야 하므로...
Title
}
\examples{
\dontrun{
 df0 <-  data.frame(word=c("온라인","진행","교육"),
                   n=c(100, 200, 300))
 df0 \%>\% gtl_df()
}


}
