% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{hist2}
\alias{hist2}
\title{histogram}
\usage{
hist2(data, color = "lightblue", linecolor = "red")
}
\arguments{
\item{data}{vector}

\item{color}{fill color}

\item{linecolor}{line color}
}
\value{
graph histogram
}
\description{
histogram
}
\examples{
 \dontrun{

set.seed(123)

data <- rnorm(100)

hist2(data)

 }

}
