% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{histogram}
\usage{
histogram(data, color = "gray60")
}
\arguments{
\item{data}{vector}

\item{color}{color}
}
\value{
plot
}
\description{
histogram
}
\examples{
\dontrun{
set.seed(123)
data <- rnorm(100)
histogram(data)

}

}
