% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC}
\alias{ICC}
\title{ICC}
\usage{
ICC(lmedata, type = "mixed", digits = 4)
}
\arguments{
\item{lmedata}{lmer result}

\item{type}{'mixed', 'logit', 'pois'}

\item{digits}{digits = 3 default}
}
\description{
ICC
}
\examples{
\dontrun{
data(sleepstudy)
sleepstudy |> str()
lmer(Reaction ~ Days + (Days | Subject), sleepstudy) |> ICC()


}
}
