% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_lme4.R
\name{icc_lme4}
\alias{icc_lme4}
\title{Calculate Intraclass Correlation Coefficient (ICC) using lme4}
\usage{
icc_lme4(y, cluster)
}
\arguments{
\item{y}{numeric vector. Outcome variable.}

\item{cluster}{vector or factor. Grouping (cluster) variable.}
}
\value{
A data.frame with the following components:
\describe{
  \item{ICC}{Intraclass correlation coefficient}
  \item{between_var}{Between-cluster variance (tau^2)}
  \item{within_var}{Within-cluster variance (sigma^2)}
  \item{J}{Number of clusters}
  \item{kbar}{Average cluster size}
}
}
\description{
This function computes the Intraclass Correlation Coefficient (ICC)
based on a random-intercept-only linear mixed-effects model.
}
\examples{
\dontrun{

# Example using a built-in R dataset
data(ToothGrowth)

# ICC of tooth length clustered by supplement type
icc_lme4(
  y = ToothGrowth$len,
  cluster = ToothGrowth$supp
)

# ICC of tooth length clustered by dose
icc_lme4(
  y = ToothGrowth$len,
  cluster = ToothGrowth$dose


# Built-in dataset
data(mtcars)

# Single ICC calculation
icc_lme4(
  y = mtcars$mpg,
  cluster = mtcars$cyl
)

# Batch ICC calculation using icc_batch()
icc_batch(
  data = mtcars,
  vars = c("mpg", "hp", "wt"),
  cluster = "cyl"
)
}
}
