% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_indicator_validity_ml.R
\name{indicator_validity_ml}
\alias{indicator_validity_ml}
\title{Indicator validity table for (Multi-level) CFA}
\usage{
indicator_validity_ml(fit, level = c("within", "between"), digits = 3)
}
\arguments{
\item{fit}{lavaan SEM / ML-SEM object}

\item{level}{"within" or "between"}

\item{digits}{numeric}
}
\value{
data.frame
}
\description{
Indicator validity table for (Multi-level) CFA
}
