% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_innermodel_sig.R
\name{innermodel_sig}
\alias{innermodel_sig}
\title{Showing significance using PLSPM regression}
\usage{
innermodel_sig(data, unite = TRUE, rhs = "endo", lhs = "exo")
}
\arguments{
\item{data}{plspm data}

\item{unite}{TRUE}

\item{rhs}{from}

\item{lhs}{to}
}
\value{
data result
}
\description{
Showing significance using PLSPM regression
}
\examples{
\dontrun{

# path matrix
IMAG = c(0,0,0,0,0,0)
EXPE = c(1,0,0,0,0,0)
QUAL = c(0,1,0,0,0,0)
VAL = c(0,1,1,0,0,0)
SAT = c(1,1,1,1,0,0)
LOY = c(1,0,0,0,1,0)
sat_path = rbind(IMAG, EXPE, QUAL, VAL, SAT, LOY)

# blocks of outer model
sat_blocks = list(1:5, 6:10, 11:15, 16:19, 20:23, 24:27)
# apply plspm
satpls = plspm(satisfaction, sat_path, sat_blocks, modes = sat_mod,
               scaled = FALSE)

#default
satpls\%>\%innermodel_sig()
satpls$inner_model \%>\%innermodel_sig()

satpls \%>\%innermodel_sig(unite=FALSE, lhs="row", rhs="col")
satpls$inner_model  \%>\%innermodel_sig(unite=FALSE, lhs="row", rhs="col")

satpls \%>\%innermodel_sig(unite=FALSE, lhs="row", rhs="col")
satpls$inner_model \%>\%innermodel_sig(unite=FALSE, lhs="row", rhs="col")

#'
}

}
