% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_prod.R
\name{interlist}
\alias{interlist}
\title{Functions that compensate for the method of input}
\usage{
interlist(...)
}
\arguments{
\item{...}{LONG LIST}
}
\value{
list
}
\description{
Functions that compensate for the method of input
}
\examples{

\dontrun{


# want list
intbasic = list(c("IMAG", 1, 2), c("SAT", 1, 2))
intbasic

inta = interlist(list(c("IMAG", 1, 2, 3), c("SAT", 1, 2)));inta
intc = interlist(list("IMAG", 1, 2, 3, "SAT", 1, 2));intc

intb = interlist(c("IMAG", 1, 2, 3), c("SAT", 1, 2));intb
intd = interlist("IMAG", 1, 2, 3, "SAT", 1, 2);intd
#'

# Partially select to create an interaction term
satisfaction\%>\% plspm_prod(sat_blocks1,
                           list(c("IMAG", 1, 2, 3), c("SAT", 1, 2, 3))) \%>\%str()

satisfaction\%>\% plspm_prod(sat_blocks1, inta) \%>\%str()
satisfaction\%>\% plspm_prod(sat_blocks1, intb) \%>\%str()
satisfaction\%>\% plspm_prod(sat_blocks1, intc) \%>\%str()



satisfaction\%>\%
plspm_prod(sat_blocks1, interlist(list("IMAG", 1, 2, 3, "SAT", 1, 2))) \%>\%str()
satisfaction\%>\%
plspm_prod(sat_blocks1, interlist(list("IMAG", 1, "SAT", 1, 2))) \%>\%str()
satisfaction\%>\%
plspm_prod(sat_blocks1, interlist(list("IMAG", 1,3,  "SAT", 2, 3))) \%>\%str()




}



}
