% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interplot.R
\name{interplot}
\alias{interplot}
\title{interplot is using data}
\usage{
interplot(
  data_long,
  type = "plot",
  size_point = 4,
  size_text = 5,
  linewidth = 1,
  vjust = -0.4,
  hjust = -2,
  title = NULL,
  xlab = "x_variable",
  ylab = "y_variable",
  legend_title = NULL,
  yadd = 0.5,
  size_element = 12,
  t = FALSE,
  xy = FALSE
)
}
\arguments{
\item{data_long}{long foramt data}

\item{size_point}{point size default 4}

\item{size_text}{text size  default 5}

\item{linewidth}{line width default 1}

\item{vjust}{vjust}

\item{hjust}{hjust}

\item{title}{title null}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{legend_title}{Change legend_title}

\item{yadd}{fig size add value}

\item{size_element}{axis element size}

\item{t}{matrix transpose efault FALSE, use when matrix data is wide format}

\item{xy}{plot ranspose default FALSE}
}
\value{
plot
}
\description{
interplot is using data
}
\examples{
\dontrun{

## interplot_data use!!
interplot_data( 6, 1, 4, 5, type = "long")\%>\%
interplot(size_element = 10)

interplot_data( 2, 1, 4, 5, type="long")\%>\%interplot()

## 2 *3 matriix
interplot_data( 4, 1, 2, 5, 6, 3, type="long", row=3, col=2)\%>\%interplot()

##3*2
interplot_data( 4, 1, 2, 5, 6, 3, type="long", row=2, col=3)\%>\%interplot()

##all data output
interplot_data( 4, 1, 2, 5, 6, 3, type="long", row=2, col=3)\%>\%interplot("all")

#generated direct data matrix and data frame
matrix(c(4, 1, 2, 5, 6, 3), nrow=2,ncol=3)  #marix

#long format
matrix(c(4, 1, 2, 5, 6, 3), nrow=2,ncol=3) \%>\%
data.frame() \%>\% rownames_to_column("v1") \%>\%
pivot_longer(cols=2:4)

##plot
matrix(c(4, 1, 2, 5, 6, 3), nrow=2,ncol=3) \%>\%
 data.frame() \%>\%
rownames_to_column("v1") \%>\% pivot_longer(cols=2:4)\%>\%
interplot()

## onestep method when use this function
matrix(c(4, 1, 2, 5, 6, 3), nrow=2, ncol=3) \%>\% interplot(t=TRUE)


#  3 * 2
matrix(c(4, 1, 2, 5, 6, 3), nrow=3, ncol=2)

matrix(c(4, 1, 2, 5, 6, 3), nrow=3, ncol=2) \%>\%
data.frame() \%>\%
rownames_to_column("v1") \%>\%
pivot_longer(cols=2:3)

matrix(c(4, 1, 2, 5, 6, 3), nrow=3, ncol=2) \%>\%
 data.frame() \%>\%
 rownames_to_column("v1") \%>\% p
 ivot_longer(cols=2:3)\%>\%
 interplot()


##one step output
matrix(c(4, 1, 2, 5, 6, 3), nrow=3, ncol=2) \%>\%
interplot(t=TRUE)


#### all data output
matrix(c(4, 1, 2, 5, 6, 3), nrow=3, ncol=2) \%>\% interplot(t=TRUE, type="all")

}

}
