% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interplot.R
\name{interplot_data}
\alias{interplot_data}
\title{interplot_data is matrix, long_data, plotting function}
\usage{
interplot_data(
  ...,
  col = 2,
  row = 2,
  type = "long",
  title = NULL,
  xlab = "x",
  ylab = "y",
  vjust = -0.4,
  hjust = 0.5,
  size_element = 12,
  byrow = TRUE,
  xy = FALSE
)
}
\arguments{
\item{...}{input data c() or as select()}

\item{col}{matrix ncol setting}

\item{row}{matrix nrow}

\item{type}{type is mat, long, all, plot}

\item{title}{plot title}

\item{xlab}{plot xlab}

\item{ylab}{plot ylab}

\item{vjust}{vjust text in plot postion}

\item{hjust}{hjust text in plot position}

\item{size_element}{plot element text size}

\item{byrow}{byrow=TRUE is default}

\item{xy}{plot transpose default FALSE}
}
\value{
data and plot
}
\description{
interplot_data is matrix, long_data, plotting function
}
\examples{
\dontrun{
 ## data output
interplot_data( 6, 1, 4, 5, type = "mat")
#' interplot_data( 6, 1, 4, 5, type = "all")
interplot_data( 6, 1, 4, 5, type = "long")
interplot_data( 6, 1, 4, 5, type = "plot")

# 2*3 data
interplot_data( c(6, 1, 4, 5,6,7), type = "mat",  col = 3,row = 2)
interplot_data( c(6, 1, 4, 5,6,7), type = "long",  col = 3,row = 2)
interplot_data( 6, 1, 4, 5,6,7, type = "all",  col = 3,row = 2)
interplot_data( 6, 1, 4, 5,6,7, type = "plot",  col = 3,row = 2)

## 3*2 data
interplot_data( 6, 1, 4, 5,6,7, type = "mat",  col = 2,row = 3)
interplot_data( 6, 1, 4, 5,6,7, type = "all",  col = 2,row = 3)
interplot_data( 6, 1, 4, 5,6,7, type = "long",  col = 2, row = 3)
interplot_data( c(6, 1, 4, 5,6,7), type = "plot",  col = 2,row = 3)

### ii is very hard example
interplot_data( 6, 1, 4, 5) \%>\%
ggplot(aes(x = v1, y = value))+
geom_point(size=4, aes(color=v2))+
geom_line(aes(group = v2, linetype = v2 ), linewidth = 1)+
geom_text(aes(label = value), vjust= -0.4, hjust= 2, size= 5)+
theme_bw()

}
}
