% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_item_CR_AVE.R
\name{item_CR_AVE}
\alias{item_CR_AVE}
\title{Item-level CR & AVE table}
\usage{
item_CR_AVE(fit, level = c("within", "between"), digits = 3)
}
\arguments{
\item{fit}{lavaan (ML-)SEM object}

\item{level}{"within" or "between"}

\item{digits}{numeric}
}
\value{
data.frame
}
\description{
Item-level CR & AVE table
}
