% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtl.R
\name{jtl}
\alias{jtl}
\title{Haehyoung translate}
\usage{
jtl(text = "", slang = "ko", elang = "en", translator = "deepl", print_url = F)
}
\arguments{
\item{text}{text to translate}

\item{slang}{starting language}

\item{elang}{ending language}

\item{translator}{translator type. lists are above."alibaba", "apertium", "argos", "baidu", "bing","caiyun", "cloudTranslation", "deepl", "elia", "google","iciba", "iflytek", "iflyrec", "itranslate", "judic", "languageWire", "lingvanex", "niutrans", "mglip", "mirai", "modernMt", "myMemory", "papago", "qqFanyi", "qqTranSmart", "reverso", "sogou", "sysTran", "tilde", "translateCom","translateMe", "utibet", "volcEngine", "yandex", "yeekit", "youdao"}

\item{print_url}{Debugging Setting}
}
\description{
Haehyoung translate
}
