% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.R
\name{k}
\alias{k}
\title{Kakao translaste}
\usage{
k(text = "", slang = "en", elang = "ko", show = "data")
}
\arguments{
\item{text}{input text}

\item{slang}{"ko"}

\item{elang}{"en"}

\item{show}{nomal, data is each sentence}
}
\value{
translate result
}
\description{
Kakao translaste
}
