% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kge_ca_analysis.R
\name{kge_ca_analysis}
\alias{kge_ca_analysis}
\title{Correspondence analysis visualization functions}
\usage{
kge_ca_analysis(
  dataset,
  typekey,
  selcol_3 = "w1f",
  type = "all",
  arrows = c(F, T),
  selcol_1 = "성조",
  selcol_2 = "onset",
  area = "부산",
  addtext = "",
  xlim = NULL,
  ylim = NULL,
  size_text = 5,
  opt = 1
)
}
\arguments{
\item{dataset}{data}

\item{typekey}{filter}

\item{selcol_3}{col3}

\item{type}{all}

\item{arrows}{c(F,T)}

\item{selcol_1}{col1}

\item{selcol_2}{col2}

\item{area}{지역}

\item{addtext}{add text}

\item{xlim}{range c(0, 0.3)}

\item{ylim}{rangec(-0.3, 0.4)}

\item{size_text}{5}

\item{opt}{opt 1~ 6}
}
\description{
Correspondence analysis visualization functions
}
