% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kge_weight.R
\name{kge_weigth}
\alias{kge_weigth}
\title{generative weight variable, kge(2024)}
\usage{
kge_weigth(df, type = "res3", pattern = "", remove = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{type}{result res, res1, res2, res3, res4}

\item{pattern}{pattern = "", If you want something different, you can specify a different pattern.}

\item{remove}{ubite option}
}
\value{
add weight
}
\description{
generative weight variable, kge(2024)
}
\examples{


\dontrun{
KGEbind3 \%>\% select(-W, -weight) \%>\% kge_weigth("res")
}
}
