% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_hypo.R
\name{lav_de_hypo}
\alias{lav_de_hypo}
\title{Hypothesis generation functions}
\usage{
lav_de_hypo(model)
}
\arguments{
\item{model}{lavaan syntax}
}
\value{
text
}
\description{
Hypothesis generation functions
}
\examples{

\dontrun{
#'
# 예제 모델
model1a <- "
EXPE ~ H1*IMAG
QUAL ~ H2*EXPE
VAL ~  H3*QUAL + H4*EXPE
SAT ~ H5*IMAG + H6*EXPE + H7*QUAL + H8*VAL
LOY ~ H9*SAT + H10*IMAG
"

# 함수 실행
lav_de_hypo(model1a)

#'
}

}
