% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_extract_ind.R
\name{lav_extract_ind}
\alias{lav_extract_ind}
\title{lav_extract_ind, indirect effect}
\usage{
lav_extract_ind(
  model,
  start_node = NULL,
  end_node = NULL,
  prefix = "a",
  cat = FALSE,
  type = "ind",
  de = FALSE,
  auto = TRUE,
  paths_name = FALSE
)
}
\arguments{
\item{model}{lavaan}

\item{start_node}{path start node}

\item{end_node}{path end node}

\item{prefix}{hypothesis "a" if U want change "H"}

\item{cat}{test show cat TRUR}

\item{type}{type "ind","new_parameters","combine"}

\item{de}{when direct paht input}

\item{auto}{TRUE path auto, FLASE all ind path}

\item{paths_name}{TRUE change paths name}
}
\value{
text
}
\description{
lav_extract_ind, indirect effect
}
\examples{


\dontrun{
#'

# Example usage
model1 <- "
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
VAL =~ val1 + val2 + val3 + val4
SAT =~ sat1 + sat2 + sat3 + sat4
LOY =~ loy1 + loy2 + loy3 + loy4

EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL+ EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"


# Extracting paths from the model with default start_node = NULL and end_node = NULL
lav_extract_mm(model1) \%>\%cat("\n")
lav_extract_sm(model1) \%>\%cat("\n")

lav_extract_ind(model1)\%>\%cat("\n")
lav_extract_ind(model1, auto=TRUE)\%>\%cat("\n")

lav_extract_ind(model1,start_node="EXPE", auto=TRUE)\%>\%cat("\n")
lav_extract_ind(model1,start_node="EXPE")\%>\%cat("\n")
lav_extract_ind(model1,start_node="QUAL")\%>\%cat("\n")
lav_extract_ind(model1,start_node="IMAG", auto=TRUE)\%>\%cat("\n")

lav_extract_ind(lav_extract_sm(model1), de=TRUE) \%>\%cat("\n")

lav_extract_ind(lav_extract_sm(model1), de=TRUE, auto=TRUE)\%>\%cat("\n")
lav_extract_ind(lav_extract_sm(model1), start_node="IMAG", de=TRUE) \%>\%cat('\n')

lav_extract_ind(lav_extract_sm(model1),"IMAG", prefix="H", de=TRUE) \%>\%cat('\n')

# Example usage
de_paths <- "
EXPE  ~ a1* IMAG
QUAL  ~ a2* EXPE
VAL  ~ a3* QUAL
VAL  ~ a4* EXPE
SAT  ~ a5* IMAG
SAT  ~ a6* EXPE
SAT  ~ a7* QUAL
SAT  ~ a8* VAL
LOY  ~ a9* SAT
LOY  ~ a10* IMAG
"

lav_extract_ind(de_paths)
#출력
#  lav_extract_ind(model1, cat=TRUE)

lav_extract_ind(de_paths, cat=TRUE)

lav_extract_ind(de_paths)\%>\%make_c_text()
}

}
