% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_extract_mm.R
\name{lav_extract_mm2}
\alias{lav_extract_mm2}
\title{extranct measurement, Extract all structural models}
\usage{
lav_extract_mm2(lav_syn, text = FALSE, interact = TRUE)
}
\arguments{
\item{lav_syn}{model lavaan syntax}

\item{text}{text output}

\item{interact}{FALSE exclude :, TRue include :}
}
\value{
text model
}
\description{
extranct measurement, Extract all structural models
}
\examples{

\dontrun{
#'

# # Example usage:
data(MarshWenHau)
str(MarshWenHau)

mwh = MarshWenHau

model2 <- "
f1 =~ x1 + x2 + x3
f1Xf2 =~ f1:f2
f2 =~ x4 + x5 + x6
f3 =~ y1 + y2 + y3

f3 ~ f1 + f2
f3 ~ f1Xf2

f1Xf2 ~~ 0*f1
f1Xf2 ~~ 0*f2
"

# interaction model
intsem =  SEM(model2, data= MarshWenHau)
summary(intsem)
intsem \%>\%lav_semPaths2(color=list(lat="darkred",man="gold"))


##2nd
#base model
lav_extract_mm(model2) \%>\%cat()
#interaction item list
lav_extract_imlist(model3)
#data
lav_latentProd(mwh, model3)\%>\%head()

#extracted interaction term
lav_extract_int(model2, lav_extract_imlist(model3))

# or interaction items ; imlist
imlist <- list(
  .f1Xf2 = c("f1", "f2"),
  .i_f1_f4 = c("f1", "f4")
)
# names(imlist)
lav_extract_int(model2, imlist)
#direct
lav_extract_int(model2, imlist= list(.f1Xf2 = c("f1", "f2"),
                                     .i_f1_f4 = c("f1", "f4")))
#estimates model
intsem =  SEM(lav_new_model(model2, lav_extract_imlist(model2)),
              data= lav_latentProd(mwh, model2))

#'
model2 <- "
 f1 =~ x1 + x2 + x3
 f1Xf2 =~ f1:f2
 f2 =~ x4 + x5 + x6
 f3 =~ y1 + y2 + y3

 f3 ~ f1 + f2
 f3 ~ f1Xf2

 f1Xf2 ~~ 0*f1
 f1Xf2 ~~ 0*f2
 "

lav_extract_mm2(model2)
lav_extract_mm2(model2, interact=F)
lav_extract_mm2(model2, interact=T)
#'
#'
}

}
