% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_extrac_paths.R
\name{lav_extract_path}
\alias{lav_extract_path}
\title{lav_extract_path}
\usage{
lav_extract_path(model)
}
\arguments{
\item{model}{model}
}
\value{
text
}
\description{
lav_extract_path
}
\examples{

\dontrun{
#'
# 모델 문자열
model1a <- "
EXPE ~ H1*IMAG
QUAL ~ H2*EXPE
VAL ~ H3*QUAL + H4*EXPE
SAT ~ H5*IMAG + H6*EXPE + H7*QUAL + H8*VAL
LOY ~ H9*SAT + H10*IMAG
"

model1 <- "
EXPE ~ IMAG
QUAL ~ EXPE
VAL ~ QUAL + EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"


# 함수 호출
lav_extract_path(model1a)

lav_extract_path(model1)

}


}
