% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_extract_sm.R
\name{lav_extract_sm}
\alias{lav_extract_sm}
\title{lav_extract_sm, Extracting a Structural Path Model from a Model}
\usage{
lav_extract_sm(model, hypo = TRUE, prefix = "a", all = FALSE)
}
\arguments{
\item{model}{lavaan model}

\item{hypo}{hypothesis, default TRUE}

\item{prefix}{"a"}

\item{all}{all paths}
}
\value{
text
}
\description{
lav_extract_sm, Extracting a Structural Path Model from a Model
}
\examples{

\dontrun{
#'
# Example usage
model1 <- "
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
VAL =~ val1 + val2 + val3 + val4
SAT =~ sat1 + sat2 + sat3 + sat4
LOY =~ loy1 + loy2 + loy3 + loy4

EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL+ EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"


lav_extract_sm(model1)\%>\%cat("\n")
lav_extract_sm(model1, hypo=FALSE)\%>\%cat("\n")
lav_extract_sm(model1, prefix="lamda")\%>\%cat("\n")
lav_extract_sm(model1, prefix="H")\%>\%cat("\n")
lav_extract_sm(model1, all=FALSE) \%>\%cat("\n")
lav_extract_sm(model1, all=TRUE) \%>\%cat("\n")
lav_extract_sm(model1,FALSE) \%>\%cat("\n")


lav_extract_sm(model1, all=TRUE) \%>\%
  diagram_model(sizeLat=6)
#'

}


}
