% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_factor_layout.R
\name{lav_factor_layout}
\alias{lav_factor_layout}
\title{자동화된 semPaths모델 layout}
\usage{
lav_factor_layout(
  sem_res,
  positions = NULL,
  nrow = 4,
  ncol = 5,
  new_position = NULL,
  ...
)
}
\arguments{
\item{sem_res}{lavaan object}

\item{positions}{각 변수의 위치를 변경 position = list("motiv" =c(1, 3))}

\item{nrow}{4}

\item{ncol}{5}

\item{new_position}{list(c(1, 1), c(1, 2), c(3, 1), c(2, 3),c(2, 4))}

\item{...}{ETC}
}
\value{
MAT
}
\description{
자동화된 semPaths모델 layout
}
\examples{

\dontrun{
#'
data(motivation)
str(motivation)
# Test the function with debug prints
motive_model <- "
achieve ~ motiv + ses + iq
motiv ~ ses + health + iq
"

# Assuming 'motivation' is your data frame
motiv_sem <- SEM(motive_model, data = motivation)

motiv_sem\%>\%summary(stand=TRUE)

# 적용 예시
motiv_sem \%>\%
  lav_semPaths2(rotation = 2, se=TRUE, edge.label.position = 0.4,
                layout = lav_factor_layout(motiv_sem,
                                           new_position = list(c(1, 1), c(1, 2), c(3, 1), c(2, 3),c(2, 4)) ),
                rotate_resid = c("motiv" =160, "achieve"=160))


# 적용 예시
motiv_sem \%>\%
  lav_semPaths2(rotation = 2, se=TRUE, edge.label.position = 0.4,
                layout = lav_factor_layout(motiv_sem,
                                           position = list("motiv" =c(1, 3))),
                rotate_resid = c("motiv" =160, "achieve"=160))

#'
}

}
