% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_htmt.R
\name{lav_htmt}
\alias{lav_htmt}
\title{lavaan htmt}
\usage{
lav_htmt(
  sem_res = NULL,
  data = NULL,
  blocks = NULL,
  htmt2 = TRUE,
  sig = TRUE,
  cut = 0.9,
  digits = 3,
  imp = ""
)
}
\arguments{
\item{sem_res}{lavaan result}

\item{data}{sem data}

\item{blocks}{sem blocks list data}

\item{htmt2}{true using htmt2}

\item{sig}{true *}

\item{cut}{cut default 0.9, roburst 0.85, general 1}

\item{digits}{3}

\item{imp}{"" upper triangle matrix}
}
\value{
data
}
\description{
lavaan htmt
}
\examples{

\dontrun{
#'
data(satisfaction)
model1 <- "
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
VAL =~ val1 + val2 + val3 + val4
SAT =~ sat1 + sat2 + sat3 + sat4
LOY =~ loy1 + loy2 + loy3 + loy4

EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL + EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"

sat_sem0 <- sem(model1, data = satisfaction)

#check
lav_htmt(sat_sem0)
# using data
lav_htmt(data= lav_return_data(sat_sem0), blocks= lav_return_blocks(sat_sem0))
#model data
lav_htmt(data= satisfaction, blocks =  lav_extract_mm(model1)\%>\% plspm_lav2blocks())
# using data
lav_htmt(data= satisfaction, blocks = model1)

#option
lav_htmt(sat_sem0, htmt2=T, cut=0.85)


}

}
