% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_modelfit.R
\name{lav_modelfit}
\alias{lav_modelfit}
\title{lavaan model fit}
\usage{
lav_modelfit(..., nice_table = FALSE)
}
\arguments{
\item{...}{sem models}

\item{nice_table}{output model fit web}
}
\value{
data
}
\description{
lavaan model fit
}
\examples{

\dontrun{
#'
example(sem)
lav_modelfit(fit)
lav_modelfit(fit, nice_table = T)

lav_modelfit("model1"=fit,"model2"= fit)

#'
}


}
