% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_pattern.R
\name{lav_Pattern}
\alias{lav_Pattern}
\title{응답패턴 분석}
\usage{
lav_Pattern(data, n = 2, dim = 0L, top = 20)
}
\arguments{
\item{data}{data}

\item{n}{구분되는 개수}

\item{dim}{0L}

\item{top}{20}
}
\value{
data
}
\description{
응답패턴 분석
}
\examples{

\dontrun{
#'
latentPattern(MarshWenHau, n=3)

HS9 <- HolzingerSwineford1939[,c("x1","x2","x3","x4","x5",
                                 "x6","x7","x8","x9")]
latentPattern(HS9)

#explain
HSbinary <- as.data.frame( lapply(HS9, cut, 2, labels=FALSE) )
HSbinary \%>\%head()
lavTables(HSbinary, dim = 0L, categorical = names(HSbinary)) \%>\%head(20)

#'
#'

x11()
latentPattern(park516_kor[,1:6], n=5) \%>\%
  ggplot(aes(x= fct_reorder(pattern, desc(obs.freq)), y= obs.freq))+
  geom_bar(stat="identity")+theme_bw()+
  theme(axis.text.x= element_text(size=14, angle=90))
}


}
