% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_remodel.R
\name{lav_remodel}
\alias{lav_remodel}
\title{lav_remodel, Models that create indirect effects}
\usage{
lav_remodel(
  model,
  prefix = "a",
  auto = TRUE,
  start_node = NULL,
  end_node = NULL,
  cat = FALSE,
  text = FALSE,
  interact = TRUE,
  paths_name = TRUE
)
}
\arguments{
\item{model}{lavaan(original model)}

\item{prefix}{model hypothesis "a", or "H"}

\item{auto}{TRUe indirect effec auto, FALSE all path}

\item{start_node}{NULL If you type it, it starts with}

\item{end_node}{NULL and that's it}

\item{cat}{text puput console}

\item{text}{Output text to save and type}

\item{interact}{TRUe :(use) using SESM, FALSE (:) using sem()}
}
\value{
text
}
\description{
lav_remodel, Models that create indirect effects
}
\examples{

\dontrun{
#'


data(satisfaction)

# Example usage
model1 <- "
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
VAL =~ val1 + val2 + val3 + val4
SAT =~ sat1 + sat2 + sat3 + sat4
LOY =~ loy1 + loy2 + loy3 + loy4

EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL+ EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"



# lav_remodel(model1)
lav_remodel(model1)\%>\%cat("\n")
lav_remodel(model1, interact=FALSE)\%>\%cat("\n")

lav_remodel(model1, auto=FALSE)\%>\%cat("\n")

lav_remodel(model1, start="EXPE", prefix="H", cat=TRUE)
lav_remodel(model1, cat=TRUE)

lav_remodel(model1, cat=TRUE, text=TRUE)
lav_remodel(model1, cat=TRUE, text=FALSE)

lav_remodel(model1, cat=FALSE, text=TRUE)
lav_remodel(model1, cat=FALSE, text=FALSE)


satis_sem = sem(lav_remodel(model1), satisfaction)
summary(satis_sem)
model1 \%>\%class()


}


}
