% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_return.R
\name{lav_return_data}
\alias{lav_return_data}
\title{lavaan결과에서 데이터 추출하는 함수}
\usage{
lav_return_data(sem_res)
}
\arguments{
\item{sem_res}{sem result}
}
\value{
data
}
\description{
lavaan결과에서 데이터 추출하는 함수
}
\examples{

\dontrun{
#'
data(satisfaction)
model1 <- "
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
VAL =~ val1 + val2 + val3 + val4
SAT =~ sat1 + sat2 + sat3 + sat4
LOY =~ loy1 + loy2 + loy3 + loy4

EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL + EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"

sat_sem0 <- sem(model1, data = satisfaction)

# 함수 결과 확인
lav_return_data(sat_sem0)
lav_return_data(sat_sem0)\%>\%head()
}


}
