% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_semPaths.R
\name{lav_semPaths2}
\alias{lav_semPaths2}
\title{Drawing Structural Equation Models Using the lavaan Syntax (park JH PhD)}
\usage{
lav_semPaths2(
  lav_obj,
  whatLabels = "est",
  rotation = 2,
  residuals = TRUE,
  residScale = 12,
  layout = "tree2",
  edge.label.cex = 0.7,
  edge.label.position = 0.55,
  edge.color = "black",
  sizeMan = 6,
  sizeMan2 = 4,
  sizeLat = 9,
  sizeLat2 = 6,
  style = "lisrel",
  exoVar = FALSE,
  exoCov = TRUE,
  curve = 1.5,
  asize = 2,
  mar = c(3, 8, 3, 8),
  nDigits = 3,
  shapeLat = "circle",
  shapeMan = "rectangle",
  shapeInt = "triangle",
  border.width = 2,
  edge.width = 1.5,
  Groups = FALSE,
  group_match = "lat",
  structural = FALSE,
  edgeLabels = NULL,
  nodeLabels = NULL,
  fixedStyle = 1,
  color = list(lat = "gray10", man = "white"),
  label.cex = 1,
  sig = TRUE,
  rotate_resid = NULL,
  se = FALSE,
  type = "plot"
)
}
\arguments{
\item{whatLabels}{"model}

\item{rotation}{rotation=2}

\item{residuals}{residauls TRUE}

\item{residScale}{residScale=12}

\item{layout}{layout="tree"}

\item{edge.label.cex}{1}

\item{edge.label.position}{0.5}

\item{edge.color}{edge.color= black}

\item{sizeMan}{8}

\item{sizeMan2}{4}

\item{sizeLat}{10}

\item{sizeLat2}{6}

\item{style}{lisrel}

\item{exoVar}{FALSE}

\item{exoCov}{TRUE}

\item{curve}{1.5}

\item{asize}{1.5}

\item{mar}{c(2,8,3,10)}

\item{nDigits}{2}

\item{shapeLat}{"circle, ellipse}

\item{shapeMan}{retangle}

\item{shapeInt}{triangle ,}

\item{border.width}{border.width=1}

\item{edge.width}{edge.width=1.5}

\item{Groups}{FALSE}

\item{group_match}{"FALSE"lat#'}

\item{structural}{structural = FALSE}

\item{edgeLabels}{edgeLabels=NULL}

\item{nodeLabels}{nodeLabels=NULL}

\item{fixedStyle}{fixedStyle=1 default rbase linetype}

\item{color}{list(lat="gray10", man="white")}

\item{label.cex}{label.cex}

\item{sig}{FALSE}

\item{rotate_resid}{c(lv= angle), c(f1 =90, f2=180)}

\item{se}{se=FALSE}

\item{type}{type ="plot, and res is lavaan simdata result}

\item{plsres_boot}{plsres_boot to model syntax}
}
\value{
plot jutpls_boot
}
\description{
Drawing Structural Equation Models Using the lavaan Syntax (park JH PhD)
}
\examples{
\dontrun{
#'

#data
data(MarshWenHau) # Marsh, Wen & Hau (2000)
# colnames(MarshWenHau) = c( paste0("x",1:6), paste0("y",1:3))
MarshWenHau \%>\%str()


model1 = "

f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
f3 =~ y1 + y2 + y3
"

mwh_sem1 = sem(model1, MarshWenHau)
mwh_sem1\%>\%summary(fit=T)


re_mwh_data = lavPredict(mwh_sem1)\%>\%data.frame()
re_mwh_data



# 이방법인 LMS방법임

Intf1f2  = re_mwh_data[,"f1"] *  re_mwh_data[,"f2"]

#fitst
MarshWenHau$f1_f2 = Intf1f2
MarshWenHau \%>\% str()

model2 = "

f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
f3 =~ y1 + y2 + y3

f1Xf2 =~ f1_f2

f3 ~ f1 + f2 + f1Xf2

f1Xf2 ~~ 0*f1
f1Xf2 ~~ 0*f2

"

mwh_sem2 = sem(model2, MarshWenHau)
mwh_sem2\%>\%summary(fit=T)
mwh_sem2\%>\% sem_effect()
x11()
mwh_sem2\%>\% lav_semPaths2(exoCov = FALSE, sig=F,
                          rotate_resid = c(f3 = 160))

mwh_sem2\%>\% lav_semPaths2(exoCov = FALSE, sig=TRUE, Groups=TRUE,
                          rotate_resid = c(f3 = 160))

mwh_sem2\%>\% lav_semPaths2(exoCov = FALSE, sig=TRUE,se=TRUE,
                          rotate_resid = c(f3 = 160))


#'

}

}
