% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LavaanPlot.R
\name{LavaanPlot}
\alias{LavaanPlot}
\title{Lavaan plot draw}
\usage{
LavaanPlot(
  x,
  rotation = "LR",
  coefs = TRUE,
  col = "gray60",
  stand = FALSE,
  label = NULL,
  covs = FALSE,
  stars = c("covs", "latent", "regress"),
  overlap = FALSE,
  layout = "dot",
  digits = 3
)
}
\arguments{
\item{x}{lavaan result}

\item{rotation}{rotation 'LR','RL', 'TB','BT'}

\item{coefs}{TRUE}

\item{col}{default 'gray60'}

\item{stand}{std coeff}

\item{label}{input name labels=c(f1_1="exp", f1_2="want",f1_3="problem") etc}

\item{covs}{TRUE}

\item{overlap}{default FALSE}

\item{layout}{'dot', 'neato', 'fdp', 'sfdp', 'circo', 'twopi', 'nop','nop2', 'osage', 'patchwork'}

\item{digits}{digits = 3#'}
}
\value{
plot
}
\description{
Lavaan plot draw
}
\examples{
# lavaan code

## Bollen (1989), page 332
model <- '
  # latent variable definitions
     ind60 =~ x1 + x2 + x3
     dem60 =~ y1 + a*y2 + b*y3 + c*y4
     dem65 =~ y5 + a*y6 + b*y7 + c*y8

  # regressions
    dem60 ~ ind60
  dem65 ~ ind60 + dem60

  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
'

fit <- sem(model, data = PoliticalDemocracy)
 ## summary
summary(fit, fit.measures = TRUE, stand=T)
##
LavaanPlot(fit, cov=F)
LavaanPlot(fit, cov=F, layout="circo")

LavaanPlot(fit, cov=F, layout="dot",
           label=c(x1="ind1",ind60="IND60"))


}
