% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semptools.R
\name{layout_mat_name}
\alias{layout_mat_name}
\title{layout_mat_name}
\usage{
layout_mat_name(plsres_boot)
}
\arguments{
\item{plsres_boot}{plsres_boot}
}
\value{
plsres_boot
}
\description{
layout_mat_name
}
\examples{
\dontrun{
layout_mat_name(nfl_pls_boot)
}
}
