% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_class_add.R
\name{lca_class_add}
\alias{lca_class_add}
\title{lca_class_add,기존데이터에 잠재클래스 추가}
\usage{
lca_class_add(data, lcadata)
}
\arguments{
\item{data}{df}

\item{lcadata}{lcadata}
}
\value{
data
}
\description{
lca_class_add,기존데이터에 잠재클래스 추가
}
\examples{
\dontrun{

jut6 \%>\% lca_class_add(jutLca3) \%>\% head()
}
}
