% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_class_freq.R
\name{lca_class_freq}
\alias{lca_class_freq}
\title{lca_class_freq cal}
\usage{
lca_class_freq(lacdata, show = "")
}
\arguments{
\item{lacdata}{lca result}

\item{show}{df, ""}
}
\value{
df
}
\description{
lca_class_freq cal
}
\examples{
\dontrun{

 jutLca3 \%>\%lca_class_freq()\%>\% data.frame()
}
}
