% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_class_plot.R
\name{lca_class_plot}
\alias{lca_class_plot}
\title{lca_class_plot,잠재클래스 패턴 그래프}
\usage{
lca_class_plot(lcadata_graph, type = "g")
}
\arguments{
\item{lcadata_graph}{lcaGraphData 데이터처리된 것}

\item{type}{g, full, line}
}
\value{
plot
}
\description{
lca_class_plot,잠재클래스 패턴 그래프
}
\examples{
\dontrun{
#'
## using
jutLca3 \%>\%lcaGraphData() \%>\% lca_class_plot()
jutLca3 \%>\%lcaGraphData() \%>\% lca_class_plot("line")

jutLca3 \%>\%lcaGraphData( "학업성적"="L05",
                         "부모지지"="L04",
                         "진로결정"="L03",
                         "준비미흡"="L02",
                         "목표확실"="L01") \%>\%
  lca_class_plot()
#full graph
jutLca3 \%>\%lcaGraphData( "학업성적"="L05",
                         "부모지지"="L04",
                         "진로결정"="L03",
                         "준비미흡"="L02",
                         "목표확실"="L01") \%>\%
  replace_df_rep(
    "L01", "L01:목표확실",
    "L02", "L02:준비미흡",
    "L03", "L03:진로결정",
    "L04", "L04:부모지지",
    "L05", "L05:학업성적")\%>\%
  replace_df_rep(
    "Class1", "Class1:적극적 진로준비형",
    "Class2", "Class2:불명확 진로인식형",
    "Class3", "Class3:조건부 진로지향형"
  )\%>\%lca_class_plot("line")

}
}
