% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_freq_plot.R
\name{lca_freq_plot}
\alias{lca_freq_plot}
\title{lca_freq_plot LCA 빈도 그래프}
\usage{
lca_freq_plot(data, lcadata)
}
\arguments{
\item{data}{df}

\item{lcadata}{lcadata}
}
\value{
plot
}
\description{
lca_freq_plot LCA 빈도 그래프
}
\examples{
\dontrun{

jut7c \%>\% lca_freq_plot(jutLca3)+ ylim(0, 380)
}
}
