% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_gamma.R
\name{lca_gamma}
\alias{lca_gamma}
\title{lca_gamma 사전확률}
\usage{
lca_gamma(lcadata)
}
\arguments{
\item{lcadata}{lcadata}
}
\value{
vedtor
}
\description{
lca_gamma 사전확률
}
\examples{
\dontrun{
jutLca3\%>\% lca_gamma()
}
}
