% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_gof.R
\name{lca_gof}
\alias{lca_gof}
\title{lca_gof}
\usage{
lca_gof(...)
}
\arguments{
\item{...}{models}
}
\value{
data
}
\description{
lca_gof
}
\examples{

\dontrun{
#'
# 함수 사용 예시
lca_gof(jutLca2, jutLca3, jutLca4)

lca_gof(jutLca3)\%>\%row2col()
}
}
