% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_result.R
\name{lca_result}
\alias{lca_result}
\title{lca_result lca분석후 결과 정리}
\usage{
lca_result(lcadata, class_name = NULL)
}
\arguments{
\item{lcadata}{lcadata}

\item{class_name}{input name}
}
\value{
data
}
\description{
lca_result lca분석후 결과 정리
}
\examples{
\fontrun{
#'
jutLca3 \%>\%lca_result() \%>\%
  replace_df_rep( "L01", "진로목표(L01)",
                  "L02", "준비여부(L02)",
                  "L03", "직업결정(L03)",
                  "L04", "부모추천(L04)",
                  "L05", "학업성적(L05)" ) #변수명을 바꾸기

jutLca3 \%>\%lca_result(class_name=c(1,2,3))

jutLca3 \%>\%lca_result(class_name=c("조건부_진로지향형\n(Conditional-Oriented)",
                                   "불명확_진로인식형\n(Unclear-Awareness)",
                                   "적극적_진로준비형\n(Active-Preparedness)")


                      jutLca3 \%>\%lca_result(class_name=c("조건부_진로지향형",
                                                         "불명확_진로인식형",
                                                         "적극적_진로준비형"))

}
}
