% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcaGraphData.R
\name{lcaGraphData}
\alias{lcaGraphData}
\title{lcaGraphData, lcadata to graph data}
\usage{
lcaGraphData(lcaResult, ..., yes_col = 3, yes = `Y = 2`)
}
\arguments{
\item{lcaResult}{lcadata}

\item{...}{iteam names, nothng auto}

\item{yes_col}{yes col , default 3}

\item{yes}{yes name default `Y = 2`}
}
\value{
data
}
\description{
lcaGraphData, lcadata to graph data
}
\examples{
\dontrun{
#'
#함수를 이용하여 그래프 데이터 생성
jutLca3 \%>\%lcaGraphData()

jutLca3 \%>\%lcaGraphData( "학업성적"="L05",
                         "부모지지"="L04",
                         "진로결정"="L03",
                         "준비미흡"="L02",
                         "목표확실"="L01")

}

}
