% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerMat.R
\name{lowerMat}
\alias{lowerMat}
\title{lower pull matrix}
\usage{
lowerMat(mat, fill = 0, diag = 0)
}
\arguments{
\item{mat}{matrix data}

\item{fill}{matrix data fill upper 0 or if that's what you want , enter it}

\item{diag}{matrix data diag enter 0}
}
\value{
lowermatrix, upper 0
}
\description{
lower pull matrix
}
