% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerMat.R
\name{lowerMat_na}
\alias{lowerMat_na}
\title{lower pull matrix fill NA}
\usage{
lowerMat_na(
  mat,
  fill = "",
  diag = 1,
  numeric = FALSE,
  digits = 3,
  justify = "right"
)
}
\arguments{
\item{mat}{matrix data}

\item{fill}{matrix data NA or .}

\item{diag}{matrix  diagonal data NA or .}
}
\value{
lowermatrix, upper 0
}
\description{
lower pull matrix fill NA
}
