% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_bic_plot.R
\name{lpa_BIC_plot}
\alias{lpa_BIC_plot}
\title{lpa_BIC_plot}
\usage{
lpa_BIC_plot(data, n_profiles_range = 1:9)
}
\arguments{
\item{data}{data}

\item{n_profiles_range}{1:9}
}
\value{
pot
}
\description{
lpa_BIC_plot
}
\examples{

\dontrun{

select(iris, -Species)\%>\%lpa_BIC_plot()
}


}
