% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_bic_plot.R
\name{lpa_BIC_plot2}
\alias{lpa_BIC_plot2}
\title{lpa_BIC_plot2}
\usage{
lpa_BIC_plot2(
  data,
  n_profiles_range = 1:9,
  modelNames = NULL,
  v = 0,
  point_size = 1.5,
  v_linewidth = 0.6,
  l_linewidth = 1,
  title = "적절한 개수의 profiles 판정 ",
  flip = TRUE,
  line = FALSE
)
}
\arguments{
\item{data}{data}

\item{n_profiles_range}{1:9}

\item{modelNames}{NULL}

\item{v}{0}

\item{point_size}{1.5}

\item{v_linewidth}{0.6, #profile cut}

\item{l_linewidth}{1, #line width}

\item{title}{"적절한 개수의 profiles 판정 "}

\item{flip}{TRUE}

\item{line}{FALSE}
}
\value{
plot
}
\description{
lpa_BIC_plot2
}
\examples{

\dontrun{

select(iris, -Species)\%>\%lpa_BIC_plot2()
}

}
