% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_bic_plot.R
\name{lpa_BIC_plot3}
\alias{lpa_BIC_plot3}
\title{lpa_BIC_plot3, all plot}
\usage{
lpa_BIC_plot3(data, n_profiles = 1:9, modelNames = NULL, v = 0)
}
\arguments{
\item{data}{data}

\item{n_profiles}{1:9}

\item{modelNames}{NULL, if you want input  c("EEV","EEI","EEE","EII"),}

\item{v}{0, abline v = 0}
}
\value{
plot
}
\description{
lpa_BIC_plot3, all plot
}
\examples{

\dontrun{

select(iris, -Species)\%>\%lpa_BIC_plot3()
}

}
