% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_bind_classdata.R
\name{lpa_bind_classdata}
\alias{lpa_bind_classdata}
\title{lpa_bind_classdata}
\usage{
lpa_bind_classdata(
  data,
  n_profiles = 3,
  model = NULL,
  package = c("mclust", "MplusAutomation"),
  type = "all",
  add_line = TRUE
)
}
\arguments{
\item{data}{data}

\item{n_profiles}{default 3}

\item{package}{mclucs , MplusAutomation}

\item{type}{all, res, fit, est, data, all, plot, group_plot}

\item{add_line}{add_line null}
}
\value{
data
}
\description{
lpa_bind_classdata
}
\examples{

\dontrun{

#'
lpa_bind_classdata(select(iris, -Species))

lpa_bind_classdata(select(iris, -Species), type="res")
lpa_bind_classdata(select(iris, -Species), type="fit")
lpa_bind_classdata(select(iris, -Species), type="est")
lpa_bind_classdata(select(iris, -Species), type="data")\%>\%Round(3)\%>\%dall
lpa_bind_classdata(select(iris, -Species), type="all")
#'

}



}
