% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_creat_profile.R
\name{lpa_create_profiles}
\alias{lpa_create_profiles}
\title{lpa_create_profiles}
\usage{
lpa_create_profiles(
  df,
  n_profiles = 3,
  variance_structure = "freed",
  covariance_structure = "freed"
)
}
\arguments{
\item{df}{df}

\item{n_profiles}{default = 3}

\item{variance_structure}{freed}

\item{covariance_structure}{freed}
}
\value{
data
}
\description{
lpa_create_profiles
}
\examples{

\dontrun{
#'
iris[,-5]\%>\% lpa_create_profiles(n_profiles =  4,
                                    variance_structure = "constrained",
                                    covariance_structure = "constrained")
}


}
