% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_creat_profile.R
\name{lpa_create_profiles2}
\alias{lpa_create_profiles2}
\title{lpa_create_profiles2}
\usage{
lpa_create_profiles2(df, n_profiles = 4, model_name = NULL, type = "mean")
}
\arguments{
\item{df}{data}

\item{n_profiles}{3, ...}

\item{model_name}{EEI, VVI, EEE, VVV}
}
\value{
data
}
\description{
lpa_create_profiles2
}
\examples{

\dontrun{
#'

iris[,-5] \%>\% lpa_create_profiles2(n_profiles = 4)

#v:constrained, Co: fixed
iris[,-5] \%>\% lpa_create_profiles2(n_profiles = 4, model_name="EEE")
#variance_structure == "fixed"
iris[,-5] \%>\% lpa_create_profiles2(n_profiles = 4, model_name="VVV")


# variance_structure == "freed" &covariance_structure == "fixed"
iris[,-5] \%>\% lpa_create_profiles2(n_profiles = 4, model_name="VVI")


#variance_structure == "constrained" & covariance_structure == "fixed"
iris[,-5] \%>\% lpa_create_profiles2(n_profiles = 4, model_name="EEI")
#'
}

}
