% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_explore_modelfit.R
\name{lpa_explore_modelfit}
\alias{lpa_explore_modelfit}
\title{lpa_explore_modelfit}
\usage{
lpa_explore_modelfit(
  df,
  n_profiles_range = 1:9,
  modelNames = c("EEV", "EEI", "EEE", "EII", "VVI", "VII", "VVV"),
  chk = FALSE,
  basic = FALSE,
  set_n_profile = NULL,
  na2zero = TRUE
)
}
\arguments{
\item{df}{data}

\item{n_profiles_range}{1:9}

\item{modelNames}{"EEV", "EEI", "EEE", "EII", "VVI", "VII", "VVV"}

\item{chk}{default FALSE, TRUE show}

\item{basic}{select EII, EEE, VII, EEI}

\item{set_n_profile}{default select row number then result show}

\item{na2zero}{na2zero is NA data to 0}
}
\value{
data
}
\description{
lpa_explore_modelfit
}
\examples{

\dontrun{

df <- select(iris, -Species)
lpa_explore_modelfit(df, 1:4)
#' lpa_explore_modelfit(df, 1:9)
lpa_explore_modelfit(pwjlpa,1:4, basic=TRUE)
lpa_explore_modelfit(pwjlpa,1:4, basic=TRUE, set_n_profile=4)
lpa_explore_modelfit(pwjlpa,1:6, basic=TRUE)

lpa_explore_modelfit(pwjlpa,1:6, basic=FALSE)
lpa_explore_modelfit(pwjlpa,1:10, basic=TRUE, set_n_profile=4)
lpa_explore_modelfit(pwjlpa,1:10, basic=TRUE, set_n_profile=4, chk=TRUE)

}

}
