% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_profile_plot_summary.R
\name{lpa_profile_plot_summary}
\alias{lpa_profile_plot_summary}
\title{lpa_profile_plot summary data}
\usage{
lpa_profile_plot_summary(resl, digits = 3)
}
\arguments{
\item{resl}{result}

\item{digits}{default 3}
}
\value{
df
}
\description{
lpa_profile_plot summary data
}
\examples{
\dontrun{
#'
lpa_profile_plot(edlcarename, n_profiles = 4, model_name = "EII", gtype="trans",
                 fct_reorder= f_order,
                 size.x=16, alpha = 0.2)\%>\%
  lpa_profile_plot_summary() \%>\%
  nice_table()
#'
}

}
