% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_LRT.R
\name{LRT_apa2}
\alias{LRT_apa2}
\title{LRT_apa}
\usage{
LRT_apa2(data, md = FALSE)
}
\arguments{
\item{data}{anova data, LRT data}

\item{md}{md=TRUE}
}
\value{
table explain
}
\description{
LRT_apa
}
\examples{

\dontrun{
#'

library(jjstat)
HS.model <- '
    visual  =~ x1 + b1*x2 + x3
    textual =~ x4 + b2*x5 + x6
    speed   =~ x7 + b3*x8 + x9
'
fit1 <- cfa(HS.model, data = HolzingerSwineford1939)
fit0 <- cfa(HS.model, data = HolzingerSwineford1939,
            orthogonal = TRUE)

fit1 \%>\% sem_effect(effect ="=~")
fit0 \%>\% sem_effect(effect ="=~")
lavTestLRT(fit1, fit0)
lavTestLRT(fit1, fit0) \%>\% LRT_apa()
}


}
