% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_c_text.R
\name{make_c_text}
\alias{make_c_text}
\title{make_c_text}
\usage{
make_c_text(data)
}
\arguments{
\item{data}{vector data}
}
\value{
text
}
\description{
make_c_text
}
\examples{


\dontrun{

# Example usage
data <- c("FieldGoals", "OtherTDs", "YardsRushAtt", "RushYards",
          "RushFirstDown", "YardsPassComp", "PassYards", "PassFirstDown",
          "YardsRushAtt", "RushYards", "RushFirstDown", "YardsPassComp",
          "PassYards", "PassFirstDown", "PointsGame", "OffensTD",
          "TDGame", "Speical", "Rushing", "Passing",
          "Offense", "Scoring", 123, 456.78)

make_c_text(data)

}
}
