% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapaste_dfTotext.R
\name{make_df_text}
\alias{make_df_text}
\title{data.frame to input text}
\usage{
make_df_text(data, rownames = FALSE, text = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{rownames}{rownames=TRUE, FALSE -> nothing rownames}

\item{text}{Convert all configurations to char, e.g. '10'}
}
\value{
text
}
\description{
data.frame to input text
}
\examples{

\dontrun{

# 데이터 프레임 생성
cho2013_cor <- data.frame(
  필요성 = c(1.000, 0.320, 0.266, 0.506, 0.429, 0.401),
  수업 = c(0.320, 1.000, 0.647, 0.463, 0.548, 0.554),
  교사 = c(0.266, 0.647, 1.000, 0.383, 0.491, 0.545),
  계획 = c(0.506, 0.463, 0.383, 1.000, 0.531, 0.491),
  영향 = c(0.429, 0.548, 0.491, 0.531, 1.000, 0.743),
  만족도 = c(0.401, 0.554, 0.545, 0.491, 0.743, 1.000)
)
cho2013_cor \%>\% make_df_text()


}

}
