% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_make_layout.R
\name{make_layout}
\alias{make_layout}
\title{Create a Custom Layout Matrix}
\usage{
make_layout(nc = 10, nr = NULL)
}
\arguments{
\item{nc}{Integer. Number of columns in the layout matrix. Default is 10.}

\item{nr}{Integer or NULL. Number of rows in the layout matrix. If NULL, the number of rows will be automatically determined. Default is NULL.}
}
\value{
A matrix representing the layout, optionally edited interactively using the `edit` function.
}
\description{
Create a Custom Layout Matrix
}
\examples{
\dontrun{
make_layout(nc = 7)
make_layout(nc = 7, nr = 6)
make_layout(3, 6)
}
}
