% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_list_text.R
\name{make_list_text}
\alias{make_list_text}
\title{make_list_text}
\usage{
make_list_text(data)
}
\arguments{
\item{data}{list data}
}
\value{
text
}
\description{
make_list_text
}
\examples{

\dontrun{
#

# Example data (characters)
data1 <- list(
  Speical = c("FieldGoals", "OtherTDs"),
  Rushing = c("YardsRushAtt", "RushYards", "RushFirstDown"),
  Passing = c("YardsPassComp", "PassYards", "PassFirstDown"),
  Offense = c("rush1", "pass1"),
  Scoring = c("PointsGame", "OffensTD", "TDGame")
)
make_list_text(data1)

# Example data (numeric)
data2 <- list(
  Speical = 7:8,
  Rushing = 1:3,
  Passing = 4:6,
  Offense = 18:19,
  Scoring = 9:11
)
make_list_text(data2)



# Example data (numeric)
list(7:8, 1:3, 4:6, 18:19, 9:11)\%>\%make_list_text()
#
#
}


}
