% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapaste_matTotext.R
\name{make_mat_text2}
\alias{make_mat_text2}
\title{첫번째 열이 존재하는 경우에 사용하는 make_mat_text2}
\usage{
make_mat_text2(data, text = TRUE, firstcol = 1)
}
\arguments{
\item{data}{data.frame}

\item{text}{text TRUE}

\item{firstcol}{ncol setting}
}
\value{
text
}
\description{
첫번째 열이 존재하는 경우에 사용하는 make_mat_text2
}
\examples{

\dontrun{
#'
data_text = "Tobacco_Dependence (0,5] (5,10] (10,15] (15,20] (20,98]
No_Nicotine_Dependence 130 210 43 114 20
Nicotine_Dependence 119 267 91 254 67"
nt=data_text\%>\% text2df(header=T)\%>\%tibble::column_to_rownames("Tobacco_Dependence")
nt \%>\%tibble()
nt=Char2num(nt,1:5)
nt
}


}
