% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnema_test.R
\name{mcnemar_test}
\alias{mcnemar_test}
\title{McNemar's Chi-squared test}
\usage{
mcnemar_test(mat, correct = TRUE, exact = TRUE, show = TRUE)
}
\arguments{
\item{mat}{matrix}

\item{correct}{continuity correction}

\item{exact}{binom.test exact p, 95%CI}

\item{show}{show TRUE message FALSE only data output}
}
\value{
report data
}
\description{
McNemar's Chi-squared test
}
\examples{

\dontrun{
#'
examtest_post  = matrix(c(6,8,1,5), nrow=2,
                        dimnames= list(c("사전검사합격","사후검사불합격"),
                                       c("사후검사합격","사후검사불합격")))
examtest_post
#base
mcnemar.test(examtest_post, correct=FALSE)
mcnemar.test(examtest_post, correct=TRUE)
# this function
binom.test(1,9)
mcnemar_test(examtest_post)
mcnemar_test(examtest_post, correct=F)


#exsmale 2
gen_table = matrix(c(c(9, 3, 13, 10)),ncol=2,
                   dimnames=list("case"=c("yes", "no"),
                                 "control"=c("yes", "no")   ))
gen_table
mcnemar_test(gen_table)
mcnemar_test(gen_table, correct = FALSE)
mcnemar_test(gen_table, correct = FALSE, exact=FALSE)


#example 3

Buying_car  = matrix(c(23,28,7,12), nrow=2,
                     dimnames= list(c("시승전_있음","시승전_없음"),
                                    c("시승후_있음","시승후_없음")))
Buying_car


mcnemar_test(Buying_car)
mcnemar_test(Buying_car, show=FALSE)

}

}
