% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_mysummary.R
\name{mean_apa}
\alias{mean_apa}
\title{mean_apa: Generate descriptive statistics in APA style}
\usage{
mean_apa(df, sd = FALSE)
}
\arguments{
\item{df}{A data frame containing descriptive statistics, including variables such as N, MEAN, SD, Skew, and Kurt.}

\item{sd}{A logical value. If TRUE, includes standard deviation (SD) along with the mean in the output. Default is FALSE.}
}
\value{
A string formatted in APA style, summarizing the descriptive statistics in order of mean values and assessing normality based on skewness and kurtosis.
}
\description{
mean_apa: Generate descriptive statistics in APA style
}
\examples{
df <- data.frame(
  var = c("Q33_1", "Q33_2", "Q33_3"),
  N = c(4542, 4542, 4542),
  MEAN = c(1.9489, 1.9337, 1.9243),
  SD = c(0.84679, 0.86568, 0.86960),
  Skew = c(0.72576, 0.63900, 0.70717),
  Kurt = c(0.63369, 0.087763, 0.27470)
)
mean_apa(df, sd = TRUE)

# Example Output:
# "The results of the descriptive statistics analysis are as follows.
# Variables ordered by mean are Q33_1 (M=1.9489, SD=0.84679) > Q33_2 (M=1.9337, SD=0.86568) >
# Q33_3 (M=1.9243, SD=0.86960). Skewness ranged from 0.63900 to 0.72576, and kurtosis ranged from
# 0.087763 to 0.63369. Following Kline (2005), skewness values exceeding an absolute value of 3
# and kurtosis values exceeding an absolute value of 8 or 10 are considered non-normal. Skewness
# and kurtosis met the criteria for normality."
}
