% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bar.R
\name{mean_bar}
\alias{mean_bar}
\title{mean_bar: Create a Bar Plot to Compare Means}
\usage{
mean_bar(df, x.title = "문항", sort = FALSE)
}
\arguments{
\item{df}{A data frame containing at least two columns:
- `var`: A categorical variable representing the grouping.
- `MEAN`: A numeric variable representing the mean values.}

\item{x.title}{text default "문항"}

\item{sort}{Logical. If TRUE, sorts the bars by mean values. Default is FALSE.}
}
\value{
A ggplot2 object displaying the bar plot.
}
\description{
This function generates a bar plot comparing the mean values of a dependent variable,
with options to sort the bars based on the mean values.
}
\examples{
\dontrun{
# Example data
data <- data.frame(var = c("A", "B", "C"),
                   MEAN = c(3.5, 2.8, 4.1))
# Plot without sorting
mean_bar(data)

# Plot with sorting
mean_bar(data, sort = TRUE)
}
}
