% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCentering_df.R
\name{meanCentering_df}
\alias{meanCentering_df}
\title{meancentering mixed model}
\usage{
meanCentering_df(data, grp_id, gv, ...)
}
\arguments{
\item{data}{data.frame}

\item{grp_id}{group id}

\item{gv}{group variable, Group Variables That Separate Groups}

\item{...}{individual measurement independent variables evaluated at individual level}
}
\description{
meancentering mixed model
}
\examples{
\dontrun{
meanCentering_df(mtcars[,c("cyl","mpg","wt","hp")],"cyl","mpg","wt","hp")|> data.frame()
}



}
