% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_effect.R
\name{med_effect}
\alias{med_effect}
\title{DE,IE,TE Effect extraction}
\usage{
med_effect(x, effect1 = "", caption = "", effect2 = ".", option = 1)
}
\arguments{
\item{x}{lavaan result}

\item{effect1}{first, you must select effect1 = IE, DE, TE}

\item{caption}{title}

\item{effect2}{'.' 2nd oprion}

\item{option}{1(p_value) or 2(ci)}
}
\value{
effect result
}
\description{
DE,IE,TE Effect extraction
}
