% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_min_max_scaler.R
\name{min_max_scaler}
\alias{min_max_scaler}
\title{비율 척도로 변환하는 함수 정의}
\usage{
min_max_scaler(data, columns = NULL)
}
\arguments{
\item{data}{data}

\item{columns}{colum all}
}
\value{
df
}
\description{
비율 척도로 변환하는 함수 정의
}
\examples{
\dontrun{
#'
 사용 예
data_frame: 데이터프레임 객체
변환할 열 이름을 명시적으로 지정하거나 NULL로 모든 열을 변환
scaled_data <- min_max_scaler(data_frame, columns = c("V1_1", "V2_1", "V3_1"))
print(scaled_data)
}
}
