% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missCheck.R
\name{missCheck}
\alias{missCheck}
\title{Check and resolve missing data-}
\usage{
missCheck(data, plot = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{plot}{missing data plot}
}
\value{
Results of checking for missingness
}
\description{
If there are missing values, print how many there are
}
\examples{
\dontrun{
 sleep \%>\% str()
sleep \%>\% missCheck()
}


}
