% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missfillin.R
\name{missfillin}
\alias{missfillin}
\title{Function to determine missing values and then fill in any missing values}
\usage{
missfillin(data, fn = mean)
}
\arguments{
\item{data}{data.frame}

\item{fn}{mean, median}
}
\value{
missing data fill in
}
\description{
Function to determine missing values and then fill in any missing values
}
\examples{
\dontrun{
sleep$NonD \%>\% missfillin()
sleep$NonD \%>\% missfillin(fn=median)
sleep$Exp \%>\% missfillin()
}

}
