% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missfillinFor.R
\name{missfillinFor}
\alias{missfillinFor}
\title{Fill in missing information at once}
\usage{
missfillinFor(data, fn = mean, range = NULL)
}
\arguments{
\item{data}{data,frame}

\item{fn}{mean, median}
}
\value{
result fill data
}
\description{
Fill in missing information at once
}
\examples{
\dontrun{
# sleep \%>\% str()
sleep \%>\% ncol()
sleep \%>\% missfillinFor()
sleep \%>\% missfillinFor(fn=median)
}
}
