% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_mlsem_fornell_larcker.R
\name{mlsem_fornell_larcker}
\alias{mlsem_fornell_larcker}
\title{Fornell–Larcker Table with Significance (Multilevel SEM)}
\usage{
mlsem_fornell_larcker(fit, level = c("within", "between"), digits = 3)
}
\arguments{
\item{fit}{lavaan multilevel SEM object}

\item{level}{"within" or "between"}

\item{digits}{numeric}
}
\description{
Fornell–Larcker Table with Significance (Multilevel SEM)
}
