% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modindices.R
\name{Modindices}
\alias{Modindices}
\title{Modification Indices}
\usage{
Modindices(x, mi.level = 10, op1 = "~~")
}
\arguments{
\item{x}{lavaan result}

\item{mi.level}{mi over 10, default}

\item{op1}{regression '~~'}
}
\value{
chisquare
}
\description{
Given a fitted lavaan object, compute the modification indices (= univariate score tests) for a selected set of fixed-to-zero parameters.
}
\examples{
\dontrun{
library(lavaan)
model1 <- '
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
 # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
'

fit1 <- sem(model1, data = PoliticalDemocracy)

 summary(fit1, standardized = TRUE)
cfa2(fit1)

Modindices(fit1)
}



}
