% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_col.R
\name{move_col}
\alias{move_col}
\title{move_col}
\usage{
move_col(data, from, to, history = FALSE)
}
\arguments{
\item{data}{df}

\item{from}{from}

\item{to}{to}

\item{history}{histor show text}
}
\value{
df
}
\description{
move_col
}
\examples{
\dontrun{
#'
# 사용 예시
df <- data.frame(
  A = 1:5,
  B = 6:10,
  C = 11:15,
  D = 16:20
)

# 2번째 열을 4번째 위치로 이동
result_df <- move_col(df, from = 2, to = 4, history = TRUE)
print(result_df)

}

}
