% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_col.R
\name{move_cols}
\alias{move_cols}
\title{move_cols replication}
\usage{
move_cols(data, ..., history = TRUE)
}
\arguments{
\item{data}{data frame}

\item{...}{from to, from to}

\item{history}{history show text}
}
\value{
df
}
\description{
move_cols replication
}
\examples{
\dontrun{
#'
# 사용 예시
df <- data.frame(
  A = 1:5,
  B = 6:10,
  C = 11:15,
  D = 16:20
)

# 여러 열을 이동: 2번째 열을 4번째로, 3번째 열을 1번째로 이동
result_df <- move_cols(df, 2, 4, 3, 1, history = TRUE)
print(result_df)

}

}
