% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_mat.R
\name{move_mat}
\alias{move_mat}
\title{move_mat matrix element coord position}
\usage{
move_mat(data, ..., imp = NA, replace = FALSE)
}
\arguments{
\item{data}{mat}

\item{...}{coord}

\item{imp}{NA}

\item{replace}{FALSE}
}
\value{
mat
}
\description{
move_mat matrix element coord position
}
\examples{

\fontrun{
#'
# example
data1 <- matrix(c(NA, NA, NA, NA, NA, NA, NA, NA, NA, "D_P3",
                  NA, NA, NA, NA, NA, NA, "진로태도", NA, NA, "D_P2",
                  NA, NA, NA, NA, NA, NA, NA, NA, NA, "D_P1",
                  "C_S5", NA, NA, NA, NA, NA, NA, NA, NA, "B14",
                  "C_S4", NA, NA, NA, NA, NA, NA, NA, NA, "B10",
                  "C_S3", NA, "자기효능감", NA, NA, NA, "진로준비", NA, NA, "B04",
                  "C_S2", NA, NA, NA, NA, NA, NA, NA, NA, "B02"))

# Test with replace = FALSE
move_mat(data, list(10, 10, 1, 2), list(9, 10, 1, 3), list(8, 10, 6,3), replace = FALSE)
move_mat(data, c(10, 10, 1, 2), c(9, 10, 1, 3), c(8, 10, 6,3), replace = FALSE)


# Test with replace = TRUE
move_mat(data, list(10, 10, 1, 2), list(9, 10, 1, 3), list(8, 10, 6,3), replace = TRUE)
move_mat(data, c(10, 10, 1, 2), c(9, 10, 1, 3), c(8, 10, 6,3), replace = TRUE)
}


}
