% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_mat_change.R
\name{move_mat_change}
\alias{move_mat_change}
\title{Functions that replace a single component}
\usage{
move_mat_change(mat, pattern, imp)
}
\arguments{
\item{mat}{matrix}

\item{pattern}{pattern}

\item{imp}{imputation}
}
\value{
matrix
}
\description{
Functions that replace a single component
}
\examples{

\dontrun{
#'
plspm_boot_factor_layout(pls_ts)\%>\%
  move_mat_cut(1,3)\%>\%
  move_mat_change("int","i_I")\%>\%make_mat_text()

layout_New = matrix(
  c( NA, 'IMA',  NA, NA, 'i_I',
     NA, 'QUA',NA,  'VAL', 'LOY',
     'EXP', NA, 'SAT',NA, NA),
  nrow = 3, ncol = 5, byrow = TRUE)
}


}
