% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_mat_change.R
\name{move_mat_nc}
\alias{move_mat_nc}
\title{move_mat_nc}
\usage{
move_mat_nc(mat, row, col, name)
}
\arguments{
\item{mat}{mat}

\item{row}{row}

\item{col}{cp;}

\item{name}{change name}
}
\value{
mat
}
\description{
move_mat_nc
}
