% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_mat_ori.R
\name{move_mat_ori}
\alias{move_mat_ori}
\title{move_mat_ori}
\usage{
move_mat_ori(mat, sort = TRUE)
}
\arguments{
\item{mat}{mat}

\item{sort}{sort =TRUE}
}
\value{
coordnate
}
\description{
move_mat_ori
}
\examples{

\dontrun{
#'
layout_jut <- matrix(
  c(NA, NA, NA, NA, NA, NA, NA, NA, NA, 'D_P3',
    NA, NA, NA, NA, NA, NA, '진로태도', NA, NA, 'D_P2',
    NA, NA, NA, NA, NA, NA, NA, NA, NA, 'D_P1',
    'C_S5', NA, NA, NA, NA, NA, NA, NA, NA, 'B14',
    'C_S4', NA, NA, NA, NA, NA, NA, NA, NA, 'B10',
    'C_S3', NA, '자기효능감', NA, NA, NA, '진로준비', NA, NA, 'B04',
    'C_S2', NA, NA, NA, NA, NA, NA, NA, NA, 'B02',
    'C_S1', NA, NA, NA, NA, NA, NA, NA, NA, 'A_M3',
    NA, NA, NA, NA, NA, NA, NA, NA, NA, 'A_M2',
    NA, NA, NA, NA, NA, NA, '진로동기', NA, NA, 'A_M1'),
  nrow = 10, ncol = 10, byrow = TRUE)
move_mat_ori(layout_jut)
}


}
