% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_mat_cut.R
\name{move_mat_substr}
\alias{move_mat_substr}
\title{move_mat_cut}
\usage{
move_mat_substr(mat, start = 1, n = 3)
}
\arguments{
\item{mat}{matrix}

\item{start}{1}

\item{n}{width = 3}
}
\value{
matrix
}
\description{
move_mat_cut
}
\examples{

\dontrun{
#'
# 테스트용 행렬 데이터
mat <- matrix(c(NA, "진로동기", NA, "자기효능감", NA,
    "진로준비", NA, "진로태도", NA), nrow = 3, byrow = TRUE)

# 함수 실행
move_mat_cut(mat, start = 1, n = 3)

}


}
