% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_matpos.R
\name{move_matpos}
\alias{move_matpos}
\title{move_matpos is element change position}
\usage{
move_matpos(data, ..., imp = NA, replace = FALSE)
}
\arguments{
\item{data}{mat}

\item{...}{position var}

\item{imp}{NA}

\item{replace}{FALSE}
}
\value{
mat
}
\description{
move_matpos is element change position
}
\examples{

\dontrun{
#'
# # Original matrix
layout_jut <- matrix(
  c(NA, NA, NA, NA, NA, NA, NA, NA, NA, 'D_P3',
    NA, NA, NA, NA, NA, NA, '진로태도', NA, NA, 'D_P2',
    NA, NA, NA, NA, NA, NA, NA, NA, NA, 'D_P1',
    'C_S5', NA, NA, NA, NA, NA, NA, NA, NA, 'B14',
    'C_S4', NA, NA, NA, NA, NA, NA, NA, NA, 'B10',
    'C_S3', NA, '자기효능감', NA, NA, NA, '진로준비', NA, NA, 'B04',
    'C_S2', NA, NA, NA, NA, NA, NA, NA, NA, 'B02',
    'C_S1', NA, NA, NA, NA, NA, NA, NA, NA, 'A_M3',
    NA, NA, NA, NA, NA, NA, NA, NA, NA, 'A_M2',
    NA, NA, NA, NA, NA, NA, '진로동기', NA, NA, 'A_M1'),
  nrow = 10, ncol = 10, byrow = TRUE)
layout_jut
# Apply the matpos function
layout_jut\%>\% move_matpos(
  list("진로태도", 3, 5),
  list("진로동기", 8, 5),
  list("B02", 1, 3),
  list("B04", 1, 5),
  list("B10", 1, 7),
  list("B14", 1, 9),
  list("D_P3", 4, 10),
  list("D_P2", 6, 10),
  list("D_P1", 8, 10),
  list("A_M1", 10, 3),
  list("A_M2", 10, 4),
  list("A_M3", 10, 5),
  replace = FALSE
)


matrix(
  c(NA, NA, NA, NA, NA, NA, NA, NA, NA, 'D_P3',
    NA, NA, NA, NA, NA, NA, '진로태도', NA, NA, 'D_P2',
    NA, NA, NA, NA, NA, NA, NA, NA, NA, 'D_P1',
    'C_S5', NA, NA, NA, NA, NA, NA, NA, NA, 'B14',
    'C_S4', NA, NA, NA, NA, NA, NA, NA, NA, 'B10',
    'C_S3', NA, '자기효능감', NA, NA, NA, '진로준비', NA, NA, 'B04',
    'C_S2', NA, NA, NA, NA, NA, NA, NA, NA, 'B02',
    'C_S1', NA, NA, NA, NA, NA, NA, NA, NA, 'A_M3',
    NA, NA, NA, NA, NA, NA, NA, NA, NA, 'A_M2',
    NA, NA, NA, NA, NA, NA, '진로동기', NA, NA, 'A_M1'),
  nrow = 10, ncol = 10, byrow = TRUE)\%>\% move_matpos(
    list("진로태도", 3, 5),
    list("진로동기", 8, 5),
    list("B02", 1, 3),
    list("B04", 1, 5),
    list("B10", 1, 7),
    list("B14", 1, 9),
    list("D_P3", 4, 10),
    list("D_P2", 6, 10),
    list("D_P1", 8, 10),
    list("A_M1", 10, 3),
    list("A_M2", 10, 4),
    list("A_M3", 10, 5),
    replace = FALSE
  )

}


}
