% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_rows.R
\name{move_row}
\alias{move_row}
\title{Functions to move rows to a desired position}
\usage{
move_row(data, from, to, history = FALSE)
}
\arguments{
\item{data}{data.fram}

\item{from}{row from}

\item{to}{row to}

\item{history}{history}
}
\value{
data
}
\description{
Functions to move rows to a desired position
}
\examples{

\dontrun{
#' #'
#' # Example data
car = data.frame(
  rownames = c(paste0("Row_",1:6)),
  mpg = c('21', '21', '22.8', '21.4', '18.7', '18.1'),
  cyl = c('6', '6', '4', '6', '8', '6'),
  disp = c('160', '160', '108', '258', '360', '225')
)

car
car\%>\%move_row(1,6)
car\%>\%move_row(6,1)
car\%>\%move_row(6,2)
car\%>\%move_row(6,3)
car\%>\%move_row(1,3)
car\%>\%move_row(6,5)

car
car\%>\%
  move_row(6,1)\%>\%
  move_row(1,3)\%>\%
  move_row(3,2)
#'
}


}
