% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_row.R
\name{move_rows}
\alias{move_rows}
\title{move_row rep function}
\usage{
move_rows(data, ...)
}
\arguments{
\item{data}{data}

\item{...}{movelist c(a,b),c(c,d)...}
}
\value{
data
}
\description{
move_row rep function
}
\examples{

\dontrun{
#'
data <- data.frame(
  source = c("A05", "A08", "A04", "A06", "A09", "A11", "A13", "A02", "A03", "A01", "A14", "A15", "A07", "A12", "A10", "eigen_value", "Proportion_Var", "Cumulative_Var"),
  MR1 = c("", "", "", "", "", "", "", "0.362", "0.449", "0.457", "0.47", "0.538", "0.62", "0.792", "0.96", "3.363", "0.224", "0.224"),
  MR2 = c("", "0.382", "", "", "0.445", "0.724", "0.906", "0.392", "", "", "0.487", "", "", "", "", "2.095", "0.14", "0.364"),
  MR3 = c("", "", "0.402", "0.963", "", "", "", "", "", "", "", "", "", "", "", "1.691", "0.113", "0.477")
)
data
#'
move_row(data, 14,1)
move_row(data, 14,1)\%>\% move_row(1,2)
move_row(data, 14,1)\%>\% move_row(1,2)\%>\% move_row(2,3)
#'
data \%>\% move_rows(c(14, 1), c(1, 2))
data \%>\% move_rows(c(14, 1), c(1, 2), c(2, 3))
}

}
