% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_est_df.R
\name{mplus_est_df}
\alias{mplus_est_df}
\title{mplus_est_df}
\usage{
mplus_est_df(text, digits = 3)
}
\arguments{
\item{text}{text est}

\item{digits}{3}
}
\value{
data.frame
}
\description{
mplus_est_df
}
\examples{


\dontrun{
#'
esttest <- "MODEL RESULTS

                                                    Two-Tailed
                    Estimate       S.E.  Est./S.E.    P-Value

 F1       BY
    X1                 1.000      0.000    999.000    999.000
    X2                 0.771      0.034     22.380      0.000
    X3                 0.683      0.033     20.510      0.000

 F2       BY
    X4                 1.000      0.000    999.000    999.000
    X5                 0.722      0.032     22.400      0.000
    X6                 0.702      0.033     21.320      0.000

 F3       BY
    Y1                 1.000      0.000    999.000    999.000
    Y2                 0.705      0.029     24.023      0.000
    Y3                 0.695      0.029     24.243      0.000

 F2       WITH
    F1                -0.482      0.082     -5.912      0.000

 F3       WITH
    F1                 0.522      0.084      6.203      0.000
    F2                -0.558      0.088     -6.315      0.000

 Variances
    F1                 1.347      0.121     11.095      0.000
    F2                 1.492      0.131     11.382      0.000
    F3                 1.599      0.135     11.868      0.000

 Residual Variances
    X1                 0.364      0.046      7.916      0.000
    X2                 0.245      0.029      8.583      0.000
    X3                 0.320      0.029     11.160      0.000
    X4                 0.347      0.048      7.224      0.000
    X5                 0.279      0.029      9.579      0.000
    X6                 0.342      0.031     10.865      0.000
    Y1                 0.302      0.044      6.931      0.000
    Y2                 0.290      0.028     10.444      0.000
    Y3                 0.269      0.026     10.230      0.000"

mplus_est_df(esttest)
#'
}

}
