% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_fit_df.R
\name{mplus_fit_df}
\alias{mplus_fit_df}
\title{mplus_fit_df: model fit from  Mplus result}
\usage{
mplus_fit_df(text, digits = 3)
}
\arguments{
\item{text}{mplut copy data}

\item{digits}{3}
}
\value{
fit data.frame
}
\description{
mplus_fit_df: model fit from  Mplus result
}
\examples{


\dontrun{


text= c("MODEL FIT INFORMATION

Number of Free Parameters                       21

Loglikelihood

          H0 Value                       -4568.339
          H1 Value                       -4551.194

Information Criteria

          Akaike (AIC)                    9178.677
          Bayesian (BIC)                  9263.623
          Sample-Size Adjusted BIC        9196.983
            (n* = (n + 2) / 24)

Chi-Square Test of Model Fit

          Value                             34.290
          Degrees of Freedom                    24
          P-Value                           0.0796

RMSEA (Root Mean Square Error Of Approximation)

          Estimate                           0.032
          90 Percent C.I.                    0.000  0.054
          Probability RMSEA <= .05           0.901

CFI/TLI

          CFI                                0.996
          TLI                                0.994

Chi-Square Test of Model Fit for the Baseline Model

          Value                           2533.266
          Degrees of Freedom                    36
          P-Value                           0.0000

SRMR (Standardized Root Mean Square Residual)

          Value                              0.022")

mplus_fit_df(text)
#' }
}
