% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_rows.R
\name{mrs}
\alias{mrs}
\title{move_row rep function}
\usage{
mrs(data, ..., history = TRUE)
}
\arguments{
\item{data}{data}

\item{...}{movelist c(a,b),c(c,d)...}
}
\value{
data
}
\description{
move_row rep function
}
\examples{

\dontrun{
#'
data <- data.frame(
  source = c("A05", "A08", "A04",
  "A06", "A09", "A11", "A13", "A02",
  "A03", "A01", "A14", "A15", "A07",
  "A12", "A10", "eigen_value", "Proportion_Var", "Cumulative_Var"),
  MR1 = c("", "", "", "", "", "", "",
  "0.362", "0.449", "0.457", "0.47",
  "0.538", "0.62", "0.792", "0.96",
  "3.363", "0.224", "0.224"),
  MR2 = c("", "0.382", "", "", "0.445",
  "0.724", "0.906", "0.392", "", "",
  "0.487", "", "", "", "", "2.095", "0.14", "0.364"),
  MR3 = c("", "", "0.402", "0.963", "",
   "", "", "", "", "", "", "", "",
    "", "", "1.691", "0.113", "0.477")
)
data
#'
move_row(data, 14,1)
move_row(data, 14,1)\%>\% move_row(1,2)
move_row(data, 14,1)\%>\% move_row(1,2)\%>\% move_row(2,3)
#'
data \%>\% move_rows(c(14, 1), c(1, 2))
data \%>\% move_rows(c(14, 1), c(1, 2), c(2, 3))
#'
car = data.frame(
  rownames = c(paste0("Row_",1:6)),
  mpg = c('21', '21', '22.8', '21.4', '18.7', '18.1'),
  cyl = c('6', '6', '4', '6', '8', '6'),
  disp = c('160', '160', '108', '258', '360', '225'))
cars

#repeated
cars
car \%>\%
  move_row(6,1)\%>\%
  move_row(1,3)\%>\%
  move_row(3,2)

#history
car \%>\%
  move_row(6,1, history=TRUE)\%>\%
  move_row(1,3, history=TRUE)\%>\%
  move_row(3,2, history=TRUE)\%>\%
  move_row(2,6, history=TRUE)

car\%>\%move_rows(6,1) #errorcheck
car\%>\%move_rows(c(6,1), c(1,2),c(2,6))
car\%>\%move_rows(c(6,1), c(1,2),c(2,6), history=T)

# simple input method
move_rows(car, 6, 1, 1, 2, 2, 6, history=TRUE)
car\%>\% move_rows( 6, 1, 1, 2, 2, 6, 6,3, history=TRUE)
# Testing the function with direct pair inputs
move_rows(car, 6, 1, 1, 2, 2, 6, history=TRUE)

# Testing the function with list of pairs inputs
move_rows(car, c(6, 1), c(1, 2), c(2, 6), history=TRUE)


# Testing the function with direct pair inputs
move_rows2(car, 6, 1,  history=FALSE)
move_rows2(car, 6, 1,1,2)
move_rows2(car, 6, 1, 1, 2, 2, 6, history=TRUE)
car\%>\%move_rows2(6, 1, 1, 2, 2, 6, history=TRUE)
move_rows2(car, 6, 1, 1, 2, 2, 6, 6,3, history=TRUE)
car\%>\%move_rows2(c(6,1), c(1,2),c(2,6),c(6,5), history=T)
car\%>\%move_rows2(6, 1, 1, 2, 2, 6, 6,3, history=T)


# Testing the function with list of pairs inputs
move_rows2(car, c(6, 1), history=TRUE)
move_rows(car, c(6, 1),c(1,2), history=TRUE)
move_rows2(car, c(6, 1), c(1, 2), c(2, 6),c(6,3), history=TRUE)
#'
#'



}

}
