% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysummary.R
\name{mydes}
\alias{mydes}
\title{mydes 'n','mean','sd','min','max'}
\usage{
mydes(myvariable, var = NULL, digits = 2)
}
\arguments{
\item{myvariable}{variable vector}

\item{var}{input variable name}

\item{digits}{digits}
}
\description{
mydes 'n','mean','sd','min','max'
}
\examples{
\dontrun{
## view variable
mydes(mtcars$mpg,"mpg")
##not see variable
mydes(mtcars$mpg)

}

}
