% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydescriptive.R
\name{mydescriptive}
\alias{mydescriptive}
\title{mydescriptive 'n','mean','sd','min','max'}
\usage{
mydescriptive(myvariable, digits = 3)
}
\arguments{
\item{myvariable}{variable vector}

\item{digits}{digits}
}
\description{
mydescriptive 'n','mean','sd','min','max'
}
\examples{
\dontrun{

mydescriptive(mtcars$mpg)

}

}
