% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysummary.R
\name{mysummary}
\alias{mysummary}
\title{my summary descriptive statistics}
\usage{
mysummary(myobject, ..., digits = 5, msdn = FALSE, na.rm = FALSE)
}
\arguments{
\item{myobject}{data.frame, matrix}

\item{...}{column variable}

\item{digits}{round default 3}

\item{msdn}{TRUE only output var, N, mean. sd}

\item{na.rm}{FALSE, TRUE is remove NA}
}
\value{
size   MEAN    SD  MIN  MAX
}
\description{
my summary descriptive statistics
}
\examples{
\dontrun{
mysummary(mtcars, "mpg","wt","hp")
## size      MEAN         SD    MIN     MAX
## 1   32  20.09062  6.0269481 10.400  33.900
## 2   32   3.21725  0.9784574  1.513   5.424
## 3   32 146.68750 68.5628685 52.000 335.000

## all variable
 jjstat::mysummary(mtcars, colnames(mtcars))

mysummary(mtcars, colnames(mtcars))

mysummary(mtcars)

mysummary(mtcars, all=FALSE)

}
}
