% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_mysummary.R
\name{mysummary_apa}
\alias{mysummary_apa}
\title{mysummary_apa: Generate descriptive statistics in APA style (English or Korean)}
\usage{
mysummary_apa(df, sd = FALSE, eng = TRUE)
}
\arguments{
\item{df}{A data frame containing descriptive statistics, including variables such as N, MEAN, SD, Skew, and Kurt.}

\item{sd}{A logical value. If TRUE, includes standard deviation (SD) along with the mean in the output. Default is FALSE.}

\item{eng}{A logical value. If TRUE, the output is in English. If FALSE, the output is in Korean. Default is TRUE.}
}
\value{
A string formatted in APA style (English or Korean), summarizing the descriptive statistics in order of mean values and assessing normality based on skewness and kurtosis.
}
\description{
mysummary_apa: Generate descriptive statistics in APA style (English or Korean)
}
\examples{
\dontrun{
df <- data.frame(
  var = c("Q33_1", "Q33_2", "Q33_3"),
  N = c(4542, 4542, 4542),
  MEAN = c(1.9489, 1.9337, 1.9243),
  SD = c(0.84679, 0.86568, 0.86960),
  Skew = c(0.72576, 0.63900, 0.70717),
  Kurt = c(0.63369, 0.087763, 0.27470)
)
mysummary_apa(df, sd = TRUE, eng = TRUE)
mysummary_apa(df, sd = TRUE, eng = FALSE)
}
}
