% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_remover.R
\name{na_remover}
\alias{na_remover}
\title{na_remover row unite}
\usage{
na_remover(df_var, pattern = "", type = "data")
}
\arguments{
\item{df_var}{df vector}

\item{pattern}{""}

\item{type}{data}
}
\value{
df
}
\description{
na_remover row unite
}
\examples{
\dontrun{


}
}
