% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_aov_df.R
\name{nice_aov_df}
\alias{nice_aov_df}
\title{Nice tables for papers in aov_df results  to  word, html}
\usage{
nice_aov_df(df, show = TRUE, tibble = FALSE, digits = 2)
}
\arguments{
\item{df}{aov_Df}

\item{show}{TRUE nice, FALSE pass}

\item{digits}{2}
}
\value{
result
}
\description{
Nice tables for papers in aov_df results  to  word, html
}
\examples{

\dontrun{
#'
aov_df(data = mtcars,
       dv_var = "mpg",
       iv_vars = c("cyl", "gear", "carb") )

## generate nice table
aov_df(data = mtcars,
       dv_var = "mpg",
       iv_vars = c("cyl", "gear", "carb") )\%>\%
  nice_aov_df()


aov_df(data = mtcars,
       dv_var = "mpg",
       iv_vars = c("cyl", "gear", "carb") )\%>\%
  nice_aov_df() \%>\%web()


}
}
