% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_alpha_table.R, R/nice_Var.R
\name{nice_Var}
\alias{nice_Var}
\title{bind_alpha_table문자 정리 함수}
\usage{
nice_Var(data)

nice_Var(data)
}
\arguments{
\item{data}{alpha data}
}
\value{
table data.frame

data.frame
}
\description{
bind_alpha_table문자 정리 함수

nice_Var alpha분석한 경우 첫열의 a1->a
}
\examples{

\dontrun{
#'
jutr2 \%>\% select(B01:B25) \%>\%
  bind_alpha_table(c(1:5),c(6:10),c(11:15),c(16:20),c(21:25))\%>\%
  nice_Var()\%>\%  # <----this
  nice_table()
}
}
