% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_normality.R
\name{Normality_test}
\alias{Normality_test}
\title{Multivariate normality tests}
\usage{
Normality_test(multiVariable_data)
}
\arguments{
\item{multiVariable_data}{data}
}
\value{
normality
}
\description{
Multivariate normality tests
}
\examples{

\dontrun{
mtcars\%>\%Normality_test()
}

}
