% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_exp_table.R
\name{obs_exp_table}
\alias{obs_exp_table}
\title{Observation Expectation Table with calculate_chi_sig}
\usage{
obs_exp_table(obs_data)
}
\arguments{
\item{obs_data}{matrix}
}
\value{
matrix
}
\description{
Observation Expectation Table with calculate_chi_sig
}
