% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway_aov.R
\name{oneway_aov}
\alias{oneway_aov}
\title{oneway ANVOA calculation}
\usage{
oneway_aov(formula, data, format = "pandoc")
}
\arguments{
\item{formula}{ex)mpg ~ cyl}

\item{data}{data.frame}

\item{format}{'markdwon', 'pandoc', 'html'}
}
\value{
anova result
}
\description{
oneway ANVOA calculation
}
\examples{
\dontrun{
oneway_aov(mpg ~ cyl, data = mtcars)
}

}
