% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sig_cal.R
\name{p_sig_cal}
\alias{p_sig_cal}
\title{observed expect table with calculate_chi_sig}
\usage{
p_sig_cal(obs_data)
}
\arguments{
\item{obs_data}{matrix}
}
\value{
matrix
}
\description{
observed expect table with calculate_chi_sig
}
