% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_cal.R
\name{p_value_cal}
\alias{p_value_cal}
\title{Observation Expectation Table with calculate_chi_sig}
\usage{
p_value_cal(obs_data)
}
\arguments{
\item{obs_data}{matrix}
}
\value{
matrix
}
\description{
Observation Expectation Table with calculate_chi_sig
}
