% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_detector.R
\name{pattern_detector}
\alias{pattern_detector}
\title{pattern_detector: missing data detect}
\usage{
pattern_detector(df, missing = TRUE, data = FALSE, pattern = "")
}
\arguments{
\item{df}{data}

\item{missing}{missing TRUE detecte missing T,F}

\item{data}{TRUE data print, FALSE rownames number}

\item{pattern}{pattern "" or  NA}
}
\value{
data
}
\description{
pattern_detector: missing data detect
}
\examples{

\dontrun{

ed_tam \%>\% pattern_detector()
ed_tam \%>\% pattern_detector(FALSE,TRUE)\%>\% head()
# missing data chek
ed_tam \%>\% pattern_detector(missing=T, data=F)
#missing data output
ed_tam \%>\% pattern_detector(missing=T, data=T) \%>\%head(30)

# not missing
ed_tam \%>\% pattern_detector(data=T,missing=F) \%>\%head()

# ot missing , not data
ed_tam \%>\% pattern_detector(data=F,missing=F)


}


}
