% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternGraph.R
\name{patternGraph}
\alias{patternGraph}
\title{patternGraph}
\usage{
patternGraph(
  data,
  size = 4,
  strip = 14,
  text = 12,
  axis = 16,
  hjust = -0.4,
  yadd = 10,
  type = "g",
  xlab = "성조형",
  ylab = "빈도",
  show = TRUE,
  tolong = TRUE
)
}
\arguments{
\item{data}{data.frame}

\item{size}{bar text size}

\item{strip}{strip size}

\item{text}{axis text test size}

\item{axis}{aixs size}

\item{hjust}{hjust}

\item{yadd}{yadd}

\item{type}{type "g", "data", "all"}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{show}{show.legend}

\item{tolong}{long format transpose}
}
\value{
graph
}
\description{
patternGraph
}
\examples{
\dontrun{
aaa=kge_bind2 \%>\% filter(지역 =="부산" & N ==1) \%>\%
  auto_pattern("고유어1") \%>\%
  kge_chisq_table("a1","성조", "고유어1 부산")

#sAVE
data_long<- aaa$chi_table \%>\%
  rownames_to_column("syllabic") \%>\%
  pivot_longer(names_to = "accent", values_to = 'ratio', cols=2:4)
data_long

data_long\%>\% ggplot(aes(x = accent, y = ratio))+
  geom_bar(stat = "identity", aes( fill = accent),
           position = "dodge", show.legend = FALSE)+
  geom_hline(yintercept = 1, linetype=2, color="gray80")+
  geom_text(aes(label = round(ratio,2)), hjust = -0.1, size = 4)+
  ylim(0,max(data1$ratio)+0.07)+
  coord_flip()+
  theme_bw()+
  theme(axis.text = element_text(size= 12),
        axis.title = element_text(size= 12),
        strip.text = element_text(size= 14)
  )+
  scale_fill_grey(start = 0, end = 0.7) +
  facet_wrap(~ syllabic )

aaa\%>\% patternGraph()

}
}
