% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternGraph_obs_exp_ko.R
\name{patternGraph_obs_exp_ko}
\alias{patternGraph_obs_exp_ko}
\title{patternGraph_obs_exp}
\usage{
patternGraph_obs_exp_ko(
  data,
  raw = TRUE,
  Ncol = NULL,
  yadd = 0.5,
  strip_size = 16,
  axis_size = 16,
  text_size = 13,
  size_bartext = 5,
  xlab = "기준변수",
  ylab = "관측기대비율",
  Colsnames = NULL,
  type = "g"
)
}
\arguments{
\item{data}{data,frame}

\item{raw}{raw data}

\item{Ncol}{grid}

\item{yadd}{up yadd}

\item{strip_size}{strip}

\item{axis_size}{axis}

\item{text_size}{text}

\item{size_bartext}{bartext}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{Colsnames}{new select col pivot long}

\item{type}{res}
}
\value{
result fill data
}
\description{
patternGraph_obs_exp
}
\examples{
\dontrun{
aaa=kge_bind2 \%>\% filter(지역 =="부산" & N ==1) \%>\%
  auto_pattern("고유어1") \%>\%
  kge_chisq_table("a1","성조", "고유어1 부산")

#sAVE
data_long<- aaa$chi_table \%>\%
  rownames_to_column("syllabic") \%>\%
  pivot_longer(names_to = "accent", values_to = 'ratio', cols=2:4)
data_long

data_long\%>\% ggplot(aes(x = accent, y = ratio))+
  geom_bar(stat = "identity", aes( fill = accent),
           position = "dodge", show.legend = FALSE)+
  geom_hline(yintercept = 1, linetype=2, color="gray80")+
  geom_text(aes(label = round(ratio,2)), hjust = -0.1, size = 4)+
  ylim(0,max(data1$ratio)+0.07)+
  coord_flip()+
  theme_bw()+
  theme(axis.text = element_text(size= 12),
        axis.title = element_text(size= 12),
        strip.text = element_text(size= 14)
  )+
  scale_fill_grey(start = 0, end = 0.7) +
  facet_wrap(~ syllabic )

aaa\%>\% patternGraph_obs_exp_ko()

#신진영(2013)
shin2013_mat =matrix(c(14,1,23,3,4,0,9,0,13,18,89,8,11,40,46,12,3,4,22,0), ncol= 5)
# colnames(shin2013_mat)=c("지능지수가높다","학업성취도가높다","창의적사고를한다","특정분야에서뛰어난능력을보인다","모든분야에서뛰어난능력을보인다")
colnames(shin2013_mat)=c("지능지수","학업성취도","창의적사고","특정분야","모든분야")
rownames(shin2013_mat)=c("고졸","초대졸","대졸","대학원이상")
shin2013_mat

shin2013_mat\%>\% data.frame() \%>\%
  long_df("학력","영재성", cols=2:6) \%>\%
  add_rows_freq("real") \%>\%
  chisq_test_kge("학력","영재성", type="data_graph") \%>\%
  patternGraph_obs_exp_ko()


select
}

}
