% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcfa.R
\name{pcfa_porp_plot}
\alias{pcfa_porp_plot}
\title{pcfa_porp_plot}
\usage{
pcfa_porp_plot(Data, digits = 2, prop = FALSE, ymax = NA)
}
\arguments{
\item{Data}{factanal result uniqueness}

\item{digits}{digits 3}

\item{prop}{FALSE}

\item{ymax}{plot y size}
}
\value{
graph
}
\description{
pcfa_porp_plot
}
\examples{
\dontrun{

library(MVT)
library(GPArotation)
data(examScor)
X = examScor
Z = scale(X, scale=T)  # scale
R = cor(X)  # correlation matrix

## factor analysis tools : factanal
mlfa <- factanal(Z, factors = 2, rotation = "varimax")

mlfa|> pcfa_porp_plot()

}
}
