% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_compare_bar.R
\name{pdata}
\alias{pdata}
\title{pdata plspm group change long data format}
\usage{
pdata(widedata, rownames_to_column=FALSE)
}
\arguments{
\item{data}{wide data}
}
\value{
long
}
\description{
pdata plspm group change long data format
}
