% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_chi_square_test.R
\name{perform_chi_square_test}
\alias{perform_chi_square_test}
\title{Various output functions after chi-square analysis}
\usage{
perform_chi_square_test(observed, type = "data", simple = FALSE)
}
\arguments{
\item{observed}{table data}

\item{type}{type data, res, chisq_test, cramersv, observed_over_expected_ratios, cell_p_values, p_sig}

\item{simple}{* or ***}
}
\value{
chisq table obs/exp
}
\description{
Various output functions after chi-square analysis
}
\examples{

\dontrun{

#' #'mat = matrix(c(36, 67, 11,
          0, 35, 44,
         41, 108, 1,
          56, 87, 54),
           nrow = 4, byrow = TRUE.
   dimnames = list(c("공명음", "마찰음",
     "유기음_경음", "평파열음_평파찰음"),  c("H", "H(H)", "L")))

 mat \%>\% perform_chi_square_test()


mtcars\%>\%select(am, cyl)\%>\%table()\%>\%perform_chi_square_test()
mtcars\%>\%select(am, cyl)\%>\%table()\%>\%perform_chi_square_test("all")
}

}
