% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_multcomp.R
\name{chisq_multcomp2}
\alias{chisq_multcomp2}
\title{Pairwise comparisons after a chi-squared goodness-of-fit test}
\usage{
chisq_multcomp2(df, p.method = "fdr", ko = TRUE)
}
\arguments{
\item{p.method}{method for p-values correction. See help of p.adjust}

\item{ko}{TRUE variable names is changed to korean language}

\item{x}{numeric vector (counts).}
}
\value{
result
}
\description{
Pairwise comparisons after a chi-squared goodness-of-fit test
}
\section{Functions}{
\itemize{
\item \code{chisq_multcomp2()}: pairwise comparisons after a global chi-squared goodness-of-fit test.

}}
\examples{
\dontrun{
#'
counts <- c(49,30,63,59)
chisq.test(counts)
chisq_multcomp(counts)
# compare
chisq.multcomp(counts)

}
}
