% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_pivot_table.R
\name{pivot_table}
\alias{pivot_table}
\title{pivot}
\usage{
pivot_table(
  data,
  formula,
  names_from = NULL,
  unite = FALSE,
  col = 1:2,
  remove = TRUE,
  change_col = TRUE
)
}
\arguments{
\item{data}{data}

\item{formula}{formul}

\item{names_from}{column}

\item{unite}{1,2 colum unite}

\item{col}{합쳐질 컬럼}

\item{remove}{기존것 지우기}

\item{change_col}{열의 위치 바꾸기}
}
\value{
데이터
}
\description{
pivot
}
\examples{

\dontrun{

#' kge_bind3a \%>\%
filter(type == "고유어1")\%>\%
  pivot_table(~ area + age + 성조, unite=TRUE)
#'
kge_bind3a \%>\%
  filter(type == "고유어1")\%>\%
  pivot_table(~  age + area+ 성조, unite=FALSE)

#고유어1 공명음  지역별 연령
kge_bind3a \%>\%
  filter(type == "고유어1" & a1=="공명음")\%>\%
  pivot_table(~ age + area + 성조, unite=T)  #\%>\% md()



kge_bind3a \%>\%
  filter(type == "고유어1")\%>\%
  pivot_table(~  age +  성조, unite=FALSE)

#고유어1 공명음  지역별 연령
kge_bind3a \%>\%
  filter(type == "고유어1" & a1=="공명음")\%>\%
  pivot_table(~  area + 성조, unite=F)  #\%>\% md()


  }
#'

}
