% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_modelfit_apa.R
\name{pls_modelfit_apa}
\alias{pls_modelfit_apa}
\title{pls_modelfit_apa}
\usage{
pls_modelfit_apa(model_fit_data)
}
\arguments{
\item{model_fit_data}{modelfit}
}
\value{
data
}
\description{
pls_modelfit_apa
}
\examples{

\dontrun{
#'
rest= plspm_modelfit(edt_pls, edblock, bootstrap = 1000, ci = 0.95, trans=FALSE)
rest
rest\%>\%t()\%>\%rowdata2col(1)\%>\%data.frame()

pls_modelfit_apa(rest)

}
}
