% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_interaction.R
\name{plspm_add_blocks}
\alias{plspm_add_blocks}
\title{plspm_add_blocks, indicator blocks}
\usage{
plspm_add_blocks(newData, blocks, interactionTerm)
}
\arguments{
\item{newData}{plspm_make_prod data}

\item{interactionTerm}{}
}
\description{
plspm_add_blocks, indicator blocks
}
\examples{
\dontrun{
#' data(satisfaction)
#'
newData1 <- plspm_make_prod(satisfaction, sat_blocks1,
                            list(intIE = interlist("IMAG",1,2, "EXPE",1,2) ) )


interactionTerm1 = list(list(from = "IMAG*EXPE", to = "LOY"))
plspm_add_blocks(newData1, sat_blocks1, interactionTerm1)

}

}
