% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_interaction.R
\name{plspm_add_inter_paths}
\alias{plspm_add_inter_paths}
\title{plspm_add_inter_paths}
\usage{
plspm_add_inter_paths(data, paths)
}
\arguments{
\item{data}{path matrix data}

\item{paths}{add path}
}
\value{
mat
}
\description{
plspm_add_inter_paths
}
\examples{


\dontrun{
#'
data <- data.frame(
  IMAG = c(0, 1, 0, 0, 1, 1),
  EXPE = c(0, 0, 1, 1, 1, 0),
  QUAL = c(0, 0, 0, 1, 1, 0),
  VAL  = c(0, 0, 0, 0, 1, 0),
  SAT  = c(0, 0, 0, 0, 0, 1),
  LOY  = c(0, 0, 0, 0, 0, 0),
  row.names = c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")
)

# 함수 호출 예시
plspm_add_inter_paths(data, list(
  int1 = list(from = "IMAG*EXPE", to = "LOY"),
  int2 = list(from = "IMAG*QUAL", to = "LOY")
))

plspm_add_inter_paths(data, list(
  list(from = "IMAG*EXPE", to = "LOY"),
  list(from = "IMAG*QUAL", to = "LOY")
))


}


}
