% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_apa.R
\name{plspm_apa_boot_paths}
\alias{plspm_apa_boot_paths}
\title{bootstrap path effect apa analysis}
\usage{
plspm_apa_boot_paths(data)
}
\arguments{
\item{data}{plspm_boot_paths_sig}
}
\value{
text
}
\description{
bootstrap path effect apa analysis
}
\examples{

\dontrun{

#'

# 예시 데이터로 함수 호출
nfl_pls_boot2 <- data.frame(
  paths = c("Speical -> Scoring", "Rushing -> Offense", "Passing -> Offense", "Offense -> Scoring"),
  Original = c(0.311, 0.535, 0.789, 0.810),
  Mean.Boot = c(0.098, 0.522, 0.756, 0.762),
  Std.Error = c(0.317, 0.044, 0.252, 0.227),
  t = c("0.310", "11.840***", "2.995**", "3.361***"),
  `95\%CI` = I(list(c(-0.484, 0.450), c(0.411, 0.592), c(0.563, 0.948), c(0.587, 0.895)))
)

result <- plspm_apa_boot_paths(nfl_pls_boot2)
cat(result)
#'
}

}
