% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_apa.R
\name{plspm_apa_grp}
\alias{plspm_apa_grp}
\title{plspm_apa_grp}
\usage{
plspm_apa_grp(data)
}
\arguments{
\item{data}{plspm_grp_summary}
}
\value{
text
}
\description{
plspm_apa_grp
}
\examples{

\dontrun{

# Example usage
data14 <- tibble::tribble(
  ~paths, ~global, ~`group.고등학교`, ~`group.중학교`, ~diff, ~t, ~p.value, ~sig,
  "자기효능감->진로동기", 0.88, 0.894, 0.81, 0.085, 0.908, 0.183, "no",
  "자기효능감->진로태도", -0.674, -0.749, 0.438, 1.19, 2.49, 0.007, "yes",
  "자기효능감->진로준비", 0.429, 0.542, 0.595, 0.053, 0.395, 0.347, "no",
  "진로동기->진로준비", 0.446, 0.431, 0.385, 0.046, 0.232, 0.408, "no",
  "진로태도->진로준비", 0.007, 0.098, -0.126, 0.224, 0.626, 0.266, "no"
)

plspm_apa_grp(data14)

}

}
