% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_apa.R
\name{plspm_apa_ind}
\alias{plspm_apa_ind}
\title{Function to interpret plspm results}
\usage{
plspm_apa_ind(data)
}
\arguments{
\item{data}{plspm_ind_effect_test data}
}
\value{
text
}
\description{
Function to interpret plspm results
}
\examples{

\dontrun{
#'
# Example data frame
data11 <- tibble::tibble(
  paths = c("IMAG -> EXPE -> QUAL -> VAL -> SAT -> LOY",
            "IMAG -> EXPE -> QUAL -> SAT -> LOY",
            "IMAG -> EXPE -> VAL -> SAT -> LOY",
            "IMAG -> EXPE -> SAT -> LOY",
            "IMAG -> SAT -> LOY"),
  ind_effect = c(0.0970, 0.0297, 0.0178, -0.000790, 0.0995),
  SE = c(4.44e-4, 4.07e-3, 2.98e-3, 1.91e-2, 3.35e-2),
  Z = c(218.39, 7.31, 5.98, -0.0413, 2.97),
  p.value = c(0, 2.76e-13, 2.26e-9, 9.67e-1, 3.00e-3),
  sig = c("***", "***", "***", "ns", "**")
)

plspm_apa_ind(data11)

## exaample
plspm_ind_effect_test(satpls_boot)\%>\%plspm_apa_ind()

}
}
