% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_apa.R
\name{plspm_apa_paths}
\alias{plspm_apa_paths}
\title{plspm_apa_paths}
\usage{
plspm_apa_paths(data)
}
\arguments{
\item{data}{plspm_paths_sig}
}
\value{
interpretation
}
\description{
plspm_apa_paths
}
\examples{

\dontrun{
#'
# 예시 데이터 프레임
jutpls_boot1 <- data.frame(
  endogenous = c("진로동기", "진로태도", "진로준비", "진로준비", "진로준비"),
  exogenous = c("자기효능감", "자기효능감", "자기효능감", "진로동기", "진로태도"),
  Est = c(0.880, -0.674, 0.429, 0.446, 0.007),
  SE = c(0.050, 0.078, 0.129, 0.121, 0.078),
  t = c(17.597, -8.656, 3.330, 3.694, 0.087),
  p.value = c("<.001", "<.001", 0.001, "<.001", 0.931),
  sig = c("***", "***", "**", "***", "ns")
)

# 함수 실행
plspm_apa_paths(jutpls_boot)

jutpls_boot \%>\% plspm_paths_sig()\%>\% plspm_apa_paths()
#'
jutpls_boot \%>\% plspm_apa_paths()
jutpls_boot \%>\% plspm_paths_sig()\%>\% plspm_apa_paths()
jutpls_boot \%>\% plspm_paths_sig(unite=T)\%>\% plspm_apa_paths()


}

}
