% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_interaction.R
\name{plspm_auto_blocks}
\alias{plspm_auto_blocks}
\title{plspm_auto_blocks}
\usage{
plspm_auto_blocks(mat, to_lower = TRUE)
}
\arguments{
\item{mat}{path matrix}

\item{to_lower}{Change to lowercase}
}
\value{
data
}
\description{
plspm_auto_blocks
}
\examples{


\dontrun{

#'
mat <- matrix(c(0, 1, 0, 0, 0, 1, 0, 0,
                0, 0, 1, 1, 1, 0, 0, 0,
                0, 0, 0, 1, 0, 0, 0, 0,
                0, 0, 0, 1, 0, 0, 0, 0,
                0, 0, 0, 0, 0, 1, 0, 0,
                0, 0, 0, 0, 0, 0, 1, 1,
                0, 0, 0, 0, 0, 0, 0, 0,
                0, 0, 0, 0, 0, 0, 0, 0),
              nrow = 8, byrow = TRUE)

rownames(mat) <- c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY", "int_IMEX", "int_IMQU")
colnames(mat) <- c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY", "int_IMEX", "int_IMQU")
mat
# 함수 호출
plspm_auto_blocks(mat)

plspm_add_inter_paths(satpls_boot$model$IDM, list(
  list(from = "IMAG*EXPE", to = "LOY"),
  list(from = "IMAG*QUAL", to = "LOY")
))\%>\%plspm_auto_blocks()

}

}
