% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_interaction.R
\name{plspm_auto_genData}
\alias{plspm_auto_genData}
\title{plspm_auto_genData}
\usage{
plspm_auto_genData(data, interactions, to_lower = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{interactions}{interaction term list data}

\item{to_lower}{Change to lowercase}
}
\value{
data
}
\description{
plspm_auto_genData
}
\examples{

\dontrun{
#'
# 예시 데이터
data1 <- data.frame(
  IMAG = c(-0.06448267, 0.83858784, 0.47542512, 0.42690176, 1.04176442, 0.27894142),
  EXPE = c(0.3442433, 0.7974660, 0.5531210, 0.1281051, 0.4492096, 0.6914954),
  QUAL = c(-0.5254008, 0.3971273, 0.1642778, -0.6892055, 0.7255953, -0.2088824),
  VAL = c(-0.07103969, 0.44321149, 0.56088128, -0.81127296, 1.21587873, -0.29482532),
  SAT = c(-0.20583024, 0.36908558, 0.83019100, -0.05045141, 0.53174538, -0.20212349),
  LOY = c(0.1664431, 0.4924859, 0.8117161, -0.7969487, 0.7348195, 0.1714641)
)

# 함수 실행 예시
plspm_auto_genData(data1, list(
  int1 = list(from = "IMAG*EXPE", to = "LOY"),
  int2 = list(from = "IMAG*QUAL", to = "LOY")
))

# 함수 실행 예시
plspm_auto_genData(satpls_boot$scores , list(
  int1 = list(from = "IMAG*EXPE", to = "LOY"),
  int2 = list(from = "IMAG*QUAL", to = "LOY")
)) \%>\%head()

# 함수 실행 예시
plspm_auto_genData(satpls_boot$scores , list(
  list(from = "IMAG*EXPE", to = "LOY"),
  list(from = "IMAG*QUAL", to = "LOY")
))\%>\%head()
#'
}


}
