% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_blocks2lav.R
\name{plspm_blocks2lav}
\alias{plspm_blocks2lav}
\title{blocks to lavaan syntax}
\usage{
plspm_blocks2lav(listdata, name = FALSE)
}
\arguments{
\item{listdata}{block list data}

\item{name}{defulat FALSE, TRUE generate blocks_model=}
}
\value{
lavaan syntax
}
\description{
blocks to lavaan syntax
}
\examples{

\dontrun{
#'
listdata <- list(
  IMAG = c("imag1", "imag2", "imag3", "imag4", "imag5"),
  EXPE = c("expe1", "expe2", "expe3", "expe4", "expe5"),
  QUAL = c("qual1", "qual2", "qual3", "qual4", "qual5"),
  VAL = c("val1", "val2", "val3", "val4"),
  SAT = c("sat1", "sat2", "sat3", "sat4"),
  LOY = c("loy1", "loy2", "loy3", "loy4")
)
listdata
# Function calls and result output
plspm_blocks2lav(listdata)
plspm_blocks2lav(listdata)\%>\%cat("\n")
}

}
