% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_boot_effect.R
\name{plspm_boot_effect}
\alias{plspm_boot_effect}
\title{plspm_boot_effect Total effect}
\usage{
plspm_boot_effect(plsres, accept = FALSE, type = "res", digits = 3)
}
\arguments{
\item{plsres}{plsdata}

\item{accept}{Hypothesis}

\item{type}{res, all}

\item{digits}{3}
}
\value{
data result
}
\description{
plspm_boot_effect Total effect
}
\examples{

\dontrun{

library(plspm)
data(satisfaction)

# path matrix
IMAG = c(0,0,0,0,0,0)
EXPE = c(1,0,0,0,0,0)
QUAL = c(0,1,0,0,0,0)
VAL = c(0,1,1,0,0,0)
SAT = c(1,1,1,1,0,0)
LOY = c(1,0,0,0,1,0)
sat_path = rbind(IMAG, EXPE, QUAL, VAL, SAT, LOY)

# blocks of outer model
sat_blocks = list(1:5, 6:10, 11:15, 16:19, 20:23, 24:27)
sat_mod=c(rep("A", ncol(sat_path)))
# apply plspm
satpls = plspm(satisfaction, sat_path, sat_blocks, modes = sat_mod,
               scaled = FALSE)



  # bootstrap data
pls_boot_effect(satpls_boot)
pls_boot_effect(satpls_boot, accept=T)

}

}
