% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semptools.R
\name{plspm_boot_factor_layout}
\alias{plspm_boot_factor_layout}
\title{plspm_boot_factor_layout}
\usage{
plspm_boot_factor_layout(plsres_boot, positions = NULL, nrow = 3, ncol = 4)
}
\arguments{
\item{plsres_boot}{plsres_boot}

\item{positions}{positions coord}

\item{nrow}{3}

\item{ncol}{3 number of lantent 4 is ncol = 3}
}
\value{
data
}
\description{
plspm_boot_factor_layout
}
\examples{

\dontrun{
#'
plspm_boot_factor_layout(jutpls_boot)
plspm_boot_factor_layout(jutpls_boot, list(진로동기 = c(2, 1), 자기효능감 = c(1, 2)))
plspm_boot_factor_layout(satpls_boot)
plspm_boot_factor_layout(satpls_boot, list( VAL=c(2, 3), SAT= c(1, 4), LOY=c(3, 4)))
}


}
