% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semptools.R
\name{plspm_boot_factor_point_to}
\alias{plspm_boot_factor_point_to}
\title{plspm_boot_factor_point_to}
\usage{
plspm_boot_factor_point_to(
  plsres_boot,
  point = NULL,
  positions = NULL,
  nrow = 4,
  ncol = 5
)
}
\arguments{
\item{plsres_boot}{plsres_boot}

\item{point}{point point = list(up = c(2,1))}

\item{positions}{positions positions = list(up = c(2,1))}

\item{nrow}{3}

\item{ncol}{3}
}
\value{
data
}
\description{
plspm_boot_factor_point_to
}
\examples{

\dontrun{
#'
plspm_boot_factor_point_to(jutpls_boot)
plspm_boot_factor_point_to(satpls_boot)
# plspm_factor_point_mat(jutpls_boot, point=list())
# point <- list(up = c(1, 2))
plspm_boot_factor_point_to(jutpls_boot, point = list(up = c(2,1)))
plspm_boot_factor_point_to(jutpls_boot, point = list(up = c(2,1), right =c(1,2)))
#'
}



}
