% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semptools.R
\name{plspm_boot_ind_factor}
\alias{plspm_boot_ind_factor}
\title{plspm_indicator_factor}
\usage{
plspm_boot_ind_factor(plsres_boot)
}
\arguments{
\item{plsres_boot}{plsres_boot}
}
\value{
plsres_boot
}
\description{
plspm_indicator_factor
}
\examples{
\dontrun{
plspm_indicator_factor(nfl_pls_boot)
}
}
