% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semptools.R
\name{plspm_boot_ind_order}
\alias{plspm_boot_ind_order}
\title{plspm_indicator_order}
\usage{
plspm_boot_ind_order(plsres_boot)
}
\arguments{
\item{plsres_boot}{plsres_boot}
}
\value{
plsres_boot
}
\description{
plspm_indicator_order
}
\examples{
\dontrun{
plspm_indicator_order(nfl_pls_boot)
}

}
