% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semptools.R
\name{plspm_boot_ind_push}
\alias{plspm_boot_ind_push}
\title{plspm_ind_push}
\usage{
plspm_boot_ind_push(pls_boot, set = 2)
}
\arguments{
\item{pls_boot}{pls_boot}

\item{set}{set=1.5 or c(set= c(2,2,1.5,1.5))}
}
\value{
vector
}
\description{
plspm_ind_push
}
\examples{

\dontrun{

# 예시 사용법
jutpls_boot <- list(model = list(gens = list(lvs_names = c("자기효능감", "진로동기", "진로태도", "진로준비"))))

# 기본값 사용하는 경우
result_default <- plspm_ind_push(jutpls_boot)
print(result_default)

# 이름과 값이 있는 경우
result_named <- plspm_ind_push(jutpls_boot, set = c(진로동기=2, 자기효능감=3, 진로준비=1.5, 진로태도=2.5))
print(result_named)

# 값만 있는 경우
result_unnamed <- plspm_ind_push(jutpls_boot, set = c(2, 2, 1.5, 1.5))
print(result_unnamed)

# 부분적으로 값이 있는 경우
result_partial <- plspm_ind_push(jutpls_boot, set = c(진로동기=2, 자기효능감=3))
print(result_partial)

}



}
