% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_boot_paths_sig.R
\name{plspm_boot_paths_sig_vec}
\alias{plspm_boot_paths_sig_vec}
\title{boot res sig table}
\usage{
plspm_boot_paths_sig_vec(plsres, type = "vec")
}
\arguments{
\item{plsres}{plspm result}

\item{type}{df, vec, vector}
}
\value{
data, vector
}
\description{
boot res sig table
}
