% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_cfa.R
\name{plspm_cfa}
\alias{plspm_cfa}
\title{plspm CFA}
\usage{
plspm_cfa(plsres_boot, type = "all", axis_x = 1.2)
}
\arguments{
\item{plsres_boot}{bootstrap plspm data}

\item{type}{all, cfa1, cfa2, item, CRAVE, fl1981, htmt, f2. total_effect}

\item{axis_x}{1.2 setting efffecbar text size}
}
\value{
multiple data
}
\description{
plspm CFA
}
\examples{

\dontrun{

data(satisfaction)
sat_path = plspm_paths(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),
  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY"))
)
sat_path

# blokcs
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)

sat_blocks1

# apply plspm
# satpls = plspm(satisfaction,
#                path_matrix = sat_path,
#                blocks = sat_blocks1,
#                scaled = FALSE)

satpls_boot = plspm_sem(satisfaction, sat_path, sat_blocks1, scaled = FALSE,
                        boot.val =TRUE, br=100)
##
satpls_boot\%>\%plspm_cfa()
satpls_boot\%>\%plspm_cfa("res")

satpls_boot\%>\%plspm_cfa(type="cfa1")
satpls_boot\%>\%plspm_cfa(type="cfa2")
satpls_boot\%>\%plspm_cfa(type="cor")
satpls_boot\%>\%plspm_cfa(type="model")
satpls_boot\%>\%plspm_cfa(type="inner_model") \%>\%plspm_inner_model_sig()
satpls_boot\%>\%plspm_cfa(type="outer_model")
satpls_boot\%>\%plspm_cfa(type="effectbar")
total
}


}
