% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_change_nodeLabels.R
\name{plspm_change_nodeLabels}
\alias{plspm_change_nodeLabels}
\title{plspm_change_nodeLabels}
\usage{
plspm_change_nodeLabels(data, ...)
}
\arguments{
\item{data}{plspm_semPaths}

\item{...}{change node from name , to name}
}
\value{
plot
}
\description{
plspm_change_nodeLabels
}
\examples{

\dontrun{


plspm_semPaths2(jutpls_boot,   curve=1, layout="tree2",rotation=1,
                edge.label.cex=0.7, mar=c(2,1,2,1),
                edge.label.position=0.6, label.cex=1,
                sizeLat= 10)\%>\%
  plspm_change_nodeLabels("B02","결정성",
                          "B04", "독립성",
                          "B10", "계획성",
                          "B14", "성향"
  )







}


}
