% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_crossloadings_plot.R
\name{plspm_crossloadings}
\alias{plspm_crossloadings}
\title{plspm_crossloadings}
\usage{
plspm_crossloadings(plsres, digits = 3)
}
\arguments{
\item{plsres}{boot result}

\item{digits}{reound 3}
}
\value{
data
}
\description{
plspm_crossloadings
}
