% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_crossloadings_plot.R
\name{plspm_crossloadings_plot}
\alias{plspm_crossloadings_plot}
\title{plspm_crossloadings_plot}
\usage{
plspm_crossloadings_plot(plsres, wide = F)
}
\arguments{
\item{plsres}{plspm result}

\item{wide}{plot type TRUE block, FALSE wide plot}
}
\value{
pplot
}
\description{
plspm_crossloadings_plot
}
\examples{


\dontrun{

data(satisfaction)
#model 1
model1 = "
EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL+ EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"
# # Generate the matrix
sat_path = plspm_lav2path(model1,
                          fct_order=c("IMAG", "EXPE","QUAL", "VAL", "SAT", "LOY" ))
# sat_path

# Method 2: paths
sat_path = plspm_paths(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),
  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY")
  )
)
# blokcs
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)
sat_blocks1

# vector of modes (reflective indicators):auto
sat_mod = rep("A", 6)
#############

# apply plspm
satpls = plspm(satisfaction,
                   path_matrix = sat_path,
                   blocks = sat_blocks1,
                   scaled = FALSE)

satpls_boot = plspm_sem(satisfaction, sat_path, sat_blocks1, scaled = FALSE,
                        boot.val =TRUE, br=100)


#'
satpls \%>\% plspm_crossloadings_plot()
satpls \%>\% plspm_crossloadings_plot(T)
satpls$crossloadings \%>\% plspm_crossloadings_plot()
satpls$crossloadings \%>\% plspm_crossloadings_plot(T)
#'
#'
}

}
