% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_draw_blocks.R
\name{plspm_drawblocks}
\alias{plspm_drawblocks}
\title{plspm block draw graph}
\usage{
plspm_drawblocks(
  data,
  latent_size = 20,
  edge_width = 1,
  edge.alpha = 0.7,
  edge_colour = "darkgreen",
  text.colour = "gray10",
  nudge_y = 0.1,
  nudge_x = -0.1,
  title = " [Figure 1] Latent and items "
)
}
\arguments{
\item{data}{blocked data}

\item{latent_size}{20}

\item{edge_width}{1}

\item{edge.alpha}{0.7}

\item{edge_colour}{darkgreen}

\item{text.colour}{gray10}

\item{nudge_y}{0.1}

\item{nudge_x}{-0.1}

\item{title}{[Figure 1] Latent and items}

\item{layout}{nicely , kk, fr, drl, mds}
}
\value{
plot
}
\description{
plspm block draw graph
}
\examples{

\dontrun{
#'
edu_blocks_named <- blocks(
  Support = item("sup.help", "sup.under", "sup.safe", "sup.conc"),
  Advising = item("adv.comp", "adv.acces", "adv.comm", "adv.qual"),
  Tutoring= item( "tut.prof","tut.sched", "tut.stud", "tut.qual" ),
  Value = item("val.devel", "val.deci","val.meet", "val.info"),
  Satisfaction = item("sat.glad","sat.expe","sat.over"),
  Loyalty =item("loy.proud", "loy.recom","loy.asha","loy.back")
)

edu_blocks_named
draw_blocks(edu_blocks_named)

sat_blocks1 <- blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)
sat_blocks1

draw_blocks(sat_blocks)


}

}
