% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_edge_position.R
\name{plspm_edge_position}
\alias{plspm_edge_position}
\title{plspm_edge_position}
\usage{
plspm_edge_position(semPaths_plot, ..., plot = TRUE)
}
\arguments{
\item{semPaths_plot}{plot data generated from a PLS-PM model.}

\item{...}{triples of arguments specifying `from` node, `to` node, and `new_position`.}

\item{plot}{logical, if `TRUE`, the plot will be displayed with updated edge positions. Default is `TRUE`.}
}
\value{
Modified plot object with updated edge positions.
}
\description{
plspm_edge_position
}
\examples{

\dontrun{
# Example 1: Adjusting edge positions
pathmodel <- "
자기효능감 =~ C_S1 +  C_S2 + C_S3 + C_S4 + C_S5
진로태도 =~ B02 + B04 + B10 + B14
진로동기 =~ A_M1 + A_M2 + A_M3
진로준비 =~ D_P1 + D_P2 + D_P3

진로동기 ~ H1*자기효능감
진로태도 ~  H2*자기효능감
진로준비 ~ H5 *진로태도
진로준비 ~  H4 * 진로동기
진로준비 ~  H3*자기효능감
"

semPaths_plot <- plspm_diagram_model(pathmodel, layout="tree2",
                    rotation = 2, residuals = TRUE, residScale = 6,
                    edge.label.position = 0.6)

semPaths_plot \%>\%
  plspm_edge_position("진로태도", "진로준비", 0.4,
                      "자기효능감", "진로동기", 0.3)

# Example 2: Adjusting multiple edge positions
semPaths_plot \%>\%
  plspm_edge_position("진로동기", "진로태도", 0.5,
                      "진로태도", "진로준비", 0.7,
                      "자기효능감", "진로준비", 0.6)
}

}
