% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_edges_values.R
\name{plspm_edge_values}
\alias{plspm_edge_values}
\title{Functions to include significance data for structural models in PLSPM}
\usage{
plspm_edge_values(plsdata, type = "vec", digits = 3, unite = TRUE)
}
\arguments{
\item{plsdata}{plsdata}

\item{type}{vec, df}

\item{digits}{round 3}

\item{unite}{data combine unite TRUE -> estimat + sig}
}
\value{
estimates vectors
}
\description{
Functions to include significance data for structural models in PLSPM
}
