% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_effect.R
\name{plspm_effect}
\alias{plspm_effect}
\title{plspm_effect}
\usage{
plspm_effect(pls_boot, digits = 3)
}
\arguments{
\item{pls_boot}{pls boot}

\item{digits}{round 3}
}
\value{
data
}
\description{
plspm_effect
}
