% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_f2.R
\name{plspm_f2}
\alias{plspm_f2}
\title{Effect Size Black}
\usage{
plspm_f2(plsres, type = "res")
}
\arguments{
\item{plsres}{pls result}

\item{type}{res, f2, R2, ceofs, inner}
}
\value{
data
}
\description{
Effect Size Black
}
\examples{

\dontrun{
#'
#' #'
sat_path = plspm_paths(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),
  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY"))
)
sat_path

# blokcs
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)

sat_blocks1

satpls = plspm(satisfaction,
               path_matrix = sat_path,
               blocks = sat_blocks1,
               scaled = FALSE)

satpls_boot = plspm_sem(satisfaction, sat_path, sat_blocks1, scaled = FALSE,
                        boot.val =TRUE, br=100)
# etc
satpls_boot\%>\%plspm_f2()
satpls_boot\%>\%plspm_f2("f2")
satpls_boot\%>\%plspm_f2("coefs")
satpls_boot\%>\%plspm_f2("r2")
satpls_boot\%>\%plspm_f2("inner")
}



}
