% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_f2_plot.R
\name{plspm_f2_plot}
\alias{plspm_f2_plot}
\title{Function to plot f2 values}
\usage{
plspm_f2_plot(plsres)
}
\arguments{
\item{plsres}{pls result}
}
\value{
plot
}
\description{
Function to plot f2 values
}
\examples{
\dontrun{

#'
# Example data for plsres
plsres <- list(
  path_coefs = data.frame(
    자기효능감 = c(0, 0.884, 0.681, 0.373),
    진로동기 = c(0, 0, 0, 0.353),
    진로태도 = c(0, 0, 0, 0.188),
    진로준비 = c(0, 0, 0, 0),
    row.names = c("자기효능감", "진로동기", "진로태도", "진로준비")
  ),
  inner_summary = data.frame(
    R2 = c(0.0000000, 0.7817016, 0.4631848, 0.7195506),
    row.names = c("자기효능감", "진로동기", "진로태도", "진로준비")
  )
)

# Run the function
plspm_f2(plsres, type="res")



plspm_f2_plot(plsres)
}


}
