% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_grp_comp.R
\name{plspm_grp_comp}
\alias{plspm_grp_comp}
\title{plspm_grp_comp:Creating multigroup list data into a data frame}
\usage{
plspm_grp_comp(lst_data)
}
\arguments{
\item{lst_data}{grp data}
}
\value{
data
}
\description{
plspm_grp_comp:Creating multigroup list data into a data frame
}
\examples{

\dontrun{
#'
plspm_grp_comp(teacher_compare) \%>\%
  nice_table(exclude=c("df", "sig"))\%>\%
  replace_df_rep(var_named) \%>\%Round(2)\%>\%pall()
}

}
