% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_grp_compare_bar.R
\name{plspm_grp_compare_bar}
\alias{plspm_grp_compare_bar}
\title{plspm_grp_compare_bar}
\usage{
plspm_grp_compare_bar(
  data,
  x = NULL,
  y = NULL,
  fill = NULL,
  size = 3,
  size_x = 12,
  angle = 90,
  position = "dodge",
  xlab = "",
  ylab = "Values",
  text = TRUE
)
}
\arguments{
\item{data}{arragne data}

\item{x}{paths}

\item{y}{values}

\item{fill}{group}

\item{size}{bargrahp text size}

\item{size_x}{textsize}

\item{angle}{xangle}

\item{position}{dodge, stack}

\item{xlab}{default ""}

\item{ylab}{values}

\item{text}{hide bartext}
}
\value{
graph
}
\description{
plspm_grp_compare_bar
}
\examples{

\dontrun{
#'
# 예시 데이터
REs <- tibble::tribble(
  ~paths, ~grp, ~values,
  "자기효능감->진로동기", "group.남자", 0.902,
  "자기효능감->진로동기", "group.여자", 0.866,
  "자기효능감->진로태도", "group.남자", -0.639,
  "자기효능감->진로태도", "group.여자", -0.672,
  "자기효능감->진로준비", "group.남자", 0.149,
  "자기효능감->진로준비", "group.여자", 0.487,
  "진로동기->진로준비", "group.남자", 0.63,
  "진로동기->진로준비", "group.여자", 0.432,
  "진로태도->진로준비", "group.남자", -0.153,
  "진로태도->진로준비", "group.여자", 0.059
)

# # 함수 호출 예시
REs \%>\% compare_bar()
REs \%>\% compare_bar(text=FALSE)
}
}
