% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_htmt.R
\name{plspm_htmt}
\alias{plspm_htmt}
\title{Define a function to calculate HTMT}
\usage{
plspm_htmt(data, blocks, htmt2 = TRUE, sig = TRUE, cut = 0.9, digits = 3)
}
\arguments{
\item{data}{data}

\item{blocks}{latent =~ item, plspl blocks}

\item{htmt2}{TRUE Apply geometric mean}

\item{sig}{sig= TRUE- > show star}

\item{cut}{= 0.9, 0.85. 1}

\item{digits}{= 3}
}
\value{
result
}
\description{
Define a function to calculate HTMT
}
\examples{
\dontrun{
#'
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)

sat_blocks1
# plspm_blocks2lav(sat_blocks1) \%>\% cat("\n")
data(satisfaction)
plspm_htmt(satisfaction, sat_blocks1)

#'
#'
plspm_htmt(satisfaction, sat_blocks1)
plspm_htmt(satisfaction, sat_blocks1, sig=F)

plspm_htmt(satisfaction, sat_blocks1,sig=F, htmt2=FALSE)
plspm_htmt(satisfaction, sat_blocks1,sig=T, cut=0.9, htmt2=FALSE)

## Reference
semTools::htmt(plspm_blocks2lav(sat_blocks1),
               satisfaction,
               htmt2 = T)
#'

}



}
