% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_ind_effect_sig.R
\name{plspm_ind_effect}
\alias{plspm_ind_effect}
\title{plspm_ind_effect}
\usage{
plspm_ind_effect(boot_data, from, through, to, type = "res", digits = 3)
}
\arguments{
\item{boot_data}{boot_data}

\item{from}{from}

\item{through}{through}

\item{to}{to}

\item{type}{res, all}

\item{digits}{3}
}
\value{
data
}
\description{
plspm_ind_effect
}
\examples{

\dontrun{
#'
# Example usage of the function with the provided PLS-PM model
plspm_ind_effec(
  boot_data = jutpls_boot,
  from = "자기효능감",
  through = c("진로동기", "진로태도"),
  to = "진로준비"
)

plspm_ind_effec(
  boot_data = jutpls_boot,
  from = "자기효능감",
  through = c("진로동기", "진로태도"),
  to = "진로준비", type="all"
)

plspm_ind_effec(
  boot_data = jutpls_boot,
  from = "자기효능감",
  through = c("진로동기"),
  to = "진로준비", type="res"
)

}


}
