% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_ind_effect_boot.R
\name{plspm_ind_effect_boot}
\alias{plspm_ind_effect_boot}
\title{plspm ind efffect bootstrap}
\usage{
plspm_ind_effect_boot(
  plsres = NULL,
  ...,
  data = NULL,
  path_matrix = NULL,
  blocks = NULL,
  modes = NULL,
  B = 100
)
}
\arguments{
\item{plsres}{plspm result boot}

\item{...}{paths}

\item{data}{data}

\item{path_matrix}{paths}

\item{blocks}{blocks}

\item{modes}{models}

\item{B}{b= 100}
}
\value{
resutl paths
}
\description{
plspm ind efffect bootstrap
}
\examples{

\dontrun{

#'
data(satisfaction)
sat_path = plspm_paths(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),
  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY"))
)
sat_path

# blokcs
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)

sat_blocks1

# vector of modes (reflective indicators):auto
sat_mod = rep("A", 6)

# satpls = plspm(satisfaction, sat_path, sat_blocks1, scaled = FALSE)
# satpls
# satpls \%>\%summary()

satpls_boot = plspm_sem(satisfaction, sat_path, sat_blocks1, scaled = FALSE,
                        boot.val =TRUE, br=100)


long_paths <- c(
  "IMAG -> EXPE -> QUAL -> VAL -> SAT -> LOY",
  "IMAG -> EXPE -> QUAL -> SAT -> LOY",
  "IMAG -> EXPE -> VAL -> SAT -> LOY",
  "IMAG -> EXPE -> SAT -> LOY",
  "IMAG -> SAT -> LOY"
)

result <- plspm_get_boot(satpls_boot, long_paths, B = 100)
result
result \%>\% Round(3)
result \%>\% unite_ci()

resultall <- plspm_get_boot(satpls_boot, find_paths(satpls_boot), B = 100)
resultall


}


}
