% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_ind_effect_sig.R
\name{plspm_ind_effect_sig}
\alias{plspm_ind_effect_sig}
\title{plspm_ind_effect_sig}
\usage{
plspm_ind_effect_sig(boot_data, from, through, to, digits = 3)
}
\arguments{
\item{boot_data}{bootstrap data}

\item{from}{from}

\item{through}{era}

\item{to}{to}

\item{digits}{3}
}
\value{
data
}
\description{
plspm_ind_effect_sig
}
\examples{

\dontrun{
#'
# 함수 호출 예시
plspm_ind_effect_sig(
  boot_data = satpls_boot,
  from = "IMAG",
  through = c("EXPE", "QUAL", "SAT"),
  to = "VAL"
)

}



}
