% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semptools.R
\name{plspm_ind_order}
\alias{plspm_ind_order}
\title{plspm_indicator_order}
\usage{
plspm_ind_order(model_syntax)
}
\arguments{
\item{model_syntax}{lavaan}
}
\value{
text
}
\description{
plspm_indicator_order
}
\examples{
\dontrun{
#'
mod <-
  'f1 =~ x01 + x02 + x03 + x06
    f3 =~ x08 + x09 + x10 + x03
    f2 =~ x04 + x05 + x06 + x07
    f4 =~ x11 + x12 + x13 + x14
    f3 ~ f1 + f2
    f4 ~ f3
   '
plspm_indicator_order(mod)

plspm_indicator_factor(mod)
}

}
