% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_indicator_test.R
\name{plspm_indicator_test}
\alias{plspm_indicator_test}
\title{지표의 균일성 평가를 위한 Permutation Test 함수 (병렬 처리 적용)}
\usage{
plspm_indicator_test(
  Data,
  path_matrix,
  blocks,
  grp,
  n_perm = 100,
  n_boot = 100,
  n_cores = 12
)
}
\arguments{
\item{Data}{data}

\item{path_matrix}{path}

\item{blocks}{measurement}

\item{grp}{grp}

\item{n_perm}{permutation number}

\item{n_boot}{boot n}

\item{n_cores}{cores=8 (recently 12~15), exceapt one !!!}
}
\value{
data
}
\description{
지표의 균일성 평가를 위한 Permutation Test 함수 (병렬 처리 적용)
}
\examples{

\dontrun{
#'
# 예시: 지표의 균일성 평가 실행 (병렬 처리 적용)
result_df2 <- plspm_indicator_test(Data = jut7c, path_matrix = jutpath1, blocks = jut_blocks, grp = "P01")
print(result_df2)
}

}
