% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_paths_sig.R
\name{plspm_inner_model_sig}
\alias{plspm_inner_model_sig}
\title{Showing significance using PLSPM regression}
\usage{
plspm_inner_model_sig(
  data,
  unite = FALSE,
  rhs = "endo",
  lhs = "exo",
  digits = 3
)
}
\arguments{
\item{data}{plspm data}

\item{unite}{TRUE}

\item{rhs}{from}

\item{lhs}{to}

\item{digits}{round 3}
}
\value{
data result
}
\description{
Showing significance using PLSPM regression
}
\examples{
\dontrun{
library(plspm)
data(satisfaction)

# path matrix
IMAG = c(0,0,0,0,0,0)
EXPE = c(1,0,0,0,0,0)
QUAL = c(0,1,0,0,0,0)
VAL = c(0,1,1,0,0,0)
SAT = c(1,1,1,1,0,0)
LOY = c(1,0,0,0,1,0)
sat_path = rbind(IMAG, EXPE, QUAL, VAL, SAT, LOY)

# blocks of outer model
sat_blocks = list(1:5, 6:10, 11:15, 16:19, 20:23, 24:27)
sat_mod=c(rep("A", ncol(sat_path)))
# apply plspm
satpls = plspm(satisfaction, sat_path, sat_blocks, modes = sat_mod,
               scaled = FALSE)

#default
satpls\%>\%plspm_paths_sig()
satpls$inner_model \%>\%plspm_paths_sig()

satpls \%>\%plspm_paths_sig(unite=FALSE, lhs="row", rhs="col")
satpls$inner_model  \%>\%innermodel_sig(unite=FALSE, lhs="row", rhs="col")

satpls \%>\%plspm_paths_sig(unite=FALSE, lhs="row", rhs="col")
satpls$inner_model \%>\%plspm_paths_sig(unite=FALSE, lhs="row", rhs="col")

#'
}

}
