% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_innermodel.R
\name{plspm_innermodel}
\alias{plspm_innermodel}
\title{Draw a Model of a Measured PLSPM Structure}
\usage{
plspm_innermodel(
  data,
  digits = 3,
  layout = "spring",
  fade = FALSE,
  border.width = 1.5,
  border.color = "gray30",
  edge.label.cex = 1,
  edge.label.position = 0.5,
  asize = 3,
  posCol = "gray20",
  negCol = "red",
  shape = "circle",
  groups = NULL,
  pastel = "pastel",
  trans = TRUE
)
}
\arguments{
\item{data}{plspm data}

\item{digits}{3}

\item{layout}{default spring}

\item{fade}{FALSE}

\item{border.width}{1.5}

\item{border.color}{gray30}

\item{edge.label.cex}{1}

\item{edge.label.position}{0.5}

\item{asize}{3}

\item{posCol}{gray20}

\item{negCol}{red}

\item{shape}{cicle}

\item{groups}{NULL}

\item{pastel}{pastel}

\item{trans}{true}
}
\value{
plot
}
\description{
Draw a Model of a Measured PLSPM Structure
}
\examples{
\dontrun{

# path matrix
IMAG = c(0,0,0,0,0,0)
EXPE = c(1,0,0,0,0,0)
QUAL = c(0,1,0,0,0,0)
VAL = c(0,1,1,0,0,0)
SAT = c(1,1,1,1,0,0)
LOY = c(1,0,0,0,1,0)
sat_path = rbind(IMAG, EXPE, QUAL, VAL, SAT, LOY)

# blocks of outer model
sat_blocks = list(1:5, 6:10, 11:15, 16:19, 20:23, 24:27)
# apply plspm
satpls = plspm(satisfaction, sat_path, sat_blocks, modes = sat_mod,
               scaled = FALSE)

#default
satpls \%>\% innermodel_plspm()



lay_p = matrix(c('IMA', NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
 'EXP', NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, 'QUA', NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, NA, NA, 'VAL', NA, NA, NA, NA,
 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'SAT',
 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, 'LOY', NA, NA, NA, NA, NA, NA, NA, NA),
 nrow = 10, ncol = 10, byrow = FALSE)

satpls \%>\% innermodel_plspm(layout=lay_p, edge.label.position = 0.6,
                            border.color="gray99", groups=1:6)

#'
}
}
