% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_lav2blocks.R
\name{plspm_lav2blocks}
\alias{plspm_lav2blocks}
\title{lavaan syntax to plspm blocks}
\usage{
plspm_lav2blocks(blocks_model)
}
\arguments{
\item{blocks_model}{lavaan syntax}
}
\value{
list
}
\description{
lavaan syntax to plspm blocks
}
\examples{
\dontrun{

# Input data
blocks_model <- "
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
VAL =~ val1 + val2 + val3 + val4
SAT =~ sat1 + sat2 + sat3 + sat4
LOY =~ loy1 + loy2 + loy3 + loy4
"

# Example function calls
plspm_lav2blocks(blocks_model)

}
}
