% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_lav2paths.R
\name{plspm_lav2paths}
\alias{plspm_lav2paths}
\title{lavaan syntax to plspm path matrix}
\usage{
plspm_lav2paths(model_syntax, fct_order = NULL)
}
\arguments{
\item{model_syntax}{lavaan syntax}

\item{fct_order}{fct_order is reorder option for rownamnes of matrix}
}
\value{
matrix
}
\description{
lavaan syntax to plspm path matrix
}
\examples{
\dontrun{
#'
# Example usage
model1 <- "
EXPE ~ IMAG; QUAL ~ EXPE
VAL ~ QUAL + EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"

 #or
model1 = "
EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL+ EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"
# Generate the matrix
plspm_lav2paths(model1)

plspm_lav2paths(model1, fct_order=c("IMAG", "EXPE","QUAL", "VAL", "SAT", "LOY" ))


}

}
