% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_loadings.R
\name{plspm_loadings}
\alias{plspm_loadings}
\title{plspm loadings output}
\usage{
plspm_loadings(plsres, digits = 3)
}
\arguments{
\item{plsres}{plspm result}

\item{digits}{round =3}
}
\value{
data table
}
\description{
plspm loadings output
}
