% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_make_layout_mat.R
\name{plspm_make_layout_mat}
\alias{plspm_make_layout_mat}
\title{plspm_make_layout_mat}
\usage{
plspm_make_layout_mat(...)
}
\arguments{
\item{...}{variable}
}
\value{
mat
}
\description{
plspm_make_layout_mat
}
\examples{

\dontrun{
#'
plspm_make_layout_mat(x1 = c(1, 1),
                      x2 = c(2, 1),
                      x3 = c(2, 2),
                      x4 = c(3, 2),
                      x5 = c(3, 5))
}


}
