% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_interaction.R
\name{plspm_make_prod}
\alias{plspm_make_prod}
\title{plspm_make_prod , Generate data for indicator interaction multiplied terms}
\usage{
plspm_make_prod(data, blocks, indicator, intname = "int", n_name = 2)
}
\arguments{
\item{data}{ori data}

\item{blocks}{ori blocks}

\item{indicator}{indicator product}

\item{intname}{"int" default}

\item{n_name}{name 2, more 3  int_AAABBB}
}
\value{
data
}
\description{
plspm_make_prod , Generate data for indicator interaction multiplied terms
}
\examples{

\dontrun{
#' data(satisfaction)
#상황이 바뀌는 경우, 데이터를 바꾸어서 생성하는 경우,
plspm_make_prod(satisfaction,
                sat_blocks1,
                indicator= list(intIE = interlist("IMAG",1,2, "EXPE",1,2),
                                intIV = interlist("IMAG",1,2, "QUAL",1)  )
) \%>\%str()
}

}
