% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_mat_lav.R
\name{plspm_mat_lav}
\alias{plspm_mat_lav}
\title{plspm_mat_lavaan syntax}
\usage{
plspm_mat_lav(data, free = FALSE)
}
\arguments{
\item{data}{matrix}

\item{free}{TRUE is free parameter}
}
\value{
syntax lavaan
}
\description{
plspm_mat_lavaan syntax
}
\examples{

\dontrun{
#'

# Example data for testing
data_b <- matrix(c(
  0.000, 0.000, 0.000, 0.000, 0.000, 0,
  0.471, 0.000, 0.000, 0.000, 0.000, 0,
  0.000, 0.834, 0.000, 0.000, 0.000, 0,
  0.000, 0.046, 0.701, 0.000, 0.000, 0,
  0.245, -0.017, 0.222, 0.527, 0.000, 0,
  0.182, 0.000, 0.000, 0.000, 0.628, 0
), nrow = 6, byrow = TRUE,
dimnames=list(
  c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY"),
  c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")
))
plspm_mat_lav(data_b)\%>\%cat()
plspm_mat_lav(data_b)\%>\%
  diagram_model(whatLabels = "est", layout="tree2", nDigits=3,
                edge.label.position=0.65, rotation=1)


data_r <- matrix(c(
  0.000, 0.000, 0.000, 0.000, 0.000, 0,
  0.471, 0.000, 0.000, 0.000, 0.000, 0,
  0.000, 0.834, 0.000, 0.000, 0.000, 0,
  0.000, 0.046, 0.701, 0.000, 0.000, 0,
  0.245, -0.017, 0.222, 0.527, 0.000, 0,
  0.182, 0.000, 0.000, 0.000, 0.628, 0
), nrow = 6, byrow = TRUE,
dimnames=list(c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")))
data_r

plspm_mat_lav(data_r)\%>\%cat()

plspm_mat_lav(data_r)\%>\%
  diagram_model(whatLabels = "est", layout="tree2", nDigits=3,
                edge.label.position=0.65, rotation=1)

data_n <- matrix(c(
  0.000, 0.000, 0.000, 0.000, 0.000, 0,
  0.471, 0.000, 0.000, 0.000, 0.000, 0,
  0.000, 0.834, 0.000, 0.000, 0.000, 0,
  0.000, 0.046, 0.701, 0.000, 0.000, 0,
  0.245, -0.017, 0.222, 0.527, 0.000, 0,
  0.182, 0.000, 0.000, 0.000, 0.628, 0
), nrow = 6, byrow = TRUE)
data_n
plspm_mat_lav(data_n)\%>\%cat()
plspm_mat_lav(data_n)\%>\%
  diagram_model(whatLabels = "est", layout="tree2", nDigits=3,
                edge.label.position=0.65, rotation=1)



data_null <- matrix(c(
  0.000, 0.000, 0.000, 0.000, 0.000, 0,
  "H1", 0.000, 0.000, 0.000, 0.000, 0,
  0.000, "H2", 0.000, 0.000, 0.000, 0,
  0.000, "H3", "H4", 0.000, 0.000, 0,
  "H5", "H6", "H7", "H8", 0.000, 0,
  "H9", 0.000, 0.000, 0.000, "H10", 0
), nrow = 6, byrow = TRUE,
dimnames=list(
  c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY"),
  c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")
))

plspm_mat_lav(data_null)\%>\%cat()

plspm_mat_lav(data_null)\%>\%
  diagram_model(whatLabels = "model", layout="tree2", nDigits=3,
                edge.label.position=0.65, rotation=1)

#'
lay_p = matrix(c('IMA', NA, NA, NA, NA, NA, NA, NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'EXP', NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'QUA',
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, 'VAL', NA, NA, NA, NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'SAT', NA, NA, NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'LOY',
                 NA, NA, NA, NA, NA, NA, NA, NA),
               nrow = 10, ncol = 10, byrow = FALSE)
lay_p

# Using digaram_model

satpls$path_coefs \%>\% plspm_mat_lav() \%>\%
  diagram_model(layout="tree2", whatLabels = "est",
                rotation =1, edge.label.position=0.6)

satpls$path_coefs \%>\% plspm_mat_lav() \%>\%
  diagram_model(layout="tree2", whatLabels = "est",
                rotation =1, edge.label.position=0.6,
                edgeLabels=satpls_fit$inner_model \%>\%plspm_edge_values())
#'
#'


}

}
