% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_path_coefs_plot.R
\name{plspm_path_coefs_plot}
\alias{plspm_path_coefs_plot}
\title{Draw a Model of a Measured PLSPM Structure}
\usage{
plspm_path_coefs_plot(
  plsdata,
  grp = TRUE,
  groups = NULL,
  digits = 3,
  layout = "spring",
  fade = FALSE,
  node = FALSE,
  border.width = 1.5,
  border.color = "gray90",
  edge.label.cex = 1,
  edge.label.position = 0.6,
  asize = 4,
  vsize = 8,
  esize = 10,
  edge_labels_sig = TRUE,
  posCol = "gray20",
  negCol = "red",
  shape = "circle",
  pastel = "pastel",
  trans = TRUE
)
}
\arguments{
\item{grp}{groups on T}

\item{groups}{NULL}

\item{digits}{3}

\item{layout}{default spring}

\item{fade}{FALSE}

\item{node}{nodeNames legeng when node =TRUE}

\item{border.width}{1.5}

\item{border.color}{gray30}

\item{edge.label.cex}{1}

\item{edge.label.position}{0.5}

\item{asize}{3 arrow size}

\item{vsize}{8 nodesixe}

\item{esize}{10 edgesize}

\item{edge_labels_sig}{Tedge_labels_sig= TRUE add sig star}

\item{posCol}{gray20}

\item{negCol}{red}

\item{shape}{cicle}

\item{pastel}{pastel}

\item{trans}{true}

\item{data}{plspm data}
}
\value{
plot
}
\description{
Draw a Model of a Measured PLSPM Structure
}
\examples{
\dontrun{

# path matrix
IMAG = c(0,0,0,0,0,0)
EXPE = c(1,0,0,0,0,0)
QUAL = c(0,1,0,0,0,0)
VAL = c(0,1,1,0,0,0)
SAT = c(1,1,1,1,0,0)
LOY = c(1,0,0,0,1,0)
sat_path = rbind(IMAG, EXPE, QUAL, VAL, SAT, LOY)

# blocks of outer model
sat_blocks = list(1:5, 6:10, 11:15, 16:19, 20:23, 24:27)

sat_mod = c(rep("A", 6))
# apply plspm
satpls = plspm(satisfaction, sat_path, sat_blocks, modes = sat_mod,
               scaled = FALSE)


 ##jjstat
data(satisfaction)
#model 1
model1 = "
EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL+ EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"
# # Generate the matrix
sat_path = plspm_lav2path(model1,
                          fct_order=c("IMAG", "EXPE","QUAL", "VAL", "SAT", "LOY" ))
# sat_path

# Method 2: paths
sat_path = plspm_paths(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),
  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY")
  )
)
# blokcs
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)
sat_blocks1

# vector of modes (reflective indicators):auto
sat_mod = rep("A", 6)
#############

# apply plspm
satpls = plspm_sem(satisfaction,
                   path_matrix = sat_path,
                   blocks = sat_blocks1,
                   scaled = FALSE)

satpls_boot = plspm_sem(satisfaction, sat_path, sat_blocks1, scaled = FALSE,
                        boot.val =TRUE, br=100)

#default
satpls \%>\% plspm_path_coefs_plot()



lay_p = matrix(c('IMA', NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
 'EXP', NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, 'QUA', NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, NA, NA, 'VAL', NA, NA, NA, NA,
 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'SAT',
 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
 NA, NA, NA, NA, 'LOY', NA, NA, NA, NA, NA, NA, NA, NA),
 nrow = 10, ncol = 10, byrow = FALSE)

satpls \%>\% plspm_path_coefs_plot(layout=lay_p, edge.label.position = 0.6,
                            border.color="gray99", groups=1:6)


# various type
satpls \%>\% plspm_path_coefs_plot()
satpls \%>\% plspm_path_coefs_plot(node =TRUE)

satpls \%>\% plspm_path_coefs_plot(layout=lay_p, edge.label.position = 0.6,
                                 border.color="gray99", groups=1:6)
satpls \%>\% plspm_path_coefs_plot(layout=lay_p, edge.label.position = 0.6,
                                 border.color="gray99", groups=1:6, node=TRUE)

satpls \%>\% plspm_path_coefs_plot(shape="square", layout=lay_p,
                                 edge.label.position = 0.6)



satpls \%>\% plspm_path_coefs_plot()
satpls \%>\% plspm_path_coefs_plot(node =TRUE)
satpls \%>\% plspm_path_coefs_plot(grp=FALSE)
satpls \%>\% plspm_path_coefs_plot(node = FALSE,grp=TRUE)
satpls \%>\% plspm_path_coefs_plot(node = TRUE,grp=FALSE)
satpls \%>\% plspm_path_coefs_plot(node = FALSE,grp=FALSE)
satpls \%>\% plspm_path_coefs_plot(node = TRUE,grp=TRUE)
#'
}
}
