% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_paths.R
\name{plspm_paths}
\alias{plspm_paths}
\title{plspm path create}
\usage{
plspm_paths(row_names = NULL, col_names = NULL, relationship = NULL)
}
\arguments{
\item{relationship}{input path}

\item{rownames}{rowname}

\item{colnames}{colname}
}
\value{
matrix
}
\description{
plspm path create
}
\examples{

\dontrun{
edu_path <- plspm_path(
  rownames = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  colnames = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  relationship = list(
    list(from = "Value",
         to = c("Support", "Advising", "Tutoring")),
    list(from = "Satisfaction",
         to = c("Support", "Advising", "Tutoring","Value")),
    list(from = "Loyalty",
         to =c("Satisfaction"))
  )
)
edu_path


edu_path <- plspm_path(
  rownames = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  colnames = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  relationship = list(

    path(form = "Support", to ="Advising"),
    path(form = "Tutoring", to ="Advising"),
    path(from = "Value",
         to = c("Support", "Advising", "Tutoring")),
    path(from = "Satisfaction",
         to = c("Support", "Advising", "Tutoring","Value")),
    path (from = "Loyalty",
         to =c("Satisfaction"))
  )
)
edu_path
#'

plspm_paths(
  relationship = list(
    list(from="IMAG", to=c("EXPE","SAT","LOY")),
    list(from="EXPE", to=c("QUAL","VAL","SAT")),
    list(from="QUAL", to=c("VAL","SAT")),
    list(from="VAL", to="SAT"),
    list(from="SAT", to="LOY")
  )
)

# 결과 출력

sat_path1 = plspm_paths(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),
  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY"))
)
sat_path1

#'

sat_path <- plspm_paths(relationship = list(
  path(from="IMAG", to=c("EXPE","SAT","LOY")),
  path(from="EXPE", to=c("QUAL","VAL","SAT")),
  path(from="QUAL", to=c("VAL","SAT")),
  path(from="VAL", to="SAT"),
  path(from="SAT", to="LOY")
) )

sat_path <- plspm_paths(relationship = list(
  path(from="IMAG", to=c("EXPE","SAT","LOY")),
  path(from="EXPE", to=c("QUAL","VAL","SAT")),
  path(from="QUAL", to=c("VAL","SAT")),
  path(from="VAL", to="SAT"),
  path(from="SAT", to="LOY")
),
row_names =  c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")
)

# 결과 출력
print(sat_path)


}
}
