% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_lav2paths.R
\name{plspm_paths2lav}
\alias{plspm_paths2lav}
\title{plspm_paths2lav}
\usage{
plspm_paths2lav(path_matrix)
}
\arguments{
\item{path_matrix}{matrix paths}
}
\value{
lavaan syntax
}
\description{
plspm_paths2lav
}
\examples{
\dontrun{
#'
# Example path_matrix
path_matrix <- matrix(c(0, 1, 0, 0, 1, 1,
                        0, 0, 1, 1, 1, 0,
                        0, 0, 0, 1, 1, 0,
                        0, 0, 0, 0, 1, 0,
                        0, 0, 0, 0, 0, 1,
                        0, 0, 0, 0, 0, 0),
                      nrow = 6, byrow = TRUE)
rownames(path_matrix) <- c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")
colnames(path_matrix) <- c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")

# Example of running a function
cat(plspm_paths2lav(path_matrix))
#'
}

}
