% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_prod.R
\name{plspm_prod}
\alias{plspm_prod}
\title{Create a multiply variable to create a moderator variable}
\usage{
plspm_prod(data, blocks, ..., intname = "int", n_name = 2, id = FALSE)
}
\arguments{
\item{data}{data}

\item{blocks}{latent and measurement blocks}

\item{...}{interation terms}

\item{id}{discriminat variable}
}
\value{
data
}
\description{
Create a multiply variable to create a moderator variable
}
\examples{

\dontrun{

# Using example data
data(satisfaction)
sat_blocks1 <- list(
  IMAG = paste0("imag", 1:5),
  EXPE = paste0("expe", 1:5),
  QUAL = paste0("qual", 1:5),
  VAL = paste0("val", 1:4),
  SAT = paste0("sat", 1:4),
  LOY = paste0("loy", 1:4)
)
# Partially select to create an interaction term
satisfaction\%>\% plspm_prod(sat_blocks1,
                           intIS = interlist(c("IMAG", 1, 2, 3), c("SAT", 1, 2, 3))) \%>\% str()

# Partially select to create an interaction term : repeated
satisfaction\%>\% plspm_prod(sat_blocks1,
                           intIS = interlist(c("IMAG", 1, 2, 3), c("SAT", 1, 2, 3))
)\%>\% plspm_prod(sat_blocks1,
                intIS = interlist(c("IMAG", 1, 2, 3), c("EXPE", 1, 2))
) \%>\%str()


# Create an interaction variable
plspm_prod(satisfaction, sat_blocks1, intIS = interlist("IMAG", "SAT"))    \%>\%str()
# Create a variable that interacts with multiple variables
plspm_prod(satisfaction, sat_blocks1, intIS = interlist("IMAG", "SAT"),
           intIE = interlist("IMAG", "EXPE"))    \%>\%str()

satisfaction\%>\% plspm_prod(sat_blocks1,
                           interlist(c("IMAG", 1, 2, 3), c("SAT", 1, 2, 3))) \%>\%str()

satisfaction\%>\% plspm_prod(sat_blocks1,
                           interlist("IMAG", 1, 2, 3, "SAT", 1, 2, 3) ) \%>\%str()

# plspm_prod(satisfaction, sat_blocks1, list("IMAG", "SAT"),
#                                        list("IMAG", "EXPE")) \%>\%str()


#  Use vector
plspm_prod(satisfaction, sat_blocks1,
           intIS = list(c("IMAG", 1, 2 ), c("SAT", 3,4)),
           intIE = list(c("IMAG", 1, 2, 3), c("EXPE" ,4,5))
) \%>\%str()

#interlist use use list
plspm_prod(satisfaction,
           sat_blocks1,
           intIS = interlist(list("IMAG", 1, 2 ,"SAT", 3,4)),
           intIE = interlist(list("IMAG", 1, 2, 3, "EXPE" ,4,5)) ) \%>\%str()


#interlist use now use list
plspm_prod(satisfaction,
           sat_blocks1,
           intIS = interlist("IMAG", 1, 2 ,"SAT", 3,4),
           intIE = interlist("IMAG", 1, 2, 3, "EXPE" ,4,5)) \%>\%str()
}



}
