% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semPaths2.R
\name{plspm_semPaths2}
\alias{plspm_semPaths2}
\title{Drawing Structural Equation Models Using the lavaan Syntax (park JH PhD)}
\usage{
plspm_semPaths2(
  plsres_boot,
  residuals = FALSE,
  whatLabels = "model",
  rotation = 2,
  residScale = 12,
  layout = "tree2",
  edge.label.cex = 0.7,
  edge.label.position = 0.55,
  edge.color = "black",
  sizeMan = 6,
  sizeMan2 = 4,
  sizeLat = 9,
  sizeLat2 = 6,
  style = "lisrel",
  sig = FALSE,
  exoVar = FALSE,
  exoCov = TRUE,
  curve = 1.5,
  asize = 2,
  mar = c(3, 8, 3, 8),
  nDigits = 3,
  shapeLat = "circle",
  shapeMan = "rectangle",
  shapeInt = "triangle",
  sample.nobs = 100,
  border.width = 2,
  edge.width = 1.5,
  Groups = FALSE,
  group_match = "lat",
  growth = FALSE,
  structural = FALSE,
  edgeLabels = NULL,
  nodeLabels = NULL,
  fixedStyle = 1,
  color = list(lat = "skyblue", man = "orange"),
  label.cex = 1,
  type = "plot"
)
}
\arguments{
\item{plsres_boot}{plsres_boot to model syntax}

\item{residuals}{residauls TRUE}

\item{whatLabels}{"model}

\item{rotation}{rotation=2}

\item{residScale}{residScale=12}

\item{layout}{layout="tree"}

\item{edge.label.cex}{1}

\item{edge.label.position}{0.5}

\item{edge.color}{edge.color= black}

\item{sizeMan}{8}

\item{sizeMan2}{4}

\item{sizeLat}{10}

\item{sizeLat2}{6}

\item{style}{lisrel}

\item{sig}{FALSE}

\item{exoVar}{FALSE}

\item{exoCov}{TRUE}

\item{curve}{1.5}

\item{asize}{1.5}

\item{mar}{c(2,8,3,10)}

\item{nDigits}{2}

\item{shapeLat}{"circle, ellipse}

\item{shapeMan}{retangle}

\item{shapeInt}{triangle ,}

\item{sample.nobs}{sample.nobs = 100}

\item{border.width}{border.width=1}

\item{edge.width}{edge.width=1.5}

\item{Groups}{FALSE}

\item{group_match}{"FALSE"lat#'}

\item{growth}{growth= FALSE, TRUE  model.type ="grouwth"}

\item{structural}{structural = FALSE}

\item{edgeLabels}{edgeLabels=NULL}

\item{nodeLabels}{nodeLabels=NULL}

\item{fixedStyle}{fixedStyle=1 default rbase linetype}

\item{color}{list(lat="skyblue", man="yellow")}

\item{label.cex}{label.cex}

\item{type}{type ="plot, and res is lavaan simdata result}
}
\value{
plot
}
\description{
Drawing Structural Equation Models Using the lavaan Syntax (park JH PhD)
}
