% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semPaths_lav.R
\name{plspm_semPaths_lav}
\alias{plspm_semPaths_lav}
\title{Generate lavaan syntax from a model or plspm result}
\usage{
plspm_semPaths_lav(plsres_boot, type = "lav", paths = NULL, blocks = NULL)
}
\arguments{
\item{plsres_boot}{plspm result}

\item{type}{lav, ,model, edge, edgeLabels}

\item{paths}{paths matrix}

\item{blocks}{blocks tree}
}
\value{
lavaan syntax
}
\description{
Generate lavaan syntax from a model or plspm result
}
\examples{

\dontrun{

library(jjstat)
data(satisfaction)
sat_path = plspm_paths(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),
  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY"))
)
sat_path

# blokcs
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)

sat_blocks1


###lav
satpls_boot\%>\%
  plspm_semPaths_lav(paths=sat_path, blocks=sat_blocks1)\%>\%cat()

## model
satpls_boot\%>\%plspm_semPaths_lav()

## edge
satpls_boot\%>\%plspm_semPaths_lav("edge")

## draw model
satpls_boot\%>\%
  plspm_semPaths_lav(paths=sat_path, blocks=sat_blocks1)\%>\%
  diagram_model(  edgeLabels= satpls_boot\%>\%plspm_semPaths_lav("edge"),
                  mar=c(2,2,2,2),
                  sizeLat=4, sizeMan=3, sizeMan2= 2, edge.label.cex=0.5, rotation=1)

###
satpls_boot\%>\%plspm_semPaths_lav()\%>\%
  diagram_model(edgeLabels=satpls_boot\%>\%plspm_semPaths_lav("edge"),
                mar=c(2,2,2,2),edge.label.cex=0.5, edge.label.position= 0.7,
                sizeLat=4, sizeMan=3, sizeMan2= 2, rotation=1)

}


}
