% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_semptools.R
\name{plspm_semptools}
\alias{plspm_semptools}
\title{plspm_semptools}
\usage{
plspm_semptools(
  semPaths_plot,
  pls_boot,
  position = NULL,
  point = NULL,
  set_push = 2,
  set_spread = 2,
  nrow = 3,
  ncol = 3,
  loading_position = 0.5
)
}
\arguments{
\item{semPaths_plot}{plot data}

\item{pls_boot}{boot data}

\item{position}{layout  position}

\item{point}{point position}

\item{set_push}{each push}

\item{set_spread}{each spread}

\item{nrow}{default 3}

\item{ncol}{default 3}

\item{loading_position}{0.5}
}
\value{
plot
}
\description{
plspm_semptools
}
\examples{

\dontrun{
#'

#########################################################
pathmodel = "
자기효능감 =~ C_S1 +  C_S2 + C_S3 + C_S4 + C_S5
진로태도 =~ B02 + B04 + B10 + B14
진로동기 =~ A_M1 + A_M2 + A_M3
진로준비 =~ D_P1 + D_P2 + D_P3

진로동기 ~ H1*자기효능감
진로태도 ~  H2*자기효능감
진로준비 ~ H5*진로태도 + H4*진로동기 + H3*자기효능감
"
pathmodel

plspm_semPaths2(jutpls_boot,   curve=1, layout="tree2",rotation=1,
                sizeMan= 10, color=list(lat="gray10",man="gray98"),
                edge.label.cex=0.7, mar=c(5,5,5,5),
                edge.label.position=0.55,
                sizeLat= 8)\%>\%
  semptools::set_sem_layout(
    indicator_order =  plspm_indicator_order(pathmodel),
    indicator_factor =  plspm_indicator_factor(pathmodel),
    factor_layout=  matrix(
      c(  NA, '진로태도', NA,
          '자기효능감', NA, '진로준비',
          NA, '진로동기', NA),
      nrow = 3, ncol = 3, byrow = TRUE),
    factor_point_to= matrix(
      c(  NA, 'up', NA,
          'left', NA, 'right',
          NA, 'down', NA),
      nrow = 3, ncol = 3, byrow = TRUE),
    indicator_push <- c(진로태도 = 2, 진로동기 = 2, 자기효능감=1.5, 진로준비 =1.5),
    indicator_spread <- c(자기효능감 = 3, 진로태도 = 2, 진로동기 =2, 진로준비 = 2)
  )\%>\%  semptools::set_curve(c("진로준비~~ 진로태도" = 0.5))\%>\%
  semptools::set_curve(c("진로준비~~ 진로동기" = 0))\%>\%
  plspm_change_nodeLabels("B02","진로결정",
                          "B04", "진로독립",
                          "B10", "진로계획",
                          "B14", "진로성향",
                          "A_M1","진로정체",
                          "A_M2","진로통찰",
                          "A_M3","진로탄력",
                          "C_S1","자기평가",
                          "C_S2","직업정보",
                          "C_S3","계획수립",
                          "C_S4","목표설정",
                          "C_S5","해결방식",
                          "D_P1","진로준비",
                          "D_P2","도구준비",
                          "D_P3","목표달성"
  )
# second ##################################
plspm_indicator_order(pathmodel)
plspm_indicator_factor(pathmodel)


#옵션함수
plspm_boot_ind_order(jutpls_boot)
plspm_boot_ind_factor(jutpls_boot)

plspm_boot_factor_layout(jutpls_boot)
plspm_boot_factor_point_to(jutpls_boot, ncol=3)

plspm_ind_push(jutpls_boot)
plspm_ind_spread(jutpls_boot, set= c(자기효능감=4,진로태도=3, 진로동기=3))

##plot
plspm_semPaths2(jutpls_boot,   curve=1, layout="tree2",rotation=1,
                sizeMan= 10, color=list(lat="gray10",man="gray98"),
                edge.label.cex=0.7, mar=c(5,5,5,5),
                edge.label.position=0.55, edge.color= "tomato4",
                sizeLat= 8)\%>\%
  semptools::set_sem_layout(
    indicator_order = plspm_boot_ind_order(jutpls_boot),
    indicator_factor = plspm_boot_ind_factor(jutpls_boot),
    factor_layout= plspm_boot_factor_layout(jutpls_boot, ncol=3),
    factor_point_to= plspm_boot_factor_point_to(jutpls_boot, ncol=3),
    indicator_push <- plspm_ind_push(jutpls_boot),
    indicator_spread <-  plspm_ind_spread(jutpls_boot, set= c(자기효능감=4,진로태도=3, 진로동기=3)))\%>\%
  plspm_curve(c("진로준비~ 진로동기" =0))\%>\%
  plspm_change_nodeLabels("B02","진로결정",
                          "B04", "진로독립",
                          "B10", "진로계획",
                          "B14", "진로성향",
                          "A_M1","진로정체",
                          "A_M2","진로통찰",
                          "A_M3","진로탄력",
                          "C_S1","자기평가",
                          "C_S2","직업정보",
                          "C_S3","계획수립",
                          "C_S4","목표설정",
                          "C_S5","해결방식",
                          "D_P1","진로준비",
                          "D_P2","도구준비",
                          "D_P3","목표달성"
  )

#'
#' #we want #########################################
plspm_semPaths2(jutpls_boot,   curve=1, layout="tree2",rotation=1,
                sizeMan= 10, color=list(lat="gray10",man="gray98"),
                edge.label.cex=0.7, mar=c(5,5,5,5),
                edge.label.position=0.55, edge.color= "tomato4",
                sizeLat= 8) \%>\%
  plspm_semptools(jutpls_boot)\%>\%
  plspm_curve(c("진로준비~ 진로동기" =0))\%>\%
  plspm_change_nodeLabels("B02","진로결정",
                          "B04", "진로독립",
                          "B10", "진로계획",
                          "B14", "진로성향",
                          "A_M1","진로정체",
                          "A_M2","진로통찰",
                          "A_M3","진로탄력",
                          "C_S1","자기평가",
                          "C_S2","직업정보",
                          "C_S3","계획수립",
                          "C_S4","목표설정",
                          "C_S5","해결방식",
                          "D_P1","진로준비",
                          "D_P2","도구준비",
                          "D_P3","목표달성"
  )
#'


}

}
