% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_duncan.R
\name{posthoc_duncan}
\alias{posthoc_duncan}
\title{posthoc_duncan}
\usage{
posthoc_duncan(aov_res, group_var)
}
\arguments{
\item{aov_res}{aov result}

\item{group_var}{group}
}
\value{
df
}
\description{
posthoc_duncan
}
\examples{
\dontrun{
#'
aov_post_eng <- pwj3_long \%>\%
  filter(time == "Post" & factor == "eng") \%>\%
  aov(formula = score ~ Eng_Class)

# Duncan 사후분석 수행
posthoc_duncan(aov_post_eng, "Eng_Class")
}


}
