% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_plot.R
\name{posthoc_plot}
\alias{posthoc_plot}
\title{Post-hoc results and plots}
\usage{
posthoc_plot(
  aov_data,
  type = "all",
  console = FALSE,
  group = TRUE,
  adjust = "none"
)
}
\arguments{
\item{aov_data}{aov data}

\item{type}{type is all, res, g, post hoc}

\item{adjust}{}
}
\value{
plot and analysis result
}
\description{
Post-hoc results and plots
}
\examples{

\dontrun{

#Need to replace cyl with factor first. Change to as_trt() first
aov(mpg ~ cyl, data=mtcars \%>\%as_trt("cyl"))\%>\%posthoc_plot()

##Error in emmeans::emmeans(., iv) :
aov(mpg ~ factor(cyl), data=mtcars)\%>\%posthoc_plot()
}

}
