% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE.R
\name{PRE}
\alias{PRE}
\title{PRE(proportional reduction in error)}
\usage{
PRE(model1, model2)
}
\arguments{
\item{model1}{model 1}

\item{model2}{model q mixed model}
}
\value{
random slope, inetcept
}
\description{
PRE(proportional reduction in error)
}
\examples{
\duntrun{
 PRE(gmlm.dich.m1b, gmlm.dich.m2c)
}
}
