% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_find_paths.R
\name{print.plspm_paths}
\alias{print.plspm_paths}
\title{Define a custom print function}
\usage{
\method{print}{plspm_paths}(x, ...)
}
\arguments{
\item{x}{ril}

\item{...}{imput}
}
\value{
print
}
\description{
Define a custom print function
}
