% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_profile_plot2.R
\name{profile_plot2}
\alias{profile_plot2}
\title{profile_plot2 Brargrahp}
\usage{
profile_plot2(data, n_profiles = 3, model_name = NULL, view = "pair")
}
\arguments{
\item{data}{data}

\item{n_profiles}{profiles number}

\item{model_name}{"EEE","EEI", "VVV","VVI",}

\item{view}{each, pair}
}
\value{
plot
}
\description{
profile_plot2 Brargrahp
}
\examples{

\dontrun{

iris[,-5] \%>\%profile_plot2(n_profiles = 4, model_name = "EEE")

}


}
