% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pull.R
\name{Pull}
\alias{Pull}
\title{Pull select data.frame col to vector}
\usage{
Pull(df, col, name = NULL, sep = "_")
}
\arguments{
\item{df}{df}

\item{col}{select col}

\item{name}{using names from rownames or col data}

\item{sep}{paste sep = "_"}
}
\value{
vector
}
\description{
Pull select data.frame col to vector
}
\examples{

\dontrun{
#'
# 예시 데이터
df <- tibble::tibble(
  term = c("학교 수업에서", "혼자서 공부할 때", "활용 안함"),
  Freq = c(2058, 1248, 646)
)

# 함수 사용 예시
Pull(df, col="Freq", name="term")
}


}
