% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_coding.R
\name{rc}
\alias{rc}
\title{REversse codings rc}
\usage{
rc(vec, n = 6, check = FALSE)
}
\arguments{
\item{vec}{input data}

\item{n}{likert max+1}

\item{check}{View the original source and reverse coded results}
}
\value{
reverse coding data
}
\description{
REversse codings rc
}
\examples{
\dontrun{
## error  :Error message output when data frames are entered
rev_coding(mtcars)

#reverse coding
rev_coding(mtcars$gear)

## View the original source and decoded results
rev_coding(mtcars$gear, check=TRUE)

###If the Likert scale is wrong, the highest scale is 5, but the highest scale is 8.
rev_coding(mtcars$cyl)
}
}
