% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_text.R
\name{read_text}
\alias{read_text}
\title{read_text: Read a Text File with Flexible Handling Using read.delim}
\usage{
read_text(file_path, header = TRUE, stringsAsFactors = FALSE, sep = "\\t")
}
\arguments{
\item{file_path}{The path of the file to read.}

\item{header}{Logical. Does the file contain a header? Defaults to TRUE.}

\item{stringsAsFactors}{Logical. Should character columns be converted to factors? Defaults to FALSE.}
}
\value{
A data frame containing the file's contents.
}
\description{
read_text: Read a Text File with Flexible Handling Using read.delim
}
\examples{
\dontrun{
# Save and read a file
teacher_question <- c("Example question 1", "Example question 2")
write_text(teacher_question, file_path = "output.txt")
result <- read_text("output.txt")
}

}
