% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_sep_vec.R
\name{recode_sep_vec}
\alias{recode_sep_vec}
\title{recode_sep_vec 벡터데이터를 리코딩}
\usage{
recode_sep_vec(vector, ranges_list)
}
\arguments{
\item{vector}{vector}

\item{ranges_list}{변환리스트 list("0~5년" = 0:5)}
}
\value{
vector
}
\description{
recode_sep_vec 벡터데이터를 리코딩
}
\examples{
\dontrun{
#'
# 예시 사용
vec <- c("53", "45", "38", "28", "41", "12", "0", "10", "30", "35")

# 함수를 사용하여 벡터 재코딩
new_vec <- recode_sep_vec(vec, list("0~5년" = 0:5,
                                    "5~10년" = 6:10,
                                    "11~15년" = 11:15,
                                    "16~20년" = 16:20,
                                    "21~25년" = 21:25,
                                    "26~30년" = 26:30,
                                    "30년이상" = 31:60))

# 결과 확인
print(new_vec)

# # # 예시 사용
# vec <- c(53, 45, 45, 38, 36, 28, 38, 28, 41, 55, 39, 29, 31, 35, 37, 34, 34, 35, 58, 46)

# # 함수를 사용하여 벡터 재코딩
new_vec <- recode_sep_vec(vec, list("20대" = 20:29, "30대" = 30:39, "40대" = 40:49, "50대" = 50:59))

# # 결과 확인
# print(new_vec)
}


}
