% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_reg.R
\name{reg_apa}
\alias{reg_apa}
\title{regression APA report}
\usage{
reg_apa(model, digits = 3)
}
\arguments{
\item{model}{lm_data}

\item{digits}{3}
}
\value{
report
}
\description{
regression APA report
}
\examples{
\dontrun{
lm(mpg ~ hp + wt + disp + qsec, data = mtcars) \%>\% reg_apa()
}

}
