% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_attr.R
\name{remove_all_attr}
\alias{remove_all_attr}
\title{Remove all ATTRs present in the data}
\usage{
remove_all_attr(df)
}
\arguments{
\item{df}{data}
}
\value{
data
}
\description{
Remove all ATTRs present in the data
}
