% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_attr.R
\name{remove_attr}
\alias{remove_attr}
\title{Remove all ATTRs present in the data, deselect}
\usage{
remove_attr(df, range = 1:ncol(df))
}
\arguments{
\item{df}{data}

\item{range}{col range}
}
\value{
data.frame
}
\description{
Remove all ATTRs present in the data, deselect
}
\examples{

\dontrun{
#'
# Example usage with your data frame jut3
# Remove attributes from all columns
# jut3_cleaned_all <- remove_attr(jut3)

# Remove attributes from columns 1 to 5
# jut3_cleaned_partial <- remove_attr(jut3, range = 1:5)

# Check the structure
# str(jut3_cleaned_all)
# str(jut3_cleaned_partial)
}


}
