% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_blank.R
\name{remove_blank}
\alias{remove_blank}
\title{데이터내에서 빈칸 제거}
\usage{
remove_blank(
  data,
  col = NULL,
  data_type = c("data.frame", "vector", "character")
)
}
\arguments{
\item{data}{df}

\item{col}{col 숫자와 열이름으로 지정}

\item{data_type}{data.frame, vector}
}
\value{
df
}
\description{
데이터내에서 빈칸 제거
}
\examples{
\dontrun{
#'
# 예제 데이터
vector_data <- c("익산 어양 초", "익산옥 야 초")
df_data <- data.frame(
  ID = c(259, 260),
  Name = c("익산 어양 초", "익산옥 야 초"),
  Value = c(1, 1),
  Score = c(0.1237624, 0.1237624)
)

# 벡터에서 공백 제거
clean_vector <- remove_blank(vector_data, data_type = "vector")
print(clean_vector)

# 데이터프레임에서 공백 제거 (열 이름으로)
clean_df <- remove_blank(df_data, col = "Name")
print(clean_df)

# 데이터프레임에서 공백 제거 (열 번호로)
clean_df <- remove_blank(df_data, col = 2)
print(clean_df)
#'

}
}
