% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_remove.R
\name{remove_col}
\alias{remove_col}
\title{Delete cols in the data}
\usage{
remove_col(data, rm_col = NULL, ...)
}
\arguments{
\item{data}{data}

\item{rm_col}{first column name or number}

\item{...}{add columns}
}
\value{
data
}
\description{
Delete cols in the data
}
\examples{
\dontrun{
#'
# 예시 사용법
col_remove(mtcars, 1, "am", 2, "vs", "wt")

# 예시 사용법
col_remove(mtcars, 1, "am", 2, "vs", "wt",4)


# 예시 사용법
col_remove(mtcars, 1, "am", 2, "vs", "wt",3)


# 예시 사용법
col_remove(mtcars, 1, "am", 2, "vs", "wt",3,4)

# 예시 사용법
col_remove(mtcars, rm_col=1:3)
col_remove(mtcars, c(1, 3))
col_remove(mtcars, 1,2, 3)
col_remove(mtcars, 1, "am",2,"vs","wt")
col_remove(mtcars, rm_col="mpg")
col_remove(mtcars, "mpg", "am","vs")

}


}
