% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_filter.R
\name{remove_filter}
\alias{remove_filter}
\title{Functions to remove selected rows}
\usage{
remove_filter(data, ...)
}
\arguments{
\item{data}{data}

\item{...}{condition filter}
}
\value{
data
}
\description{
Functions to remove selected rows
}
\examples{

\dontrun{
#'
mtcars \%>\% remove_filter(cyl == 6 )
mtcars \%>\% remove_filter(vs == 1)
mtcars \%>\% remove_filter(am==0)
mtcars \%>\% remove_filter(vs == 1)
mtcars \%>\% remove_filter(cyl == 6 | vs == 1)
mtcars \%>\% remove_filter(cyl == 6 & vs == 1)
mtcars \%>\% remove_filter(cyl == 6 | cyl == 4 | am == 0)
}

}
