% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rows.R
\name{remove_rows}
\alias{remove_rows}
\title{remove rows}
\usage{
remove_rows(data, ..., filter = FALSE)
}
\arguments{
\item{data}{data}

\item{...}{rows}

\item{filter}{TRUE possible filter}
}
\value{
data
}
\description{
remove rows
}
\examples{

\dontrun{
#'
# 1~3행 제거
remove_rows(mtcars,  1:30)
remove_rows(mtcars, 1,2,3,6,5,7,8,9,30)
remove_rows(mtcars, rm = c(1,3,4,5,6,7,11:30))
remove_rows(mtcars, cyl == 4, filter=T)

# 1행과 3행 제거
print(remove_rows(mtcars, rm_row = c(1, 3)))

# 필터 조건 적용 (cyl == 6)
remove_rows(mtcars,rm=NULL, cyl == 6) #error
remove_rows(mtcars, cyl == 6 |cyl ==4, filter=T)

# 필터 조건 적용 (cyl == 6 또는 vs == 1)
remove_rows(mtcars, cyl == 6 | vs == 1, filter=T)

# 필터 조건 적용 (cyl == 6 그리고 vs == 1)
remove_rows(mtcars, cyl == 6 | vs == 1, filter=T)

# 필터 조건 적용 후 특정 행 제거 (필터가 우선 적용)
remove_rows(mtcars, 1:3, cyl == 6 | vs == 1)

#'
}


}
