% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace.R
\name{Replace}
\alias{Replace}
\title{Create a function to replace NA with an empty string in an R vector data}
\usage{
Replace(vector, pattern = NA, imp = "")
}
\arguments{
\item{vector}{input data}

\item{pattern}{dectect pattern}

\item{imp}{imputation}
}
\description{
Create a function to replace NA with an empty string in an R vector data
}
\examples{

\dontrun{
# Replace(c("A", NA, "B", "", "C", "C"), "C", imp="T")
}

}
