% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_df_rep.R
\name{replace_dataframe}
\alias{replace_dataframe}
\title{replace_dataframe}
\usage{
replace_dataframe(data, pattern, imp)
}
\arguments{
\item{data}{data}

\item{pattern}{patern}

\item{imp}{imputation}
}
\value{
data
}
\description{
replace_dataframe
}
