% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_df.R
\name{replace_df}
\alias{replace_df}
\title{Search for and fill in empty data}
\usage{
replace_df(df, pattern = NA, imp = "")
}
\arguments{
\item{df}{data.frame}

\item{pattern}{default ""}

\item{imp}{iimputation , defalut NA}
}
\value{
data.frame
}
\description{
Search for and fill in empty data
}
