% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_df.R
\name{replace_df2}
\alias{replace_df2}
\title{Search for and fill in empty data}
\usage{
replace_df2(df, pattern = NA, imp = "", cat = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{pattern}{default ""}

\item{imp}{iimputation , defalut NA}
}
\value{
data.frame
}
\description{
Search for and fill in empty data
}
