% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_df_rep.R
\name{replace_in_dataframe}
\alias{replace_in_dataframe}
\title{replace_in_dataframe}
\usage{
replace_in_dataframe(data, changes)
}
\arguments{
\item{data}{data}

\item{changes}{change}
}
\value{
data
}
\description{
replace_in_dataframe
}
