% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_df.R
\name{replace_in_df}
\alias{replace_in_df}
\title{replace_in_df}
\usage{
replace_in_df(data, pattern, imp)
}
\arguments{
\item{data}{data}

\item{pattern}{pattern}

\item{imp}{""}
}
\value{
data
}
\description{
replace_in_df
}
\examples{
\dontrun{
#'
data <- data.frame(
  Question = c("나는 나의 세 가지 가장 큰 약점을 열거할 수 있다.", "나의 행동은 나의 핵심 가치를 반영한다.", "나는 스스로 결정하기 전에 다른 사람들의 의견을 구한다."),
  factor = c("자기", "자기", "자기"),
  stringsAsFactors = FALSE
)

data
data \%>\% replace_in_df("약점", "강점")
}
}
