% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_aov.R
\name{report_aov}
\alias{report_aov}
\title{anova report ,report_aov}
\usage{
report_aov(data, digits = 2, trans = FALSE)
}
\arguments{
\item{data}{aov data}

\item{digits}{default 3}

\item{trans}{TRUE english to korea}
}
\value{
report
}
\description{
anova report ,report_aov
}
\examples{
\dontrun{
aov(len ~ supp*factor(dose), data= ToothGrowth)|>report_aov()
aov(len ~ supp*factor(dose), data= ToothGrowth)|>report_aov(trans=TRUE)


}

}
