% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_report.R
\name{report_lme}
\alias{report_lme}
\title{linear mixed model report}
\usage{
report_lme(
  lmedata,
  type = "basic",
  form = "lmer",
  apa = FALSE,
  fit_more = FALSE,
  ranef_sig = FALSE,
  show.effect = FALSE,
  show.ci = FALSE
)
}
\arguments{
\item{lmedata}{lmedata is lmer() function result}

\item{type}{= 'all' is res, 'Fixed_effect','Random_effect','ICC', 'ConfidenceInterval_95','Satterthwaite_method','FIT','APA'}

\item{form}{'lmer','glmer' = 'logit'='dich', 'poisson' ='pois'}

\item{apa}{= default FALSE, if you  REML is FALSE}

\item{fit_more}{= default FALSE, TRUE detailed report}

\item{ranef_sig}{ranef_sig =TRUE  random effect test}

\item{show.effect}{if you show.effect = TRUE -> show ranef, fixef, coef}

\item{show.ci}{TRUE calculation 95percent CI}
}
\description{
Mixed model summary
}
\examples{
\dontrun{
data(sleepstudy)

sleepstudy |> str()

lmer(Reaction ~ Days + (Days | Subject), sleepstudy) |> lme_report()

lmer(Reaction ~ Days + (Days | Subject), sleepstudy) |> lme_report(apa=TRUE)


book.url <- "https://stat.ethz.ch/~meier/teaching/book-anova"
quality <- readRDS(url(file.path(book.url, "data/quality.rds")))
str(quality)
fit.quality <- lmer(score ~ (1 | employee) + (1 | batch) +
                     (1 | employee:batch), data = quality)
summary(fit.quality)
fit.quality|> lme_report()
fit.quality|> lme_report(apa= TRUE) #error the app is FALSE
fit.quality|> lme_report(apa=FALSE) #error the app is FALSE
fit.quality|> ICC()

lmer(Reaction ~ Days + (Days | Subject), sleepstudy) |>
SEM212::lme_report(apa=TRUE, fit_more = T)
lmer(Reaction ~ Days + (Days | Subject), sleepstudy) |>
SEM212::lme_report(apa=TRUE, fit_more = F)  #default FALSE, it is show that AIC and BIC

}

}
