% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ttest.R
\name{report_ttest}
\alias{report_ttest}
\title{report_ttest}
\usage{
report_ttest(ttest_result, trans = FALSE, type = "report")
}
\arguments{
\item{ttest_result}{t.test result}

\item{trans}{true english to korean}

\item{type}{default 'report', 'all' is t.table and APA}
}
\value{
APA report
}
\description{
report_ttest
}
\examples{
\dontrun{
## english report
t.test(mpg ~ vs, mtcars, var.equal=TRUE) \%>\%
        report_ttest()

## translation
t.test(mpg ~ vs, mtcars, var.equal=TRUE) \%>\%
report_ttest(trans=T)

t.test(mpg ~ vs, mtcars) \%>\% report_ttest("all")
}
}
