% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stat.R
\name{report_vartest}
\alias{report_vartest}
\title{var.test report}
\usage{
report_vartest(var_test, trans = FALSE)
}
\arguments{
\item{var_test}{var.test result}

\item{trans}{tans true english to korean}
}
\value{
explain result apa
}
\description{
var.test report
}
\examples{
\dontrun{
var.test(mpg ~ vs, data = mtcars)|>
     report_vartest()

var.test(mpg ~ vs, data = mtcars)|>
     report_vartest() |>
     jjstat::kakaoi()

var.test(mpg ~ vs, data = mtcars)|>
report_vartest(trans=TRUE)

}
}
