% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_coding_extract_only.R
\name{rev_coding_extract_only}
\alias{rev_coding_extract_only}
\title{Extract only reverse-coded variables from a data.frame}
\usage{
rev_coding_extract_only(data, drop_attr = FALSE)
}
\arguments{
\item{data}{A data.frame of class 'revcoded_df'}

\item{drop_attr}{Logical. If TRUE, removes attributes and classes (default = FALSE)}
}
\value{
A data.frame containing only reverse-coded variables
}
\description{
Extract only reverse-coded variables from a data.frame
}
\examples{
df <- data.frame(A = 1:5, B = 5:1)
res <- rev_coding_rep(df, 6, A, B)
rev_coding_extract_only(res)
}
