% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_coding_map.R
\name{rev_coding_map}
\alias{rev_coding_map}
\title{Create a mapping table between original and reverse-coded variables}
\usage{
rev_coding_map(data)
}
\arguments{
\item{data}{A data.frame of class 'revcoded_df'}
}
\value{
A data.frame with original and reverse-coded variable name pairs
}
\description{
Create a mapping table between original and reverse-coded variables
}
\examples{
df <- data.frame(A = 1:5, B = 5:1)
res <- rev_coding_rep(df, 6, A, B)
rev_coding_map(res)
}
