% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_coding_plot_var.R
\name{rev_coding_plot_var}
\alias{rev_coding_plot_var}
\title{Plot reverse-coded variable against original values}
\usage{
rev_coding_plot_var(x, data = NULL, ...)
}
\arguments{
\item{x}{A reverse-coded vector with class 'reverse_var'}

\item{data}{Optional: original data.frame containing the original variable}

\item{...}{Additional graphical arguments passed to plot}
}
\value{
A scatter plot of original vs. reverse-coded values
}
\description{
Displays a comparison between the original and reverse-coded values.
Works only when original variable is still available in parent environment or in the given data.
}
\examples{
\dontrun{
df <- data.frame(A = 1:5)
res <- rev_coding_rep(df, 6, A)
rev_coding_plot_var(res$A_r, data = df)
}
}
