% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_coding_restore_original.R
\name{rev_coding_restore_original}
\alias{rev_coding_restore_original}
\title{Remove reverse-coded variables, keeping only original variables}
\usage{
rev_coding_restore_original(data)
}
\arguments{
\item{data}{A data.frame of class 'revcoded_df'}
}
\value{
A data.frame without reverse-coded variables
}
\description{
Remove reverse-coded variables, keeping only original variables
}
\examples{
df <- data.frame(A = 1:5, B = 5:1)
res <- rev_coding_rep(df, 6, A, B)
rev_coding_restore_original(res)
}
