% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcode.R
\name{revcode}
\alias{revcode}
\title{Custom function to perform reverse coding for multiple variables with custom n values}
\usage{
revcode(data, n = 8, ..., postfix = "_r", replace = FALSE)
}
\arguments{
\item{data}{Data frame containing the variables}

\item{n}{Vector of maximum scale values (one for each vadata:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAbElEQVR4Xs2RQQrAMAgEfZgf7W9LAguybljJpR3wEse5JOL3ZObDb4x1loDhHbBOFU6i2Ddnw2KNiXcdAXygJlwE8OFVBHDgKrLgSInN4WMe9iXiqIVsTMjH7z/GhNTEibOxQswcYIWYOR/zAjBJfiXh3jZ6AAAAAElFTkSuQmCCriable), c("v1","v2",...)}

\item{...}{Variable names to be reverse coded}

\item{postfix}{postfix is reverse variable name default '_r'}

\item{replace}{How to enter and replace the reverse coding results}
}
\value{
add reverse data in data.frame
}
\description{
Custom function to perform reverse coding for multiple variables with custom n values
}
\examples{
\dontrun{

### How to decode all variables by applying only one n
revcode(stat_onl, 8, "upgrade", "satisfy", "fatigue" )|>
select(upgrade, satisfy, fatigue, upgrade_r, satisfy_r, fatigue_r)

### the method for the reverse coding variable being applied separately
### to each variable and decoding
revcode(stat_onl, n= c(8,8,8), "upgrade", "satisfy", "fatigue" )|>
select(upgrade, satisfy, fatigue, upgrade_r, satisfy_r, fatigue_r)

## mtcars data
## Since the inverse coding value of the cyl variable is 4,6,8,
## n enters 4 + 8 = 12; the minimum + maximum = n value.
revcode(mtcars[,c("cyl", "gear","carb")], 12, "cyl")


revcode(mtcars[,c("cyl", "gear","carb")],
            n = c(12, 6, 9),"cyl","gear","carb")
#'
df <- data.frame(
  A = c(1, 2, 3, 4, 5),
  B = c(5, 4, 3, 2, 1)
)

# Passing variable names as symbols
df_new1 <- revcode(df, n = 6, A, B)
print(df_new1)
str(df_new1)
print(attr(df_new1$A_r, "reverse")) # Output properties for verification
print(attr(df_new1$B_r, "reverse"))

# pass the variable name as a string
df_new2 <- revcode(df, n = 6, "A", "B")
print(df_new2)
str(df_new2)
print(attr(df_new2$A_r, "reverse"))  # Output properties for verification
print(attr(df_new2$B_r, "reverse"))

# Reverse-code to replace original column
df_replaced <- revcode(df, n = 6, A, B, replace = TRUE)
print(df_replaced)
str(df_replaced)
print(attr(df_replaced$A, "reverse"))  # Output properties for verification
print(attr(df_replaced$B, "reverse"))

#'

}
}
