% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rload.R
\name{Rload}
\alias{Rload}
\title{Rload}
\usage{
Rload(file, verbose = TRUE)
}
\arguments{
\item{file}{position}

\item{verbose}{default TRUR}
}
\value{
datafile
}
\description{
Rload
}
\examples{

\dontrun{
#'
#'
#'
Rload("pwj2.RData", verbose = TRUE)
pwj2 \%>\%head()
Rload(file="F:/Rwork/02_OutAna/Parkwonjun_yonsei_2024/pwj2.RData")
}


}
