% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round.R
\name{Round}
\alias{Round}
\title{Find and round numeric variables only}
\usage{
Round(data, digits = 2, exclude = NULL, type = "tibble")
}
\arguments{
\item{data}{data.frame}

\item{digits}{default 2}

\item{exclude}{exclude variable not adapted}

\item{type}{data type tibble or data.frame , matrix}
}
\value{
rounding data
}
\description{
Find and round numeric variables only
}
\examples{
\dontrun{

# vector
Round(12.345654)

##data.frame
mtcars\%>\% Round(1)

 ##
dff = data.frame(A = c(1.23456, 2.34567, 3.45678),
                 B = c(4.56789, 5.67890, 6.78901),
                C = c("a", "b", "c"))
 ## rounding 1
 dff\%>\%Round(1)

#exclue variable B
 dff\%>\%Round(1, exclude ="B")

}
}
