% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_col.R
\name{round_col}
\alias{round_col}
\title{round_col, 한개의 열만 round처리}
\usage{
round_col(df, col = ncol(df), digits = 3)
}
\arguments{
\item{df}{df}

\item{col}{select col , default last column}

\item{digits}{default 3}
}
\value{
data
}
\description{
round_col, 한개의 열만 round처리
}
\examples{
\dontrun{



}
}
