% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row2col.R
\name{row2col}
\alias{row2col}
\title{rownames to column}
\usage{
row2col(data, colname = "rownames")
}
\arguments{
\item{data}{data.fram}

\item{colname}{new column default rownames}
}
\value{
data
}
\description{
rownames to column
}
\examples{

\dontrun{
mtcars \%>\%row2col("car")
}

}
