% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row2col.R
\name{row2cols}
\alias{row2cols}
\title{rownames to columns}
\usage{
row2cols(data, colname = "rows")
}
\arguments{
\item{data}{data.fram}

\item{colname}{new column default rownames}
}
\value{
data
}
\description{
rownames to columns
}
\examples{

\dontrun{
mtcars \%>\%row2cols("car")
}

}
