% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_row.R
\name{row_filter}
\alias{row_filter}
\title{select rows filter function}
\usage{
row_filter(data, ...)
}
\arguments{
\item{data}{data}

\item{...}{row condition}
}
\value{
data
}
\description{
select rows filter function
}
\examples{

\dontrun{
#'
mtcars \%>\% row_filter(cyl == 6 )
mtcars \%>\% row_filter(vs == 1)
mtcars \%>\% row_filter(cyl == 6 | vs == 1)
mtcars \%>\% row_filter(cyl == 6 & vs == 1)
}



}
