% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pattern_detector.R
\name{row_pattern_detector}
\alias{row_pattern_detector}
\title{row pattern detector}
\usage{
row_pattern_detector(df, pattern = "", rev = FALSE)
}
\arguments{
\item{df}{df}

\item{pattern}{pattern =""}

\item{rev}{Find the opposite row of the pattern}
}
\value{
data
}
\description{
row pattern detector
}
\examples{
\dontrun{
#'
# 샘플 데이터프레임
df <- data.frame(
  A = c(1, NA, 3, 4),
  B = c("Yes", "No", NA, "Yes"),
  C = c(NA, 2, 3, 4)
)

# NA 값을 포함한 행을 찾음
na_rows <- pattern_detector(df, pattern = NA)
print(na_rows)

# "No" 값을 포함한 행을 찾음
pattern_rows <- pattern_detector(df, pattern = "No")
print(pattern_rows)

# "No" 값을 포함하지 않은 행을 찾음 (rev = TRUE)
non_pattern_rows <- pattern_detector(df, pattern = "No", rev = TRUE)
print(non_pattern_rows)
}
}
