% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames2colnames.R
\name{rownames2colnames}
\alias{rownames2colnames}
\title{rownames2colnames}
\usage{
rownames2colnames(df, reverse = FALSE)
}
\arguments{
\item{df}{df}

\item{reverse}{reverse = FALSE}
}
\value{
df
}
\description{
rownames2colnames
}
\examples{
\dontrun{
#'

# 예시 데이터프레임
df <- data.frame(a = 1:3, b = 4:6)
rownames(df) <- c("row1", "row2", "row3")

# rownames를 colnames로 변환
new_df <- rownames2colnames(df)
print(new_df)

# colnames를 rownames로 변환
df_with_colnames_as_rownames <- rownames2colnames(new_df, reverse = TRUE)
print(df_with_colnames_as_rownames)
}
}
