% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsave.R
\name{Rsave}
\alias{Rsave}
\title{RData fast save}
\usage{
Rsave(df, file = "")
}
\arguments{
\item{df}{data to save}

\item{file}{filename}
}
\value{
data
}
\description{
RData fast save
}
\examples{
\dontrun{

Rsave(Kge_jbs)           # 기본 파일 이름: Kge_jbs.RData
Rsave(Kge_jbs, file = "my_custom_name")  # 사용자 지정 파일 이름: my_custom_name.RData
}
}
