% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsave_xlsx.R
\name{Rsave_xlsx}
\alias{Rsave_xlsx}
\title{엑셀파일로 저장}
\usage{
Rsave_xlsx(data)
}
\arguments{
\item{data}{df}
}
\value{
excel data
}
\description{
엑셀파일로 저장
}
\examples{
\dontrun{
#'
Rsave(jen_binddata)
Rload(file="F:/Rwork/02_OutAna/Joeunnarae_yonsei_2024/jen_binddata.RData")
}

}
