% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_format.R
\name{select_format}
\alias{select_format}
\title{select_format}
\usage{
select_format(data, ..., nsmall = 3, digits = 3)
}
\arguments{
\item{data}{data}

\item{...}{col selection}

\item{nsmall}{default 3}

\item{digits}{round 3}
}
\value{
data
}
\description{
select_format
}
\examples{


\dontrun{


# Usage example
data2 <- data.frame(a = rnorm(5), b = rnorm(5), c = rnorm(5), d = rnorm(5), e = rnorm(5))
select_format(data2, 1, 2)
select_format(data2, 1:5)
select_format(data2, a)
select_format(data2, a:e)
select_format(data2, "a")
}
}
