% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_sem.R
\name{sem_apa}
\alias{sem_apa}
\title{SEM direct effect interpretation}
\usage{
sem_apa(data, md = FALSE)
}
\arguments{
\item{data}{DE_effect obj}

\item{md}{markdown_table output}
}
\value{
report apa
}
\description{
SEM direct effect interpretation
}
\examples{

\dontrun{

 #library(lavaan) # only needed once per session
 models <- '
   # measurement model
     ind60 =~ x1 + x2 + x3
     dem60 =~ y1 + y2 + y3 + y4
     dem65 =~ y5 + y6 + y7 + y8
 # regressions
     dem60 ~ ind60
     dem65 ~ ind60 + dem60
   # residual correlations
     y1 ~~ y5
     y2 ~~ y4 + y6
     y3 ~~ y7
     y4 ~~ y8
    y6 ~~ y8
 '
 fits <- sem(models, data=PoliticalDemocracy)
 summary(fits, standardized=TRUE)

 ##Direct effect
 DE_effect(fits)
 DE_effect(fits)\%>\% report_sem_de()

}

}
