% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_sem.R
\name{sem_apa_ie}
\alias{sem_apa_ie}
\title{sem_apa_ie}
\usage{
sem_apa_ie(data, msg = "IE", md = FALSE, est = 3, std = 4, sig = 7, p = 6)
}
\arguments{
\item{data}{data}

\item{msg}{msg="IE"}

\item{md}{md = FALSE}

\item{est}{est=3}

\item{std}{std=4}

\item{sig}{sig = 7}

\item{p}{p=6}
}
\value{
ie report
}
\description{
sem_apa_ie
}
\examples{

\dontrun{
example(sem)
IE_effect(fit)
IE_effect(fit) \%>\% sem_apa_ie()
}
}
