% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_sem.R
\name{sem_apa_ie_CI}
\alias{sem_apa_ie_CI}
\title{sem_apa_ie_CI}
\usage{
sem_apa_ie_CI(
  data,
  caption = "매개효과 분석 결과 ",
  md = FALSE,
  est = 2,
  std = 3,
  sig = 6,
  CI = 7,
  Z = FALSE,
  tot_filter = "총",
  ind_filter = "IE"
)
}
\arguments{
\item{data}{lavaan object}

\item{caption}{"매개효과 분석 결과",}

\item{md}{markdown}

\item{est}{est sel 2}

\item{std}{std sel 3}

\item{sig}{sig sel 6}

\item{CI}{CI seledct 7}

\item{Z}{Z=FALSE}

\item{tot_filter}{tot_filter="총",}

\item{ind_filter}{ind_filter="IE"}
}
\value{
table
}
\description{
sem_apa_ie_CI
}
