% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_count_df.R
\name{sem_count_df}
\alias{sem_count_df}
\title{Functions for calculating degrees of freedom in SEM}
\usage{
sem_count_df(
  model,
  type = "fit",
  growth = FALSE,
  sample.nobs = 100,
  whatLabels = "model",
  sig = FALSE,
  plot = FALSE,
  rotation = 2
)
}
\arguments{
\item{model}{lavaan syntax}

\item{type}{fit, all}

\item{growth}{grouwth =FALSE}

\item{sample.nobs}{100(default)}

\item{whatLabels}{model, If it is an actual measurement, est}

\item{sig}{FALSE not outpur significant}

\item{plot}{plot FALSE}

\item{rotation}{rotation 4}
}
\value{
table and plot
}
\description{
Functions for calculating degrees of freedom in SEM
}
\examples{

\dontrun{
#'
m1 <- "
y1 =~ x1 + x2 +x5
y2 =~ x3 + x4 +x6
y1 ~~ y2"
sem_count_df(m1)
sem_count_df(m1,"fit")

#그래프를 보고 싶을 때
diagram_model(m1)
#이함수에서 보려면
sem_count_df(m1, plot=TRUE)
#전체를 보여주는 함수
sem_count_df(m1,"all")
sem_count_df(m1,"all", plot=TRUE )
#'
#'
m1a <- "
y1 =~ x1 + x2 +x5
y2 =~ x3 + x4 +x6
y1 ~~ 0*y2"
sem_count_df(m1a)
sem_count_df(m1a,"all")

#이함수에서 보려면
sem_count_df(m1, plot=TRUE)

#simulation data lavaan
sem_count_df(m1,"lav")

}
}
