% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_effect.R
\name{sem_effect}
\alias{sem_effect}
\title{SEM direct or indirect  effect result}
\usage{
sem_effect(model_sem, effect = "~", effect2 = NULL, ci = FALSE)
}
\arguments{
\item{model_sem}{lavaan object}

\item{effect}{select op}

\item{effect2}{select op 2nd option}
}
\value{
repot
}
\description{
SEM direct or indirect  effect result
}
\examples{
\dontrun{

 #library(lavaan) # only needed once per session

models1 <- '
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
    # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
      # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
'
models2 ='
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
     # regressions
    dem60 ~ a1*ind60
    dem65 ~ a2*ind60 + a3*dem60
 #NEw
   ie := a1*a3
   te := a2 +a1*a3
'
fits1 <- sem(models1, data=PoliticalDemocracy)
fits2 <- sem(models2, data=PoliticalDemocracy)
summary(fits1, standardized=TRUE)
summary(fits2, standardized=TRUE)

parameterEstimates(fits1, ci=F, stand=T)
parameterEstimates(fits2, ci=F, stand=T)
sem_effect(fits1)
sem_effect(fits1, effect="=~")
sem_effect(fits1, effect="=~", effect2="~")
sem_effect(fits1, effect="~~")
sem_effect(fits1, effect=":=")

sem_effect(fits2)
sem_effect(fits2, effect="=~")
sem_effect(fits2, effect="~~")
sem_effect(fits2, effect=":=")
sem_effect(fits2, effect=":=", effect2="~")
 sem_effect(fits)\%>\% interpretation_de()
 # same method
 sem_effect(fits)\%>\% report_sem_de()

}


}
