% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_effect.R
\name{sem_effect_ci}
\alias{sem_effect_ci}
\title{SEM indirect effect}
\usage{
sem_effect_ci(
  model_sem,
  effect1 = "DE",
  effect2 = "IE",
  effect3 = "TE",
  label = FALSE,
  ci = TRUE
)
}
\arguments{
\item{model_sem}{lavaan objgect}

\item{effect1}{"DE"}

\item{effect2}{"IE"}

\item{effect3}{"TE"}

\item{label}{label}

\item{ci}{confidence interval}
}
\value{
parameterEstimates()
}
\description{
SEM indirect effect
}
\examples{

\dontrun{
#library(lavaan) # only needed once per session

models1 <- '
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
    # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
      # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8

fits1 <- sem(models1, data=PoliticalDemocracy)


models2 ='
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
     # regressions
    dem60 ~ a1*ind60
    dem65 ~ a2*ind60 + a3*dem60
 #NEw
   ie := a1*a3
   te := a2 +a1*a3
   diff := a1 - a3

fits2 <- sem(models2, data=PoliticalDemocracy)

summary(fits1, standardized=TRUE)
summary(fits2, standardized=TRUE)
fits2 \%>\% sem_effect_ie()
}

}
