% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_plot.R
\name{semplot}
\alias{semplot}
\title{Fast sem plot draw}
\usage{
semplot(data, layout = "tree3", opt = 1, mar = c(1, 5, 1, 5))
}
\arguments{
\item{data}{lavaaan object}

\item{layout}{layY(opt =2, you make layout), or tree3}

\item{opt}{opt =1 tree3}

\item{mar}{margin}
}
\value{
plot
}
\description{
Fast sem plot draw
}
\examples{

\dontrun{

# library(lavaan)
#import data
example(sem)

semplot(fit)

example(cfa)
semplot(fit)


}


}
