% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro_test_df.R
\name{shapiro_test_all}
\alias{shapiro_test_all}
\title{Custom function to perform Shapiro-Wilk test on all columns of a data frame}
\usage{
shapiro_test_all(data, digits = 3)
}
\arguments{
\item{data}{data.frame}

\item{digits}{default 3}
}
\value{
nomalitytest result
}
\description{
Custom function to perform Shapiro-Wilk test on all columns of a data frame
}
\examples{
\dontrun{
data <- mtcars[,c("mpg","disp","hp","drat","wt","qsec")]

## Shapiro-Wilke Test Results for All Columns
shapiro_test_all(data)

}

}
