% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro_test_df.R
\name{shapiro_test_df}
\alias{shapiro_test_df}
\title{Calculate Shapiro-Wilke test results for all columns}
\usage{
shapiro_test_df(data, digits = 3)
}
\arguments{
\item{data}{data.frame}

\item{digits}{default 3}
}
\value{
nomalitytest result
}
\description{
Calculate Shapiro-Wilke test results for all columns
}
\examples{
\dontrun{
data <- mtcars[,c("mpg","disp","hp","drat","wt","qsec")]

## Shapiro-Wilke Test Results for All Columns
shapiro_test_df(data)

}
}
