% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_cov.R
\name{simdata_cov}
\alias{simdata_cov}
\title{Regression with covariance, generate structural equation data restoration}
\usage{
simdata_cov(
  covmat = NULL,
  n = 300,
  dv_mean = 0,
  iv_means = rep(0, ncol(covmat) - 1),
  means = NULL,
  cormat = NULL,
  sds = rep(1, ncol(cormat)),
  seed = NULL
)
}
\arguments{
\item{covmat}{cov matrix}

\item{n}{sample size default = 300}

\item{dv_mean}{dv mean}

\item{iv_means}{iv means}

\item{means}{total means default null}

\item{cormat}{correlation matrix}

\item{sds}{sds is need cor to cov}

\item{seed}{set seed today Sys.Data}
}
\value{
sim data
}
\description{
Regression with covariance, generate structural equation data restoration
}
\examples{

\dontrun{
#'
#simdata mean setting 0
#set.seed(todaty)
cov(mtcars[,c("mpg","hp","wt")]) \%>\%cov_simdata(n=32)
cov(mtcars[,c("mpg","hp","wt")]) \%>\%cov_simdata(n=32, means=c(0,0,0))

cov(mtcars[,c("mpg","hp","wt")]) \%>\%
  cov_simdata(n=32, means=c(20.09, 146.69 ,3.22))

cov(mtcars[,c("mpg","hp","wt")]) \%>\%
  cov_simdata(n=32, dv_mean=20.09, iv_means=c(146.69 ,3.22))

#simdata
cov(mtcars[,c("mpg","hp","wt")]) \%>\%
  cov_simdata(n=32) \%>\%
  lm(formula = mpg ~ hp + wt)\%>\%summary()

#revised mean data Closer to real-world data
cov(mtcars[,c("mpg","hp","wt")]) \%>\%
  cov_simdata(n=32, dv_mean=20.09, iv_means=c(146.69 ,3.22)) \%>\%
  lm(formula = mpg ~ hp + wt)\%>\%summary()


#real data
lm(formula = mpg ~ hp +wt,data=mtcars)\%>\%summary()



# define sample statistics
x <- MASS::Cars93[,c("EngineSize", "Horsepower", "RPM")]
covx <- cov(x)
covx <- matrix(c(1.076122, 39.776999, -339.163745,
                 39.776999, 2743.079, 1146.634,
                 -339.163745, 1146.634, 356088.7097), nrow = 3,
               dimnames = list(c("EngineSize", "Horsepower", "RPM"),
                               c("EngineSize", "Horsepower", "RPM")))
covx
# 함수 실행
sim_data <- simdata_cov(covmat = covx, n = 100)
head(sim_data)
# check
sim_data <- simdata_cov(covmat = covx, n = 100)
lm( EngineSize~ Horsepower+RPM, data = sim_data) \%>\% summary()
#real check
lm( EngineSize ~ Horsepower +RPM , data=x)\%>\% summary()



#'
}
}
