% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_bar.R
\name{simple_bar}
\alias{simple_bar}
\title{Tsimple_bar}
\usage{
simple_bar(
  data,
  add_y = 10,
  color = 2:17,
  vjust = 1,
  cex = 1,
  line = FALSE,
  main = NULL,
  xlab = "x",
  ylab = "y"
)
}
\arguments{
\item{data}{table}

\item{add_y}{max 30}

\item{color}{color=1:16}

\item{vjust}{vjust =10}

\item{cex}{cex=1}

\item{line}{T/F}

\item{main}{main sistitle}

\item{xlab}{xlab}

\item{ylab}{ylab}
}
\value{
bar graph
}
\description{
Tsimple_bar
}
\examples{
\dontrun{
table(mtcars$cyl) \%>\% simple_bar(cex=1.5)

data(bioChemists, package = "pscl")
bioChemists <- bioChemists \%>\%
rename(
  Num_Article = art, #articles in last 3 years of PhD
   Sex = fem, #coded 1 if female
   Married = mar, #coded 1 if married
   Num_Kid5 = kid5, #number of childeren under age 6
   PhD_Quality = phd, #prestige of PhD program
   Num_MentArticle = ment #articles by mentor in last 3 years
   )
  table(bioChemists$Num_Article) \%>\%
      simple_bar(cex=1.5, add_y= 30, vjust=10)

}


}
