% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobel_test.R
\name{sobel_test_general}
\alias{sobel_test_general}
\title{Generalized Sobel Test function}
\usage{
sobel_test_general(a, SE, digits = 5)
}
\arguments{
\item{a}{est}

\item{SE}{se}

\item{digits}{5}
}
\value{
result
}
\description{
Generalized Sobel Test function
}
\examples{

\dontrun{
#'

general_Sobel_test(a = c(0.5, 0.3), SE = c(0.1, 0.05))

Sobel_test_2(a=0.5, b=0.3, sa=0.1, sb =0.05)

# Input data
a <- c(0.2, 0.3, 0.4, 0.5)
SE <- c(0.05, 0.04, 0.03, 0.02)

# Calculate results
sobel_test_general(a = c(0.2, 0.3, 0.4, 0.5), SE = c(0.05, 0.04, 0.03, 0.02))
aroian_sobel_test_general(a = c(0.2, 0.3, 0.4, 0.5), SE = c(0.05, 0.04, 0.03, 0.02))
Goodman_sobel_test_general(a = c(0.2, 0.3, 0.4, 0.5), SE = c(0.05, 0.04, 0.03, 0.02))
general_Sobel_test(a = c(0.2, 0.3, 0.4, 0.5), SE = c(0.05, 0.04, 0.03, 0.02))
#'
}


}
