% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_kword.R
\name{split_korean_word}
\alias{split_korean_word}
\title{split_korean_word- Separating vowels and consonants in Korean}
\usage{
split_korean_word(word, paste = TRUE)
}
\arguments{
\item{word}{input word, one or more}

\item{paste}{Whether you want the outputs to be separate or combined into one}
}
\value{
text data
}
\description{
split_korean_word- Separating vowels and consonants in Korean
}
\examples{
\dontrun{
split_korean_word("책")
## [1] "ㅊ,ㅐ,ㄱ"
split_korean_word("채")
## [1] "ㅊ,ㅐ,"
split_korean_word("고양이")
#' ##[1] "ㄱ,ㅗ,,ㅇ,ㅑ,ㅇ,ㅇ,ㅣ,"
split_korean_word("곱창밥")
## [1] "ㄱ,ㅗ,ㅂ,ㅊ,ㅏ,ㅇ,ㅂ,ㅏ,ㅂ"
}

}
