% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_kword.R
\name{split_korean_word_check}
\alias{split_korean_word_check}
\title{Tsplit_korean_word_check is datacheck}
\usage{
split_korean_word_check(worddata, type = "all")
}
\arguments{
\item{worddata}{worddata}

\item{type}{type defaul all, text, df , all is input and output, text is result , df is data.frame}
}
\value{
split word
}
\description{
Tsplit_korean_word_check is datacheck
}
\examples{
\dontrun{
split_korean_word_check("호랑이")
}

}
