% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_kword.R
\name{split_kw_df}
\alias{split_kw_df}
\title{split_kw_df Create a data column by separating the initial, neutral, and final Korean characters.}
\usage{
split_kw_df(df, sel_col = "word", remove = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{sel_col}{select word column}

\item{remove}{remove False is overlab, TRUE is generate new data}
}
\value{
data.frame
}
\description{
split_kw_df Create a data column by separating the initial, neutral, and final Korean characters.
}
\examples{
\dontrun{
wd <- data.frame(
id =  paste0("BY", 1:5),
word = c("토끼", "휘파람", "뻐스", "우리냥","우리낙"),
tone = c("HH","LHH","HL","LHH","LHL" ),
gender =c("F","F","M","M","M"),
age = c(20,60,20,60, 30),
area= c("n","s","n","s","s"),
mix= c(FALSE,FALSE,FALSE,TRUE,TRUE),
stringsAsFactors = FALSE)

wd\%>\%split_kw_df("word")
wd\%>\%split_kw_df()

}

}
