% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_kw_df_match.R
\name{split_kw_df_match}
\alias{split_kw_df_match}
\title{split_kw_df_match, using Goeun Kim, PhD (Seoul National University 2024)}
\usage{
split_kw_df_match(df, sel_col = "word", remove = TRUE, type = "res")
}
\arguments{
\item{df}{data.frame}

\item{sel_col}{word column}

\item{remove}{TRUE is newdata add}

\item{type}{type is res and all}
}
\value{
data.frame
}
\description{
split_kw_df_match, using Goeun Kim, PhD (Seoul National University 2024)
}
\examples{

\dontrun{

#' wd <- data.frame(
id =  paste0("BY", 1:5),
word = c("토끼", "휘파람", "뻐스", "우리냥","우리낙"),
tone = c("HH","LHH","HL","LHH","LHL" ),
gender =c("F","F","M","M","M"),
age = c(20,60,20,60, 30),
area= c("n","s","n","s","s"),
mix= c(FALSE,FALSE,FALSE,TRUE,TRUE),
stringsAsFactors = FALSE)

# Create new data
wd\%>\%split_kw_df_match("word", type="res")

wd\%>\%split_kw_df_match("word", type="all")

wd2 <- data.frame(id =  paste0("BY", 1:5),
word = c("토끼", "휘파람", "뻐스", "우리냥","우리낙망"))

wd2\%>\%split_kw_df_match("word", type="res")

wd2\%>\%split_kw_df_match("word", type="all")


}



}
