% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_word.R
\name{split_word}
\alias{split_word}
\title{Separate words into syllables}
\usage{
split_word(word)
}
\arguments{
\item{word}{Words being typed}
}
\value{
Syllable-separated data
}
\description{
Separate words into syllables
}
\examples{

\dontrun{
split_word("HLL")
split_word("호랑이")

}

}
