% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_cut.R
\name{str_cut}
\alias{str_cut}
\title{str_cut function}
\usage{
str_cut(df, col = "term", sep = "\\\\(")
}
\arguments{
\item{df}{A data frame.}

\item{col}{The name of the column to modify in the data frame. Default is "term".}

\item{sep}{The separator used to cut the string. Default is "\\(".}
}
\value{
A modified data frame.
}
\description{
A function that modifies the values of a specified column in a data frame by cutting the string based on a specific separator (sep).
}
\examples{
\dontrun{
# Example usage
data <- data.frame(
  term = c("AI 학습 진단 및 문제은행 (예: AI 펭톡, EBS 단추)",
           "기타(직접입력)",
           "메타버스/게임화 도구 (예: Zep, 띵커벨, 퀴즈앤, Kahoot! 등)")
)

# Using the function
result <- str_cut(data, "term")
print(result)
}
}
