% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fa.R
\name{summary_fa}
\alias{summary_fa}
\title{FA,PCA result summary}
\usage{
summary_fa(fa_data, digits = 3)
}
\arguments{
\item{fa_data}{fa, factanal, pricipla result}

\item{digits}{roundign defaylt 4}
}
\value{
table
}
\description{
FA,PCA result summary
}
\examples{

library(MVT)

#  Mardia, K.V., Kent, J.T., and Bibby, J.M. (1979).  _Multivariate
#  Analysis_. Academic Press, London.
data(examScor)
factanal(scale(examScor), factors = 2, rotation = "varimax")\%>\% summary_fa()

#
psych::principal(cor(examScor), nfactor = 2) \%>\%summary_fa()

psych::prcomp(cor(examScor), nfactor = 2) \%>\%summary_fa()

factanal(scale(examScor), factors = 2) \%>\%summary_fa()

}
