% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_kmo.R
\name{summary_kmo}
\alias{summary_kmo}
\title{Kaiser-Meyer-Olkin factor adequacy (KMO)}
\usage{
summary_kmo(data, test = "none", cat = TRUE)
}
\arguments{
\item{data}{data, cor data, result data}

\item{test}{none, cor, df}

\item{cat}{TRUE}
}
\value{
result
}
\description{
Kaiser-Meyer-Olkin factor adequacy (KMO)
}
\examples{
\dontrun{
#'
## Example usage
New = data.frame(
  A01 = c(4, 4, 5, 4, 5, 5, 5, 4, 4, 2, 4, 4, 5, 4, 4, 4, 4, 5, 1, 3),
  A02 = c(2, 3, 3, 5, 5, 4, 5, 5, 4, 4, 4, 4, 4, 3, 3, 5, 4, 4, 3, 1),
  A03 = c(3, 5, 4, 5, 4, 4, 4, 4, 4, 4, 5, 4, 5, 4, 4, 4, 4, 5, 3, 3),
  A04 = c(4, 4, 4, 4, 5, 4, 5, 4, 4, 3, 5, 4, 5, 5, 5, 4, 5, 5, 1, 5),
  A05 = c(3, 2, 3, 5, 4, 4, 4, 4, 4, 4, 4, 3, 5, 2, 3, 4, 5, 5, 4, 2)
)
New \%>\%summary_kmo("df")
New \%>\%summary_kmo("df", cat=FALSE)
New \%>\%cor()\%>\% summary_kmo("cor")
New \%>\%cor()\%>\%psych::KMO()\%>\% summary_kmo()
}

}
