% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_text2df.R
\name{t_sep}
\alias{t_sep}
\title{t_sep separate t value and star}
\usage{
t_sep(data)
}
\arguments{
\item{data}{data}
}
\value{
add column
}
\description{
t_sep separate t value and star
}
\examples{

\dontrun{
#'
reg_text = "var B se b t
성취감 -.03 .09 -.03 -.33
인정감 .24 .09 .20 2.50
책임감 .30 .08 .23 3.56***
발전성 -.08 .06 -.10 -1.29
감독 .06 .05 .08 1.35
보수 -.14 .07 -.18 -2.01*
대인관계 .06 .06 .06 .89
작업환경 .33 .09 .34 3.82***
안정성 .17 .05 .26 3.42**
성별 -.05 .09 -.03 -.52
연령 -.20 .07 -.22 -3.01**
학력 .05 .07 .04 .70
소득 .00 .04 .01 .12
종사기간 .17 .06 .23 2.71**
기관규모 -.10 .03 -.20 -3.36**"


text2df(reg_text)

text2df(reg_text) \%>\%t_sep()

}

}
