% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_report.R
\name{t_test_report}
\alias{t_test_report}
\title{Advanceed t test report function}
\usage{
t_test_report(
  data,
  iv = NULL,
  dv = NULL,
  type = "all",
  xlab = "independent Variable",
  ylab = "dependent Variable",
  translate = FALSE,
  mu = 0,
  paired = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{data}{data.frame}

\item{iv}{independent variable, if you want to change the name, you can change the setting}

\item{dv}{dependent variable, if you want to change the name, you can change the setting}

\item{type}{type option default is 'all'. and  'var.test', 'var.test.report', 't.test','t.test.report', 'boxplot', 'var.test.full', 't.test.full', 'descriptive' output each result}

\item{xlab}{graph name dependent variable}

\item{ylab}{graph name dependent variable}

\item{translate}{report translate defualt FALSE, TRUE is korean}

\item{mu}{a number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{paired}{a logical indicating whether you want a paired t-test.
var.equal a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{conf.level}{confidence level of the interval.}
}
\description{
Advanceed t test report function
}
\examples{
# mtcars data
\dontrun{
##default analysis
t_test_report(mtcars, iv="vs",dv="mpg")
##$var_test_report

##
##$t_test_result_tibble
## A tibble: 1 × 7
##IV    DV       df t.value   p.value sig   test_type
##<chr> <chr> <dbl>   <dbl>     <dbl> <chr> <chr>
##  1 vs    mpg      30   -4.86 0.0000342 ***   student t.test
##
##$t_test_report
## New namaing xlab and ylab
t_test_report(mtcars, iv="vs",dv="mpg",xlab="transmission", ylab= "males per galon")
}
}
