% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_df_colnames.R
\name{table_df_colnames}
\alias{table_df_colnames}
\title{table_df_colnames}
\usage{
table_df_colnames(df, input = "구분")
}
\arguments{
\item{df}{df}

\item{input}{"구분}
}
\value{
df
}
\description{
table_df_colnames
}
\examples{
\dontrun{


}

}
