% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_sum.R
\name{table_sum}
\alias{table_sum}
\title{Functions to sum parts of a table}
\usage{
table_sum(data, ..., trans = FALSE, sep = "_")
}
\arguments{
\item{data}{table,data.frame}

\item{...}{contrast excon= c(1, 1, -1, -1)}

\item{trans}{data transpose, con must be entered to match the row in the data}

\item{sep}{combine sep ="/'}
}
\value{
table
}
\description{
Functions to sum parts of a table
}
\examples{

\dontrun{
#'
#'
data <- data.frame(T = c(30, 40, 60, 120),
                   S = c(50, 70, 60, 80))
rownames(data) <- LETTERS[1:4]

data = as.matrix(data)


table_sum(data, con= c(1, 1, -1, -1))
table_sum(data, con= c(1, 1, 2, -1))
table_sum(data, con= c(1, 1, 2, 3))
table_sum(data, con= c(0, 1, 2, 1))
table_sum(data, con= c(0, 1, 0, 2))

New = data.frame(
  row.names= c('No_Nicotine_Dependence', 'Nicotine_Dependence'),

  "A" = c(130, 119),
  "B" = c(210, 267),
  "C" = c(43, 91),
  "D" = c(114, 254),
  "E" = c(20, 67)
);New

t(New)\%>\%table_sum(c(1,1,3,2,2))

New \%>\%table_sum(c(1,1))
t(New)
t(New)\%>\%table_sum(c(1,1,1,2,2))

New\%>\%table_sum(c(1,1,3,2,2), trans=TRUE)
New\%>\%table_sum(1,1,3,2,2, trans=TRUE)
New\%>\%table_sum(c(1,1), trans=FALSE)

}


}
