% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_effect_ie.R
\name{TE_effect}
\alias{TE_effect}
\title{Total effect}
\usage{
TE_effect(model_sem, effect1 = "TE", effect2 = "te", label = FALSE, ci = F)
}
\arguments{
\item{model_sem}{lavaan objgect}

\item{effect1}{"IE"}

\item{effect2}{"IE"}

\item{label}{label}

\item{ci}{confidence interval}
}
\value{
parameterEstimates()
}
\description{
Total effect
}
\examples{

\dontrun{
#library(lavaan) # only needed once per session

models1 <- '
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
    # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
      # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8

models2 ='
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
     # regressions
    dem60 ~ a1*ind60
    dem65 ~ a2*ind60 + a3*dem60
 #NEw
   ie := a1*a3
   te := a2 +a1*a3
   diff := a1 - a3




fits1 <- sem(models1, data=PoliticalDemocracy)
fits2 <- sem(models2, data=PoliticalDemocracy)
summary(fits1, standardized=TRUE)
summary(fits2, standardized=TRUE)
fits2 \%>\% TE_effect()
}

}
