% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_df.R
\name{to_long}
\alias{to_long}
\title{longdata transfomation}
\usage{
to_long(
  data,
  cols = NULL,
  names_to = "names",
  rowname = "rows",
  values_to = "vlaues",
  fix = NULL,
  rownames_to_column = FALSE
)
}
\arguments{
\item{data}{widedata}

\item{cols}{colrange}

\item{names_to}{name}

\item{rowname}{defai;t acce}

\item{values_to}{value}

\item{fix}{column fixed (thereofore not using cols )}
}
\value{
longdata
}
\description{
longdata transfomation
}
\examples{
\dontrun{

mtcars[,1:4] \%>\%to_long("car", "value")
mtcars[,1:4] \%>\%to_long("car", "value", fix="cyl")
mtcars[,1:4] \%>\%to_long("car", "value", fix= 1:2)
mtcars[,1:4] \%>\%to_long("car", "value", fix= c(1))
 mtcars[,1:4] \%>\%to_long("car", "value", fix= c(2))
mtcars[,1:4] \%>\%to_long("car", "value", fix=c("cyl","mpg"))
 mtcars[,1:6] \%>\%long_df("car", "value", rownames_to_column = F, cols = c(1,3:6))






}
}
