% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_table.R
\name{to_table}
\alias{to_table}
\title{data.frame to table}
\usage{
to_table(data, sel = ncol(data) - 1, value = ncol(data), type = "mat")
}
\arguments{
\item{data}{data.frame}

\item{sel}{select column (names_to)}

\item{value}{values_to}

\item{type}{mat(matrix), df(data.frame)}
}
\value{
table
}
\description{
data.frame to table
}
\examples{
\dontrun{
data <- data.frame(am = c(0, 1, 0, 1),
                   vs = c(0, 0, 1, 1),
                   Freq = c(12, 6, 7, 7))
data
#make  table
to_table(data)

mtcars \%>\%select(am, vs) \%>\% table()

mtcars \%>\%select(am, vs) \%>\% table() \%>\%data.frame()

mtcars \%>\%select(am, vs) \%>\% table() \%>\%data.frame() \%>\%to_table()

mtcars \%>\%select(am, vs) \%>\% table() \%>\%data.frame() \%>\%to_table()\%>\%addmargins()

mtcars \%>\%select(am, vs) \%>\% table() \%>\%data.frame() \%>\%to_table()\%>\%add_ratio()

mtcars \%>\%select(am, vs) \%>\% table() \%>\%data.frame() \%>\%to_table()\%>\%add_ratio()\%>\%as.matrix()\%>\%addmargins()

mtcars \%>\%select(am, vs) \%>\% table() \%>\%data.frame() \%>\%to_table()\%>\%add_ratio_df()

mtcars \%>\%select(am, vs) \%>\% table() \%>\%data.frame() \%>\%to_table()\%>\%add_sum()

mtcars \%>\%select(am, vs) \%>\% table() \%>\%data.frame() \%>\%to_table(sel="vs")

mtcars \%>\%select(am, cyl) \%>\% table()

mtcars \%>\%select(am, cyl) \%>\% table() \%>\%data.frame()

mtcars \%>\%select(am, cyl) \%>\% table() \%>\%data.frame() \%>\%to_table()

mtcars \%>\%select(am, cyl) \%>\% table() \%>\%data.frame() \%>\%to_table()\%>\%add_ratio()

mtcars \%>\%select(am, cyl) \%>\% table() \%>\%data.frame() \%>\%to_table()\%>\%add_ratio_df()

mtcars \%>\%select(cyl, am) \%>\% table() \%>\%data.frame() \%>\%to_table() \%>\%add_sum()


# mean data to table
aggregate(len ~ ., data = ToothGrowth, mean) \%>\% xtabs(formula = len ~.)
#this is jjstat method
aggregate(len ~ ., data = ToothGrowth, mean) \%>\% to_table()

aggregate(len ~ ., data = ToothGrowth, mean) \%>\% to_table("dose","len")

}
}
