% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_df.R
\name{ttest_df}
\alias{ttest_df}
\title{Repeat t.test to create a table}
\usage{
ttest_df(
  data,
  dv_var = NULL,
  iv_vars = NULL,
  grp_mean = TRUE,
  unite_t = FALSE,
  unite_p = FALSE,
  digits = 2,
  var.equal = TRUE,
  paired = FALSE,
  sig = FALSE
)
}
\arguments{
\item{data}{data. fram}

\item{dv_var}{ddpendent variable}

\item{iv_vars}{Independent variables}

\item{grp_mean}{mean each level}

\item{unite_t}{tvalue+sig}

\item{unite_p}{p_vakye+sig}

\item{digits}{digits 2}

\item{var.equal}{variance equal -student ttest}

\item{paired}{paired t.test (TRUE)}

\item{sig}{star}
}
\value{
t.test table
}
\description{
Repeat t.test to create a table
}
\examples{

\dontrun{

ttest_df(data = mtcars, dv_var = "mpg", iv_vars = c("am", "vs"))

ttest_df(data = mtcars, dv_var = "mpg", iv_vars = c("am", "vs"), sig=T)


ttest_df(data = mtcars, dv_var = "mpg", iv_vars = c("am", "vs"), sig=T, unite_t=TRUE)

ttest_df(data = mtcars, dv_var = "mpg", iv_vars = c("am", "vs"), sig=T, unite_p=TRUE)

ttest_df(data = mtcars, dv_var = "mpg", iv_vars = c("am", "vs"), grp_mean=FALSE)


}


}
