% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unCombind_data.R
\name{unCombind_data}
\alias{unCombind_data}
\title{Functions to uncombind combind_data again}
\usage{
unCombind_data(
  data,
  into = c("value", "sig"),
  sep = " ",
  pattern = NULL,
  imp = ""
)
}
\arguments{
\item{data}{conbind data}

\item{into}{column name}

\item{sep}{separate sep= " "}

\item{pattern}{remove tail}

\item{imp}{imputation tail}
}
\value{
results  two data
}
\description{
Functions to uncombind combind_data again
}
\examples{
\dontrun{

Generate example data2
combined_data2 <- data.frame(
  "H" = c("1.0587162 ***", "1.1269248 ***", "0.9163776 ***", "0.9530237 ***"),
  "H(H)" = c("1.1477347 ***", "0.5743712 ***", "1.4060606 ***", "0.8624314 ***"),
  "L" = c("0.50877193 ***", "1.94957983 *", "0.03515152 ***", "1.44531611 **")
)
# Generate example data2
combined_data <- data.frame(
  "H" = c("1.0587162(***)", "1.1269248(***)", "0.9163776(***)", "0.9530237(***)"),
  "H(H)" = c("1.1477347(***)", "0.5743712(***)", "1.4060606(***)", "0.8624314(***)"),
  "L" = c("0.50877193(***)", "1.94957983(*)", "0.03515152(***)", "1.44531611(**)")
)

combined_data2
unCombind(combined_data2,sep=" ")

combined_data
unCombind(combined_data, sep="\\\\(", pattern=")")


}
}
