% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unCombine_data.R
\name{unCombine}
\alias{unCombine}
\title{Functions to uncombind combind_data again}
\usage{
unCombine(
  data,
  into = c("value", "sig"),
  sep = " ",
  type = "all",
  pattern = NULL,
  imp = ""
)
}
\arguments{
\item{data}{conbind data}

\item{into}{column name}

\item{sep}{separate sep= " "}

\item{type}{"all", res1 = m1, res2=m2}

\item{pattern}{remove tail}

\item{imp}{imputation tail}
}
\value{
results  two data
}
\description{
Functions to uncombind combind_data again
}
\examples{
\dontrun{

Generate example data2
combined_data2 <- data.frame(
  "H" = c("1.0587162 ***", "1.1269248 ***", "0.9163776 ***", "0.9530237 ***"),
  "H(H)" = c("1.1477347 ***", "0.5743712 ***", "1.4060606 ***", "0.8624314 ***"),
  "L" = c("0.50877193 ***", "1.94957983 *", "0.03515152 ***", "1.44531611 **")
)
# Generate example data2
combined_data <- data.frame(
  "H" = c("1.0587162(***)", "1.1269248(***)", "0.9163776(***)", "0.9530237(***)"),
  "H(H)" = c("1.1477347(***)", "0.5743712(***)", "1.4060606(***)", "0.8624314(***)"),
  "L" = c("0.50877193(***)", "1.94957983(*)", "0.03515152(***)", "1.44531611(**)")
)

combined_data2
unCombind(combined_data2,sep=" ")

combined_data
unCombind(combined_data, sep="\\\\(", pattern=")")

#iris cor
iris_cor = iris \%>\% select(-Species) \%>\%psych::corr.test()
iris_cor$r
iris_cor$p

combine_data(iris_cor$r\%>\%round(2),
             iris_cor$p\%>\%round(5),
             left="(", right=")")
 sig = add_significance_symbols(iris_cor$p)
 sig

iris_cor_combine =combine_data( round(iris_cor$r, 3) , sig)
iris_cor_combine

 combine_data( round(iris_cor$r, 3) , sig) \%>\%
 lowerMat(diag = 1, fill="")






}
}
