% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows_freq.R
\name{unCount}
\alias{unCount}
\title{Functions that generate rows with a frequency}
\usage{
unCount(data, type = "res", sel = ncol(data), fix = NULL)
}
\arguments{
\item{data}{data}

\item{type}{res, all}

\item{sel}{frequency column}

\item{fix}{final select column}
}
\value{
data
}
\description{
Functions that generate rows with a frequency
}
