% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_ci.R
\name{unite_ci}
\alias{unite_ci}
\title{Organizing and displaying confidence intervals}
\usage{
unite_ci(
  data,
  col1 = ncol(data) - 1,
  col2 = ncol(data),
  left = "[",
  right = "]",
  digits = 3,
  colname = "95\%CI",
  remove = TRUE
)
}
\arguments{
\item{data}{data}

\item{col1}{lower col}

\item{col2}{upper col}

\item{left}{left Parentheses}

\item{right}{right Parentheses}

\item{digits}{round}

\item{colname}{"95%CI"}

\item{remove}{True to remove existing columns, False to keep existing columns}
}
\value{
data
}
\description{
Organizing and displaying confidence intervals
}
\examples{

\dontrun{

#'
# Sample data to simulate the scenario
data <- data.frame(
  paths = c('자기효능감-C_S1', '자기효능감-C_S2', '자기효능감-C_S3'),
  Original = c(0.883, 0.826, 0.794),
  Mean.Boot = c(0.885, 0.833, 0.8),
  Std.Error = c(0.022, 0.041, 0.039),
  t = c('40.771***', '20.22***', '20.612***'),
  perc.025 = c(0.837, 0.733, 0.725),
  perc.975 = c(0.923, 0.893, 0.872)
)

# Apply the function to the dataframe with remove = TRUE
unite_ci(data, col1 = 6, col2 = 7, remove = TRUE)

# Apply the function to the dataframe with remove = FALSE
unite_ci(data, col1 = 'perc.025', col2 = 'perc.975', remove = FALSE)

}

}
