% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test.R
\name{wald_test}
\alias{wald_test}
\title{wald_test}
\usage{
wald_test(
  b1,
  se1,
  b2,
  se2,
  est1 = NULL,
  est2 = NULL,
  digits = 3,
  type = "res2"
)
}
\arguments{
\item{b1}{fires coef}

\item{se1}{firtse se}

\item{b2}{secoond coef}

\item{se2}{second se}

\item{est1}{est1 name}

\item{est2}{est2 name}

\item{digits}{round}

\item{type}{res1 res2,}
}
\value{
test result
}
\description{
wald_test
}
\examples{
\dontrun{
##Clogg, C. C., Petkova, E., & Haritou, A. (1995).
Statistical methods for comparing regression coefficients between models.
American Journal of Sociology, 100(5), 1261-1293.
##https://blog.naver.com/shoutjoy/223294372137

 #' wald_test(-0.795,-0.795/-1.138, -0.795+0.149,  (-0.795+0.149)/1.344)

wald_test(-0.795,-0.795/-1.138,-0.795+0.149,
          (-0.795+0.149)/1.344,
           est1 = "남성",
           est2 = "여성")

}


wald_test <- function(b1, se1, b2, se2, est1=NULL, est2=NULL){
  library(broom)
  t1 = b1/se1
  t2 = b2/se2

  z = (b1 - b2)/ sqrt(se1^2 + se2^2)
  p = 2*(1-pnorm(abs(z)))

  var1 = paste0(est1,"_est1 = ", b1,", se = ", round(se1, 3), ", t = ", t1,"." )
  var2 = paste0(est2,"_est2 = ", b2,", se = ", round(se2, 3), ", t = ", t2,"." )

  res = broom::tidy(c(z.value = z,  pvalue = p))
  colnames(res)=c("statistics","value")

  res = res\%>\% dplry::mutate(coef=c(var1, var2))
  res
  # list(res, var1, var2)
}


}
