% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test.R
\name{wald_test2}
\alias{wald_test2}
\title{wald_test2 input t value}
\usage{
wald_test2(
  b1,
  t1,
  b2,
  t2,
  est1 = NULL,
  est2 = NULL,
  gender = FALSE,
  digits = 3,
  type = "res1"
)
}
\arguments{
\item{b1}{fires coef}

\item{t1}{firtse t}

\item{b2}{secoond coef}

\item{t2}{second t}

\item{est1}{est1 name}

\item{est2}{est2 name}

\item{gender}{For dichotomous variables with intercepts, the Gender}

\item{digits}{round}

\item{type}{res=res1. res2 = all}
}
\value{
ztest
}
\description{
wald_test2 input t value
}
\examples{
\dontrun{

wald_test2(-0.795,8.245, -0.795+0.149,  (-0.795+0.149)/1.344)

wald_test2(-0.795,-0.795/-1.138,-0.795+0.149,
           (-0.795+0.149)/1.344,est1="남성",est2="여성")


##gender variable
##first method
wald_test2(1.311, 1.628, 1.311-0.366,  -2.732)
##second method
wald_test2(1.311, 1.628, -0.366,  -2.732, gender=T)
wald_test2(1.311, 1.628, -0.366,  -2.732,"male","female", gender=T)

}



}
