% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web}
\alias{web}
\title{Functions that output to a web browser}
\usage{
web(data, digits = 3, font_size = 14, file = "output.html")
}
\arguments{
\item{data}{data}

\item{digits}{default 3}

\item{font_size}{default 14}

\item{file}{file name default "output.html}
}
\value{
web output
}
\description{
Functions that output to a web browser
}
\examples{

\dontrun{

}

}
