% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webr.R
\name{webr}
\alias{webr}
\title{data to web}
\usage{
webr(
  data,
  show = TRUE,
  digits = 3,
  font_size = 18,
  border_width = 1,
  border_color = "#ccc",
  file = "output.html"
)
}
\arguments{
\item{data}{data.frame}

\item{show}{TRUE: web FALSE" console}

\item{digits}{round 3}

\item{font_size}{14}

\item{border_width}{1}

\item{border_color}{"gray"}

\item{file}{file = "output.html"}
}
\value{
web
}
\description{
data to web
}
\examples{

\dontrun{
#'
# Example usage
webr(mtcars)
webr(mtcars, font_size=20)
}

}
