% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_se.R
\name{white_se}
\alias{white_se}
\title{White의 수정된 표준오차를 계산하는 함수 정의}
\usage{
white_se(original_se, n = NULL, k = 2, type = "white")
}
\arguments{
\item{original_se}{original_se}

\item{n}{sample size}

\item{k}{number of regression coefs}

\item{type}{all, white, roburst}
}
\value{
se data
}
\description{
White의 수정된 표준오차를 계산하는 함수 정의
}
\examples{
\dontrun{
#'
# 함수 사용 예시
original_se <- 0.009619  # 예시 표준오차
n <- 100  # 예시 표본 크기
k <- 2  # 예시 회귀 계수 개수 (절편 포함)

# 함수 호출 및 결과 출력
white_se(original_se= 0.009619, n= 100, k=2)
}

}
