% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_df.R
\name{wide_df}
\alias{wide_df}
\title{pivot_wider}
\usage{
wide_df(
  data,
  names_from = "names",
  values_from = "values",
  names_prefix = "",
  names_sep = "_",
  names_sort = FALSE,
  add_id = FALSE
)
}
\arguments{
\item{data}{long format data}

\item{names_from}{names_from}

\item{values_from}{values_from}

\item{names_prefix}{names_prefix +variable}

\item{names_sep}{sep}

\item{names_sort}{colnames sort}

\item{add_id}{default FALSE}
}
\value{
wide data
}
\description{
pivot_wider
}
\examples{

\dontrun{
mtcars[1:3, 1:7]
mtcars[1:3, 1:7] \%>\% long_df()
mtcars[1:3, 1:7] \%>\% long_df(cols= 3:7)
mtcars[1:3, 1:7] \%>\% long_df(fix=1:2)
mtcars[1:3, 1:7] \%>\% long_df(fix= 1:3, rownames_to_column=F)

mtcars[1:3, 1:7] \%>\% long_df() \%>\% mutate(ik = "bc", io = "a")
mtcars[1:3, 1:7] \%>\% long_df() \%>\% mutate(ik = "bc", io = "a")\%>\% wide_df()
mtcars[1:3, 1:7] \%>\% long_df() \%>\%
  mutate(ik = "bc", io = "a")\%>\%
  wide_df(names_from = c("names","ik"))
mtcars[1:3, 1:7] \%>\% long_df() \%>\%
  mutate(ik = rownames, io = "a")\%>\%
  wide_df(names_from = c("ik"))\%>\%pall()

mtcars[1:3, 1:7] \%>\% long_df() \%>\%
  mutate(ik = "bc", io = "a")\%>\%
  wide_df(names_from = 2)

}

}
