% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest.R
\name{ztest}
\alias{ztest}
\title{Z test compare coefficient}
\usage{
ztest(b1, se1, b2, se2, type = "res")
}
\arguments{
\item{b1}{b1 coeff}

\item{se1}{se1 standard error}

\item{b2}{b2 coeff}

\item{se2}{se2 se}

\item{type}{result type res1, res2}
}
\value{
z value and p value
}
\description{
Z test compare coefficient
}
\examples{
\dontrun{
# hp, wt
ztest(b1= -0.032230,se1= 0.00892,   b2=-3.23 , se2= 0.796 )
}


}
