% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest_df.R
\name{ztest_df}
\alias{ztest_df}
\title{Z-test using data.frame(or tibble)}
\usage{
ztest_df(data, row1 = 2, row2 = 3, rname1 = "", rname2 = "", sel = "row")
}
\arguments{
\item{data}{lm() and data.frame}

\item{row1}{row1 = 2}

\item{row2}{row2 = 3}

\item{rname1}{select first row name}

\item{rname2}{select second row name}

\item{sel}{row is number, term rowname}
}
\value{
diff, z, p value
}
\description{
Z-test using data.frame(or tibble)
}
\examples{
\dontrun{
 lm(mpg ~ hp + wt + drat, mtcars) \%>\%
 ztest_df()

# #selecting variable
lm(mpg ~ hp + wt + drat, mtcars) \%>\%
   ztest_df(row1=2, row2=3)
#
# # When other variables are selected and analyzed
 lm(mpg ~ hp + wt + drat, mtcars) \%>\%
   ztest_df(row1=2, row2=4)
#
# # When variable-names are selected
 lm(mpg ~ hp + wt + drat, mtcars) \%>\%
   ztest_df(sel="term", rname1="hp", rname2="wt")
#
 lm(mpg ~ hp + wt + drat, mtcars) \%>\%
   ztest_df(sel="term", rname1="hp", rname2="drat")
 }


}
