% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kge_chisq_table.R
\name{accent_table}
\alias{accent_table}
\title{accent_table}
\usage{
accent_table(data, cols = "a1", rows = "성조", trans = TRUE, type = "res")
}
\arguments{
\item{data}{data.frame}

\item{cols}{v1}

\item{rows}{v2}

\item{trans}{transpose}
}
\value{
cross table df
}
\description{
accent_table
}
