% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_table.R
\name{accent_table}
\alias{accent_table}
\title{accent_table}
\usage{
accent_table(data, Var1 = "a1", Var2 = "accent", trans = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{Var1}{v1}

\item{Var2}{v2}

\item{trans}{transpose}
}
\value{
cross table df
}
\description{
accent_table
}
