% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_normality.R, R/mydescriptive.R
\name{add_normality}
\alias{add_normality}
\title{Add skewness and kurtosis normality checks to descriptive statistics results}
\usage{
add_normality(data, skew = "Skew", kurt = "Kurt")

add_normality(data, skew = "Skew", kurt = "Kurt")
}
\arguments{
\item{data}{mysummarydata}

\item{skew}{select skew}

\item{kurt}{select kurt}
}
\value{
normalitydata

add normality
}
\description{
Add skewness and kurtosis normality checks to descriptive statistics results

add_normality
}
\examples{

\dontrun{

mysummary(mtcars)
mysummary(mtcars)\%>\%add_normality()


}


\dontrun{

jjstat::mysummary(mtcars, colnames(mtcars)) \%>\% add_nomality()

}


}
