% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydescriptive.R
\name{add_normality}
\alias{add_normality}
\title{add_normality}
\usage{
add_normality(data, skew = "Skew", kurt = "Kurt")
}
\arguments{
\item{data}{mysummarydata}
}
\value{
add normality
}
\description{
add_normality
}
\examples{

\dontrun{

jjstat::mysummary(mtcars, colnames(mtcars)) \%>\% add_nomality()

}


}
