% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prop.R
\name{add_prop}
\alias{add_prop}
\title{add_prop}
\usage{
add_prop(df, freq = "Freq")
}
\arguments{
\item{df}{df}

\item{freq}{col}
}
\value{
data
}
\description{
add_prop
}
\examples{
\dontrun{

mtcars$am \%>\% Freq_table() \%>\% add_prop()

}


}
