% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows_freq.R
\name{add_rows_freq}
\alias{add_rows_freq}
\title{Functions that generate rows with a frequency}
\usage{
add_rows_freq(data, type = "res", sel = ncol(data), fix = NULL)
}
\arguments{
\item{data}{data}

\item{type}{res, all}

\item{sel}{frequency column}

\item{fix}{final select column}
}
\value{
data
}
\description{
Functions that generate rows with a frequency
}
\examples{

\dontrun{
dff = data.frame(
  trt = c("r","s","z"),
  con = c("f","m","l"),
  num = c(10,0,2),
  fre = c(4,0,5) )
dff
dff \%>\%add_rows_freq()
dff \%>\%add_rows_freq()
dff \%>\%add_rows_freq(fix=1)
dff \%>\%add_rows_freq("res", fix=1:2)
dff \%>\%add_rows_freq("res", fix="con")
dff \%>\%add_rows_freq("all")
dff \%>\%add_rows_freq("real")
dff \%>\%add_rows_freq("zero")
dff \%>\%add_rows_freq("zero", fix="con")

}


}
