% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sig.R, R/p_star_add.R
\name{add_sig}
\alias{add_sig}
\title{mutate data generate}
\usage{
add_sig(var)

add_sig(var)
}
\arguments{
\item{p.value}{pvalue}
}
\value{
in data

indata
}
\description{
mutate data generate

mutate data generate
}
\examples{

\dontrun{

aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = T, unite_F=TRUE) \%>\%  mutate(STAR= add_sig(p_value))

}

\dontrun{

#' aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = T, unite_F=TRUE) \%>\%  mutate(STAR= add_star(p_value))

}
}
