% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_aov.R
\name{anova_apa}
\alias{anova_apa}
\title{Report aov data (format APA)}
\usage{
anova_apa(aov_post_data, md = FALSE)
}
\arguments{
\item{aov_post_data}{aov_post data}

\item{md}{md = TRUE is markdown outpput}
}
\value{
report data
}
\description{
Report aov data (format APA)
}
\examples{
\dontrun{
Mtcars <- mtcars\%>\%as_trt("cyl")
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post() \%>\% anova_apa()
# This is error
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="contrast") \%>\% anova_apa()

# aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="all")

aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="all") \%>\% anova_apa()
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post(type="res") \%>\% anova_apa()

}

}
