% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_post.R
\name{aov_post}
\alias{aov_post}
\title{ANOVA posthoc and plot, Contrast data}
\usage{
aov_post(aov_data, type = "df", adjust = "none", lsd = FALSE)
}
\arguments{
\item{aov_data}{aov data}

\item{type}{type is all, res, contras, df, raq(plotdata), g(plot), aov, anova}

\item{adjust}{adjust ="none", bonferroni, sidak etc}

\item{lsd}{plosthoc LSD}
}
\value{
contrast and g
}
\description{
ANOVA posthoc and plot, Contrast data
}
\examples{

\dontrun{
Mtcars <- mtcars\%>\%as_trt("cyl")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post()
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="all")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="res")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="df")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="call")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="formula")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="g")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="contrast")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="aov")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="anova")
aov(mpg ~ cyl , data=Mtcars)\%>\% aov_post(type="posthoc", lsd=TRUE)

#report
aov(mpg ~ cyl , data=Mtcars)\%>\%aov_post()\%>\% anova_apa()
}


}
