% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_table.R
\name{aov_table}
\alias{aov_table}
\title{Generate an AOV results table for multiple independent variables}
\usage{
aov_table(
  data,
  dv_var = NULL,
  iv_vars = NULL,
  grp_mean = TRUE,
  unite_p = FALSE,
  unite_F = FALSE,
  digits = 2,
  posthoc = "lsd",
  sig = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{dv_var}{dv}

\item{iv_vars}{c(iv1, iv2....)}

\item{unite_p}{pvalue unite}

\item{unite_F}{Fvalue unite}

\item{posthoc}{'tukey, scheffe, duncan}

\item{sig}{'sig = T' is add star}

\item{mean}{Option to average by group and add}
}
\value{
table
}
\description{
Generate an AOV results table for multiple independent variables
}
\examples{

\dontrun{
aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"))

aov_table(data = mtcars, dv_var = "mpg",iv_vars = c("cyl", "gear", "carb"),
           mean = TRUE)

aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = FALSE)

aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = TRUE)

aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = TRUE, unite_F = TRUE)

aov_table(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear", "carb"), grp_mean = TRUE, unite_F = TRUE) \%>\% markdown_table()
}



}
