% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_aov_df.R
\name{apa_aov_df}
\alias{apa_aov_df}
\title{APA style ANOVA summary}
\usage{
apa_aov_df(data, show = TRUE, dv_var = NULL, iv_vars = NULL)
}
\arguments{
\item{data}{A data frame. This can be either the result of `aov_df()` / `aov_table()`,
or raw data to be analyzed if `dv_var` and `iv_vars` are provided.}

\item{show}{Logical. Whether to print the ANOVA table and APA-style summaries to the console. Default is TRUE.}

\item{dv_var}{A string. The name of the dependent variable (DV) to be used in the ANOVA.}

\item{iv_vars}{A character vector. The names of the independent variables (IVs) to be included in the ANOVA.}
}
\value{
Prints APA-style one-way ANOVA summaries for each IV, and (invisibly) returns the input data frame.
}
\description{
APA style ANOVA summary
}
\examples{
\dontrun{
# Example 1: Analyze raw data
apa_aov_df(data = mtcars, dv_var = "mpg", iv_vars = c("cyl", "gear"))

# Example 2: Pipe analysis results into APA report
mtcars \%>\%
  aov_df(dv_var = "mpg", iv_vars = c("cyl", "gear")) \%>\%
  apa_aov_df()




}
}
