% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_as_trt.R
\name{as_trt}
\alias{as_trt}
\title{Treat selected variables as factors}
\usage{
as_trt(data, ..., fun = as.factor)
}
\arguments{
\item{data}{data.frame}

\item{...}{variables}

\item{fun}{funtion default as.factor}
}
\value{
data treatment
}
\description{
Treat selected variables as factors
}
\examples{
\dontrun{
 str(mrcars)

 mtcars1 = mtcars \%>\% sel_as_trt("cyl", "am","vs", "gear", "carb")
 str(mtcars1)
}

}
