% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_add_stat.R
\name{bind_add_stat}
\alias{bind_add_stat}
\title{Functions that generate multiple pieces of data}
\usage{
bind_add_stat(data, ..., fun = mean)
}
\arguments{
\item{data}{data.fram}

\item{...}{colname, cols}

\item{fun}{fun  mean sd}
}
\value{
data.fram
}
\description{
Functions that generate multiple pieces of data
}
\examples{
\dontrun{


mtcars \%>\% mutate(
Mean_col1 = (mpg + cyl)/2 ,
Mean_col2 = (disp + hp)/2 ,
Mean_col3 = (drat + wt )/2 ,
Mean_col4 = (qsec + vs + am)/3 )


bind_add_stat(mtcars,
c("Mean_col1","mpg", "cyl"),
c("Mean_col2", "disp", "hp"),
c("Mean_col3", "drat", "wt"),
c("Mean_col4", "qsec", "vs", "am") )



bind_add_stat(mtcars,
   c("Mean_col1", 1:2),
   c("Mean_col2", 3:4),
   c("Mean_col3", 5:6),
   c("Mean_col4", 7:9))

}

}
