% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_mysummaryBy.R
\name{bind_mysummaryBy}
\alias{bind_mysummaryBy}
\title{bind_mysummaryBy is group mean of mean}
\usage{
bind_mysummaryBy(data, ..., unite = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{...}{fomula data ...}

\item{unite}{... bind persional data and group data  TRUE}
}
\value{
Use group data obtained from group descriptive statistics to obtain group-level descriptive statistics (means, standard deviations, etc.) and to analyze a mixed model
}
\description{
bind_mysummaryBy is group mean of mean
}
\examples{
\dontrun{
##
bind_mysummaryBy(mtcars, mpg ~ am)

##compare
mysummaryBy(data=mtcars, mpg ~ am)

group_by(mtcars, am) \%>\%  summarise(mpg = mean(mpg))

 ## A tibble: 2 × 2
 ##    am   mpg
 ##<dbl> <dbl>
 ## 1     0  17.1
 ##2     1  24.4




bind_mysummaryBy(mtcars, mpg ~ am, mpg ~ vs)


## grp   dv        N  MEAN    SD   MIN   MAX
##  <chr> <chr> <int> <dbl> <dbl> <dbl> <dbl>
##1 am    mpg       2  20.8  5.12  17.1  24.4
##2 vs    mpg       2  20.6  5.61  16.6  24.6
#

bind_mysummaryBy(mtcars, mpg ~ am, mpg ~ vs, mpg ~cyl)

# A tibble: 3 × 9
# grp   dv        N  MEAN    SD   MIN   MAX      Skew  Kurt
# <chr> <chr> <int> <dbl> <dbl> <dbl> <dbl>     <dbl> <dbl>
#   1 am    mpg       2  20.8  5.12  17.1  24.4 -5.28e-16 -2.75
# 2 vs    mpg       2  20.6  5.61  16.6  24.6  0        -2.75
# 3 cyl   mpg       3  20.5  5.82  15.1  26.7  1.28e- 1 -2.33


jjstat::bind_mysummaryBy(mtcars, mpg ~ am, mpg ~ vs, mpg ~ cyl, unite=TRUE)

# A tibble: 3 × 8
# var         N  MEAN    SD   MIN   MAX      Skew  Kurt
# <chr>   <int> <dbl> <dbl> <dbl> <dbl>     <dbl> <dbl>
#   1 am_mpg      2  20.8  5.12  17.1  24.4 -5.28e-16 -2.75
# 2 vs_mpg      2  20.6  5.61  16.6  24.6  0        -2.75
# 3 cyl_mpg     3  20.5  5.82  15.1  26.7  1.28e- 1 -2.33




}
}
