% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Char2num.R
\name{Char2num}
\alias{Char2num}
\title{Functions to convert character to numeric}
\usage{
Char2num(data, auto = TRUE, iter = 1:ncol(data), remove = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{auto}{If TRUE, converts all variables to}

\item{iter}{Select the range of variables to convert}
}
\description{
Functions to convert character to numeric
}
\examples{

\dontrun{

## Replacing a character variable with numeric
mysummary(sats, "sux1") \%>\%
    mutate(across(everything(), ~ format(., digits = 4))) \%>\%
    tibble() \%>\% Char2num()


## Set remove=TRUE if you only want to process the selected variable and output only that.
  mysummary(sats, "sux1") \%>\%
  mutate(across(everything(), ~ format(., digits = 5))) \%>\%
    tibble() \%>\%Char2num(auto=FALSE, iter= 4:ncol(.), remove= TRUE)

##
mysummary(sats, "sux1") \%>\%
  mutate(across(everything(), ~ format(., digits = 5))) \%>\%
    tibble() \%>\%Char2num(auto=FALSE, iter= 4:ncol(.), remove= FALSE)
## same
  mysummary(sats, "sux1") \%>\%
     mutate(across(everything(), ~ format(., digits = 5))) \%>\%
     tibble() \%>\%Char2num(auto=FALSE, iter= 4:ncol(.))


}


}
