% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_gof_posthoc.R
\name{chisq_each}
\alias{chisq_each}
\title{chisq_gof_posthoc}
\usage{
chisq_each(Xs, type = "all", method = "fdr")
}
\arguments{
\item{Xs}{data}

\item{type}{all, res, p, chisq}

\item{method}{p.adust}
}
\value{
data
}
\description{
chisq_gof_posthoc
}
\examples{

\dontrun{
#'
#'
X <- c(49, 30, 63, 59)

chisq_each(X, "res")
chisq_each(X,"all")$gof_test
chisq_each(X,"all")$p
chisq_each(X,"all")$chisq
chisq_each(X, typ="all")
chisq_each(X, typ="res")
chisq_each(X, type="p")
chisq_each(X, type="chisq")

chisq_each(X, method="bonferroni")

X1 <- c(4, 3, 6, 19)
chisq_each(X1)
chisq_each(X1, typ="all")
chisq_each(X1, type="p")
chisq_each(X1, type="chisq")
chisq_each(X1, method="bonferroni")


chisq_each(X, method="bonferroni")

chisq_each(c(49, 30, 63, 59, 40, 60))

#'
}



}
