% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chisq_retest.R
\name{Chisq_retest}
\alias{Chisq_retest}
\title{Functions to check the chi-square test of a paper}
\usage{
Chisq_retest(
  input,
  ncol = 2,
  colname = NULL,
  rowname = NULL,
  simulate.p.value = FALSE,
  digits = 3,
  type = "all",
  correct = TRUE,
  B = 2000
)
}
\arguments{
\item{input}{vectro}

\item{ncol}{matrix col}

\item{colname}{input}

\item{rowname}{input}

\item{simulate.p.value}{default FaLSE fisher exact test}

\item{digits}{round 3}

\item{type}{all, chisq}

\item{correct}{default is Pearson’s Chi-squared test with Yates’ continuity correction (If you have something with a small sample count, you should use it. ), FALSE Pearson's Chi-squared test}

\item{B}{B is bootstrap default 2000}
}
\value{
chisquare data
}
\description{
Functions to check the chi-square test of a paper
}
\examples{

\dontrun{
##Mina Choi. (2007). Analysis of learning effects according to online and offline learning environments in the same class by the same instructor. Educational Information Media Research, 13(4>  5-23.
## jjpaper
## https://blog.naver.com/shoutjoy/223005233887

Chisq_retest(c(111, 33, 64, 58), ncol=2 )
## 논문에 나온 값이 잘못됨 :chisq = 17.793
## this chisq = 16.7

Chisq_retest(c(111, 33, 64, 58), ncol=2, type="chisq" )

Chisq_retest(c(111, 33, 64, 58), ncol=2,
colname = c("experience", "Non-experience"),
rowname =c("E-Leanring","FaceToFace"))

Chisq_retest(c(4,10, 78, 46, 6, 0,5,79,36,2), ncol=2 )

Chisq_retest(c(4,10, 78, 46, 6, 0,5,79,36,2), ncol=2,
colname = c("experience", "Non-experience"),
rowname = c("maxlow","low","mid","high","maxhigh") )


##  Comparison of gender learning methods --> An invalid value was calculated.
## chisq = 5.238, p =0.015 But recalculate chisq = 4.69, p = 0.0303
Chisq_retest(c(84, 60, 54, 68), ncol=2 ,
             colname = c("experience", "Non-experience"),
                          rowname = c("male", "female"))

## Comparing Learner Characteristics by Grade Level, Mina Choi (2007)

Chisq_retest(c(18,38, 40, 48, 52,35,19,16), ncol=2 ,
colname = c("experience", "Non-experience"),
rowname = c("1학년","2학년","3학년","4학년")


}


}
