% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test_kge.R
\name{chisq_test_kge}
\alias{chisq_test_kge}
\title{kge_chisq_table chisq table observed/Expected table-}
\usage{
chisq_test_kge(
  dataset,
  v1 = "var1",
  v2 = "var2",
  title = "Table",
  type = "result",
  digits = 3,
  yadd = 0.4,
  Ncol = NULL,
  trans = FALSE,
  simple = FALSE,
  ko = TRUE,
  simulate.p.value = FALSE,
  correct = FALSE,
  size_bartext = 5,
  strip_size = 16,
  axis_size = 16,
  text_size = 13,
  xlab = "관측기대비율",
  ylab = "var2",
  cramer = "adjust"
)
}
\arguments{
\item{dataset}{data.frame}

\item{v1}{col}

\item{v2}{row}

\item{title}{title}

\item{type}{res2 default}

\item{digits}{3}

\item{yadd}{up yadd}

\item{Ncol}{grid}

\item{trans}{transpose}

\item{simple}{* style}

\item{ko}{graph korean}

\item{simulate.p.value}{exact}

\item{correct}{exactt}

\item{size_bartext}{graph text}

\item{strip_size}{strip}

\item{axis_size}{axis_size}

\item{text_size}{text_size}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{cramer}{cramer cor}
}
\value{
multiple data
}
\description{
kge_chisq_table chisq table observed/Expected table-
}
\examples{

\dontrun{
kge_chisq_table(mtcars,"am","cyl")


#chisq test result
vote_chi <- Chisq_retest(c(183, 35, 824, 50), ncol=2,
                        colname = c("Did not vote", "Voted"),
                         rowname =c("All other ages","Youngest"))

vote_chi


# generate matrix
voteMat = matrix(c(183, 35, 824, 50), ncol=2)
colnames(voteMat) = c("Did not vote", "Voted")
rownames(voteMat) = c("All other ages","Youngest")

voteMat\%>\% data.frame() \%>\%
  long_df("voted", rowname="youth")\%>\% #long format
  add_rows_freq() \%>\%     # add rows
  kge_chisq_table("voted","youth")    #chisq test


#신진영(2013)
shin2013_mat =matrix(c(14,1,23,3,4,0,9,0,13,18,89,8,11,40,46,12,3,4,22,0), ncol= 5)
# colnames(shin2013_mat)=c("지능지수가높다","학업성취도가높다","창의적사고를한다","특정분야에서뛰어난능력을보인다","모든분야에서뛰어난능력을보인다")
colnames(shin2013_mat)=c("지능지수","학업성취도","창의적사고","특정분야","모든분야")
rownames(shin2013_mat)=c("고졸","초대졸","대졸","대학원이상")
shin2013_mat

shin2013_mat \%>\% data.frame() \%>\%
  rownames_to_column("학력") \%>\%
  to_long(names_to = "영재성",cols=2:6) \%>\%
  unCount() \%>\%
  chisq_test_kge(v1="영재성", v2="학력")


shin2013_mat\%>\%  data.frame() \%>\%
  long_df("학력","영재성") \%>\%
  unCount() \%>\%
  chisq_test_kge("학력","영재성")

}
}
