% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMH_test.R
\name{cmh_test}
\alias{cmh_test}
\title{cochran_mantel_haenszel_test}
\usage{
cmh_test(
  x,
  y = NULL,
  z = NULL,
  alternative = c("two.sided", "less", "greater"),
  correct = TRUE,
  exact = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{x}{array}

\item{y}{NULL}

\item{z}{NULL}

\item{correct}{TRUE}

\item{exact}{FALSE}

\item{conf.level}{0.95}

\item{alternativec("two.sided", }{"less", "greater")}
}
\value{
result
}
\description{
cochran_mantel_haenszel_test
}
\examples{

\dontrun{

CAD <-array(c(8, 4, 10, 11, 21, 6, 9, 9), dim = c(2, 2, 2),
            dimnames = list(
              EKG = c(">=0.1 ST Dep", "< 0.1 ST Dep"),
              Response = c("Case", "Control"),
              Penicillin.Level = c("Female", "Male")))
CAD

#Example: Coronary Artery Disease
mantelhaen.test(CAD,correct=FALSE)

cmh_test(CAD,correct=FALSE)\%>\%tidy()

}
}
