% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_plot_lm.R
\name{confint_plot_lm}
\alias{confint_plot_lm}
\title{lm() confint plot}
\usage{
confint_plot_lm(
  lm_data,
  size_text = 12,
  color = "steelblue",
  linewidth = 1,
  alpha = 1,
  intercept = FALSE,
  type = "g"
)
}
\arguments{
\item{lm_data}{lm(formula. data )}

\item{size_text}{text size 12}

\item{color}{color = "steelblue"}

\item{linewidth}{linewidth =1}

\item{alpha}{alpha = 0.9}

\item{intercept}{default FALSE}

\item{type}{graph is g, data =df , all is res}
}
\value{
plot and data
}
\description{
lm() confint plot
}
\examples{
\dontrun{

lm(mpg ~ hp+ wt+ disp, mtcars) \%>\% confint_plot_lm()

lm(mpg ~ hp+ wt+ disp, mtcars) \%>\% confint_plot_lm(intercept=T)

}

}
