% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_plot_sem.R
\name{confint_plot_sem}
\alias{confint_plot_sem}
\title{lavaan data confint plot}
\usage{
confint_plot_sem(
  lavaan_obj,
  type = "g",
  effect = "~",
  size_axis = 13,
  size_text = 4
)
}
\arguments{
\item{lavaan_obj}{lavaan}

\item{type}{g, res, all}

\item{size_axis}{axis test size}

\item{size_text}{size ext}

\item{op}{op}
}
\value{
plot
}
\description{
lavaan data confint plot
}
\examples{

\dontrun{



}

}
