% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_plot_sem.R
\name{confint_plot_sem}
\alias{confint_plot_sem}
\title{lavaan data confin tplot}
\usage{
confint_plot_sem(
  data,
  y = "lhs",
  x = "est",
  lower = "ci.lower",
  upper = "ci.upper",
  size_text = 12,
  color = "steelblue",
  linewidth = 1,
  alpha = 0.8,
  vlinewidth = 1.2,
  vlinecolor = "red"
)
}
\arguments{
\item{data}{lavaan data parameterEstimates %>% filter(op==":=")}

\item{y}{lhs}

\item{x}{est}

\item{lower}{ci.lower}

\item{upper}{ci.upper}

\item{size_text}{12}

\item{color}{steelblue}

\item{linewidth}{1}

\item{alpha}{0.8}

\item{vlinewidth}{1,2}

\item{vlinecolor}{red}
}
\value{
plot
}
\description{
lavaan data confin tplot
}
\examples{

\dontrun{

# Create a tibble named semdata with the given data

semdata\%>\%confint_plot_sem()

}
}
