% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2cor.R
\name{Cor2cov}
\alias{Cor2cov}
\title{cor2cov}
\usage{
Cor2cov(cor_matrix, sds = rep(1, ncol(cor_matrix)))
}
\arguments{
\item{cor_matrix}{cor}

\item{sds}{sds =c(...), default sds 1}
}
\value{
cov
}
\description{
cor2cov
}
\examples{

\dontrun{
# correlation matrix
mtcars[1:6] \%>\%cor()

#change cor to cov
mtcars[1:6] \%>\%cor() \%>\% cor2cov(sds = mtcars[1:6] \%>\% sapply(sd))

#real cov matrix
mtcars[1:6] \%>\%cov()

}


}
