% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_print.R
\name{cut_print}
\alias{cut_print}
\title{Functions that cutoff a given value}
\usage{
cut_print(
  data,
  cut = 0,
  digits = 3,
  rm_row = FALSE,
  rm_cut = FALSE,
  star = FALSE,
  each = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{cut}{cutoff  default 0}

\item{digits}{round 3}

\item{rm_row}{remove row}

\item{rm_cut}{remove cutoff rows}

\item{star}{over cut marked star pasted}

\item{each}{each col using cut}
}
\value{
data
}
\description{
Functions that cutoff a given value
}
\examples{

\dontrun{
#'
mtcars\%>\% cut_print(cut = 280)
mtcars\%>\% cut_print(cut = 280, rm_row = T)
mtcars\%>\% cut_print(cut = 280, rm_row = T,rm_cut=T)
#'
}
}
