% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_freq.R
\name{Freq_table}
\alias{Freq_table}
\title{Frequency analysis table}
\usage{
Freq_table(
  data,
  ...,
  prop = FALSE,
  plot = FALSE,
  angle = 0,
  size_text = 4,
  size_axis = 10,
  legend.position = "",
  reorder = FALSE,
  type = "res"
)
}
\arguments{
\item{data}{data.frame}

\item{...}{vars If you input varabale names "var1", "var2" ,...}

\item{prop}{proportion value, default TRUE}

\item{size_text}{default 3, bargraph text}

\item{size_axis}{default 12, axis test}

\item{legend.position}{legend.position="" none. top, bottom, left, right}

\item{type}{all, res=data, g=plot=graph}
}
\value{
Freqency table
}
\description{
Frequency analysis table
}
\examples{
\dontrun{
Mtcars = mtcars
Mtcars$am = factor(Mtcars$am, levels=c(0,1), labels= c("automatic","manual" ))
Mtcars$vs  = factor(Mtcars$vs, levels=c(0,1), labels= c("V-shaped","straight" ))
Mtcars$cyl  = factor(Mtcars$cyl, levels=c(4,6,8), labels= c("cyl-4","cyl-6","cyl-8"))

#vector variables
Freq_table(c("a","b","c","a","b","a","c","c","c"))
Freq_table(c("a","b","c","a","b","a","c","c","c"), prop=TRUE)
Freq_table(c("a","b","c","a","b","a","c","c","c"), plot=T)


Mtcars \%>\% Freq_table("am")
Mtcars \%>\% Freq_table("am", prop=F)
Mtcars \%>\% Freq_table("vs", plot=FALSE)

Mtcars \%>\% Freq_table("vs")
Mtcars \%>\% Freq_table("vs","am")

Mtcars \%>\% Freq_table("vs","am","cyl", type="all")

#graph
Mtcars \%>\% Freq_table("vs","am","cyl", type="g")

##only graph
Mtcars \%>\% Freq_table("vs","am","cyl", angle=90, type="g")
Mtcars \%>\% Freq_table("vs","am","cyl", angle=90, reorder = TRUE, type="g")

Mtcars \%>\% Freq_table("vs","am","cyl", angle=90, type="g")+ylim(0,14)
Mtcars \%>\% Freq_table("vs","am","cyl", angle=90, reorder = TRUE, type="g")+ylim(0,14)

Mtcars \%>\% Freq_table("vs","am") \%>\% arrange(am)
Mtcars \%>\% Freq_table("vs","am") \%>\% arrange(vs)
Mtcars \%>\% Freq_table("vs","am") \%>\% arrange(Freq)

}



}
