% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq_table.R
\name{Freq_table}
\alias{Freq_table}
\title{Frequency analysis table}
\usage{
Freq_table(
  data,
  ...,
  prop = FALSE,
  plot = FALSE,
  angle = 0,
  size_text = 4,
  size_axis = 10,
  vjust = -0.05,
  hjust = 0.5,
  yadd = 100,
  legend.position = "",
  reorder = FALSE,
  wt = NULL,
  remove_rows = NULL,
  add_row_sum = FALSE,
  xlab = "Category",
  flip = FALSE,
  size_axis_title = 12,
  sort_bar = TRUE,
  type = "res"
)
}
\arguments{
\item{data}{data.frame}

\item{...}{vars If you input varabale names "var1", "var2" ,...}

\item{prop}{proportion value, default TRUE}

\item{plot}{plot defualt false}

\item{angle}{angle = 0}

\item{size_text}{default 3, bargraph text}

\item{size_axis}{default 12, axis test}

\item{vjust}{default -0.5}

\item{hjust}{default -0.5}

\item{yadd}{default 0.8}

\item{legend.position}{legend.position="" none. top, bottom, left, right}

\item{reorder}{reorder Sorting Graph Arrays}

\item{wt}{weight}

\item{remove_rows}{Enter the number of rows to hide and they will be removed}

\item{add_row_sum}{default FALSE, Sum excluding removed rows}

\item{xlab}{xalb = "Category",}

\item{flip}{FALSE}

\item{size_axis_title}{default 12,}

\item{sort_bar}{TRUE(default)}

\item{type}{default  "res",  all, res=data, g=plot=graph}
}
\value{
Freqency table
}
\description{
Frequency analysis table
}
\examples{
\dontrun{
Mtcars = mtcars
Mtcars$am = factor(Mtcars$am, levels=c(0,1), labels= c("automatic","manual" ))
Mtcars$vs  = factor(Mtcars$vs, levels=c(0,1), labels= c("V-shaped","straight" ))
Mtcars$cyl  = factor(Mtcars$cyl, levels=c(4,6,8), labels= c("cyl-4","cyl-6","cyl-8"))

#vector variables
Freq_table(c("a","b","c","a","b","a","c","c","c"))
Freq_table(c("a","b","c","a","b","a","c","c","c"), prop=TRUE)
Freq_table(c("a","b","c","a","b","a","c","c","c"), plot=T)


Mtcars \%>\% Freq_table("am")
Mtcars \%>\% Freq_table("am", prop=F)
Mtcars \%>\% Freq_table("vs", plot=FALSE)

Mtcars \%>\% Freq_table("vs")
Mtcars \%>\% Freq_table("vs","am")

Mtcars \%>\% Freq_table("vs","am","cyl", type="all")

#graph
Mtcars \%>\% Freq_table("vs","am","cyl", type="g")

##only graph
Mtcars \%>\% Freq_table("vs","am","cyl", angle=90, type="g")
Mtcars \%>\% Freq_table("vs","am","cyl", angle=90, reorder = TRUE, type="g")

Mtcars \%>\% Freq_table("vs","am","cyl", angle=90, type="g")+ylim(0,14)
Mtcars \%>\% Freq_table("vs","am","cyl", angle=90, reorder = TRUE, type="g")+ylim(0,14)

Mtcars \%>\% Freq_table("vs","am") \%>\% arrange(am)
Mtcars \%>\% Freq_table("vs","am") \%>\% arrange(vs)
Mtcars \%>\% Freq_table("vs","am") \%>\% arrange(Freq)

##when the data has already been aggregated once
df_sample <- tribble(
  ~name,    ~gender,   ~ sum,
  "Max",    "male",       10,
  "Sandra", "female",      1,
  "Susan",  "female",      4)
df_sample

#weight frequency
df_sample \%>\% unCount(sel = "sum") \%>\% Freq_table("gender")

# this function
df_sample \%>\% Freq_table("gender", wt="sum")
#'
# 데이터 생성
df <- data.frame(
  term = c("", "10-20분", "20-30분", "30분 이상", "5-10분", "5분 이내"),
  Freq = c(728, 222, 77, 120, 319, 135)
)

# 기본 그래프
Freq_table(df, term, prop = TRUE, remove_rows = 1, add_row_sum = FALSE, plot = TRUE)

# xlab과 flip 옵션 사용
Freq_table(df, term, prop = TRUE, remove_rows = 1, add_row_sum = TRUE, xlab = "응답 시간", flip = TRUE, plot = TRUE)
#'

#빈도분석 및 그래프
Freq_table(eduteck2$Q9, prop=TRUE, remove_rows=1, add_row_sum=TRUE, vjust=-0.5,
           plot=T, size_axis = 14, flip=FALSE, yadd=130, xlab="Eduteck Time", size_text = 6) #'
}


}
