% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parceling.R
\name{item_parceling}
\alias{item_parceling}
\title{item_parceling: Mutate-like Item Parceling Function}
\usage{
item_parceling(data, ..., fun = mean, type = "data")
}
\arguments{
\item{data}{A data.frame.}

\item{...}{Named arguments. The name is the new parcel variable,
and the value specifies columns using names, indices, or ranges.}

\item{fun}{A function to apply across selected columns (default = mean).}

\item{type}{Output type:
\describe{
  \item{data}{Return data.frame with parcel variables (default).}
  \item{res}{Same as data.}
  \item{new}{Return only newly created parcel variables.}
  \item{cols_check}{Return columns used for each parcel.}
  \item{all}{Return list(data, cols_check).}
}}
}
\value{
A data.frame or list depending on `type`.
}
\description{
This function creates composite (parcel) variables in a data.frame
similar to `dplyr::mutate()`. The left-hand side specifies the name
of the new parcel variable, and the right-hand side specifies the
columns to be aggregated.
}
\examples{
\dontrun{

############################################################
## Part 1. mtcars examples
############################################################

## 1. Create parcels and return full data
item_parceling(
  mtcars,
  Mean_col1 = c("mpg", "cyl"),
  Mean_col2 = c("disp", "hp")
)


## 2. Return only newly created parcel variables
item_parceling(
  mtcars,
  Mean_col1 = c("mpg", "cyl"),
  Mean_col2 = c("disp", "hp"),
  type = "new"
)


############################################################
## Part 2. stat_onl examples (survey data)
############################################################

## 3. Satisfaction-related parcels
stat_onl <- item_parceling(
  stat_onl,
  Satisfy_Mean = c("satisfy", "On_Satisfy", "On_Joy"),
  Learning_Engage = c("learning:S_Focus_on"),
  fun  = mean,
  type = "data"
)


## 4. Extract only parcel scores for SEM / CFA
parcel_scores <- item_parceling(
  stat_onl,
  Satisfy_Mean = c("satisfy", "On_Satisfy", "On_Joy"),
  Learning_Engage = c("learning:S_Focus_on"),
  Online_Exp = c("On_Easy:Intension_use"),
  type = "new"
)

}
}
