% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kge_chisq_table.R
\name{kge_chisq_table}
\alias{kge_chisq_table}
\title{hisq table observed/Expected table}
\usage{
kge_chisq_table(
  data,
  v1 = "a1",
  v2 = "성조",
  title = "Table",
  type = "res2",
  digits = 3,
  yadd = 0.1,
  ncol = NULL,
  trans = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{v1}{v1}

\item{v2}{v1}

\item{title}{title name}

\item{type}{outtype ct: crosstable, df: data.frame. 'margin','chisq_test','chitable','res1','res2'}

\item{digits}{round 3}

\item{trans}{transpose}

\item{simulate.p.value}{Chi-squared approximation may be incorrect, default= FALSE}

\item{warn}{warning msg
#'}
}
\value{
output
}
\description{
hisq table observed/Expected table
}
