% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kge_chisq_table.R
\name{kge_chisq_table}
\alias{kge_chisq_table}
\title{kge_chisq_table chisq table observed/Expected table-}
\usage{
kge_chisq_table(
  dataset,
  v1 = "a1",
  v2 = "성조",
  title = "Table",
  type = "res2",
  digits = 3,
  yadd = 0.4,
  Ncol = NULL,
  trans = FALSE,
  simple = FALSE,
  ko = TRUE,
  simulate.p.value = FALSE,
  correct = FALSE,
  size_bartext = 5,
  strip_size = 16,
  axis_size = 16,
  text_size = 13,
  xlab = "관측기대비율",
  ylab = "성조형",
  cramer = "adjust"
)
}
\arguments{
\item{dataset}{data.frame}

\item{v1}{col}

\item{v2}{row}

\item{title}{title}

\item{type}{res2 default}

\item{digits}{3}

\item{yadd}{up yadd}

\item{Ncol}{grid}

\item{trans}{transpose}

\item{simple}{* style}

\item{ko}{graph korean}

\item{simulate.p.value}{exact}

\item{correct}{exactt}

\item{size_bartext}{graph text}

\item{strip_size}{strip}

\item{axis_size}{axis_size}

\item{text_size}{text_size}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{cramer}{cramer cor}
}
\value{
multiple data
}
\description{
kge_chisq_table chisq table observed/Expected table-
}
\examples{

\dontrun{
kge_chisq_table(mtcars,"am","cyl")

}
}
