% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_extrac_paths.R
\name{lav_extract_ind_paths}
\alias{lav_extract_ind_paths}
\title{lav_extract_ind_paths}
\usage{
lav_extract_ind_paths(paths_data, allpaths = FALSE)
}
\arguments{
\item{paths_data}{direct path}

\item{allpaths}{all and ind(FALSE)}
}
\value{
text
}
\description{
lav_extract_ind_paths
}
\examples{

\dontrun{
#'
# 예시 데이터
paths_data <- data.frame(
  hypo = c("H1", "H2", "H3", "H4", "H5", "H6", "H7", "H8", "H9", "H10"),
  paths = c("IMAG->EXPE", "EXPE->QUAL", "QUAL->VAL", "EXPE->VAL", "IMAG->SAT", "EXPE->SAT", "QUAL->SAT", "VAL->SAT", "SAT->LOY", "IMAG->LOY")
)
# 함수 실행
lav_extract_ind_paths(paths_data, allpaths = TRUE)
lav_extract_ind_paths(paths_data, allpaths = FALSE)



lav_extract_path(model1a)\%>\%lav_extract_ind_paths()

lav_extract_path(model1)\%>\%lav_extract_ind_paths()
}

}
