% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_df.R
\name{long_df}
\alias{long_df}
\title{longdata transfomation}
\usage{
long_df(
  data,
  rowname = "accent",
  names_to = "speaker",
  values_to = "freq",
  cols = NULL,
  fix = NULL,
  rownames_to_column = TRUE
)
}
\arguments{
\item{data}{widedata}

\item{rowname}{defailt accent}

\item{names_to}{name}

\item{values_to}{value}

\item{cols}{colrange}

\item{fix}{column fixed (thereofore not using cols )}
}
\value{
longdata
}
\description{
longdata transfomation
}
\examples{
\dontrun{

mtcars[,1:4] \%>\%long_df("car", "value")

 mtcars[,1:6] \%>\%long_df("car", "value", rownames_to_column = F, cols = c(1,3:6))

}
}
