% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_bic_plot.R
\name{lpa_BIC_plot}
\alias{lpa_BIC_plot}
\title{BIC Plot for Latent Profile Analysis (LPA)}
\usage{
lpa_BIC_plot(
  data,
  n_profiles_range = 1:6,
  type = "all",
  legend_position = "top",
  covar_model = c("EII", "EEE", "VII", "EEV", "EEI", "VVI", "VVV"),
  set_n_profile = 4,
  size.text = 4,
  flip = TRUE,
  na2zero = FALSE,
  basic = TRUE
)
}
\arguments{
\item{data}{A data frame for LPA modeling.}

\item{n_profiles_range}{A numeric vector specifying the range of profiles to test (default: 1:6).}

\item{type}{A string specifying the output type: "all", "data", "long_data", "wide_data", "graph".}

\item{legend_position}{A string specifying legend position in the plot (default: "top").}

\item{covar_model}{A character vector specifying covariance models (default: c("EII", "EEE", "VII", "EEV", "EEI", "VVI", "VVV")).}

\item{set_n_profile}{A numeric value specifying the selected number of profiles (default: 4).}

\item{size.text}{Numeric, size of text annotations in the plot (default: 4).}

\item{flip}{Logical, whether to flip BIC values for plotting (default: TRUE).}

\item{na2zero}{Logical, whether to replace NA values with zeros in BIC results (default: FALSE).}

\item{basic}{Logical, whether to use a basic set of covariance models (default: TRUE).}
}
\value{
A list containing the wide data, long data, and the generated ggplot object.
}
\description{
This function generates a BIC plot for comparing covariance structures
             across different numbers of profiles in Latent Profile Analysis (LPA).
}
\examples{
\dontrun{
lpa_BIC_plot(edLca, 1:8, basic = FALSE, set_n_profile = 4)
}
}
