% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_bic_plot.R
\name{lpa_BIC_plot2}
\alias{lpa_BIC_plot2}
\title{lpa_BIC_plot2}
\usage{
lpa_BIC_plot2(
  data,
  n_profiles_range = 1:6,
  type = "all",
  legend_position = "top",
  covar_model = c("EII", "EEE", "VII", "EEV", "EEI", "VVI", "VVV"),
  set_n_profile = 4,
  size.text = 4,
  flip = FALSE,
  na2zero = FALSE,
  basic = TRUE
)
}
\arguments{
\item{data}{data}

\item{n_profiles_range}{1:9}

\item{type}{"all", res, data, long_data, wide_data, graph, gg}

\item{legend_position}{top}

\item{covar_model}{"EEI", "EEE", "VVI", "VVV", "EEV", "EII", "VII",VVV}

\item{set_n_profile}{model 4 default}

\item{size.text}{size.text 4}

\item{flip}{y value <0 default}

\item{na2zero}{na2zero=FALSE na to 0 change}

\item{basic}{basic is "EII", "EEE", "VII", "EEI"}
}
\value{
pot
}
\description{
lpa_BIC_plot2
}
\examples{

\dontrun{

select(iris, -Species)\%>\%lpa_BIC_plot()
}

}
