% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_bind_classdata.R
\name{lpa_bind_classdata}
\alias{lpa_bind_classdata}
\title{lpa_bind_classdata}
\usage{
lpa_bind_classdata(
  data,
  n_profiles = 3,
  models = NULL,
  package = c("mclust", "MplusAutomation"),
  type = "all",
  add_line = TRUE,
  var_model = "EEE",
  Class_colors = 1:4,
  linewidth = 0.9,
  sd = TRUE
)
}
\arguments{
\item{data}{data}

\item{n_profiles}{default 3}

\item{package}{mclucs , MplusAutomation}

\item{type}{all, res, fit, est, data, all, plot, group_plot}

\item{add_line}{add_lineTRUE}

\item{var_model}{"EEE","EEI,"VVV","EII"}

\item{Class_colors}{1:4}

\item{linewidth}{linewidth 0.9}

\item{sd}{TRUE graph}
}
\value{
data
}
\description{
lpa_bind_classdata
}
\examples{

\dontrun{

#'
lpa_bind_classdata(select(iris, -Species))

lpa_bind_classdata(select(iris, -Species), type="res")
lpa_bind_classdata(select(iris, -Species), type="fit")
lpa_bind_classdata(select(iris, -Species), type="est")
lpa_bind_classdata(select(iris, -Species), type="data")\%>\%Round(3)\%>\%dall
lpa_bind_classdata(select(iris, -Species), type="all")
#'
# "EEE": variances = "equal", covariances = "equal"
# 프로파일 간 분산과 공분산이 모두 동일하다고 가정.
# "EEV": variances = "equal", covariances = "varying"
# 분산은 동일하지만, 공분산은 프로파일마다 다르게 추정.
# "EEI": variances = "equal", covariances = "zero"
# 분산은 동일하지만, 공분산은 추정하지 않음 (공변량 간 독립성 가정).
# "EII": variances = "equal", covariances = "zero"
# 동일한 가정으로 분산이 동일하고, 공분산은 추정되지 않음.
}



}
