% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_explore_modelfit.R
\name{lpa_explore_modelfit}
\alias{lpa_explore_modelfit}
\title{lpa_explore_modelfit}
\usage{
lpa_explore_modelfit(
  df,
  n_profiles_range = 1:9,
  modelNames = c("EEV", "EEI", "EEE", "EII")
)
}
\arguments{
\item{df}{data}

\item{n_profiles_range}{1:9}

\item{modelNames}{"EEV","EEI","EEE","EII"}
}
\description{
lpa_explore_modelfit
}
\examples{

\dontrun{

df <- select(iris, -Species)
lpa_explore_modelfit(df, 1:4)
lpa_explore_modelfit(df, 1:9)

}

}
