% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_gmm.R
\name{lpa_gmm}
\alias{lpa_gmm}
\title{잠재프로파일 분석 변수별 군집분포와 결과데이터}
\usage{
lpa_gmm(df, model = 4, modelNames = "EII")
}
\arguments{
\item{df}{df}

\item{model}{G, model N = 4}

\item{modelNames}{model ="EEE", EII}
}
\value{
df, graph
}
\description{
잠재프로파일 분석 변수별 군집분포와 결과데이터
}
\examples{
\dontrun{
lpa_gmm(edlcarename, model=4, "EII")
}
}
