% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa_profiles_plot.R
\name{lpa_profile_plot}
\alias{lpa_profile_plot}
\title{lpa_profile_plot}
\usage{
lpa_profile_plot(data, n_profiles = 3, model_name = "EEE", view = "pair")
}
\arguments{
\item{data}{data}

\item{n_profiles}{3}

\item{model_name}{EEE,EEI,VVI,VVV}

\item{view}{pair, each}
}
\value{
plot
}
\description{
lpa_profile_plot
}
\examples{

\dontrun{

}


}
