% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapaste_matTotext.R
\name{make_mat_text}
\alias{make_mat_text}
\title{Functions that let you input data in matrix form}
\usage{
make_mat_text(data, text = FALSE)
}
\arguments{
\item{data}{input dat a}
}
\value{
make test
}
\description{
Functions that let you input data in matrix form
}
\examples{

\dontrun{
lay
V1      V2 V3        V4        V5       V6        V7        V8 V9
[1,] "ceo_8" NA "satis_5" "satis_4" ""       "satis_3" "satis_1" NA ""
[2,] ""      NA ""        ""        ""       ""        ""        NA ""
[3,] "ceo_7" NA ""        ""        ""       ""        ""        NA ""
[4,] ""      NA ""        ""        "SATISF" ""        ""        NA ""
[5,] "ceo_6" NA ""        ""        ""       ""        ""        NA "stay_2"
[6,] ""      NA ""        ""        ""       ""        ""        NA ""
[7,] "ceo_4" NA "CEOCOM"  ""        ""       ""        "STAYON"  NA ""
[8,] ""      NA ""        ""        ""       ""        ""        NA ""
[9,] "ceo_3" NA ""        ""        ""       ""        ""        NA "stay_1"
[10,] ""      NA ""        ""        ""       ""        ""        NA ""
[11,] "ceo_2" NA ""        ""        ""       ""        ""        NA ""
[12,] ""      NA ""        ""        ""       ""        ""        NA ""
[13,] "ceo_1" NA ""        ""        ""       ""        ""        NA ""

lay \%>\% make_mat_text()


NewMat = matrix(c('ceo_8', NA, 'ceo_7', NA, 'ceo_6', NA, 'ceo_4', NA, 'ceo_3', NA, 'ceo_2', NA, 'ceo_1', NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'satis_5', NA, NA, NA, NA, NA, 'CEOCOM', NA, NA, NA, NA, NA, NA, 'satis_4', NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'SATISF', NA, NA, NA, NA, NA, NA, NA, NA, NA, 'satis_3', NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'satis_1', NA, NA, NA, NA, NA, 'STAYON', NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'stay_2', NA, NA, NA, 'stay_1', NA, NA, NA, NA),
                nrow = 13, ncol = 9)


colnames(NewMat) <- c("V1", "V2", "V3", "V4", "V5", "V6", "V7", "V8", "V9")

#as.numeric
# NewMat <- apply(NewMat, 2, function(x) as.numeric(gsub("'", "", x)))

V1      V2 V3        V4        V5       V6        V7        V8 V9
[1,] "ceo_8" NA "satis_5" "satis_4" NA       "satis_3" "satis_1" NA NA
[2,] NA      NA NA        NA        NA       NA        NA        NA NA
[3,] "ceo_7" NA NA        NA        NA       NA        NA        NA NA
[4,] NA      NA NA        NA        "SATISF" NA        NA        NA NA
[5,] "ceo_6" NA NA        NA        NA       NA        NA        NA "stay_2"
[6,] NA      NA NA        NA        NA       NA        NA        NA NA
[7,] "ceo_4" NA "CEOCOM"  NA        NA       NA        "STAYON"  NA NA
[8,] NA      NA NA        NA        NA       NA        NA        NA NA
[9,] "ceo_3" NA NA        NA        NA       NA        NA        NA "stay_1"
[10,] NA      NA NA        NA        NA       NA        NA        NA NA
[11,] "ceo_2" NA NA        NA        NA       NA        NA        NA NA
[12,] NA      NA NA        NA        NA       NA        NA        NA NA
[13,] "ceo_1" NA NA        NA        NA       NA        NA        NA NA
#'
}


}
