% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemanr_test.R
\name{mcnemar_test}
\alias{mcnemar_test}
\title{mcnemar_test}
\usage{
mcnemar_test(mat, correct = TRUE, exact = TRUE)
}
\arguments{
\item{mat}{matrix}

\item{correct}{McNemar's Chi-squared test with continuity correction}

\item{exact}{CI check binom.test}
}
\value{
reusult
}
\description{
mcnemar_test
}
\examples{

\dontrun{
#'
Buying_car  = matrix(c(23,28,7,12), nrow=2,
                     dimnames= list(c("시승전_있음","시승전_없음"),
                                    c("시승후_있음","시승후_없음")))
Buying_car
mcnemar_test(Buying_car)
mcnemar_test(Buying_car, correct=FALSE)

gen_table = matrix(c(c(9, 3, 13, 10)),ncol=2,
                   dimnames=list("case"=c("yes", "no"),
                                 "control"=c("yes", "no")   ))
gen_table
#Rbase
mcnemar.test(gen_table, correct = FALSE)
mcnemar_test(gen_table, correct = FALSE)



#모든 종류의 분석을 시행한 경우
bind_cols(
  "TEST"= c("correct and exact", "not Correct"," Not exact", "Not correct, exact"),
  bind_rows(
    mcnemar_test(Buying_car),  # 디폴트ㅡ
    mcnemar_test(Buying_car, correct=FALSE), # 연속보정 중지
    mcnemar_test(Buying_car,exact=FALSE), # 정확도 중지
    mcnemar_test(Buying_car,exact=FALSE, correct=FALSE)# 연속성 보정, 정확도 중지,SAS
  )
)\%>\%select(-b,-c,-odds_ratio)
}
}
