% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_col.R
\name{mutate_col}
\alias{mutate_col}
\title{Create, modify, and delete columns col position}
\usage{
mutate_col(data, ..., col = ncol(data))
}
\arguments{
\item{data}{data.frame}

\item{...}{creat name}

\item{col}{position}
}
\value{
data
}
\description{
Create, modify, and delete columns col position
}
\examples{

\dontrun{
#'
fm <- data.frame(
  age = c("20대", "20대", "20대", "30대", "30대", "30대", "40대", "40대", "40대"),
  manufacturer = c("google", "apple", "samsung", "google", "apple", "samsung", "google", "apple", "samsung"),
  values = c(794, 804, 1243, 987, 1055, 2517, 1528, 1519, 2255)
)
fm

#data
fm \%>\% mutate_col(gender= rep("a",9))
fm \%>\% mutate_col(gender= rep("a",9), col = 2)

#position
fm \%>\% mutate_col(gender = rep("female", nrow(fm)), col = 2)
fm \%>\% mutate_col(gender = rep("female", nrow(fm)), col = 1)
fm \%>\% mutate_col(gender = rep("female", nrow(fm)), col = 3)
fm \%>\% mutate_col(gender = rep("female", nrow(fm)), col = 4)

#Build and adjust with data from within
fm \%>\% mutate(gender = age)
fm \%>\% mutate_col(age_int = substring(fm$age,1,2))
fm \%>\% mutate_col(gender = paste(age,values,sep="//"))

# Double data generation
fm \%>\%
  mutate_col(insert_2 = rep("female", nrow(fm)), col = 2) \%>\%
  mutate_col( insert_4 = rep("a",9), col = 4)
}

}
