% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na2zero.R
\name{na2zero}
\alias{na2zero}
\title{NA replace to value default 0}
\usage{
na2zero(data, imp = 0)
}
\arguments{
\item{data}{data.frame or matrix}

\item{imp}{imputation value default 0}
}
\value{
dat
}
\description{
NA replace to value default 0
}
\examples{

\dontrun{

data \%>\% na2zero()
data \%>\% na2zero(0)
data \%>\% na2zero(1)
}

}
