% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_mark_sig.R
\name{p_mark_sig}
\alias{p_mark_sig}
\title{p_mark_sig is Function that displays significance in data}
\usage{
p_mark_sig(data, col = "p.value")
}
\arguments{
\item{data}{input data.frame or matrix}

\item{col}{'col' specifies the name of the column representing the p value in the data. By default, it is set to p.value. For example, if it is written as p, pvalue, p_value, etc., you can change the name.}
}
\description{
p_mark_sig is Function that displays significance in data
}
\examples{

\dontrun{
## data(ToothGrowth)
library(dplyr)
library(rstatix)

 ToothGrowth \%>\%
  group_by(dose) \%>\%
  t_test(data =., len ~ supp) \%>\%
  p_mark_sig("p")

}


}
