% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_blocks.R
\name{plspm_blocks}
\alias{plspm_blocks}
\title{Function to create blocks------------}
\usage{
plspm_blocks(..., dataset = NULL, output = "match")
}
\arguments{
\item{...}{list data}

\item{dataset}{data.frame}

\item{output}{'match', 'match_colnames' is compare data.frame}
}
\value{
list blocks
}
\description{
Function to create blocks------------
}
\examples{

\dontrun{
edu_blocks_named2 <- plspm_blocks(
  Support = item(1:4),
  Advising = item(5:8),
  Tutoring= item( 9:12),
  Value = item(13:16),
  Satisfaction = item(17:19),
  Loyalty =item(20:23),
  output = "match_colnames"
edu_blocks_named2

# to lavaan syntax
edu_blocks_named2 \%>\% plspm_blocks2lav()\%>\%cat("\n")

# # Example usage
edu_blocks_named1 <- plspm_blocks(
  Support = item("sup.help", "sup.under", "sup.safe", "sup.conc"),
  Advising = item("adv.comp", "adv.acces", "adv.comm", "adv.qual"),
  Tutoring = item("tut.prof", "tut.sched", "tut.stud", "tut.qual"),
  Value = item("val.devel", "val.deci", "val.meet", "val.info"),
  Satisfaction = item("sat.glad", "sat.expe", "sat.over"),
  Loyalty = item("loy.proud", "loy.recom", "loy.asha", "loy.back"),
  dataset = education , # Provide your dataset here
  output = "match_colnames"
)
#'
sat_blocks1 <- plspm_blocks(
  IMAG = item(paste0("imag",1:5)),
  EXPE = item(paste0("expe", 1:5)),
  QUAL = item( paste0("qual", 1:5)),
  VAL = item(paste0("val", 1:4)),
  SAT = item(paste0("sat", 1:4)),
  LOY = item(paste0("loy", 1:4))
)
sat_blocks1
##to lavaan syntax
sat_blocks1 \%>\% plspm_blocks2lav()\%>\%cat("\n")
}

}
