% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_boot_summary.R
\name{plspm_boot_summary}
\alias{plspm_boot_summary}
\title{plspm_boot_summary}
\usage{
plspm_boot_summary(pls_boot)
}
\arguments{
\item{pls_boot}{pls boot data}
}
\value{
data
}
\description{
plspm_boot_summary
}
\examples{

\dontrun{

jutpls_boot\%>\%plspm_boot_summary()
jutpls_boot$boot\%>\%plspm_boot_summary()

}


}
