% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_change_nodeLabels.R
\name{plspm_change_nodeLabels}
\alias{plspm_change_nodeLabels}
\title{plspm_change_nodeLabels}
\usage{
plspm_change_nodeLabels(
  data,
  ...,
  label.cex = 1.2,
  label.scale = TRUE,
  label.prop = 1,
  label.norm = "width",
  plot = TRUE
)
}
\arguments{
\item{data}{plspm_semPaths}

\item{...}{change node from name , to name}

\item{label.cex}{label.cex}

\item{label.scale}{label.scale}

\item{label.prop}{label.prop}

\item{label.norm}{label.norm}

\item{plot}{TRUE}
}
\value{
plot
}
\description{
plspm_change_nodeLabels
}
\examples{

\dontrun{


pathmodel = "
자기효능감 =~ C_S1 +  C_S2 + C_S3 + C_S4 + C_S5
진로태도 =~ B02 + B04 + B10 + B14
진로동기 =~ A_M1 + A_M2 + A_M3
진로준비 =~ D_P1 + D_P2 + D_P3

진로동기 ~ H1*자기효능감
진로태도 ~  H2*자기효능감
진로준비 ~ H5 *진로태도
진로준비 ~  H4 * 진로동기
진로준비 ~  H3*자기효능감
"

plspm_semPaths2(jutpls_boot,   curve=1, layout="tree2",rotation=1,
                edge.label.cex=0.7, mar=c(2,1,2,1),
                edge.label.position=0.6, label.cex=1,
                sizeLat= 10)\%>\%
  plspm_change_nodeLabels("B02","결정성",
                          "B04", "독립성",
                          "B10", "계획성",
                          "B14", "성향"
  )







}


}
