% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_dULS_modelfit.R
\name{plspm_dULS_modelfit}
\alias{plspm_dULS_modelfit}
\title{Bollen-Stine Bootstrap based d_ULS and d_G calculation}
\usage{
plspm_dULS_modelfit(pls_model, blocks, bootstrap = 1000, ci = 0.95)
}
\arguments{
\item{pls_model}{plsreslut}

\item{blocks}{model block}

\item{bootstrap}{default 1000}

\item{ci}{95% ci}
}
\value{
data
}
\description{
Bollen-Stine Bootstrap based d_ULS and d_G calculation
}
\examples{
\dontrun{
# d_ULS 및 d_G 지표에 대한 신뢰구간을 계산하려면,
 #Bollen-Stine 부트스트랩 절차를 통해 신뢰구간을
 #제시해야 합니다. 이 부트스트랩 절차는 일반적인
 #부트스트랩 방식과는 다르게 관측된 데이터를 여러
 #번 재추출하여 신뢰구간을 계산하는 방식입니다.
 #PLS-SEM에서 Bollen-Stine 부트스트랩을 사용하는
 #이유는 모형의 정확도를 평가하고, 특정 지표(d_ULS, d_G, SRMR 등)의
 #신뢰구간을 도출하기 위해서입니다.
#d_ULS는 마치 지도에서 두 지점 사이의 직선 거리를 측정하는 것과 비슷하다.
#즉, 각 변수 간의 차이를 개별적으로 계산해, 그 차이를 더하는 방식이다.
 # d_G는 산을 넘어가는 경로의 거리를 계산하는 것과 비슷하다.
  # 두 지점 사이의 거리를 단순히 직선으로 잴 수 없는 경우, 산의
  # 기울기와 복잡한 지형을 고려하여 최단 경로를 계산하는 방식이다.
  # 즉, 단순한 값 차이가 아니라, 데이터가 가지고 있는 전체적인
  # 구조적 특성을 고려하여 평가하는 것이다#'
#'
# Example call to the dULS_modelfit function with 1000 bootstrap samples
plspm_dULS_modelfit(edt_pls, edblock, bootstrap = 1000, ci = 0.95)
#'
  }
}
