% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_edges_values.R
\name{plspm_edge_values}
\alias{plspm_edge_values}
\title{Functions to include significance data for structural models in PLSPM}
\usage{
plspm_edge_values(plsdata, digits = 3, unite = TRUE, type = "vec")
}
\arguments{
\item{plsdata}{plsdata}

\item{digits}{round 3}

\item{unite}{data combine unite TRUE -> estimat + sig}

\item{type}{vec, df}
}
\value{
estimates vectors
}
\description{
Functions to include significance data for structural models in PLSPM
}
