% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_extract_blocks.R
\name{plspm_extract_blocks}
\alias{plspm_extract_blocks}
\title{olsom boot result model}
\usage{
plspm_extract_blocks(plsres)
}
\arguments{
\item{plsres}{result model}
}
\value{
lsit
}
\description{
olsom boot result model
}
\examples{

\dontrun{
# using htmt
}


}
