% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_f2.R
\name{plspm_f2}
\alias{plspm_f2}
\title{Effect Size Black}
\usage{
plspm_f2(plsres, type = "res")
}
\arguments{
\item{plsres}{pls result}

\item{type}{res, f2, R2, ceofs, inner}
}
\value{
data
}
\description{
Effect Size Black
}
\examples{

\dontrun{
#'
#'
# Example data for plsres
plsres <- list(
  path_coefs = data.frame(
    자기효능감 = c(0, 0.884, 0.681, 0.373),
    진로동기 = c(0, 0, 0, 0.353),
    진로태도 = c(0, 0, 0, 0.188),
    진로준비 = c(0, 0, 0, 0),
    row.names = c("자기효능감", "진로동기", "진로태도", "진로준비")
  ),
  inner_summary = data.frame(
    R2 = c(0.0000000, 0.7817016, 0.4631848, 0.7195506),
    row.names = c("자기효능감", "진로동기", "진로태도", "진로준비")
  )
)
plspm_f2(plsres, type="res")
plspm_f2(plsres, type="f2")
# etc
satpls_boot\%>\%plspm_f2()
satpls_boot\%>\%plspm_f2("f2")
satpls_boot\%>\%plspm_f2("coefs")
satpls_boot\%>\%plspm_f2("r2")
satpls_boot\%>\%plspm_f2("inner")
}



}
