% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_innerplot.R
\name{plspm_innerplot}
\alias{plspm_innerplot}
\title{inner_plot(Pre-draw a mockup )}
\usage{
plspm_innerplot(
  path_mat,
  layout = "spring",
  posCol = "black",
  rotation = 1,
  negCol = "red",
  fade = FALSE,
  edge.labels = FALSE,
  edge.label.cex = 1.2,
  edge.label.position = 0.5,
  edge.width = 2,
  vsize = 8,
  asize = 4,
  borders = TRUE,
  border.width = 2,
  border.color = "gray40",
  labels = NULL,
  label.cex = 1.2,
  label.color = "gray30",
  groups = NULL,
  palette = "pastel",
  curve = 0.2,
  width = 1
)
}
\arguments{
\item{path_mat}{mati}

\item{layout}{spring or layout}

\item{posCol}{black}

\item{rotation}{1}

\item{negCol}{red}

\item{fade}{false}

\item{edge.labels}{FALSE}

\item{edge.label.cex}{1.2}

\item{edge.label.position}{0.5}

\item{edge.width}{2}

\item{vsize}{8}

\item{asize}{4}

\item{borders}{TRUE}

\item{border.width}{2}

\item{border.color}{gray40}

\item{labels}{NULL}

\item{label.cex}{1.2}

\item{label.color}{gray30}

\item{groups}{NULL}

\item{palette}{pastel}

\item{curve}{0.2}

\item{width}{1}
}
\value{
graph
}
\description{
inner_plot(Pre-draw a mockup )
}
\examples{

\dontrun{

# path matrix
IMAG = c(0,0,0,0,0,0)
EXPE = c(1,0,0,0,0,0)
QUAL = c(0,1,0,0,0,0)
VAL = c(0,1,1,0,0,0)
SAT = c(1,1,1,1,0,0)
LOY = c(1,0,0,0,1,0)
sat_path = rbind(IMAG, EXPE, QUAL, VAL, SAT, LOY)

sat_path\%>\%plspm_innerplot()
#'
path1 = plspm_path(
  row_names = c("IMAG","EXPE","QUAL","VAL","SAT","LOY"),

  relationship = list(
    path(from="IMAG", to=c("EXPE","SAT","LOY")),
    path("EXPE", c("QUAL","VAL","SAT")),
    path("QUAL", c("VAL","SAT")),
    path("VAL",c("SAT")),
    path("SAT","LOY")
  )
)
path1
path1\%>\%plspm_innerplot()

#'
model1 = "
EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL+ EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"
# Generate the matrix
lavpath = plspm_lav2path(model1)


lavpath\%>\%plspm_innerplot()

#color
sat_path \%>\%plspm_innerplot(groups=1:6)



lay_p = matrix(c('IMA', NA, NA, NA, NA, NA, NA, NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'EXP', NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'QUA',
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, 'VAL', NA, NA, NA, NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'SAT', NA, NA, NA,
                 NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 'LOY',
                 NA, NA, NA, NA, NA, NA, NA, NA),
               nrow = 10, ncol = 10, byrow = FALSE)
lay_p
#'
#'
}


}
