% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_loadings.R
\name{plspm_loadings}
\alias{plspm_loadings}
\title{plspm loadings output}
\usage{
plspm_loadings(plsres, digits = 3, unite = TRUE)
}
\arguments{
\item{plsres}{plspm result}

\item{digits}{round =3}

\item{unite}{unite t+star}
}
\value{
data table
}
\description{
plspm loadings output
}
\examples{
\dontrun{

plsres \%>\% plspm_loadings()
plsres$boot \%>\% plspm_loadings()
plsres$boot$paths \%>\% plspm_loadings()


}

}
