% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_modelfit.R
\name{plspm_modelfit}
\alias{plspm_modelfit}
\title{plspm_modelfit}
\usage{
plspm_modelfit(
  pls_model,
  blocks,
  bootstrap = 1000,
  ci = 0.95,
  digits = 5,
  trans = TRUE
)
}
\arguments{
\item{pls_model}{plsresul}

\item{blocks}{model block}

\item{bootstrap}{bootstrap 1000}

\item{ci}{ci 95%}

\item{digits}{defaul 5}

\item{trans}{trans result}
}
\value{
datatable
}
\description{
plspm_modelfit
}
\examples{
\dontrun{

#'
#meausrement
edblock = plspm_blocks(
  SUX = item(paste0("sux",1:6)),
  FLOW = item(paste0("flow",1:5)),
  PEOU = item(paste0("peou",1:5)),
  PU = item(paste0("pu",1:5)),
  ATT = item(paste0("att",1:3)),
  USE = item(paste0("use",1:3))
  # dataset = ed_tam2
)
edblock
#Seed value: 19995
edt_pls = plspm_sem(Data=ed_tam2, edmodel, edblock)


# Example call to the plspm_modelfit function
plspm_modelfit(edt_pls, edblock)
#' # Example call to the plspm_modelfit function with 1000 bootstrap samples
rest= plspm_modelfit(edt_pls, edblock, bootstrap = 1000, ci = 0.95, trans=FALSE)
rest
rest\%>\%t()\%>\%rowdata2col(1)\%>\%data.frame()

rest2= plspm_modelfit(edt_pls, edblock, bootstrap = 1000, ci = 0.95, trans=TRUE)
rest2
}


}
