% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_path.R
\name{plspm_paths}
\alias{plspm_paths}
\title{plspm path create}
\usage{
plspm_paths(
  rownames = c("r1", "r2", "r3"),
  colnames = NULL,
  relationship = NULL
)
}
\arguments{
\item{rownames}{rowname}

\item{colnames}{colname}

\item{relationship}{input path}
}
\value{
matrix
}
\description{
plspm path create
}
\examples{

\dontrun{
edu_path <- plspm_path(
  rownames = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  colnames = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  relationship = list(
    list(from = "Value",
         to = c("Support", "Advising", "Tutoring")),
    list(from = "Satisfaction",
         to = c("Support", "Advising", "Tutoring","Value")),
    list(from = "Loyalty",
         to =c("Satisfaction"))
  )
)
edu_path


edu_path <- plspm_path(
  rownames = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  colnames = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  relationship = list(

    path(form = "Support", to ="Advising"),
    path(form = "Tutoring", to ="Advising"),
    path(from = "Value",
         to = c("Support", "Advising", "Tutoring")),
    path(from = "Satisfaction",
         to = c("Support", "Advising", "Tutoring","Value")),
    path (from = "Loyalty",
         to =c("Satisfaction"))
  )
)
edu_path

}
}
