% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm_paths2lav.R
\name{plspm_paths2lav}
\alias{plspm_paths2lav}
\title{plspm_paths2lav}
\usage{
plspm_paths2lav(path_matrix, name = FALSE, hypo = FALSE)
}
\arguments{
\item{path_matrix}{matrix paths}

\item{name}{default FALSE, TRUE : paths_model=}
}
\value{
lavaan syntax
}
\description{
plspm_paths2lav
}
\examples{
\dontrun{
#'
#' # Example path_matrix
path_matrix <- matrix(c(0, 1, 0, 0, 1, 1,
                        0, 0, 1, 1, 1, 0,
                        0, 0, 0, 1, 1, 0,
                        0, 0, 0, 0, 1, 0,
                        0, 0, 0, 0, 0, 1,
                        0, 0, 0, 0, 0, 0),
                      nrow = 6, ncol = 6, byrow = FALSE)

rownames(path_matrix) <- c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")
colnames(path_matrix) <- c("IMAG", "EXPE", "QUAL", "VAL", "SAT", "LOY")

cat(plspm_paths2lav(path_matrix))

#Add hypothesis markers
cat(plspm_paths2lav(path_matrix, hypo=TRUE))
#'
}
}
