% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace.R
\name{Replace_na}
\alias{Replace_na}
\title{Create a function to replace NA with an empty string in an R vector data}
\usage{
Replace_na(vector, pattern = NA, imp = "")
}
\arguments{
\item{vector}{input data}

\item{pattern}{dectect pattern}

\item{imp}{imputation}
}
\value{
imputation data
}
\description{
Create a function to replace NA with an empty string in an R vector data
}
\examples{

\dontrun{
# Replace(c("A", NA, "B", "", "C", "C"), "C", imp="T")
}

}
