% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcode.R
\name{revcode}
\alias{revcode}
\title{Custom function to perform reverse coding for multiple variables with custom n values}
\usage{
revcode(data, n = 8, ..., postfix = "_r")
}
\arguments{
\item{data}{Data frame containing the variables}

\item{n}{Vector of maximum scale values (one for each variable), c("v1","v2",...)}

\item{...}{Variable names to be reverse coded}

\item{postfix}{postfix is reverse variable name default '_r'}
}
\value{
add reverse data in data.frame
}
\description{
Custom function to perform reverse coding for multiple variables with custom n values
}
\examples{
\dontrun{

### How to decode all variables by applying only one n
revcode(stat_onl, 8, "upgrade", "satisfy", "fatigue" )|>
select(upgrade, satisfy, fatigue, upgrade_r, satisfy_r, fatigue_r)

### the method for the reverse coding variable being applied separately
### to each variable and decoding
revcode(stat_onl, n= c(8,8,8), "upgrade", "satisfy", "fatigue" )|>
select(upgrade, satisfy, fatigue, upgrade_r, satisfy_r, fatigue_r)

## mtcars data
## Since the inverse coding value of the cyl variable is 4,6,8,
## n enters 4 + 8 = 12; the minimum + maximum = n value.
revcode(mtcars[,c("cyl", "gear","carb")], 12, "cyl")


revcode(mtcars[,c("cyl", "gear","carb")],
            n = c(12, 6, 9),"cyl","gear","carb")
}
}
