% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round.R
\name{Round}
\alias{Round}
\title{Find and round numeric variables only}
\usage{
Round(data, digits = 2)
}
\arguments{
\item{data}{data.frame}

\item{digits}{default 2}
}
\value{
rounding data
}
\description{
Find and round numeric variables only
}
\examples{
\dontrun{

# vector
Round(12.345654)

##data.frame
mtcars\%>\% Round(1)

}
}
