% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_lav_hypo.R
\name{sem_lav_hypo}
\alias{sem_lav_hypo}
\title{Hypothesis generation functions}
\usage{
sem_lav_hypo(model)
}
\arguments{
\item{model}{lavaan syntax}
}
\value{
text
}
\description{
Hypothesis generation functions
}
\examples{

\dontrun{
#'
# 예제 모델
model1a <- "
EXPE ~ H1*IMAG
QUAL ~ H2*EXPE
VAL ~  H3*QUAL + H4*EXPE
SAT ~ H5*IMAG + H6*EXPE + H7*QUAL + H8*VAL
LOY ~ H9*SAT + H10*IMAG
"

# 함수 실행
sem_lav_hypo(model1a)
#'

# Example usage with the provided model
model10 <- "
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
VAL =~ val1 + val2 + val3 + val4
SAT =~ sat1 + sat2 + sat3 + sat4
LOY =~ loy1 + loy2 + loy3 + loy4

EXPE ~ IMAG
QUAL ~ EXPE
VAL ~  QUAL+ EXPE
SAT ~ IMAG + EXPE + QUAL + VAL
LOY ~ SAT + IMAG
"
lav_extract_sm(model10)\%>\%sem_lav_hypo()
lav_extract_sm(model10, prefix="H")\%>\%sem_lav_hypo()



#'
}

}
