% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_plot.R
\name{sem_plot}
\alias{sem_plot}
\title{Fast sem plot draw}
\usage{
sem_plot(
  data,
  whatLabels = "std",
  type = "res1",
  opt = 1,
  layout = "tree3",
  curve = 2,
  rotation = 2,
  sizeLat = 10,
  sizeLat2 = 6,
  sizeMan = 8,
  sizeMan2 = 4,
  edge.label.cex = 0.9,
  edge.label.position = 0.5,
  mar = c(1, 5, 1, 5),
  groups = "lat",
  pastel = TRUE,
  fade = FALSE,
  se = FALSE,
  residuals = FALSE,
  residScale = 12,
  intercepts = FALSE,
  semptools = TRUE
)
}
\arguments{
\item{data}{lavaaan object}

\item{whatLabels}{ste}

\item{opt}{1, 2, 3,}

\item{layout}{layY(opt =2, you make layout), or tree3}

\item{curve}{-2}

\item{rotation}{rotation = 2}

\item{sizeLat}{8}

\item{sizeLat2}{6}

\item{sizeMan}{6}

\item{sizeMan2}{3}

\item{edge.label.cex}{0.7}

\item{edge.label.position}{0.6}

\item{mar}{c(1,5,1,5)}

\item{groups}{groups}

\item{pastel}{pastel}

\item{fade}{default FALSE when opt =3}

\item{se}{se false}

\item{residuals}{residuals F}

\item{residScale}{12}

\item{intercepts}{intercepts =FALSE becouse semptools}

\item{semptools}{using semptools TRUE}
}
\value{
plot
}
\description{
Fast sem plot draw
}
\examples{

\dontrun{

# library(lavaan)
#import data
example(sem)

sem_plot(fit)

example(cfa)
sem_plot(fit)
#'
HS.model <- '
    visual  =~ x1 + b1*x2 + x3
    textual =~ x4 + b2*x5 + x6
    speed   =~ x7 + b3*x8 + x9
speed ~ visual + textual

'
fit1 <- sem(HS.model, data = HolzingerSwineford1939)

fit1 \%>\%sem_plot(residuals=T, curve=2)
}
}
