% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_plot.R
\name{sem_plot}
\alias{sem_plot}
\title{Fast sem plot draw}
\usage{
sem_plot(
  data,
  layout = "tree3",
  whatLabels = "std",
  opt = 1,
  curve = 2,
  sizeLat = 10,
  sizeLat2 = 6,
  sizeMan = 8,
  sizeMan2 = 4,
  edge.label.cex = 0.7,
  edge.label.position = 0.6,
  mar = c(1, 5, 1, 5),
  sig = TRUE
)
}
\arguments{
\item{data}{lavaaan object}

\item{layout}{layY(opt =2, you make layout), or tree3}

\item{whatLabels}{ste}

\item{opt}{1, 2, 3,}

\item{curve}{-2}

\item{sizeLat}{8}

\item{sizeLat2}{6}

\item{sizeMan}{6}

\item{sizeMan2}{3}

\item{edge.label.cex}{0.7}

\item{edge.label.position}{0.6}

\item{mar}{c(1,5,1,5)}

\item{isg}{TRUE}
}
\value{
plot
}
\description{
Fast sem plot draw
}
\examples{

\dontrun{

# library(lavaan)
#import data
example(sem)

sem_plot(fit)

example(cfa)
sem_plot(fit)


}
}
