% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobel_test.R
\name{Sobel_test}
\alias{Sobel_test}
\title{sobel test}
\usage{
Sobel_test(a, b, sa, sb, type = "Sobel")
}
\arguments{
\item{a}{est1}

\item{b}{est2}

\item{sa}{se1}

\item{sb}{se2}

\item{type}{Sobel, Aroian, Goodman}
}
\value{
test result
}
\description{
sobel test
}
\examples{

\dontrun{
#'
# Example usage
a <- 0.5  # Example coefficient a
b <- 0.3  # Example coefficient b
sa <- 0.1 # Standard error of a
sb <- 0.05 # Standard error of b

Sobel_test(a, b, sa, sb)
Sobel_test(a, b, sa, sb, "all")$Sobel
Sobel_test(a, b, sa, sb, "Sobel")
Sobel_test(a, b, sa, sb,"all")$Aroian
Sobel_test(a, b, sa, sb,"Aroian")
Sobel_test(a, b, sa, sb,"all")$Goodman
Sobel_test(a, b, sa, sb,"Goodman")
}
}
