% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobel_test.R
\name{sobel_test_general}
\alias{sobel_test_general}
\title{Generalized Sobel Test function}
\usage{
sobel_test_general(
  a,
  SE,
  roburst = FALSE,
  n = NULL,
  k = 2,
  digits = 5,
  conf = 0.95
)
}
\arguments{
\item{a}{est}

\item{SE}{se}

\item{roburst}{roburst T whide correction}

\item{n}{sample size}

\item{k}{number of regressin coefs, default 2}

\item{digits}{5}

\item{conf}{0.90, 0.95, 0.99, 0.999}
}
\value{
result
}
\description{
Generalized Sobel Test function
}
\examples{

\dontrun{
#'

general_Sobel_test(a = c(0.5, 0.3), SE = c(0.1, 0.05))

Sobel_test_2(a=0.5, b=0.3, sa=0.1, sb =0.05)

# Input data
a <- c(0.2, 0.3, 0.4, 0.5)
SE <- c(0.05, 0.04, 0.03, 0.02)

# Calculate results
sobel_test_general(a = c(0.2, 0.3, 0.4, 0.5), SE = c(0.05, 0.04, 0.03, 0.02))
aroian_sobel_test_general(a = c(0.2, 0.3, 0.4, 0.5), SE = c(0.05, 0.04, 0.03, 0.02))
Goodman_sobel_test_general(a = c(0.2, 0.3, 0.4, 0.5), SE = c(0.05, 0.04, 0.03, 0.02))
general_Sobel_test(a = c(0.2, 0.3, 0.4, 0.5), SE = c(0.05, 0.04, 0.03, 0.02))
#'
# 예제 데이터
a <- c(0.5, 0.3, 0.4)
SE <- c(0.1, 0.1, 0.1)

# 함수 실행 (기본 95\% 신뢰구간)
sobel_test_general(a, SE)

# 함수 실행 (90\% 신뢰구간)
result_90 <- sobel_test_general(a, SE, conf=0.90)
print(result_90)

# 함수 실행 (99\% 신뢰구간)
result_99 <- sobel_test_general(a, SE, conf=0.99)
print(result_99)

# 함수 실행 (99.9\% 신뢰구간)
result_999 <- sobel_test_general(a, SE, conf=0.999)
print(result_999)

}


}
