% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sem.R
\name{summary_sem}
\alias{summary_sem}
\title{summary sem result and apa report}
\usage{
summary_sem(
  lav_obj,
  whatLabels = "est",
  caption = "Result of path Coefficient",
  effect = "~",
  effect2 = NULL,
  md = TRUE,
  type = "all",
  opt = 1,
  sizeMan = 8,
  sizeMan2 = 5,
  layout = "tree2",
  mar = c(1, 5, 1, 5),
  intercepts = FALSE
)
}
\arguments{
\item{lav_obj}{lavaan object}

\item{whatLabels}{whatLabels est}

\item{caption}{caption possible}

\item{effect}{select "~"}

\item{effect2}{select 2nd}

\item{md}{markdown default TRUEz}

\item{type}{all, effect, apa, plot}

\item{opt}{opt=1~4}

\item{sizeMan}{sizeMan = 10}

\item{sizeMan2}{sizeMan2 =6}

\item{layout}{layout ="tree2"}

\item{mar}{mar = c(1,5,1,5)}

\item{intercepts}{intercept=FALSE semptools}
}
\value{
result and md
}
\description{
summary sem result and apa report
}
\examples{

\dontrun{
example(sem)
fit \%>\% summary_sem()

}
}
