% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sem.R
\name{summary_sem}
\alias{summary_sem}
\title{summary sem result and apa report}
\usage{
summary_sem(
  lav_obj,
  effect = "~",
  effect2 = NULL,
  md = TRUE,
  type = "all",
  opt = 1
)
}
\arguments{
\item{lav_obj}{lavaan object}

\item{effect}{select "~"}

\item{effect2}{select 2nd}

\item{md}{markdown default TRUE}

\item{type}{all, effect, apa, plot}

\item{opt}{opt=1~4}
}
\value{
result and md
}
\description{
summary sem result and apa report
}
\examples{

\dontrun{
example(sem)
fit \%>\% summary_sem()

}
}
