% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_df_multi.R
\name{table_df_multi}
\alias{table_df_multi}
\title{여러행으로 중복응답한 자료 빈도분석}
\usage{
table_df_multi(df, type = "res", digits = 4)
}
\arguments{
\item{df}{여려행이 있는 데이터}

\item{type}{res = freq_table=summary, missing = missing_list, data(remove missing)}

\item{digits}{4 default}
}
\value{
data
}
\description{
여러행으로 중복응답한 자료 빈도분석
}
\examples{
\dontrun{
}



}
