% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_freq.R
\name{table_freq}
\alias{table_freq}
\title{Frequency analysis table}
\usage{
table_freq(
  data,
  ...,
  prop = TRUE,
  plot = FALSE,
  angle = 0,
  size_text = 4,
  size_axis = 12,
  legend.position = "top",
  reorder = FALSE,
  type = "res"
)
}
\arguments{
\item{data}{data.frame}

\item{...}{vars If you input varabale names "var1", "var2" ,...}

\item{prop}{proportion value, default TRUE}

\item{plot}{plot}

\item{angle}{axis text angle}

\item{size_text}{default 3, bargraph text}

\item{size_axis}{default 12, axis test}

\item{legend.position}{legend.position="" none. top, bottom, left, right}

\item{reorder}{TRUE is plot x-variable reoder}

\item{type}{all, res=data, g=plot=graph}
}
\value{
Freqency table
}
\description{
Frequency analysis table
}
\examples{
\dontrun{
Mtcars = mtcars
Mtcars$am = factor(Mtcars$am, levels=c(0,1), labels= c("automatic","manual" ))
Mtcars$vs  = factor(Mtcars$vs, levels=c(0,1), labels= c("V-shaped","straight" ))
Mtcars$cyl  = factor(Mtcars$cyl, levels=c(4,6,8), labels= c("cyl-4","cyl-6","cyl-8"))

Mtcars \%>\% table_freq("am")
Mtcars \%>\% table_freq("am", prop=F)
Mtcars \%>\% table_freq("vs", plot=FALSE)

Mtcars \%>\% table_freq("vs")
Mtcars \%>\% table_freq("vs","am")

Mtcars \%>\% table_freq("vs","am","cyl")

Mtcars \%>\% table_freq("vs","am","cyl", angle=90)
Mtcars \%>\% table_freq("vs","am","cyl", angle=90, reorder = TRUE)

Mtcars \%>\% table_freq("vs","am","cyl", angle=90)

##error we need option  type = "g"
Mtcars \%>\% table_freq("vs","am","cyl", angle=90, reorder = TRUE)+ylim(0,14)
Mtcars \%>\% table_freq("vs","am","cyl", angle=90, reorder = TRUE, type="g")+ylim(0,14)

Mtcars \%>\% table_freq("vs","am") \%>\% arrange(am)
Mtcars \%>\% table_freq("vs","am") \%>\% arrange(vs)
Mtcars \%>\% table_freq("vs","am") \%>\% arrange(Freq)

#ONly data : $result  $g -> NULL
 mtcars \%>\% table_freq("vs","am", type = "all", plot = F)

# this NULL
mtcars \%>\% table_freq("vs","am", plot = F, type="g")
}



}
