% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test.R
\name{wald_test}
\alias{wald_test}
\title{wald_test}
\usage{
wald_test(
  b1,
  se1,
  b2,
  se2,
  est1 = NULL,
  est2 = NULL,
  digits = 3,
  type = "res2"
)
}
\arguments{
\item{b1}{fires coef}

\item{se1}{firtse se}

\item{b2}{secoond coef}

\item{se2}{second se}

\item{est1}{est1 name}

\item{est2}{est2 name}

\item{digits}{round}
}
\value{
test result
}
\description{
wald_test
}
\examples{
\dontrun{
##Clogg, C. C., Petkova, E., & Haritou, A. (1995).
Statistical methods for comparing regression coefficients between models.
American Journal of Sociology, 100(5), 1261-1293.
##https://blog.naver.com/shoutjoy/223294372137

 #' wald_test(-0.795,-0.795/-1.138, -0.795+0.149,  (-0.795+0.149)/1.344)

wald_test(-0.795,-0.795/-1.138,-0.795+0.149,
          (-0.795+0.149)/1.344,
           est1 = "남성",
           est2 = "여성")

}


}
