% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test_df.R
\name{wald_test_df}
\alias{wald_test_df}
\title{wald-test using data.frame(or tibble)}
\usage{
wald_test_df(data, row1 = 2, row2 = 3, rname1 = "", rname2 = "", sel = "row")
}
\arguments{
\item{data}{lm() and data.frame}

\item{row1}{row1 = 2}

\item{row2}{row2 = 3}

\item{rname1}{select first row name}

\item{rname2}{select second row name}

\item{sel}{row is number, term rowname}
}
\value{
diff, z, p value
}
\description{
wald-test using data.frame(or tibble)
}
\examples{
\dontrun{
 lm(mpg ~ hp + wt + drat, mtcars) \%>\%wald_test_df()

# #selecting variable
lm(mpg ~ hp + wt + drat, mtcars) \%>\%  wald_test_df(row1=2, row2=3)
#
# # When other variables are selected and analyzed
 lm(mpg ~ hp + wt + drat, mtcars) \%>\%
   wald_test_df(row1=2, row2=4)
#
# # When variable-names are selected
 lm(mpg ~ hp + wt + drat, mtcars) \%>\%
   wald_test_df(sel="term", rname1="hp", rname2="wt")
#
 lm(mpg ~ hp + wt + drat, mtcars) \%>\%
   wald_test_df(sel="term", rname1="hp", rname2="drat")
 }


}
